Make tsup build to dist outdir
