--- packages/kobber-icons/react-ssr-safe/index.d.ts
@@ -1,325 +1,0 @@
-import { FunctionComponent, HTMLProps } from 'react';
-
-declare const Add: FunctionComponent<HTMLProps<SVGElement>>;
-
-declare const AddFigure: FunctionComponent<HTMLProps<SVGElement>>;
-
-declare const AddSquare: FunctionComponent<HTMLProps<SVGElement>>;
-
-declare const AlarmBell: FunctionComponent<HTMLProps<SVGElement>>;
-
-declare const AlarmClock: FunctionComponent<HTMLProps<SVGElement>>;
-
-declare const AlarmTimer: FunctionComponent<HTMLProps<SVGElement>>;
-
-declare const AnalyticBars: FunctionComponent<HTMLProps<SVGElement>>;
-
-declare const ArrowBack: FunctionComponent<HTMLProps<SVGElement>>;
-
-declare const ArrowDown: FunctionComponent<HTMLProps<SVGElement>>;
-
-declare const ArrowLeft: FunctionComponent<HTMLProps<SVGElement>>;
-
-declare const ArrowRight: FunctionComponent<HTMLProps<SVGElement>>;
-
-declare const ArrowUp: FunctionComponent<HTMLProps<SVGElement>>;
-
-declare const Assets: FunctionComponent<HTMLProps<SVGElement>>;
-
-declare const AudioRecording: FunctionComponent<HTMLProps<SVGElement>>;
-
-declare const Backspace: FunctionComponent<HTMLProps<SVGElement>>;
-
-declare const Banned: FunctionComponent<HTMLProps<SVGElement>>;
-
-declare const Binocular: FunctionComponent<HTMLProps<SVGElement>>;
-
-declare const BookFlipPage: FunctionComponent<HTMLProps<SVGElement>>;
-
-declare const BookOpen: FunctionComponent<HTMLProps<SVGElement>>;
-
-declare const Bookmark: FunctionComponent<HTMLProps<SVGElement>>;
-
-declare const BulletList: FunctionComponent<HTMLProps<SVGElement>>;
-
-declare const Calendar: FunctionComponent<HTMLProps<SVGElement>>;
-
-declare const Camera: FunctionComponent<HTMLProps<SVGElement>>;
-
-declare const Chat: FunctionComponent<HTMLProps<SVGElement>>;
-
-declare const Check: FunctionComponent<HTMLProps<SVGElement>>;
-
-declare const CheckCircle: FunctionComponent<HTMLProps<SVGElement>>;
-
-declare const Checklist: FunctionComponent<HTMLProps<SVGElement>>;
-
-declare const ChevronDown: FunctionComponent<HTMLProps<SVGElement>>;
-
-declare const ChevronUp: FunctionComponent<HTMLProps<SVGElement>>;
-
-declare const Circle: FunctionComponent<HTMLProps<SVGElement>>;
-
-declare const CircleShapeAdd: FunctionComponent<HTMLProps<SVGElement>>;
-
-declare const Clock: FunctionComponent<HTMLProps<SVGElement>>;
-
-declare const ClockHand: FunctionComponent<HTMLProps<SVGElement>>;
-
-declare const Close: FunctionComponent<HTMLProps<SVGElement>>;
-
-declare const ColorBrush: FunctionComponent<HTMLProps<SVGElement>>;
-
-declare const CopyPaste: FunctionComponent<HTMLProps<SVGElement>>;
-
-declare const CursorPointer: FunctionComponent<HTMLProps<SVGElement>>;
-
-declare const CursorPointerCircle: FunctionComponent<HTMLProps<SVGElement>>;
-
-declare const Dice: FunctionComponent<HTMLProps<SVGElement>>;
-
-declare const Dictionary: FunctionComponent<HTMLProps<SVGElement>>;
-
-declare const Download: FunctionComponent<HTMLProps<SVGElement>>;
-
-declare const DrawTablet: FunctionComponent<HTMLProps<SVGElement>>;
-
-declare const Elements: FunctionComponent<HTMLProps<SVGElement>>;
-
-declare const Envelope: FunctionComponent<HTMLProps<SVGElement>>;
-
-declare const Eraser: FunctionComponent<HTMLProps<SVGElement>>;
-
-declare const Expand: FunctionComponent<HTMLProps<SVGElement>>;
-
-declare const ExpandResize: FunctionComponent<HTMLProps<SVGElement>>;
-
-declare const ExternalLink: FunctionComponent<HTMLProps<SVGElement>>;
-
-declare const ExternalLinkArrow: FunctionComponent<HTMLProps<SVGElement>>;
-
-declare const FemaleHeadset: FunctionComponent<HTMLProps<SVGElement>>;
-
-declare const FlipRight: FunctionComponent<HTMLProps<SVGElement>>;
-
-declare const FolderBookmark: FunctionComponent<HTMLProps<SVGElement>>;
-
-declare const FontExpand: FunctionComponent<HTMLProps<SVGElement>>;
-
-declare const FormTemplate: FunctionComponent<HTMLProps<SVGElement>>;
-
-declare const FrameExpand: FunctionComponent<HTMLProps<SVGElement>>;
-
-declare const FrameShrink: FunctionComponent<HTMLProps<SVGElement>>;
-
-declare const GameController: FunctionComponent<HTMLProps<SVGElement>>;
-
-declare const Glasses: FunctionComponent<HTMLProps<SVGElement>>;
-
-declare const GraphStats: FunctionComponent<HTMLProps<SVGElement>>;
-
-declare const Headphones: FunctionComponent<HTMLProps<SVGElement>>;
-
-declare const Heart: FunctionComponent<HTMLProps<SVGElement>>;
-
-declare const Home: FunctionComponent<HTMLProps<SVGElement>>;
-
-declare const Hyperlink: FunctionComponent<HTMLProps<SVGElement>>;
-
-declare const Image: FunctionComponent<HTMLProps<SVGElement>>;
-
-declare const ImageFlower: FunctionComponent<HTMLProps<SVGElement>>;
-
-declare const IndentDecrease: FunctionComponent<HTMLProps<SVGElement>>;
-
-declare const IndentIncrease: FunctionComponent<HTMLProps<SVGElement>>;
-
-declare const Information: FunctionComponent<HTMLProps<SVGElement>>;
-
-declare const Keyboard: FunctionComponent<HTMLProps<SVGElement>>;
-
-declare const Layers: FunctionComponent<HTMLProps<SVGElement>>;
-
-declare const LayersBack: FunctionComponent<HTMLProps<SVGElement>>;
-
-declare const LayoutLeft: FunctionComponent<HTMLProps<SVGElement>>;
-
-declare const LayoutModule: FunctionComponent<HTMLProps<SVGElement>>;
-
-declare const LayoutRight: FunctionComponent<HTMLProps<SVGElement>>;
-
-declare const Library: FunctionComponent<HTMLProps<SVGElement>>;
-
-declare const LightBulb: FunctionComponent<HTMLProps<SVGElement>>;
-
-declare const LockLocked: FunctionComponent<HTMLProps<SVGElement>>;
-
-declare const LockUnlocked: FunctionComponent<HTMLProps<SVGElement>>;
-
-declare const Login: FunctionComponent<HTMLProps<SVGElement>>;
-
-declare const Logout: FunctionComponent<HTMLProps<SVGElement>>;
-
-declare const Magnet: FunctionComponent<HTMLProps<SVGElement>>;
-
-declare const Menu: FunctionComponent<HTMLProps<SVGElement>>;
-
-declare const Message: FunctionComponent<HTMLProps<SVGElement>>;
-
-declare const MessageQuestion: FunctionComponent<HTMLProps<SVGElement>>;
-
-declare const MessageStar: FunctionComponent<HTMLProps<SVGElement>>;
-
-declare const MessageTyping: FunctionComponent<HTMLProps<SVGElement>>;
-
-declare const MessageWarning: FunctionComponent<HTMLProps<SVGElement>>;
-
-declare const ModuleEdit: FunctionComponent<HTMLProps<SVGElement>>;
-
-declare const ModuleFour: FunctionComponent<HTMLProps<SVGElement>>;
-
-declare const MultipleUsers: FunctionComponent<HTMLProps<SVGElement>>;
-
-declare const Newspaper: FunctionComponent<HTMLProps<SVGElement>>;
-
-declare const NotebookPencil: FunctionComponent<HTMLProps<SVGElement>>;
-
-declare const NumberList: FunctionComponent<HTMLProps<SVGElement>>;
-
-declare const Option: FunctionComponent<HTMLProps<SVGElement>>;
-
-declare const Palette: FunctionComponent<HTMLProps<SVGElement>>;
-
-declare const Paragraph: FunctionComponent<HTMLProps<SVGElement>>;
-
-declare const ParagraphCenter: FunctionComponent<HTMLProps<SVGElement>>;
-
-declare const ParagraphLeft: FunctionComponent<HTMLProps<SVGElement>>;
-
-declare const ParagraphRight: FunctionComponent<HTMLProps<SVGElement>>;
-
-declare const PartyBalloon: FunctionComponent<HTMLProps<SVGElement>>;
-
-declare const Pause: FunctionComponent<HTMLProps<SVGElement>>;
-
-declare const PenTools: FunctionComponent<HTMLProps<SVGElement>>;
-
-declare const PenWrite: FunctionComponent<HTMLProps<SVGElement>>;
-
-declare const Pencil: FunctionComponent<HTMLProps<SVGElement>>;
-
-declare const Pin: FunctionComponent<HTMLProps<SVGElement>>;
-
-declare const PinRemove: FunctionComponent<HTMLProps<SVGElement>>;
-
-declare const Play: FunctionComponent<HTMLProps<SVGElement>>;
-
-declare const Printer: FunctionComponent<HTMLProps<SVGElement>>;
-
-declare const QuestionCircle: FunctionComponent<HTMLProps<SVGElement>>;
-
-declare const RankingStar: FunctionComponent<HTMLProps<SVGElement>>;
-
-declare const RatingStarCheck: FunctionComponent<HTMLProps<SVGElement>>;
-
-declare const Read: FunctionComponent<HTMLProps<SVGElement>>;
-
-declare const ReadHome: FunctionComponent<HTMLProps<SVGElement>>;
-
-declare const ReadIdea: FunctionComponent<HTMLProps<SVGElement>>;
-
-declare const Redo: FunctionComponent<HTMLProps<SVGElement>>;
-
-declare const Refresh: FunctionComponent<HTMLProps<SVGElement>>;
-
-declare const Reload: FunctionComponent<HTMLProps<SVGElement>>;
-
-declare const Remove: FunctionComponent<HTMLProps<SVGElement>>;
-
-declare const RetouchGraph: FunctionComponent<HTMLProps<SVGElement>>;
-
-declare const RibbonStar: FunctionComponent<HTMLProps<SVGElement>>;
-
-declare const Rocket: FunctionComponent<HTMLProps<SVGElement>>;
-
-declare const Search: FunctionComponent<HTMLProps<SVGElement>>;
-
-declare const Settings: FunctionComponent<HTMLProps<SVGElement>>;
-
-declare const SettingsSlider: FunctionComponent<HTMLProps<SVGElement>>;
-
-declare const Share: FunctionComponent<HTMLProps<SVGElement>>;
-
-declare const Shrink: FunctionComponent<HTMLProps<SVGElement>>;
-
-declare const Shuffle: FunctionComponent<HTMLProps<SVGElement>>;
-
-declare const SocialInstagram: FunctionComponent<HTMLProps<SVGElement>>;
-
-declare const StudyMath: FunctionComponent<HTMLProps<SVGElement>>;
-
-declare const Subscript: FunctionComponent<HTMLProps<SVGElement>>;
-
-declare const Subtract: FunctionComponent<HTMLProps<SVGElement>>;
-
-declare const Superscript: FunctionComponent<HTMLProps<SVGElement>>;
-
-declare const SynchronizeArrowsSquare: FunctionComponent<HTMLProps<SVGElement>>;
-
-declare const SynchronizeArrowClock: FunctionComponent<HTMLProps<SVGElement>>;
-
-declare const SynchronizeArrowsWarning: FunctionComponent<HTMLProps<SVGElement>>;
-
-declare const TabletTouch: FunctionComponent<HTMLProps<SVGElement>>;
-
-declare const TaskList: FunctionComponent<HTMLProps<SVGElement>>;
-
-declare const TaskListClock: FunctionComponent<HTMLProps<SVGElement>>;
-
-declare const Teacher: FunctionComponent<HTMLProps<SVGElement>>;
-
-declare const TeacherCorrect: FunctionComponent<HTMLProps<SVGElement>>;
-
-declare const TextBold: FunctionComponent<HTMLProps<SVGElement>>;
-
-declare const TextFormat: FunctionComponent<HTMLProps<SVGElement>>;
-
-declare const TextItalic: FunctionComponent<HTMLProps<SVGElement>>;
-
-declare const TextStyle: FunctionComponent<HTMLProps<SVGElement>>;
-
-declare const TextSync: FunctionComponent<HTMLProps<SVGElement>>;
-
-declare const TextUnderline: FunctionComponent<HTMLProps<SVGElement>>;
-
-declare const Transfer: FunctionComponent<HTMLProps<SVGElement>>;
-
-declare const Trash: FunctionComponent<HTMLProps<SVGElement>>;
-
-declare const Undo: FunctionComponent<HTMLProps<SVGElement>>;
-
-declare const User: FunctionComponent<HTMLProps<SVGElement>>;
-
-declare const UserEdit: FunctionComponent<HTMLProps<SVGElement>>;
-
-declare const UserQuestion: FunctionComponent<HTMLProps<SVGElement>>;
-
-declare const Users: FunctionComponent<HTMLProps<SVGElement>>;
-
-declare const VideoFileAdd: FunctionComponent<HTMLProps<SVGElement>>;
-
-declare const VideoPlayer: FunctionComponent<HTMLProps<SVGElement>>;
-
-declare const View: FunctionComponent<HTMLProps<SVGElement>>;
-
-declare const ViewOff: FunctionComponent<HTMLProps<SVGElement>>;
-
-declare const VolumeOff: FunctionComponent<HTMLProps<SVGElement>>;
-
-declare const VolumeOn: FunctionComponent<HTMLProps<SVGElement>>;
-
-declare const Warning: FunctionComponent<HTMLProps<SVGElement>>;
-
-declare const Youtube: FunctionComponent<HTMLProps<SVGElement>>;
-
-export { Add, AddFigure, AddSquare, AlarmBell, AlarmClock, AlarmTimer, AnalyticBars, ArrowBack, ArrowDown, ArrowLeft, ArrowRight, ArrowUp, Assets, AudioRecording, Backspace, Banned, Binocular, BookFlipPage, BookOpen, Bookmark, BulletList, Calendar, Camera, Chat, Check, CheckCircle, Checklist, ChevronDown, ChevronUp, Circle, CircleShapeAdd, Clock, ClockHand, Close, ColorBrush, CopyPaste, CursorPointer, CursorPointerCircle, Dice, Dictionary, Download, DrawTablet, Elements, Envelope, Eraser, Expand, ExpandResize, ExternalLink, ExternalLinkArrow, FemaleHeadset, FlipRight, FolderBookmark, FontExpand, FormTemplate, FrameExpand, FrameShrink, GameController, Glasses, GraphStats, Headphones, Heart, Home, Hyperlink, Image, ImageFlower, IndentDecrease, IndentIncrease, Information, Keyboard, Layers, LayersBack, LayoutLeft, LayoutModule, LayoutRight, Library, LightBulb, LockLocked, LockUnlocked, Login, Logout, Magnet, Menu, Message, MessageQuestion, MessageStar, MessageTyping, MessageWarning, ModuleEdit, ModuleFour, MultipleUsers, Newspaper, NotebookPencil, NumberList, Option, Palette, Paragraph, ParagraphCenter, ParagraphLeft, ParagraphRight, PartyBalloon, Pause, PenTools, PenWrite, Pencil, Pin, PinRemove, Play, Printer, QuestionCircle, RankingStar, RatingStarCheck, Read, ReadHome, ReadIdea, Redo, Refresh, Reload, Remove, RetouchGraph, RibbonStar, Rocket, Search, Settings, SettingsSlider, Share, Shrink, Shuffle, SocialInstagram, StudyMath, Subscript, Subtract, Superscript, SynchronizeArrowClock, SynchronizeArrowsSquare, SynchronizeArrowsWarning, TabletTouch, TaskList, TaskListClock, Teacher, TeacherCorrect, TextBold, TextFormat, TextItalic, TextStyle, TextSync, TextUnderline, Transfer, Trash, Undo, User, UserEdit, UserQuestion, Users, VideoFileAdd, VideoPlayer, View, ViewOff, VolumeOff, VolumeOn, Warning, Youtube };
