--- packages/kobber-icons/react-ssr-safe/index.js
@@ -1,5179 +1,0 @@
-// src/icon/icons/add/index.react.tsx
-import { jsx, jsxs } from "react/jsx-runtime";
-var Add = (props) => {
-  let _ariaLabel = props["aria-label"];
-  _ariaLabel = _ariaLabel && _ariaLabel.trim();
-  const _ariaHidden = !_ariaLabel;
-  const _role = _ariaHidden ? "presentation" : "img";
-  return /* @__PURE__ */ jsxs(
-    "svg",
-    {
-      className: props.className,
-      viewBox: "0 0 20 20",
-      "aria-label": _ariaLabel,
-      "aria-hidden": _ariaHidden,
-      role: _role,
-      style: {
-        height: "var(--icon-height)",
-        width: "var(--icon-width)",
-        display: "flex",
-        alignItems: "center",
-        justifyContent: "center"
-      },
-      children: [
-        /* @__PURE__ */ jsx("g", { "clip-path": "url(#aaa)", children: /* @__PURE__ */ jsx("path", { fill: "currentColor", d: "M10 20a.625.625 0 0 1-.625-.625v-8.75H.625a.625.625 0 1 1 0-1.25h8.75V.625a.625.625 0 0 1 1.25 0v8.75h8.75a.625.625 0 1 1 0 1.25h-8.75v8.75A.624.624 0 0 1 10 20Z" }) }),
-        /* @__PURE__ */ jsx("defs", { children: /* @__PURE__ */ jsx("clipPath", { id: "aaa", children: /* @__PURE__ */ jsx("path", { fill: "currentColor", d: "M0 0h20v20H0z" }) }) })
-      ]
-    }
-  );
-};
-
-// src/icon/icons/add_figure/index.react.tsx
-import { jsx as jsx2, jsxs as jsxs2 } from "react/jsx-runtime";
-var AddFigure = (props) => {
-  let _ariaLabel = props["aria-label"];
-  _ariaLabel = _ariaLabel && _ariaLabel.trim();
-  const _ariaHidden = !_ariaLabel;
-  const _role = _ariaHidden ? "presentation" : "img";
-  return /* @__PURE__ */ jsxs2(
-    "svg",
-    {
-      className: props.className,
-      viewBox: "0 0 20 20",
-      "aria-label": _ariaLabel,
-      "aria-hidden": _ariaHidden,
-      role: _role,
-      style: {
-        height: "var(--icon-height)",
-        width: "var(--icon-width)",
-        display: "flex",
-        alignItems: "center",
-        justifyContent: "center"
-      },
-      children: [
-        /* @__PURE__ */ jsx2("g", { "clip-path": "url(#aba)", children: /* @__PURE__ */ jsx2("path", { fill: "currentColor", d: "M2.5 5A2.503 2.503 0 0 1 0 2.5C0 1.122 1.122 0 2.5 0S5 1.122 5 2.5 3.878 5 2.5 5Zm0-3.75a1.25 1.25 0 1 0 0 2.5 1.25 1.25 0 0 0 0-2.5ZM2.5 20A2.503 2.503 0 0 1 0 17.5C0 16.122 1.122 15 2.5 15S5 16.122 5 17.5 3.878 20 2.5 20Zm0-3.75a1.25 1.25 0 1 0 0 2.5 1.25 1.25 0 0 0 0-2.5ZM17.5 5A2.503 2.503 0 0 1 15 2.5C15 1.122 16.122 0 17.5 0S20 1.122 20 2.5 18.878 5 17.5 5Zm0-3.75a1.25 1.25 0 1 0 0 2.5 1.25 1.25 0 0 0 0-2.5Zm0 18.75a2.503 2.503 0 0 1-2.5-2.5c0-1.378 1.122-2.5 2.5-2.5s2.5 1.122 2.5 2.5-1.122 2.5-2.5 2.5Zm0-3.75a1.25 1.25 0 1 0 0 2.5 1.25 1.25 0 0 0 0-2.5ZM6.875 2.5a.625.625 0 0 1 0-1.25h6.25a.625.625 0 1 1 0 1.25h-6.25Zm11.25 11.25a.624.624 0 0 1-.625-.625v-6.25a.625.625 0 1 1 1.25 0v6.25a.624.624 0 0 1-.625.625Zm-11.25 5a.625.625 0 1 1 0-1.25h6.25a.624.624 0 1 1 0 1.25h-6.25Zm-5-5a.625.625 0 0 1-.625-.625v-6.25a.625.625 0 0 1 1.25 0v6.25a.625.625 0 0 1-.625.625Z" }) }),
-        /* @__PURE__ */ jsx2("defs", { children: /* @__PURE__ */ jsx2("clipPath", { id: "aba", children: /* @__PURE__ */ jsx2("path", { fill: "currentColor", d: "M0 0h20v20H0z" }) }) })
-      ]
-    }
-  );
-};
-
-// src/icon/icons/add_square/index.react.tsx
-import { jsx as jsx3, jsxs as jsxs3 } from "react/jsx-runtime";
-var AddSquare = (props) => {
-  let _ariaLabel = props["aria-label"];
-  _ariaLabel = _ariaLabel && _ariaLabel.trim();
-  const _ariaHidden = !_ariaLabel;
-  const _role = _ariaHidden ? "presentation" : "img";
-  return /* @__PURE__ */ jsxs3(
-    "svg",
-    {
-      className: props.className,
-      viewBox: "0 0 20 20",
-      "aria-label": _ariaLabel,
-      "aria-hidden": _ariaHidden,
-      role: _role,
-      style: {
-        height: "var(--icon-height)",
-        width: "var(--icon-width)",
-        display: "flex",
-        alignItems: "center",
-        justifyContent: "center"
-      },
-      children: [
-        /* @__PURE__ */ jsxs3("g", { "clip-path": "url(#aca)", children: [
-          /* @__PURE__ */ jsx3("path", { fill: "currentColor", d: "M10 14.375a.625.625 0 0 1-.625-.625v-3.125H6.25a.625.625 0 1 1 0-1.25h3.125V6.25a.625.625 0 0 1 1.25 0v3.125h3.125a.625.625 0 1 1 0 1.25h-3.125v3.125a.624.624 0 0 1-.625.625Z" }),
-          /* @__PURE__ */ jsx3("path", { fill: "currentColor", d: "M3.125 20A3.128 3.128 0 0 1 0 16.875V3.125A3.128 3.128 0 0 1 3.125 0h13.75A3.128 3.128 0 0 1 20 3.125v13.75A3.128 3.128 0 0 1 16.875 20H3.125Zm0-18.75A1.877 1.877 0 0 0 1.25 3.125v13.75a1.877 1.877 0 0 0 1.875 1.875h13.75a1.876 1.876 0 0 0 1.875-1.875V3.125a1.876 1.876 0 0 0-1.875-1.875H3.125Z" })
-        ] }),
-        /* @__PURE__ */ jsx3("defs", { children: /* @__PURE__ */ jsx3("clipPath", { id: "aca", children: /* @__PURE__ */ jsx3("path", { fill: "currentColor", d: "M0 0h20v20H0z" }) }) })
-      ]
-    }
-  );
-};
-
-// src/icon/icons/alarm_bell/index.react.tsx
-import { jsx as jsx4, jsxs as jsxs4 } from "react/jsx-runtime";
-var AlarmBell = (props) => {
-  let _ariaLabel = props["aria-label"];
-  _ariaLabel = _ariaLabel && _ariaLabel.trim();
-  const _ariaHidden = !_ariaLabel;
-  const _role = _ariaHidden ? "presentation" : "img";
-  return /* @__PURE__ */ jsxs4(
-    "svg",
-    {
-      className: props.className,
-      viewBox: "0 0 20 20",
-      "aria-label": _ariaLabel,
-      "aria-hidden": _ariaHidden,
-      role: _role,
-      style: {
-        height: "var(--icon-height)",
-        width: "var(--icon-width)",
-        display: "flex",
-        alignItems: "center",
-        justifyContent: "center"
-      },
-      children: [
-        /* @__PURE__ */ jsx4("g", { "clip-path": "url(#ada)", children: /* @__PURE__ */ jsx4("path", { fill: "currentColor", d: "M9.992 20a2.362 2.362 0 0 1-2.265-1.698.625.625 0 0 1 1.199-.353 1.113 1.113 0 0 0 2.137.001.63.63 0 0 1 .992-.311.626.626 0 0 1 .207.663A2.37 2.37 0 0 1 9.992 20Zm-7.499-3.749a.63.63 0 0 1-.623-.69.623.623 0 0 1 .13-.32c.005-.008 1.118-1.616 1.118-6.49a6.864 6.864 0 0 1 6.25-6.847V.626a.625.625 0 1 1 1.25 0v1.278a6.864 6.864 0 0 1 6.25 6.847c0 5.253 1.038 6.405 1.048 6.416a.63.63 0 0 1-.071.973.621.621 0 0 1-.352.111h-15Zm13.925-1.25c-.402-1.027-.801-2.892-.801-6.25a5.631 5.631 0 0 0-5.625-5.625 5.631 5.631 0 0 0-5.625 5.625c0 3.298-.484 5.23-.872 6.25h12.922Z" }) }),
-        /* @__PURE__ */ jsx4("defs", { children: /* @__PURE__ */ jsx4("clipPath", { id: "ada", children: /* @__PURE__ */ jsx4("path", { fill: "currentColor", d: "M0 0h20v20H0z" }) }) })
-      ]
-    }
-  );
-};
-
-// src/icon/icons/alarm_clock/index.react.tsx
-import { jsx as jsx5, jsxs as jsxs5 } from "react/jsx-runtime";
-var AlarmClock = (props) => {
-  let _ariaLabel = props["aria-label"];
-  _ariaLabel = _ariaLabel && _ariaLabel.trim();
-  const _ariaHidden = !_ariaLabel;
-  const _role = _ariaHidden ? "presentation" : "img";
-  return /* @__PURE__ */ jsxs5(
-    "svg",
-    {
-      className: props.className,
-      viewBox: "0 0 20 20",
-      "aria-label": _ariaLabel,
-      "aria-hidden": _ariaHidden,
-      role: _role,
-      style: {
-        height: "var(--icon-height)",
-        width: "var(--icon-width)",
-        display: "flex",
-        alignItems: "center",
-        justifyContent: "center"
-      },
-      children: [
-        /* @__PURE__ */ jsxs5("g", { "clip-path": "url(#aea)", children: [
-          /* @__PURE__ */ jsx5("path", { fill: "currentColor", d: "M15 20a.622.622 0 0 1-.56-.346l-.866-1.734a8.125 8.125 0 0 1-7.148 0l-.868 1.734a.62.62 0 0 1-.837.28.625.625 0 0 1-.28-.838l.907-1.814a8.105 8.105 0 0 1-3.473-6.657C1.875 6.145 5.52 2.5 10 2.5s8.125 3.645 8.125 8.125a8.102 8.102 0 0 1-3.473 6.656l.907 1.814A.625.625 0 0 1 15 20ZM10 3.75a6.883 6.883 0 0 0-6.875 6.875A6.883 6.883 0 0 0 10 17.5a6.883 6.883 0 0 0 6.875-6.875A6.883 6.883 0 0 0 10 3.75Z" }),
-          /* @__PURE__ */ jsx5("path", { fill: "currentColor", d: "M7.174 11.25a.625.625 0 1 1 0-1.25h2.201V6.25a.625.625 0 0 1 1.25 0v4.375a.625.625 0 0 1-.625.625H7.174ZM1.25 3.75a.625.625 0 0 1-.39-1.113l3.124-2.5a.628.628 0 0 1 .88.098.625.625 0 0 1-.098.879l-3.125 2.5a.628.628 0 0 1-.391.136Zm17.5 0a.628.628 0 0 1-.39-.137l-3.125-2.5a.626.626 0 0 1 .78-.976l3.126 2.5a.625.625 0 0 1-.391 1.113Z" })
-        ] }),
-        /* @__PURE__ */ jsx5("defs", { children: /* @__PURE__ */ jsx5("clipPath", { id: "aea", children: /* @__PURE__ */ jsx5("path", { fill: "currentColor", d: "M0 0h20v20H0z" }) }) })
-      ]
-    }
-  );
-};
-
-// src/icon/icons/alarm_timer/index.react.tsx
-import { jsx as jsx6, jsxs as jsxs6 } from "react/jsx-runtime";
-var AlarmTimer = (props) => {
-  let _ariaLabel = props["aria-label"];
-  _ariaLabel = _ariaLabel && _ariaLabel.trim();
-  const _ariaHidden = !_ariaLabel;
-  const _role = _ariaHidden ? "presentation" : "img";
-  return /* @__PURE__ */ jsxs6(
-    "svg",
-    {
-      className: props.className,
-      viewBox: "0 0 20 20",
-      "aria-label": _ariaLabel,
-      "aria-hidden": _ariaHidden,
-      role: _role,
-      style: {
-        height: "var(--icon-height)",
-        width: "var(--icon-width)",
-        display: "flex",
-        alignItems: "center",
-        justifyContent: "center"
-      },
-      children: [
-        /* @__PURE__ */ jsxs6("g", { "clip-path": "url(#afa)", children: [
-          /* @__PURE__ */ jsx6("path", { fill: "currentColor", d: "M.625 16.251a.626.626 0 0 1-.526-.963c.01-.015 1.15-1.929 1.15-6.537A6.863 6.863 0 0 1 7.5 1.904V.626a.625.625 0 0 1 1.25 0v1.875c0 .345-.28.625-.624.625A5.631 5.631 0 0 0 2.5 8.751c0 3.187-.506 5.178-.903 6.25h12.93a6.736 6.736 0 0 1-.553-1.662.622.622 0 0 1 .614-.744c.299 0 .557.212.614.505.15.774.472 1.521.931 2.162a.623.623 0 0 1-.507.99H.625Z" }),
-          /* @__PURE__ */ jsx6("path", { fill: "currentColor", d: "M14.375 11.251A5.631 5.631 0 0 1 8.75 5.626 5.631 5.631 0 0 1 14.375.001 5.631 5.631 0 0 1 20 5.626a5.631 5.631 0 0 1-5.625 5.625Zm0-10A4.38 4.38 0 0 0 10 5.626a4.38 4.38 0 0 0 4.375 4.375 4.38 4.38 0 0 0 4.375-4.375 4.38 4.38 0 0 0-4.375-4.375Z" }),
-          /* @__PURE__ */ jsx6("path", { fill: "currentColor", d: "M14.375 7.5a.625.625 0 0 1-.625-.624V3.75a.625.625 0 0 1 1.25 0v2.5h1.25a.625.625 0 0 1 0 1.25h-1.875ZM8.122 20a2.363 2.363 0 0 1-2.265-1.698.626.626 0 0 1 1.2-.353 1.108 1.108 0 0 0 1.066.801 1.112 1.112 0 0 0 1.072-.8.629.629 0 0 1 .775-.425.626.626 0 0 1 .425.776 2.355 2.355 0 0 1-1.603 1.602c-.22.065-.445.098-.67.098Z" })
-        ] }),
-        /* @__PURE__ */ jsx6("defs", { children: /* @__PURE__ */ jsx6("clipPath", { id: "afa", children: /* @__PURE__ */ jsx6("path", { fill: "currentColor", d: "M0 0h20v20H0z" }) }) })
-      ]
-    }
-  );
-};
-
-// src/icon/icons/analytic_bars/index.react.tsx
-import { jsx as jsx7 } from "react/jsx-runtime";
-var AnalyticBars = (props) => {
-  let _ariaLabel = props["aria-label"];
-  _ariaLabel = _ariaLabel && _ariaLabel.trim();
-  const _ariaHidden = !_ariaLabel;
-  const _role = _ariaHidden ? "presentation" : "img";
-  return /* @__PURE__ */ jsx7(
-    "svg",
-    {
-      className: props.className,
-      viewBox: "0 0 20 20",
-      "aria-label": _ariaLabel,
-      "aria-hidden": _ariaHidden,
-      role: _role,
-      style: {
-        height: "var(--icon-height)",
-        width: "var(--icon-width)",
-        display: "flex",
-        alignItems: "center",
-        justifyContent: "center"
-      },
-      children: /* @__PURE__ */ jsx7("path", { fill: "currentColor", d: "M.625 18.75a.625.625 0 1 1 0-1.25h.625V10c0-.69.56-1.25 1.25-1.25H5c.69 0 1.25.56 1.25 1.25v7.5H7.5v-15c0-.69.56-1.25 1.25-1.25h2.5c.69 0 1.25.56 1.25 1.25v15h1.25V6.25c0-.69.56-1.25 1.25-1.25h2.5c.69 0 1.25.56 1.25 1.25V17.5h.625a.624.624 0 1 1 0 1.25H.625ZM17.5 17.5V6.25H15V17.5h2.5Zm-6.25 0v-15h-2.5v15h2.5ZM5 17.5V10H2.5v7.5H5Z" })
-    }
-  );
-};
-
-// src/icon/icons/arrow_back/index.react.tsx
-import { jsx as jsx8, jsxs as jsxs7 } from "react/jsx-runtime";
-var ArrowBack = (props) => {
-  let _ariaLabel = props["aria-label"];
-  _ariaLabel = _ariaLabel && _ariaLabel.trim();
-  const _ariaHidden = !_ariaLabel;
-  const _role = _ariaHidden ? "presentation" : "img";
-  return /* @__PURE__ */ jsxs7(
-    "svg",
-    {
-      className: props.className,
-      viewBox: "0 0 20 20",
-      "aria-label": _ariaLabel,
-      "aria-hidden": _ariaHidden,
-      role: _role,
-      style: {
-        height: "var(--icon-height)",
-        width: "var(--icon-width)",
-        display: "flex",
-        alignItems: "center",
-        justifyContent: "center"
-      },
-      children: [
-        /* @__PURE__ */ jsx8("g", { "clip-path": "url(#aha)", children: /* @__PURE__ */ jsx8("path", { fill: "currentColor", d: "M19.376 17.503a.625.625 0 0 1-.625-.625V9.375a.625.625 0 0 0-.625-.625H2.134l3.309 3.31a.62.62 0 0 1 0 .883.624.624 0 0 1-.884 0L.183 8.567a.617.617 0 0 1-.135-.204L.039 8.34a.625.625 0 0 1 .001-.434l.006-.017a.625.625 0 0 1 .138-.208L4.56 3.307A.619.619 0 0 1 5 3.123c.167 0 .325.065.442.184a.62.62 0 0 1 .183.442.62.62 0 0 1-.183.441L2.134 7.5h15.992A1.876 1.876 0 0 1 20 9.375v7.503a.626.626 0 0 1-.625.625Z" }) }),
-        /* @__PURE__ */ jsx8("defs", { children: /* @__PURE__ */ jsx8("clipPath", { id: "aha", children: /* @__PURE__ */ jsx8("path", { fill: "currentColor", d: "M0 0h20v20H0z" }) }) })
-      ]
-    }
-  );
-};
-
-// src/icon/icons/arrow_down/index.react.tsx
-import { jsx as jsx9, jsxs as jsxs8 } from "react/jsx-runtime";
-var ArrowDown = (props) => {
-  let _ariaLabel = props["aria-label"];
-  _ariaLabel = _ariaLabel && _ariaLabel.trim();
-  const _ariaHidden = !_ariaLabel;
-  const _role = _ariaHidden ? "presentation" : "img";
-  return /* @__PURE__ */ jsxs8(
-    "svg",
-    {
-      className: props.className,
-      viewBox: "0 0 20 20",
-      "aria-label": _ariaLabel,
-      "aria-hidden": _ariaHidden,
-      role: _role,
-      style: {
-        height: "var(--icon-height)",
-        width: "var(--icon-width)",
-        display: "flex",
-        alignItems: "center",
-        justifyContent: "center"
-      },
-      children: [
-        /* @__PURE__ */ jsx9("g", { "clip-path": "url(#aia)", children: /* @__PURE__ */ jsx9("path", { fill: "currentColor", d: "M10.042 20H9.79c-.084 0-.168-.083-.168-.167l-8.789-8.75a.635.635 0 0 1-.167-.416c0-.167.083-.334.167-.417a.46.46 0 0 1 .419-.25.64.64 0 0 1 .418.167l7.701 7.666V.667c0-.417.335-.667.67-.667s.67.25.67.667v17.25l7.7-7.667a.641.641 0 0 1 .419-.167.64.64 0 0 1 .418.167c.084.083.168.25.168.417a.635.635 0 0 1-.168.416l-8.789 8.75c-.084.084-.167.084-.167.167h-.251Z" }) }),
-        /* @__PURE__ */ jsx9("defs", { children: /* @__PURE__ */ jsx9("clipPath", { id: "aia", children: /* @__PURE__ */ jsx9("path", { fill: "currentColor", d: "M0 0h20v20H0z" }) }) })
-      ]
-    }
-  );
-};
-
-// src/icon/icons/arrow_left/index.react.tsx
-import { jsx as jsx10, jsxs as jsxs9 } from "react/jsx-runtime";
-var ArrowLeft = (props) => {
-  let _ariaLabel = props["aria-label"];
-  _ariaLabel = _ariaLabel && _ariaLabel.trim();
-  const _ariaHidden = !_ariaLabel;
-  const _role = _ariaHidden ? "presentation" : "img";
-  return /* @__PURE__ */ jsxs9(
-    "svg",
-    {
-      className: props.className,
-      viewBox: "0 0 20 20",
-      "aria-label": _ariaLabel,
-      "aria-hidden": _ariaHidden,
-      role: _role,
-      style: {
-        height: "var(--icon-height)",
-        width: "var(--icon-width)",
-        display: "flex",
-        alignItems: "center",
-        justifyContent: "center"
-      },
-      children: [
-        /* @__PURE__ */ jsx10("g", { "clip-path": "url(#aja)", children: /* @__PURE__ */ jsx10("path", { fill: "currentColor", d: "M9.375 19.375a.619.619 0 0 1-.442-.183l-8.75-8.75a.618.618 0 0 1-.136-.204l-.008-.022a.625.625 0 0 1 .146-.659L8.933.808a.62.62 0 0 1 .884 0A.619.619 0 0 1 10 1.25a.619.619 0 0 1-.183.442L2.133 9.375h17.242a.625.625 0 0 1 0 1.25H2.135l7.683 7.683a.627.627 0 0 1-.443 1.067Z" }) }),
-        /* @__PURE__ */ jsx10("defs", { children: /* @__PURE__ */ jsx10("clipPath", { id: "aja", children: /* @__PURE__ */ jsx10("path", { fill: "currentColor", d: "M0 0h20v20H0z" }) }) })
-      ]
-    }
-  );
-};
-
-// src/icon/icons/arrow_right/index.react.tsx
-import { jsx as jsx11, jsxs as jsxs10 } from "react/jsx-runtime";
-var ArrowRight = (props) => {
-  let _ariaLabel = props["aria-label"];
-  _ariaLabel = _ariaLabel && _ariaLabel.trim();
-  const _ariaHidden = !_ariaLabel;
-  const _role = _ariaHidden ? "presentation" : "img";
-  return /* @__PURE__ */ jsxs10(
-    "svg",
-    {
-      className: props.className,
-      viewBox: "0 0 20 20",
-      "aria-label": _ariaLabel,
-      "aria-hidden": _ariaHidden,
-      role: _role,
-      style: {
-        height: "var(--icon-height)",
-        width: "var(--icon-width)",
-        display: "flex",
-        alignItems: "center",
-        justifyContent: "center"
-      },
-      children: [
-        /* @__PURE__ */ jsx11("g", { "clip-path": "url(#aka)", children: /* @__PURE__ */ jsx11("path", { fill: "currentColor", d: "M10.625 19.375a.627.627 0 0 1-.441-1.067l7.683-7.684H.625a.624.624 0 1 1 0-1.249h17.241l-7.683-7.683a.621.621 0 0 1 0-.884.619.619 0 0 1 .442-.183c.167 0 .324.065.442.183l8.75 8.75a.616.616 0 0 1 .136.204l.008.022a.609.609 0 0 1 0 .431l-.005.013a.629.629 0 0 1-.14.215l-8.75 8.749a.62.62 0 0 1-.44.183Z" }) }),
-        /* @__PURE__ */ jsx11("defs", { children: /* @__PURE__ */ jsx11("clipPath", { id: "aka", children: /* @__PURE__ */ jsx11("path", { fill: "currentColor", d: "M0 0h20v20H0z" }) }) })
-      ]
-    }
-  );
-};
-
-// src/icon/icons/arrow_up/index.react.tsx
-import { jsx as jsx12, jsxs as jsxs11 } from "react/jsx-runtime";
-var ArrowUp = (props) => {
-  let _ariaLabel = props["aria-label"];
-  _ariaLabel = _ariaLabel && _ariaLabel.trim();
-  const _ariaHidden = !_ariaLabel;
-  const _role = _ariaHidden ? "presentation" : "img";
-  return /* @__PURE__ */ jsxs11(
-    "svg",
-    {
-      className: props.className,
-      viewBox: "0 0 20 20",
-      "aria-label": _ariaLabel,
-      "aria-hidden": _ariaHidden,
-      role: _role,
-      style: {
-        height: "var(--icon-height)",
-        width: "var(--icon-width)",
-        display: "flex",
-        alignItems: "center",
-        justifyContent: "center"
-      },
-      children: [
-        /* @__PURE__ */ jsx12("g", { "clip-path": "url(#ala)", children: /* @__PURE__ */ jsx12("path", { fill: "currentColor", d: "M10 20a.625.625 0 0 1-.625-.625V2.135L1.692 9.817a.622.622 0 0 1-.884-.001.619.619 0 0 1-.183-.442c0-.167.065-.324.183-.442l8.75-8.75a.617.617 0 0 1 .204-.135l.02-.008a.62.62 0 0 1 .435 0l.018.006c.08.033.15.08.208.138l8.749 8.75a.619.619 0 0 1 .183.441.619.619 0 0 1-.386.577.625.625 0 0 1-.68-.135l-7.684-7.684v17.242A.625.625 0 0 1 10 20Z" }) }),
-        /* @__PURE__ */ jsx12("defs", { children: /* @__PURE__ */ jsx12("clipPath", { id: "ala", children: /* @__PURE__ */ jsx12("path", { fill: "currentColor", d: "M0 0h20v20H0z" }) }) })
-      ]
-    }
-  );
-};
-
-// src/icon/icons/assets/index.react.tsx
-import { jsx as jsx13, jsxs as jsxs12 } from "react/jsx-runtime";
-var Assets = (props) => {
-  let _ariaLabel = props["aria-label"];
-  _ariaLabel = _ariaLabel && _ariaLabel.trim();
-  const _ariaHidden = !_ariaLabel;
-  const _role = _ariaHidden ? "presentation" : "img";
-  return /* @__PURE__ */ jsxs12(
-    "svg",
-    {
-      className: props.className,
-      viewBox: "0 0 20 20",
-      "aria-label": _ariaLabel,
-      "aria-hidden": _ariaHidden,
-      role: _role,
-      style: {
-        height: "var(--icon-height)",
-        width: "var(--icon-width)",
-        display: "flex",
-        alignItems: "center",
-        justifyContent: "center"
-      },
-      children: [
-        /* @__PURE__ */ jsxs12("g", { "clip-path": "url(#ama)", children: [
-          /* @__PURE__ */ jsx13("path", { fill: "currentColor", d: "M1.251 15c-.69 0-1.25-.56-1.25-1.25V1.25C.001.56.561 0 1.251 0h11.25c.69 0 1.25.56 1.25 1.25v12.5c0 .69-.56 1.25-1.25 1.25H1.251Zm0-1.25h11.25v-2.5H1.251v2.5ZM12.501 10v-.39l-3.17-3.623a.622.622 0 0 0-.983.053L5.576 10h6.925Zm-8.45 0 3.272-4.677a1.862 1.862 0 0 1 1.54-.801 1.857 1.857 0 0 1 1.41.642l2.229 2.548V1.25H1.252V10H4.05Z" }),
-          /* @__PURE__ */ jsx13("path", { fill: "currentColor", d: "M14.797 20c-.134 0-.267-.022-.395-.064l-8.97-2.991a.312.312 0 0 1-.197-.395l.198-.593a.313.313 0 0 1 .395-.198l8.97 2.991L18.75 6.892l-3.318-1.107a.625.625 0 0 1 .396-1.185l3.318 1.106c.316.106.573.328.722.626.15.3.174.639.068.955l-3.953 11.859a1.249 1.249 0 0 1-1.186.854ZM4.22 6.25a2.034 2.034 0 0 1-2.032-2.03c0-1.12.912-2.032 2.032-2.032s2.032.911 2.032 2.031S5.34 6.251 4.22 6.251Zm0-2.813a.783.783 0 0 0 0 1.564.783.783 0 0 0 0-1.564Z" })
-        ] }),
-        /* @__PURE__ */ jsx13("defs", { children: /* @__PURE__ */ jsx13("clipPath", { id: "ama", children: /* @__PURE__ */ jsx13("path", { fill: "currentColor", d: "M0 0h20v20H0z" }) }) })
-      ]
-    }
-  );
-};
-
-// src/icon/icons/audio_recording/index.react.tsx
-import { jsx as jsx14, jsxs as jsxs13 } from "react/jsx-runtime";
-var AudioRecording = (props) => {
-  let _ariaLabel = props["aria-label"];
-  _ariaLabel = _ariaLabel && _ariaLabel.trim();
-  const _ariaHidden = !_ariaLabel;
-  const _role = _ariaHidden ? "presentation" : "img";
-  return /* @__PURE__ */ jsxs13(
-    "svg",
-    {
-      className: props.className,
-      viewBox: "0 0 20 20",
-      "aria-label": _ariaLabel,
-      "aria-hidden": _ariaHidden,
-      role: _role,
-      style: {
-        height: "var(--icon-height)",
-        width: "var(--icon-width)",
-        display: "flex",
-        alignItems: "center",
-        justifyContent: "center"
-      },
-      children: [
-        /* @__PURE__ */ jsx14("path", { fill: "currentColor", d: "M7.5 20a.625.625 0 1 1 0-1.25h1.875V15H8.75a5.632 5.632 0 0 1-5.625-5.625v-2.5a.625.625 0 0 1 1.25 0v2.5A4.38 4.38 0 0 0 8.75 13.75h2.5a4.38 4.38 0 0 0 4.375-4.375v-2.5a.625.625 0 1 1 1.25 0v2.5A5.632 5.632 0 0 1 11.25 15h-.625v3.75H12.5a.624.624 0 1 1 0 1.25h-5Z" }),
-        /* @__PURE__ */ jsx14("path", { fill: "currentColor", d: "M8.75 12.5c-1.833-.11-3.226-1.668-3.124-3.473l-.001-5.59C5.523 1.661 6.908.112 8.712.002L11.25 0c1.833.11 3.226 1.668 3.124 3.473l.001 5.59c.102 1.776-1.283 3.326-3.088 3.436l-.037.001h-2.5ZM6.875 9.062c-.064 1.147.794 2.12 1.913 2.189l2.453-.001c1.094-.072 1.949-1.037 1.886-2.152V7.5H11.25a.625.625 0 1 1 0-1.25h1.875V5H11.25a.625.625 0 1 1 0-1.25h1.875v-.313c.064-1.146-.794-2.12-1.913-2.188L8.75 1.25c-1.09.067-1.94 1.033-1.876 2.152v.348H8.75a.625.625 0 0 1 0 1.25H6.875v1.25H8.75a.625.625 0 0 1 0 1.25H6.875v1.563Z" })
-      ]
-    }
-  );
-};
-
-// src/icon/icons/backspace/index.react.tsx
-import { jsx as jsx15, jsxs as jsxs14 } from "react/jsx-runtime";
-var Backspace = (props) => {
-  let _ariaLabel = props["aria-label"];
-  _ariaLabel = _ariaLabel && _ariaLabel.trim();
-  const _ariaHidden = !_ariaLabel;
-  const _role = _ariaHidden ? "presentation" : "img";
-  return /* @__PURE__ */ jsxs14(
-    "svg",
-    {
-      className: props.className,
-      viewBox: "0 0 20 20",
-      "aria-label": _ariaLabel,
-      "aria-hidden": _ariaHidden,
-      role: _role,
-      style: {
-        height: "var(--icon-height)",
-        width: "var(--icon-width)",
-        display: "flex",
-        alignItems: "center",
-        justifyContent: "center"
-      },
-      children: [
-        /* @__PURE__ */ jsxs14("g", { "clip-path": "url(#aoa)", children: [
-          /* @__PURE__ */ jsx15("path", { fill: "currentColor", d: "M3.75 19.373a3.128 3.128 0 0 1-3.125-3.125v-12.5A3.128 3.128 0 0 1 3.75.623h12.5a3.128 3.128 0 0 1 3.125 3.125v12.5a3.128 3.128 0 0 1-3.125 3.125H3.75Zm0-17.5a1.877 1.877 0 0 0-1.875 1.875v12.5a1.877 1.877 0 0 0 1.875 1.875h12.5a1.877 1.877 0 0 0 1.875-1.875v-12.5a1.876 1.876 0 0 0-1.875-1.875H3.75Z" }),
-          /* @__PURE__ */ jsx15("path", { fill: "currentColor", d: "M13.75 12.498a.617.617 0 0 1-.442-.184l-1.433-1.433-1.433 1.433a.622.622 0 0 1-1.066-.442c0-.165.065-.324.182-.442l1.434-1.433-1.434-1.434a.619.619 0 0 1-.183-.441c0-.167.065-.324.183-.442A.619.619 0 0 1 10 7.497c.167 0 .324.065.442.183l1.433 1.433 1.433-1.433a.62.62 0 0 1 .884 0 .619.619 0 0 1 .183.442.619.619 0 0 1-.183.441l-1.434 1.434 1.434 1.433a.627.627 0 0 1-.442 1.068Z" }),
-          /* @__PURE__ */ jsx15("path", { fill: "currentColor", d: "M9.267 15.625a1.858 1.858 0 0 1-1.325-.55l-4.437-4.18A1.256 1.256 0 0 1 3.22 9.52c.062-.152.154-.29.27-.406l4.466-4.207c.343-.341.81-.533 1.31-.533H15a1.877 1.877 0 0 1 1.875 1.875v7.5A1.877 1.877 0 0 1 15 15.625H9.267Zm-.455-1.446a.647.647 0 0 0 .455.196H15a.625.625 0 0 0 .625-.625v-7.5A.625.625 0 0 0 15 5.625H9.26a.63.63 0 0 0-.435.182l-4.45 4.19 4.437 4.182Z" })
-        ] }),
-        /* @__PURE__ */ jsx15("defs", { children: /* @__PURE__ */ jsx15("clipPath", { id: "aoa", children: /* @__PURE__ */ jsx15("path", { fill: "currentColor", d: "M0 0h20v20H0z" }) }) })
-      ]
-    }
-  );
-};
-
-// src/icon/icons/banned/index.react.tsx
-import { jsx as jsx16, jsxs as jsxs15 } from "react/jsx-runtime";
-var Banned = (props) => {
-  let _ariaLabel = props["aria-label"];
-  _ariaLabel = _ariaLabel && _ariaLabel.trim();
-  const _ariaHidden = !_ariaLabel;
-  const _role = _ariaHidden ? "presentation" : "img";
-  return /* @__PURE__ */ jsxs15(
-    "svg",
-    {
-      className: props.className,
-      viewBox: "0 0 20 20",
-      "aria-label": _ariaLabel,
-      "aria-hidden": _ariaHidden,
-      role: _role,
-      style: {
-        height: "var(--icon-height)",
-        width: "var(--icon-width)",
-        display: "flex",
-        alignItems: "center",
-        justifyContent: "center"
-      },
-      children: [
-        /* @__PURE__ */ jsx16("g", { "clip-path": "url(#apa)", children: /* @__PURE__ */ jsx16("path", { fill: "currentColor", d: "M10 20C4.486 20 0 15.514 0 10a9.933 9.933 0 0 1 2.89-7.025.348.348 0 0 1 .083-.082A9.93 9.93 0 0 1 10 0c5.514 0 10 4.486 10 10a9.933 9.933 0 0 1-2.89 7.025.304.304 0 0 1-.082.082A9.93 9.93 0 0 1 10 20ZM3.39 4.275A8.71 8.71 0 0 0 1.25 10c0 4.825 3.925 8.75 8.75 8.75a8.707 8.707 0 0 0 5.725-2.14L3.391 4.274Zm13.22 11.45A8.706 8.706 0 0 0 18.75 10c0-4.825-3.925-8.75-8.75-8.75a8.711 8.711 0 0 0-5.725 2.14l12.334 12.335Z" }) }),
-        /* @__PURE__ */ jsx16("defs", { children: /* @__PURE__ */ jsx16("clipPath", { id: "apa", children: /* @__PURE__ */ jsx16("path", { fill: "currentColor", d: "M0 0h20v20H0z" }) }) })
-      ]
-    }
-  );
-};
-
-// src/icon/icons/binocular/index.react.tsx
-import { jsx as jsx17, jsxs as jsxs16 } from "react/jsx-runtime";
-var Binocular = (props) => {
-  let _ariaLabel = props["aria-label"];
-  _ariaLabel = _ariaLabel && _ariaLabel.trim();
-  const _ariaHidden = !_ariaLabel;
-  const _role = _ariaHidden ? "presentation" : "img";
-  return /* @__PURE__ */ jsxs16(
-    "svg",
-    {
-      className: props.className,
-      viewBox: "0 0 20 20",
-      "aria-label": _ariaLabel,
-      "aria-hidden": _ariaHidden,
-      role: _role,
-      style: {
-        height: "var(--icon-height)",
-        width: "var(--icon-width)",
-        display: "flex",
-        alignItems: "center",
-        justifyContent: "center"
-      },
-      children: [
-        /* @__PURE__ */ jsx17("g", { "clip-path": "url(#aqa)", children: /* @__PURE__ */ jsx17("path", { fill: "currentColor", d: "M15.625 19.587a4.38 4.38 0 0 1-4.375-4.375v-4.375h-2.5v4.375a4.38 4.38 0 0 1-4.375 4.375A4.38 4.38 0 0 1 0 15.212c0-.355.048-.72.14-1.085a.636.636 0 0 1 .016-.077L3.225 2.9A2.82 2.82 0 0 1 5.937.833 2.817 2.817 0 0 1 8.749 3.65v3.437h2.5V3.649A2.816 2.816 0 0 1 14.057.833 2.822 2.822 0 0 1 16.773 2.9l3.07 11.147a.514.514 0 0 1 .017.09c.093.361.14.723.14 1.075a4.38 4.38 0 0 1-4.375 4.375Zm0-7.5a3.128 3.128 0 0 0-3.125 3.125 3.128 3.128 0 0 0 3.125 3.125 3.128 3.128 0 0 0 3.125-3.125 3.128 3.128 0 0 0-3.125-3.125Zm-11.25 0a3.128 3.128 0 0 0-3.125 3.125 3.128 3.128 0 0 0 3.125 3.125A3.128 3.128 0 0 0 7.5 15.212a3.128 3.128 0 0 0-3.125-3.125ZM14.06 2.083a1.555 1.555 0 0 0-1.103.46 1.551 1.551 0 0 0-.457 1.105v8.508a4.341 4.341 0 0 1 3.125-1.32c.78 0 1.538.21 2.203.601L15.57 3.232a1.567 1.567 0 0 0-1.507-1.149l-.002-.312v.312Zm-9.685 8.754A4.34 4.34 0 0 1 7.5 12.156V3.649c0-.417-.162-.81-.457-1.106a1.553 1.553 0 0 0-2.05-.14c-.273.208-.47.499-.561.829l-2.259 8.205a4.336 4.336 0 0 1 2.202-.6Zm6.875-1.25v-1.25h-2.5v1.25h2.5Z" }) }),
-        /* @__PURE__ */ jsx17("defs", { children: /* @__PURE__ */ jsx17("clipPath", { id: "aqa", children: /* @__PURE__ */ jsx17("path", { fill: "currentColor", d: "M0 0h20v20H0z" }) }) })
-      ]
-    }
-  );
-};
-
-// src/icon/icons/book_flip_page/index.react.tsx
-import { jsx as jsx18, jsxs as jsxs17 } from "react/jsx-runtime";
-var BookFlipPage = (props) => {
-  let _ariaLabel = props["aria-label"];
-  _ariaLabel = _ariaLabel && _ariaLabel.trim();
-  const _ariaHidden = !_ariaLabel;
-  const _role = _ariaHidden ? "presentation" : "img";
-  return /* @__PURE__ */ jsxs17(
-    "svg",
-    {
-      className: props.className,
-      viewBox: "0 0 20 20",
-      "aria-label": _ariaLabel,
-      "aria-hidden": _ariaHidden,
-      role: _role,
-      style: {
-        height: "var(--icon-height)",
-        width: "var(--icon-width)",
-        display: "flex",
-        alignItems: "center",
-        justifyContent: "center"
-      },
-      children: [
-        /* @__PURE__ */ jsx18("g", { "clip-path": "url(#arclip0_304_7)", children: /* @__PURE__ */ jsx18("path", { d: "M10 19.9992C9.91439 19.999 9.82974 19.9813 9.75121 19.9472C9.67269 19.9131 9.60195 19.8633 9.54333 19.8009C9.52917 19.7867 7.4675 17.8026 1.03333 17.7226C0.760332 17.7183 0.499744 17.6078 0.306835 17.4146C0.113925 17.2213 0.00386654 16.9606 0 16.6876L0 3.5409C0 3.26256 0.114167 2.99173 0.3125 2.79756C0.514167 2.60423 0.776667 2.49923 1.05417 2.49923C5.81917 2.56006 8.36917 3.58423 9.6025 4.33173C9.86208 3.50059 10.3104 2.74077 10.9125 2.11173C11.802 1.12675 12.9499 0.411026 14.2258 0.0458948C14.3789 0.000708485 14.5399 -0.011292 14.6979 0.0106952C14.856 0.0326823 15.0076 0.0881508 15.1425 0.173395C15.4367 0.355062 15.6175 0.673395 15.625 1.02339V2.73423C16.7237 2.58673 17.8306 2.50852 18.9392 2.50006C19.0769 2.49852 19.2137 2.52414 19.3416 2.57544C19.4694 2.62675 19.586 2.70274 19.6845 2.79909C19.783 2.89543 19.8616 3.01023 19.9157 3.13694C19.9698 3.26365 19.9985 3.39978 20 3.53756V16.6809C19.9968 16.9552 19.8868 17.2174 19.6934 17.4119C19.5 17.6064 19.2384 17.7178 18.9642 17.7226C12.5275 17.8026 10.47 19.7876 10.45 19.8076L10.4242 19.8326C10.3092 19.9401 10.1575 19.9997 10 19.9992ZM1.25 16.4759C5.63583 16.5567 8.11083 17.4709 9.375 18.1734V5.6884C8.775 5.2334 6.5275 3.85423 1.25 3.75256V16.4759ZM15.625 14.0376C15.6248 14.1812 15.5751 14.3204 15.4844 14.4318C15.3936 14.5432 15.2673 14.62 15.1267 14.6492C13.8568 14.91 12.7007 15.5623 11.8208 16.5142C11.3993 16.9533 11.073 17.4748 10.8625 18.0459C12.1792 17.3709 14.6142 16.5509 18.7492 16.4751V3.75173C17.7035 3.76877 16.6599 3.85003 15.6242 3.99506L15.625 14.0376ZM10.625 15.9826C10.7167 15.8701 10.8125 15.7609 10.9125 15.6576C11.8487 14.6466 13.0475 13.9158 14.375 13.5467V1.30756C13.3955 1.62652 12.517 2.19724 11.8275 2.96256C11.1711 3.64591 10.7499 4.52082 10.625 5.46006V15.9826Z", fill: "currentColor" }) }),
-        /* @__PURE__ */ jsx18("defs", { children: /* @__PURE__ */ jsx18("clipPath", { id: "arclip0_304_7", children: /* @__PURE__ */ jsx18("rect", { width: "20", height: "20", fill: "currentColor" }) }) })
-      ]
-    }
-  );
-};
-
-// src/icon/icons/book_open/index.react.tsx
-import { jsx as jsx19, jsxs as jsxs18 } from "react/jsx-runtime";
-var BookOpen = (props) => {
-  let _ariaLabel = props["aria-label"];
-  _ariaLabel = _ariaLabel && _ariaLabel.trim();
-  const _ariaHidden = !_ariaLabel;
-  const _role = _ariaHidden ? "presentation" : "img";
-  return /* @__PURE__ */ jsxs18(
-    "svg",
-    {
-      className: props.className,
-      viewBox: "0 0 20 20",
-      "aria-label": _ariaLabel,
-      "aria-hidden": _ariaHidden,
-      role: _role,
-      style: {
-        height: "var(--icon-height)",
-        width: "var(--icon-width)",
-        display: "flex",
-        alignItems: "center",
-        justifyContent: "center"
-      },
-      children: [
-        /* @__PURE__ */ jsx19("g", { "clip-path": "url(#asa)", children: /* @__PURE__ */ jsx19("path", { fill: "currentColor", d: "M9.968 18.427a1.923 1.923 0 0 1-.631-.114 10.767 10.767 0 0 0-3.976-.755c-.984 0-1.958.133-2.895.395A1.96 1.96 0 0 1 0 16.052V4.108a1.958 1.958 0 0 1 1.37-1.87 12.105 12.105 0 0 1 3.955-.66c1.53 0 3.028.283 4.452.84a.61.61 0 0 0 .155.03.605.605 0 0 1 .136 0 .686.686 0 0 0 .174-.035 12.092 12.092 0 0 1 4.43-.835c1.356 0 2.693.224 3.973.665A1.95 1.95 0 0 1 20 4.106v11.947a1.954 1.954 0 0 1-.963 1.692 1.96 1.96 0 0 1-1.49.212 10.783 10.783 0 0 0-2.909-.398c-1.367 0-2.699.252-3.956.748a1.99 1.99 0 0 1-.714.12Zm-4.61-2.12c1.383 0 2.731.23 4.017.686V3.601a10.954 10.954 0 0 0-7.612-.177.718.718 0 0 0-.513.683v11.944a.713.713 0 0 0 .892.693 11.971 11.971 0 0 1 3.216-.438Zm5.267.686a11.956 11.956 0 0 1 4.013-.685c1.1 0 2.187.148 3.233.441a.715.715 0 0 0 .878-.696V4.108a.71.71 0 0 0-.497-.678 10.963 10.963 0 0 0-7.627.172v13.39Z" }) }),
-        /* @__PURE__ */ jsx19("defs", { children: /* @__PURE__ */ jsx19("clipPath", { id: "asa", children: /* @__PURE__ */ jsx19("path", { fill: "currentColor", d: "M0 0h20v20H0z" }) }) })
-      ]
-    }
-  );
-};
-
-// src/icon/icons/bookmark/index.react.tsx
-import { jsx as jsx20 } from "react/jsx-runtime";
-var Bookmark = (props) => {
-  let _ariaLabel = props["aria-label"];
-  _ariaLabel = _ariaLabel && _ariaLabel.trim();
-  const _ariaHidden = !_ariaLabel;
-  const _role = _ariaHidden ? "presentation" : "img";
-  return /* @__PURE__ */ jsx20(
-    "svg",
-    {
-      className: props.className,
-      viewBox: "0 0 20 20",
-      "aria-label": _ariaLabel,
-      "aria-hidden": _ariaHidden,
-      role: _role,
-      style: {
-        height: "var(--icon-height)",
-        width: "var(--icon-width)",
-        display: "flex",
-        alignItems: "center",
-        justifyContent: "center"
-      },
-      children: /* @__PURE__ */ jsx20("path", { fill: "currentColor", d: "M11.875 19.75c-.234 0-.463-.066-.662-.19L7.5 17.238l-3.712 2.32a1.248 1.248 0 0 1-1.913-1.055V4.625A4.38 4.38 0 0 1 6.25.25h7.5a4.38 4.38 0 0 1 4.375 4.375V7.75c0 .69-.56 1.25-1.25 1.25h-3.75v9.5c0 .334-.13.648-.367.884a1.241 1.241 0 0 1-.883.366ZM7.5 15.875a.63.63 0 0 1 .332.095l4.044 2.53-.001-13.875A1.876 1.876 0 0 1 13.75 2.75a.625.625 0 1 1 0 1.25.625.625 0 0 0-.625.625V7.75h3.75V4.625A3.128 3.128 0 0 0 13.75 1.5h-7.5a3.128 3.128 0 0 0-3.125 3.125V18.5l4.044-2.53a.621.621 0 0 1 .331-.095Z" })
-    }
-  );
-};
-
-// src/icon/icons/bullet_list/index.react.tsx
-import { jsx as jsx21, jsxs as jsxs19 } from "react/jsx-runtime";
-var BulletList = (props) => {
-  let _ariaLabel = props["aria-label"];
-  _ariaLabel = _ariaLabel && _ariaLabel.trim();
-  const _ariaHidden = !_ariaLabel;
-  const _role = _ariaHidden ? "presentation" : "img";
-  return /* @__PURE__ */ jsxs19(
-    "svg",
-    {
-      className: props.className,
-      viewBox: "0 0 20 20",
-      "aria-label": _ariaLabel,
-      "aria-hidden": _ariaHidden,
-      role: _role,
-      style: {
-        height: "var(--icon-height)",
-        width: "var(--icon-width)",
-        display: "flex",
-        alignItems: "center",
-        justifyContent: "center"
-      },
-      children: [
-        /* @__PURE__ */ jsx21("g", { "clip-path": "url(#aua)", children: /* @__PURE__ */ jsx21("path", { fill: "currentColor", d: "M2.5 4.998a2.503 2.503 0 0 1-2.5-2.5c0-1.378 1.122-2.5 2.5-2.5S5 1.12 5 2.498c0 1.379-1.122 2.5-2.5 2.5Zm0-3.75a1.251 1.251 0 0 0 0 2.5 1.25 1.25 0 0 0 0-2.5Zm0 11.25a2.503 2.503 0 0 1-2.5-2.5c0-1.378 1.122-2.5 2.5-2.5S5 8.62 5 9.998c0 1.379-1.122 2.5-2.5 2.5Zm0-3.75a1.251 1.251 0 0 0 0 2.5 1.25 1.25 0 0 0 0-2.5Zm0 11.25a2.503 2.503 0 0 1-2.5-2.5c0-1.378 1.122-2.5 2.5-2.5S5 16.12 5 17.498c0 1.379-1.122 2.5-2.5 2.5Zm0-3.75a1.251 1.251 0 0 0 0 2.5 1.25 1.25 0 0 0 0-2.5Zm4.375-12.5a.625.625 0 0 1 0-1.25h12.5a.625.625 0 1 1 0 1.25h-12.5Zm0 7.5a.625.625 0 0 1 0-1.25h12.5a.625.625 0 1 1 0 1.25h-12.5Zm0 7.5a.625.625 0 0 1 0-1.25h12.5a.625.625 0 1 1 0 1.25h-12.5Z" }) }),
-        /* @__PURE__ */ jsx21("defs", { children: /* @__PURE__ */ jsx21("clipPath", { id: "aua", children: /* @__PURE__ */ jsx21("path", { fill: "currentColor", d: "M0 0h20v20H0z" }) }) })
-      ]
-    }
-  );
-};
-
-// src/icon/icons/calendar/index.react.tsx
-import { jsx as jsx22, jsxs as jsxs20 } from "react/jsx-runtime";
-var Calendar = (props) => {
-  let _ariaLabel = props["aria-label"];
-  _ariaLabel = _ariaLabel && _ariaLabel.trim();
-  const _ariaHidden = !_ariaLabel;
-  const _role = _ariaHidden ? "presentation" : "img";
-  return /* @__PURE__ */ jsxs20(
-    "svg",
-    {
-      className: props.className,
-      viewBox: "0 0 20 20",
-      "aria-label": _ariaLabel,
-      "aria-hidden": _ariaHidden,
-      role: _role,
-      style: {
-        height: "var(--icon-height)",
-        width: "var(--icon-width)",
-        display: "flex",
-        alignItems: "center",
-        justifyContent: "center"
-      },
-      children: [
-        /* @__PURE__ */ jsx22("g", { "clip-path": "url(#ava)", children: /* @__PURE__ */ jsx22("path", { fill: "currentColor", d: "M1.877 20a1.876 1.876 0 0 1-1.875-1.875V4.375A1.877 1.877 0 0 1 1.877 2.5h3.125V.625a.625.625 0 0 1 1.25 0V2.5h7.5V.625a.625.625 0 1 1 1.25 0V2.5h3.125a1.877 1.877 0 0 1 1.875 1.875v13.75A1.877 1.877 0 0 1 18.127 20H1.877Zm-.625-1.875c0 .345.28.625.625.625h16.25a.625.625 0 0 0 .625-.625V8.75h-17.5v9.375ZM18.752 7.5V4.375a.625.625 0 0 0-.625-.625h-3.125V5a.625.625 0 0 1-1.25 0V3.75h-7.5V5a.625.625 0 0 1-1.25 0V3.75H1.877a.625.625 0 0 0-.625.625V7.5h17.5Z" }) }),
-        /* @__PURE__ */ jsx22("defs", { children: /* @__PURE__ */ jsx22("clipPath", { id: "ava", children: /* @__PURE__ */ jsx22("path", { fill: "currentColor", d: "M0 0h20v20H0z" }) }) })
-      ]
-    }
-  );
-};
-
-// src/icon/icons/camera/index.react.tsx
-import { jsx as jsx23, jsxs as jsxs21 } from "react/jsx-runtime";
-var Camera = (props) => {
-  let _ariaLabel = props["aria-label"];
-  _ariaLabel = _ariaLabel && _ariaLabel.trim();
-  const _ariaHidden = !_ariaLabel;
-  const _role = _ariaHidden ? "presentation" : "img";
-  return /* @__PURE__ */ jsxs21(
-    "svg",
-    {
-      className: props.className,
-      viewBox: "0 0 20 20",
-      "aria-label": _ariaLabel,
-      "aria-hidden": _ariaHidden,
-      role: _role,
-      style: {
-        height: "var(--icon-height)",
-        width: "var(--icon-width)",
-        display: "flex",
-        alignItems: "center",
-        justifyContent: "center"
-      },
-      children: [
-        /* @__PURE__ */ jsx23("path", { fill: "currentColor", d: "M11.875 15.625c-2.757 0-5-2.243-5-5s2.243-5 5-5 5 2.243 5 5-2.243 5-5 5Zm0-8.75a3.754 3.754 0 0 0-3.75 3.75 3.755 3.755 0 0 0 3.75 3.75 3.755 3.755 0 0 0 3.75-3.75 3.755 3.755 0 0 0-3.75-3.75Z" }),
-        /* @__PURE__ */ jsx23("path", { fill: "currentColor", d: "M1.875 18.125A1.876 1.876 0 0 1 0 16.25v-10a1.877 1.877 0 0 1 1.875-1.875H3.75V3.75a.625.625 0 0 1 1.25 0v.625h2.113l.732-1.463a1.867 1.867 0 0 1 1.678-1.037h4.704c.716 0 1.358.397 1.678 1.037l.732 1.463h1.488A1.877 1.877 0 0 1 20 6.25v10a1.876 1.876 0 0 1-1.875 1.875H1.875Zm0-12.5a.625.625 0 0 0-.625.625v10c0 .345.28.625.625.625h16.25a.624.624 0 0 0 .625-.625v-10a.625.625 0 0 0-.625-.625H16.25a.621.621 0 0 1-.56-.346l-.903-1.809a.623.623 0 0 0-.56-.345H9.523a.622.622 0 0 0-.56.346L8.06 5.279a.622.622 0 0 1-.559.346H1.875Z" }),
-        /* @__PURE__ */ jsx23("path", { fill: "currentColor", d: "M3.438 8.75a.937.937 0 1 0 0-1.875.937.937 0 0 0 0 1.875Z" })
-      ]
-    }
-  );
-};
-
-// src/icon/icons/chat/index.react.tsx
-import { jsx as jsx24, jsxs as jsxs22 } from "react/jsx-runtime";
-var Chat = (props) => {
-  let _ariaLabel = props["aria-label"];
-  _ariaLabel = _ariaLabel && _ariaLabel.trim();
-  const _ariaHidden = !_ariaLabel;
-  const _role = _ariaHidden ? "presentation" : "img";
-  return /* @__PURE__ */ jsxs22(
-    "svg",
-    {
-      className: props.className,
-      viewBox: "0 0 20 20",
-      "aria-label": _ariaLabel,
-      "aria-hidden": _ariaHidden,
-      role: _role,
-      style: {
-        height: "var(--icon-height)",
-        width: "var(--icon-width)",
-        display: "flex",
-        alignItems: "center",
-        justifyContent: "center"
-      },
-      children: [
-        /* @__PURE__ */ jsxs22("g", { "clip-path": "url(#axa)", children: [
-          /* @__PURE__ */ jsx24("path", { fill: "currentColor", d: "M16.872 20a.62.62 0 0 1-.441-.183l-3.567-3.567h-.366c-2.756 0-5-2.243-5-5s2.244-5 5-5h2.5c2.758 0 5 2.243 5 5a4.978 4.978 0 0 1-2.5 4.316v3.809a.625.625 0 0 1-.626.625ZM12.498 7.5a3.754 3.754 0 0 0-3.75 3.75 3.755 3.755 0 0 0 3.75 3.75h.625c.165 0 .325.067.441.183l2.683 2.684v-2.678a.63.63 0 0 1 .358-.565 3.759 3.759 0 0 0 2.142-3.374A3.753 3.753 0 0 0 15 7.5h-2.502Z" }),
-          /* @__PURE__ */ jsx24("path", { fill: "currentColor", d: "M3.122 13.75a.623.623 0 0 1-.625-.625V9.317A4.973 4.973 0 0 1 .29 6.674 4.957 4.957 0 0 1 .475 2.86 5.012 5.012 0 0 1 4.987 0h2.51a5.033 5.033 0 0 1 4.794 3.572.625.625 0 0 1-1.198.356A3.776 3.776 0 0 0 7.497 1.25h-2.5a3.76 3.76 0 0 0-3.392 2.143 3.717 3.717 0 0 0-.14 2.86 3.717 3.717 0 0 0 1.924 2.12.628.628 0 0 1 .358.565v2.679l1.434-1.434a.622.622 0 0 1 1.066.442.627.627 0 0 1-.183.443l-2.5 2.5a.62.62 0 0 1-.442.182Z" })
-        ] }),
-        /* @__PURE__ */ jsx24("defs", { children: /* @__PURE__ */ jsx24("clipPath", { id: "axa", children: /* @__PURE__ */ jsx24("path", { fill: "currentColor", d: "M0 0h20v20H0z" }) }) })
-      ]
-    }
-  );
-};
-
-// src/icon/icons/check/index.react.tsx
-import { jsx as jsx25 } from "react/jsx-runtime";
-var Check = (props) => {
-  let _ariaLabel = props["aria-label"];
-  _ariaLabel = _ariaLabel && _ariaLabel.trim();
-  const _ariaHidden = !_ariaLabel;
-  const _role = _ariaHidden ? "presentation" : "img";
-  return /* @__PURE__ */ jsx25(
-    "svg",
-    {
-      className: props.className,
-      viewBox: "0 0 20 20",
-      "aria-label": _ariaLabel,
-      "aria-hidden": _ariaHidden,
-      role: _role,
-      style: {
-        height: "var(--icon-height)",
-        width: "var(--icon-width)",
-        display: "flex",
-        alignItems: "center",
-        justifyContent: "center"
-      },
-      children: /* @__PURE__ */ jsx25("path", { fill: "currentColor", d: "M7.75 14.917h-.333c-.417-.084-.75-.334-.917-.667l-2-2.917c-.167-.25-.167-.666.167-.833.083-.083.25-.083.333-.083.167 0 .417.083.5.25l2.083 2.916c.084.084.084.084.167.084h.083c.084 0 .084 0 .167-.084 0 0 .083 0 .083-.083l6.584-8.25c.083-.167.333-.25.5-.25.166 0 .25.083.416.167.167.083.25.25.25.416 0 .167 0 .334-.166.5l-6.5 8.25-.334.334c-.5.166-.75.25-1.083.25Z" })
-    }
-  );
-};
-
-// src/icon/icons/check_circle/index.react.tsx
-import { jsx as jsx26, jsxs as jsxs23 } from "react/jsx-runtime";
-var CheckCircle = (props) => {
-  let _ariaLabel = props["aria-label"];
-  _ariaLabel = _ariaLabel && _ariaLabel.trim();
-  const _ariaHidden = !_ariaLabel;
-  const _role = _ariaHidden ? "presentation" : "img";
-  return /* @__PURE__ */ jsxs23(
-    "svg",
-    {
-      className: props.className,
-      viewBox: "0 0 20 20",
-      "aria-label": _ariaLabel,
-      "aria-hidden": _ariaHidden,
-      role: _role,
-      style: {
-        height: "var(--icon-height)",
-        width: "var(--icon-width)",
-        display: "flex",
-        alignItems: "center",
-        justifyContent: "center"
-      },
-      children: [
-        /* @__PURE__ */ jsxs23("g", { "clip-path": "url(#aza)", children: [
-          /* @__PURE__ */ jsx26("path", { fill: "currentColor", d: "M7.77 14.93a1.49 1.49 0 0 1-1.248-.666L4.49 11.38a.625.625 0 1 1 1.023-.72l2.041 2.898a.255.255 0 0 0 .218.123.245.245 0 0 0 .196-.1l6.544-8.278a.626.626 0 1 1 .979.775l-6.536 8.269a1.483 1.483 0 0 1-1.183.584Z" }),
-          /* @__PURE__ */ jsx26("path", { fill: "currentColor", d: "M10 19.999c-5.514 0-10-4.486-10-10s4.486-10 10-10 10 4.486 10 10-4.486 10-10 10Zm0-18.75c-4.825 0-8.75 3.925-8.75 8.75s3.925 8.75 8.75 8.75 8.75-3.925 8.75-8.75-3.925-8.75-8.75-8.75Z" })
-        ] }),
-        /* @__PURE__ */ jsx26("defs", { children: /* @__PURE__ */ jsx26("clipPath", { id: "aza", children: /* @__PURE__ */ jsx26("path", { fill: "currentColor", d: "M0 0h20v20H0z" }) }) })
-      ]
-    }
-  );
-};
-
-// src/icon/icons/checklist/index.react.tsx
-import { jsx as jsx27, jsxs as jsxs24 } from "react/jsx-runtime";
-var Checklist = (props) => {
-  let _ariaLabel = props["aria-label"];
-  _ariaLabel = _ariaLabel && _ariaLabel.trim();
-  const _ariaHidden = !_ariaLabel;
-  const _role = _ariaHidden ? "presentation" : "img";
-  return /* @__PURE__ */ jsxs24(
-    "svg",
-    {
-      className: props.className,
-      viewBox: "0 0 20 20",
-      "aria-label": _ariaLabel,
-      "aria-hidden": _ariaHidden,
-      role: _role,
-      style: {
-        height: "var(--icon-height)",
-        width: "var(--icon-width)",
-        display: "flex",
-        alignItems: "center",
-        justifyContent: "center"
-      },
-      children: [
-        /* @__PURE__ */ jsxs24("g", { "clip-path": "url(#baa)", children: [
-          /* @__PURE__ */ jsx27("path", { fill: "currentColor", d: "M1.875 19.999A1.877 1.877 0 0 1 0 18.124V1.874A1.877 1.877 0 0 1 1.875-.001h16.25A1.877 1.877 0 0 1 20 1.874v16.25a1.876 1.876 0 0 1-1.875 1.875H1.875Zm0-18.75a.625.625 0 0 0-.625.625v16.25c0 .345.28.625.625.625h16.25a.625.625 0 0 0 .625-.625V1.874a.625.625 0 0 0-.625-.625H1.875Z" }),
-          /* @__PURE__ */ jsx27("path", { fill: "currentColor", d: "M6.25 9.374a.62.62 0 0 1-.442-.183l-2.5-2.5a.619.619 0 0 1-.183-.442c0-.167.065-.324.183-.441a.619.619 0 0 1 .442-.184c.167 0 .324.065.442.184l1.99 1.99L9.5 3.373a.628.628 0 0 1 1.12.287.622.622 0 0 1-.12.463l-3.75 5a.625.625 0 0 1-.456.248l-.044.002ZM11.875 7.5a.625.625 0 1 1 0-1.25h3.75a.625.625 0 1 1 0 1.25h-3.75ZM6.25 16.874a.62.62 0 0 1-.442-.183l-2.5-2.5a.627.627 0 0 1 .442-1.068c.167 0 .324.065.442.184l1.99 1.99L9.5 10.873a.63.63 0 0 1 1.038.057.625.625 0 0 1-.038.693l-3.75 5a.625.625 0 0 1-.5.251ZM11.875 15a.625.625 0 1 1 0-1.25h3.75a.625.625 0 0 1 0 1.25h-3.75Z" })
-        ] }),
-        /* @__PURE__ */ jsx27("defs", { children: /* @__PURE__ */ jsx27("clipPath", { id: "baa", children: /* @__PURE__ */ jsx27("path", { fill: "currentColor", d: "M0 0h20v20H0z" }) }) })
-      ]
-    }
-  );
-};
-
-// src/icon/icons/chevron_down/index.react.tsx
-import { jsx as jsx28, jsxs as jsxs25 } from "react/jsx-runtime";
-var ChevronDown = (props) => {
-  let _ariaLabel = props["aria-label"];
-  _ariaLabel = _ariaLabel && _ariaLabel.trim();
-  const _ariaHidden = !_ariaLabel;
-  const _role = _ariaHidden ? "presentation" : "img";
-  return /* @__PURE__ */ jsxs25(
-    "svg",
-    {
-      className: props.className,
-      viewBox: "0 0 20 20",
-      "aria-label": _ariaLabel,
-      "aria-hidden": _ariaHidden,
-      role: _role,
-      style: {
-        height: "var(--icon-height)",
-        width: "var(--icon-width)",
-        display: "flex",
-        alignItems: "center",
-        justifyContent: "center"
-      },
-      children: [
-        /* @__PURE__ */ jsx28("g", { "clip-path": "url(#bba)", children: /* @__PURE__ */ jsx28("path", { fill: "currentColor", d: "M10 15.835c-.334 0-.647-.13-.883-.365L.183 6.534a.62.62 0 0 1 0-.884.62.62 0 0 1 .884 0l8.932 8.935 8.933-8.935a.62.62 0 0 1 .883 0 .619.619 0 0 1 0 .884l-8.933 8.935a1.242 1.242 0 0 1-.883.366Z" }) }),
-        /* @__PURE__ */ jsx28("defs", { children: /* @__PURE__ */ jsx28("clipPath", { id: "bba", children: /* @__PURE__ */ jsx28("path", { fill: "currentColor", d: "M0 0h20v20H0z" }) }) })
-      ]
-    }
-  );
-};
-
-// src/icon/icons/chevron_up/index.react.tsx
-import { jsx as jsx29, jsxs as jsxs26 } from "react/jsx-runtime";
-var ChevronUp = (props) => {
-  let _ariaLabel = props["aria-label"];
-  _ariaLabel = _ariaLabel && _ariaLabel.trim();
-  const _ariaHidden = !_ariaLabel;
-  const _role = _ariaHidden ? "presentation" : "img";
-  return /* @__PURE__ */ jsxs26(
-    "svg",
-    {
-      className: props.className,
-      viewBox: "0 0 20 20",
-      "aria-label": _ariaLabel,
-      "aria-hidden": _ariaHidden,
-      role: _role,
-      style: {
-        height: "var(--icon-height)",
-        width: "var(--icon-width)",
-        display: "flex",
-        alignItems: "center",
-        justifyContent: "center"
-      },
-      children: [
-        /* @__PURE__ */ jsx29("g", { "clip-path": "url(#bca)", children: /* @__PURE__ */ jsx29("path", { fill: "currentColor", d: "M.625 14.953a.627.627 0 0 1-.442-1.068L9.12 4.95a1.243 1.243 0 0 1 1.766-.002l8.936 8.937a.627.627 0 0 1-.681 1.021.619.619 0 0 1-.203-.136l-8.935-8.936-8.935 8.936a.618.618 0 0 1-.442.183Z" }) }),
-        /* @__PURE__ */ jsx29("defs", { children: /* @__PURE__ */ jsx29("clipPath", { id: "bca", children: /* @__PURE__ */ jsx29("path", { fill: "currentColor", d: "M0 0h20v20H0z" }) }) })
-      ]
-    }
-  );
-};
-
-// src/icon/icons/circle/index.react.tsx
-import { jsx as jsx30, jsxs as jsxs27 } from "react/jsx-runtime";
-var Circle = (props) => {
-  let _ariaLabel = props["aria-label"];
-  _ariaLabel = _ariaLabel && _ariaLabel.trim();
-  const _ariaHidden = !_ariaLabel;
-  const _role = _ariaHidden ? "presentation" : "img";
-  return /* @__PURE__ */ jsxs27(
-    "svg",
-    {
-      className: props.className,
-      viewBox: "0 0 20 20",
-      "aria-label": _ariaLabel,
-      "aria-hidden": _ariaHidden,
-      role: _role,
-      style: {
-        height: "var(--icon-height)",
-        width: "var(--icon-width)",
-        display: "flex",
-        alignItems: "center",
-        justifyContent: "center"
-      },
-      children: [
-        /* @__PURE__ */ jsx30("g", { "clip-path": "url(#bda)", children: /* @__PURE__ */ jsx30("path", { fill: "currentColor", d: "M10 20C4.486 20 0 15.514 0 10S4.486 0 10 0s10 4.486 10 10-4.486 10-10 10Zm0-18.75c-4.825 0-8.75 3.925-8.75 8.75s3.925 8.75 8.75 8.75 8.75-3.925 8.75-8.75S14.825 1.25 10 1.25Z" }) }),
-        /* @__PURE__ */ jsx30("defs", { children: /* @__PURE__ */ jsx30("clipPath", { id: "bda", children: /* @__PURE__ */ jsx30("path", { fill: "currentColor", d: "M0 0h20v20H0z" }) }) })
-      ]
-    }
-  );
-};
-
-// src/icon/icons/circle_shape_add/index.react.tsx
-import { jsx as jsx31, jsxs as jsxs28 } from "react/jsx-runtime";
-var CircleShapeAdd = (props) => {
-  let _ariaLabel = props["aria-label"];
-  _ariaLabel = _ariaLabel && _ariaLabel.trim();
-  const _ariaHidden = !_ariaLabel;
-  const _role = _ariaHidden ? "presentation" : "img";
-  return /* @__PURE__ */ jsxs28(
-    "svg",
-    {
-      className: props.className,
-      viewBox: "0 0 20 20",
-      "aria-label": _ariaLabel,
-      "aria-hidden": _ariaHidden,
-      role: _role,
-      style: {
-        height: "var(--icon-height)",
-        width: "var(--icon-width)",
-        display: "flex",
-        alignItems: "center",
-        justifyContent: "center"
-      },
-      children: [
-        /* @__PURE__ */ jsx31("g", { "clip-path": "url(#bea)", children: /* @__PURE__ */ jsx31("path", { fill: "currentColor", d: "M16.872 19.997a.624.624 0 0 1-.625-.625v-1.875h-1.875a.624.624 0 1 1 0-1.25h1.875v-1.875a.625.625 0 0 1 1.25 0v1.875h1.875a.625.625 0 0 1 0 1.25h-1.875v1.875a.624.624 0 0 1-.625.625ZM8.75 17.495a8.729 8.729 0 0 1-2.912-.497A8.687 8.687 0 0 1 .86 12.523 8.687 8.687 0 0 1 .504 5.84 8.689 8.689 0 0 1 4.98.861a8.662 8.662 0 0 1 3.766-.86c.99 0 1.973.17 2.918.503a8.692 8.692 0 0 1 4.979 4.475 8.687 8.687 0 0 1 .356 6.685.625.625 0 0 1-1.18-.417 7.449 7.449 0 0 0-.305-5.73 7.446 7.446 0 0 0-4.266-3.834 7.43 7.43 0 0 0-5.73.306 7.442 7.442 0 0 0-3.834 4.266 7.448 7.448 0 0 0 .305 5.73 7.448 7.448 0 0 0 4.267 3.834c.8.283 1.64.426 2.496.426.855 0 1.695-.143 2.496-.426a.626.626 0 0 1 .416 1.18c-.936.33-1.92.497-2.912.496Z" }) }),
-        /* @__PURE__ */ jsx31("defs", { children: /* @__PURE__ */ jsx31("clipPath", { id: "bea", children: /* @__PURE__ */ jsx31("path", { fill: "currentColor", d: "M0 0h20v20H0z" }) }) })
-      ]
-    }
-  );
-};
-
-// src/icon/icons/clock/index.react.tsx
-import { jsx as jsx32, jsxs as jsxs29 } from "react/jsx-runtime";
-var Clock = (props) => {
-  let _ariaLabel = props["aria-label"];
-  _ariaLabel = _ariaLabel && _ariaLabel.trim();
-  const _ariaHidden = !_ariaLabel;
-  const _role = _ariaHidden ? "presentation" : "img";
-  return /* @__PURE__ */ jsxs29(
-    "svg",
-    {
-      className: props.className,
-      viewBox: "0 0 20 20",
-      "aria-label": _ariaLabel,
-      "aria-hidden": _ariaHidden,
-      role: _role,
-      style: {
-        height: "var(--icon-height)",
-        width: "var(--icon-width)",
-        display: "flex",
-        alignItems: "center",
-        justifyContent: "center"
-      },
-      children: [
-        /* @__PURE__ */ jsxs29("g", { "clip-path": "url(#bfa)", children: [
-          /* @__PURE__ */ jsx32("path", { fill: "currentColor", d: "M10 19.375C4.83 19.375.625 15.169.625 10 .625 4.83 4.831.625 10 .625c5.17 0 9.375 4.206 9.375 9.375 0 5.17-4.206 9.375-9.375 9.375Zm0-17.5C5.52 1.875 1.875 5.52 1.875 10S5.52 18.125 10 18.125 18.125 14.48 18.125 10 14.48 1.875 10 1.875Z" }),
-          /* @__PURE__ */ jsx32("path", { fill: "currentColor", d: "M13.906 14.532a.619.619 0 0 1-.442-.184l-3.906-3.906a.618.618 0 0 1-.136-.204l-.007-.021a.614.614 0 0 1-.04-.217V6.875a.625.625 0 0 1 1.25 0v2.866l3.723 3.723a.62.62 0 0 1 0 .883.626.626 0 0 1-.442.185Z" })
-        ] }),
-        /* @__PURE__ */ jsx32("defs", { children: /* @__PURE__ */ jsx32("clipPath", { id: "bfa", children: /* @__PURE__ */ jsx32("path", { fill: "currentColor", d: "M0 0h20v20H0z" }) }) })
-      ]
-    }
-  );
-};
-
-// src/icon/icons/clock_hand/index.react.tsx
-import { jsx as jsx33, jsxs as jsxs30 } from "react/jsx-runtime";
-var ClockHand = (props) => {
-  let _ariaLabel = props["aria-label"];
-  _ariaLabel = _ariaLabel && _ariaLabel.trim();
-  const _ariaHidden = !_ariaLabel;
-  const _role = _ariaHidden ? "presentation" : "img";
-  return /* @__PURE__ */ jsxs30(
-    "svg",
-    {
-      className: props.className,
-      viewBox: "0 0 20 20",
-      "aria-label": _ariaLabel,
-      "aria-hidden": _ariaHidden,
-      role: _role,
-      style: {
-        height: "var(--icon-height)",
-        width: "var(--icon-width)",
-        display: "flex",
-        alignItems: "center",
-        justifyContent: "center"
-      },
-      children: [
-        /* @__PURE__ */ jsxs30("g", { "clip-path": "url(#bga)", children: [
-          /* @__PURE__ */ jsx33("path", { fill: "currentColor", d: "M.625 20.003A.625.625 0 0 1 0 19.378v-7.5a.625.625 0 0 1 1.25 0v.625H5c1.5 0 2.77 1.057 3.063 2.5h2.562a3.128 3.128 0 0 1 3.125 3.125.625.625 0 0 1-.625.625H1.25v.625a.625.625 0 0 1-.625.625Zm11.768-2.5a1.886 1.886 0 0 0-1.768-1.25H5a.625.625 0 1 1 0-1.25h1.765A1.885 1.885 0 0 0 5 13.753H1.25v3.75h11.143Zm1.791-2.708a.624.624 0 0 1-.153-1.23 6.258 6.258 0 0 0 4.01-8.95 6.239 6.239 0 0 0-5.553-3.36 6.258 6.258 0 0 0-2.879.708 6.258 6.258 0 0 0-2.651 8.432.62.62 0 0 1 .041.477.62.62 0 0 1-.92.347.626.626 0 0 1-.23-.246C3.936 7.306 5.363 2.767 9.03.855a7.51 7.51 0 0 1 7.352.232 7.487 7.487 0 0 1 2.766 2.95 7.45 7.45 0 0 1 .501 5.718 7.45 7.45 0 0 1-3.683 4.4 7.504 7.504 0 0 1-1.63.622.663.663 0 0 1-.153.018Z" }),
-          /* @__PURE__ */ jsx33("path", { fill: "currentColor", d: "M11.875 8.753a.625.625 0 0 1-.625-.625v-5a.625.625 0 1 1 1.25 0v4.375H15a.625.625 0 1 1 0 1.25h-3.125Z" })
-        ] }),
-        /* @__PURE__ */ jsx33("defs", { children: /* @__PURE__ */ jsx33("clipPath", { id: "bga", children: /* @__PURE__ */ jsx33("path", { fill: "currentColor", d: "M0 0h20v20H0z" }) }) })
-      ]
-    }
-  );
-};
-
-// src/icon/icons/close/index.react.tsx
-import { jsx as jsx34, jsxs as jsxs31 } from "react/jsx-runtime";
-var Close = (props) => {
-  let _ariaLabel = props["aria-label"];
-  _ariaLabel = _ariaLabel && _ariaLabel.trim();
-  const _ariaHidden = !_ariaLabel;
-  const _role = _ariaHidden ? "presentation" : "img";
-  return /* @__PURE__ */ jsxs31(
-    "svg",
-    {
-      className: props.className,
-      viewBox: "0 0 20 20",
-      "aria-label": _ariaLabel,
-      "aria-hidden": _ariaHidden,
-      role: _role,
-      style: {
-        height: "var(--icon-height)",
-        width: "var(--icon-width)",
-        display: "flex",
-        alignItems: "center",
-        justifyContent: "center"
-      },
-      children: [
-        /* @__PURE__ */ jsx34("g", { "clip-path": "url(#bha)", children: /* @__PURE__ */ jsx34("path", { fill: "currentColor", d: "M19.375 20a.619.619 0 0 1-.442-.184L10 10.883l-8.933 8.933a.62.62 0 0 1-.884 0 .627.627 0 0 1 0-.884l8.934-8.934L.183 1.066A.619.619 0 0 1 0 .624C0 .457.065.3.183.182A.619.619 0 0 1 .625 0c.167 0 .324.065.442.183L10 9.115 18.933.182a.62.62 0 0 1 .884 0 .622.622 0 0 1 0 .884l-8.934 8.933 8.934 8.934a.627.627 0 0 1-.442 1.066Z" }) }),
-        /* @__PURE__ */ jsx34("defs", { children: /* @__PURE__ */ jsx34("clipPath", { id: "bha", children: /* @__PURE__ */ jsx34("path", { fill: "currentColor", d: "M0 0h20v20H0z" }) }) })
-      ]
-    }
-  );
-};
-
-// src/icon/icons/color_brush/index.react.tsx
-import { jsx as jsx35, jsxs as jsxs32 } from "react/jsx-runtime";
-var ColorBrush = (props) => {
-  let _ariaLabel = props["aria-label"];
-  _ariaLabel = _ariaLabel && _ariaLabel.trim();
-  const _ariaHidden = !_ariaLabel;
-  const _role = _ariaHidden ? "presentation" : "img";
-  return /* @__PURE__ */ jsxs32(
-    "svg",
-    {
-      className: props.className,
-      viewBox: "0 0 20 20",
-      "aria-label": _ariaLabel,
-      "aria-hidden": _ariaHidden,
-      role: _role,
-      style: {
-        height: "var(--icon-height)",
-        width: "var(--icon-width)",
-        display: "flex",
-        alignItems: "center",
-        justifyContent: "center"
-      },
-      children: [
-        /* @__PURE__ */ jsx35("g", { "clip-path": "url(#bia)", children: /* @__PURE__ */ jsx35("path", { fill: "currentColor", d: "M2.09 19.856c-.46 0-.92-.04-1.374-.12a.62.62 0 0 1-.503-.754.622.622 0 0 1 .159-.293c.006-.007 1.017-1.149 1.128-3.936a3.785 3.785 0 0 1 1.086-2.699 3.762 3.762 0 0 1 3.557-.893l8.491-9.987A2.918 2.918 0 0 1 16.867.145a2.914 2.914 0 0 1 2.236 1.032 2.91 2.91 0 0 1 .69 2.136 2.91 2.91 0 0 1-1.022 2L8.79 13.85a3.762 3.762 0 0 1-.883 3.484 8.014 8.014 0 0 1-5.816 2.52Zm-.2-1.254a6.757 6.757 0 0 0 5.103-2.122 2.508 2.508 0 0 0-.001-3.407 2.508 2.508 0 0 0-1.83-.794c-.635 0-1.24.238-1.704.67a2.547 2.547 0 0 0-.709 1.817c-.073 1.863-.5 3.106-.858 3.836Zm5.498-6.975c.374.261.695.587.952.964l1.956-1.673L9.05 9.671l-1.662 1.955Zm3.861-1.525 6.71-5.74c.343-.293.552-.7.588-1.149a1.674 1.674 0 0 0-.49-1.326 1.673 1.673 0 0 0-1.19-.493 1.671 1.671 0 0 0-1.281.59L9.862 8.716l1.387 1.386Z" }) }),
-        /* @__PURE__ */ jsx35("defs", { children: /* @__PURE__ */ jsx35("clipPath", { id: "bia", children: /* @__PURE__ */ jsx35("path", { fill: "currentColor", d: "M0 0h20v20H0z" }) }) })
-      ]
-    }
-  );
-};
-
-// src/icon/icons/copy_paste/index.react.tsx
-import { jsx as jsx36, jsxs as jsxs33 } from "react/jsx-runtime";
-var CopyPaste = (props) => {
-  let _ariaLabel = props["aria-label"];
-  _ariaLabel = _ariaLabel && _ariaLabel.trim();
-  const _ariaHidden = !_ariaLabel;
-  const _role = _ariaHidden ? "presentation" : "img";
-  return /* @__PURE__ */ jsxs33(
-    "svg",
-    {
-      className: props.className,
-      viewBox: "0 0 20 20",
-      "aria-label": _ariaLabel,
-      "aria-hidden": _ariaHidden,
-      role: _role,
-      style: {
-        height: "var(--icon-height)",
-        width: "var(--icon-width)",
-        display: "flex",
-        alignItems: "center",
-        justifyContent: "center"
-      },
-      children: [
-        /* @__PURE__ */ jsxs33("g", { "clip-path": "url(#bja)", children: [
-          /* @__PURE__ */ jsx36("path", { fill: "currentColor", d: "M1.875 19.998A1.876 1.876 0 0 1 0 18.123V4.373a1.877 1.877 0 0 1 1.875-1.875h2.466l-.284-.855a1.237 1.237 0 0 1 .066-.952c.149-.3.404-.522.721-.628a1.18 1.18 0 0 1 .373-.065h7.04c.335 0 .647.13.883.364a1.242 1.242 0 0 1 .303 1.284l-.285.852h2.467A1.877 1.877 0 0 1 17.5 4.373v2.5a.625.625 0 1 1-1.25 0v-2.5a.625.625 0 0 0-.625-.625h-2.883l-.132.395a1.25 1.25 0 0 1-1.186.855H6.076a1.25 1.25 0 0 1-1.186-.854l-.132-.396H1.875a.625.625 0 0 0-.625.625v13.75c0 .345.28.625.625.625h5a.625.625 0 1 1 0 1.25h-5Zm3.366-18.75.834 2.5h5.35l.832-2.5H5.242Z" }),
-          /* @__PURE__ */ jsx36("path", { fill: "currentColor", d: "M10.625 19.998a1.877 1.877 0 0 1-1.875-1.875v-7.5a1.877 1.877 0 0 1 1.875-1.875h7.5A1.877 1.877 0 0 1 20 10.623v7.5a1.876 1.876 0 0 1-1.875 1.875h-7.5Zm0-10a.625.625 0 0 0-.625.625v7.5c0 .345.28.625.625.625h7.5a.625.625 0 0 0 .625-.625v-7.5a.625.625 0 0 0-.625-.625h-7.5Z" }),
-          /* @__PURE__ */ jsx36("path", { fill: "currentColor", d: "M11.875 12.498a.625.625 0 1 1 0-1.25h5a.625.625 0 0 1 0 1.25h-5Zm0 2.5a.625.625 0 1 1 0-1.25h5a.625.625 0 0 1 0 1.25h-5Zm0 2.5a.625.625 0 1 1 0-1.25h1.875a.625.625 0 0 1 0 1.25h-1.875Z" })
-        ] }),
-        /* @__PURE__ */ jsx36("defs", { children: /* @__PURE__ */ jsx36("clipPath", { id: "bja", children: /* @__PURE__ */ jsx36("path", { fill: "currentColor", d: "M0 0h20v20H0z" }) }) })
-      ]
-    }
-  );
-};
-
-// src/icon/icons/cursor_pointer/index.react.tsx
-import { jsx as jsx37 } from "react/jsx-runtime";
-var CursorPointer = (props) => {
-  let _ariaLabel = props["aria-label"];
-  _ariaLabel = _ariaLabel && _ariaLabel.trim();
-  const _ariaHidden = !_ariaLabel;
-  const _role = _ariaHidden ? "presentation" : "img";
-  return /* @__PURE__ */ jsx37(
-    "svg",
-    {
-      className: props.className,
-      viewBox: "0 0 20 20",
-      "aria-label": _ariaLabel,
-      "aria-hidden": _ariaHidden,
-      role: _role,
-      style: {
-        height: "var(--icon-height)",
-        width: "var(--icon-width)",
-        display: "flex",
-        alignItems: "center",
-        justifyContent: "center"
-      },
-      children: /* @__PURE__ */ jsx37("path", { fill: "currentColor", d: "M8.805 19.375a4.384 4.384 0 0 1-3.557-1.832l-.915-1.278A6.433 6.433 0 0 1 3.125 12.5c0-1.5 1.057-2.77 2.5-3.063V2.5c0-1.034.84-1.875 1.875-1.875 1.034 0 1.875.84 1.875 1.875v3.232c.198-.071.41-.107.625-.107.643 0 1.238.336 1.58.865.278-.158.595-.243.92-.243a1.88 1.88 0 0 1 1.81 1.382 1.877 1.877 0 0 1 2.565 1.744v5.107a4.9 4.9 0 0 1-4.893 4.895H8.806Zm-3.18-8.64a1.884 1.884 0 0 0-1.25 1.765c0 1.096.337 2.146.973 3.037l.916 1.279a3.131 3.131 0 0 0 2.541 1.309h3.177a3.647 3.647 0 0 0 3.642-3.643v-5.11a.625.625 0 0 0-1.25 0l-.002.036a.596.596 0 0 1-.024.138l-.012.042-.012.031a.804.804 0 0 1-.065.115.662.662 0 0 1-.126.13.617.617 0 0 1-.513.122l-.018-.006a.634.634 0 0 1-.364-.247l-.015-.021a.81.81 0 0 1-.06-.121l-.014-.048a.65.65 0 0 1-.024-.169v-1.25a.625.625 0 1 0-1.25 0v1.25a.625.625 0 1 1-1.25 0V7.5a.625.625 0 1 0-1.25 0 .62.62 0 0 1-.082.31.626.626 0 0 1 .082.313v1.25a.625.625 0 0 1-1.25 0v-1.25c0-.114.03-.22.083-.312a.62.62 0 0 1-.083-.311v-5a.625.625 0 0 0-1.25 0v9.373a.625.625 0 1 1-1.25 0v-1.138Z" })
-    }
-  );
-};
-
-// src/icon/icons/cursor_pointer_circle/index.react.tsx
-import { jsx as jsx38, jsxs as jsxs34 } from "react/jsx-runtime";
-var CursorPointerCircle = (props) => {
-  let _ariaLabel = props["aria-label"];
-  _ariaLabel = _ariaLabel && _ariaLabel.trim();
-  const _ariaHidden = !_ariaLabel;
-  const _role = _ariaHidden ? "presentation" : "img";
-  return /* @__PURE__ */ jsxs34(
-    "svg",
-    {
-      className: props.className,
-      viewBox: "0 0 20 20",
-      "aria-label": _ariaLabel,
-      "aria-hidden": _ariaHidden,
-      role: _role,
-      style: {
-        height: "var(--icon-height)",
-        width: "var(--icon-width)",
-        display: "flex",
-        alignItems: "center",
-        justifyContent: "center"
-      },
-      children: [
-        /* @__PURE__ */ jsxs34("g", { "clip-path": "url(#bla)", children: [
-          /* @__PURE__ */ jsx38("path", { fill: "currentColor", d: "M8.547 20.003a.627.627 0 0 1-.5-.25L4.84 15.476a1.908 1.908 0 0 1-.062-2.197 1.904 1.904 0 0 1 2.722-.468l.419.314v-6.87a1.877 1.877 0 0 1 3.754 0v4.498l2.988.665a3.133 3.133 0 0 1 2.435 3.334l-.426 4.685a.622.622 0 0 1-.622.568l-7.5-.002Zm6.929-1.248.374-4.117a1.878 1.878 0 0 0-1.46-2l-3.48-.773a.62.62 0 0 1-.488-.61v-5a.627.627 0 0 0-1.254 0v8.122a.621.621 0 0 1-.625.625.63.63 0 0 1-.376-.125L6.75 13.813l-.03-.02a.65.65 0 0 0-.882.934l3.021 4.027 6.617.002Z" }),
-          /* @__PURE__ */ jsx38("path", { fill: "currentColor", d: "M5.422 11.963a.621.621 0 0 1-.439-.179c-2.705-2.655-2.745-7.017-.09-9.722A6.823 6.823 0 0 1 9.8.006a6.873 6.873 0 0 1 5.952 10.312.63.63 0 0 1-.854.228.626.626 0 0 1-.228-.854 5.625 5.625 0 0 0-.93-6.83A5.593 5.593 0 0 0 9.8 1.255a5.582 5.582 0 0 0-4.014 1.682 5.632 5.632 0 0 0 .075 7.956.624.624 0 0 1-.44 1.07Z" })
-        ] }),
-        /* @__PURE__ */ jsx38("defs", { children: /* @__PURE__ */ jsx38("clipPath", { id: "bla", children: /* @__PURE__ */ jsx38("path", { fill: "currentColor", d: "M0 0h20v20H0z" }) }) })
-      ]
-    }
-  );
-};
-
-// src/icon/icons/dice/index.react.tsx
-import { jsx as jsx39, jsxs as jsxs35 } from "react/jsx-runtime";
-var Dice = (props) => {
-  let _ariaLabel = props["aria-label"];
-  _ariaLabel = _ariaLabel && _ariaLabel.trim();
-  const _ariaHidden = !_ariaLabel;
-  const _role = _ariaHidden ? "presentation" : "img";
-  return /* @__PURE__ */ jsxs35(
-    "svg",
-    {
-      className: props.className,
-      viewBox: "0 0 20 20",
-      "aria-label": _ariaLabel,
-      "aria-hidden": _ariaHidden,
-      role: _role,
-      style: {
-        height: "var(--icon-height)",
-        width: "var(--icon-width)",
-        display: "flex",
-        alignItems: "center",
-        justifyContent: "center"
-      },
-      children: [
-        /* @__PURE__ */ jsxs35("g", { "clip-path": "url(#bma)", children: [
-          /* @__PURE__ */ jsx39("path", { fill: "currentColor", d: "M3.438 11.875a.938.938 0 0 1-.928-.828.614.614 0 0 1 0-.219.939.939 0 0 1 1.036-.819c.427.05.769.392.819.82a.598.598 0 0 1 0 .218.937.937 0 0 1-.82.818.58.58 0 0 1-.107.01Zm3.437 0a.938.938 0 0 1-.928-.828.598.598 0 0 1 0-.218.939.939 0 0 1 .82-.819.583.583 0 0 1 .216 0c.428.05.77.39.82.818a.597.597 0 0 1 0 .219.937.937 0 0 1-.82.818.588.588 0 0 1-.108.01Zm3.438 0a.938.938 0 0 1-.928-.828.597.597 0 0 1 0-.218.939.939 0 0 1 .82-.819.581.581 0 0 1 .217 0 .938.938 0 0 1 .819 1.037.938.938 0 0 1-.928.828Zm0-6.875a.597.597 0 0 1-.11-.01.938.938 0 0 1-.818-.818.598.598 0 0 1 0-.219.937.937 0 0 1 .818-.818.597.597 0 0 1 .219 0 .939.939 0 0 1 .819 1.035.939.939 0 0 1-.82.82.533.533 0 0 1-.108.01Zm5.625 0a.938.938 0 0 1-.928-.828.6.6 0 0 1 0-.218.938.938 0 0 1 .818-.819.598.598 0 0 1 .218 0 .939.939 0 0 1 .819 1.036.939.939 0 0 1-.82.819.584.584 0 0 1-.108.01Zm-.001 5.625a.938.938 0 0 1-.928-1.046.938.938 0 0 1 .818-.819.598.598 0 0 1 .219 0 .939.939 0 0 1 .819 1.036.939.939 0 0 1-.82.819.588.588 0 0 1-.108.01ZM3.438 16.25a.938.938 0 0 1-.928-.828.614.614 0 0 1 0-.219.939.939 0 0 1 1.035-.819.94.94 0 0 1 .82.82.598.598 0 0 1 0 .218.937.937 0 0 1-.82.818.57.57 0 0 1-.107.01Zm3.437 0a.938.938 0 0 1-.928-.828.597.597 0 0 1 0-.218.939.939 0 0 1 1.036-.819c.428.05.77.392.82.82a.597.597 0 0 1 0 .217.937.937 0 0 1-.82.819.656.656 0 0 1-.108.009Zm3.438 0a.938.938 0 0 1-.928-.828.597.597 0 0 1 0-.218.939.939 0 0 1 1.037-.819.938.938 0 0 1 .819 1.037.938.938 0 0 1-.928.828Z" }),
-          /* @__PURE__ */ jsx39("path", { fill: "currentColor", d: "M1.458 20A1.46 1.46 0 0 1 0 18.542V7.708A1.46 1.46 0 0 1 1.458 6.25h10.834a1.46 1.46 0 0 1 1.458 1.458v10.834A1.46 1.46 0 0 1 12.292 20H1.458Zm0-12.5a.208.208 0 0 0-.208.208v10.834c0 .115.093.208.208.208h10.834a.208.208 0 0 0 .208-.208V7.708a.208.208 0 0 0-.208-.208H1.458Z" }),
-          /* @__PURE__ */ jsx39("path", { fill: "currentColor", d: "M15.625 13.75a.624.624 0 1 1 0-1.25h2.5a.624.624 0 0 0 .625-.625v-10a.625.625 0 0 0-.625-.625h-10a.625.625 0 0 0-.625.625v2.5a.625.625 0 0 1-1.25 0v-2.5A1.877 1.877 0 0 1 8.125 0h10A1.877 1.877 0 0 1 20 1.875v10a1.876 1.876 0 0 1-1.875 1.875h-2.5Z" })
-        ] }),
-        /* @__PURE__ */ jsx39("defs", { children: /* @__PURE__ */ jsx39("clipPath", { id: "bma", children: /* @__PURE__ */ jsx39("path", { fill: "currentColor", d: "M0 0h20v20H0z" }) }) })
-      ]
-    }
-  );
-};
-
-// src/icon/icons/dictionary/index.react.tsx
-import { jsx as jsx40, jsxs as jsxs36 } from "react/jsx-runtime";
-var Dictionary = (props) => {
-  let _ariaLabel = props["aria-label"];
-  _ariaLabel = _ariaLabel && _ariaLabel.trim();
-  const _ariaHidden = !_ariaLabel;
-  const _role = _ariaHidden ? "presentation" : "img";
-  return /* @__PURE__ */ jsxs36(
-    "svg",
-    {
-      className: props.className,
-      viewBox: "0 0 20 20",
-      "aria-label": _ariaLabel,
-      "aria-hidden": _ariaHidden,
-      role: _role,
-      style: {
-        height: "var(--icon-height)",
-        width: "var(--icon-width)",
-        display: "flex",
-        alignItems: "center",
-        justifyContent: "center"
-      },
-      children: [
-        /* @__PURE__ */ jsxs36("g", { "clip-path": "url(#bna)", children: [
-          /* @__PURE__ */ jsx40("path", { fill: "currentColor", d: "M2.5 20A2.503 2.503 0 0 1 0 17.5V3.125A3.128 3.128 0 0 1 3.125 0h12.5a.625.625 0 1 1 0 1.25h-12.5A1.877 1.877 0 0 0 1.25 3.125v12.21c.38-.221.811-.336 1.25-.335h13.125a.624.624 0 1 1 0 1.25H15v2.5h.625a.624.624 0 1 1 0 1.25H2.5Zm0-3.75a1.25 1.25 0 0 0 0 2.5h11.25v-2.5H2.5Z" }),
-          /* @__PURE__ */ jsx40("path", { fill: "currentColor", d: "M6.875 11.25a.625.625 0 0 1-.625-.625V8.75H5v1.875a.625.625 0 1 1-1.25 0v-5c0-1.034.84-1.875 1.875-1.875 1.034 0 1.875.84 1.875 1.875v5a.625.625 0 0 1-.625.625ZM6.25 7.5V5.625a.625.625 0 0 0-1.25 0V7.5h1.25Zm4.375 6.25a.624.624 0 0 1-.625-.625v-6.25a.625.625 0 0 1 .625-.625h1.25a1.877 1.877 0 0 1 1.875 1.875v.625c0 .47-.172.91-.48 1.25.31.34.48.78.48 1.25v.625a1.876 1.876 0 0 1-1.875 1.875h-1.25Zm1.25-1.25a.624.624 0 0 0 .625-.625v-.625a.624.624 0 0 0-.625-.625h-.625l.625 1.875Zm0-3.125a.625.625 0 0 0 .625-.625v-.625a.625.625 0 0 0-.625-.625h-.625v1.875h.625Zm7.5 1.875a3.128 3.128 0 0 1-3.125-3.125v-1.25a3.128 3.128 0 0 1 3.125-3.125.625.625 0 1 1 0 1.25A1.877 1.877 0 0 0 17.5 6.875v1.25A1.876 1.876 0 0 0 19.375 10a.624.624 0 1 1 0 1.25Z" })
-        ] }),
-        /* @__PURE__ */ jsx40("defs", { children: /* @__PURE__ */ jsx40("clipPath", { id: "bna", children: /* @__PURE__ */ jsx40("path", { fill: "currentColor", d: "M0 0h20v20H0z" }) }) })
-      ]
-    }
-  );
-};
-
-// src/icon/icons/download/index.react.tsx
-import { jsx as jsx41, jsxs as jsxs37 } from "react/jsx-runtime";
-var Download = (props) => {
-  let _ariaLabel = props["aria-label"];
-  _ariaLabel = _ariaLabel && _ariaLabel.trim();
-  const _ariaHidden = !_ariaLabel;
-  const _role = _ariaHidden ? "presentation" : "img";
-  return /* @__PURE__ */ jsxs37(
-    "svg",
-    {
-      className: props.className,
-      viewBox: "0 0 20 20",
-      "aria-label": _ariaLabel,
-      "aria-hidden": _ariaHidden,
-      role: _role,
-      style: {
-        height: "var(--icon-height)",
-        width: "var(--icon-width)",
-        display: "flex",
-        alignItems: "center",
-        justifyContent: "center"
-      },
-      children: [
-        /* @__PURE__ */ jsxs37("g", { "clip-path": "url(#boa)", children: [
-          /* @__PURE__ */ jsx41("path", { fill: "currentColor", d: "M10 13.75a.629.629 0 0 1-.216-.04l-.018-.006a.61.61 0 0 1-.208-.137L5.81 9.817a.619.619 0 0 1-.183-.442c0-.167.065-.324.183-.442a.619.619 0 0 1 .442-.183c.167 0 .324.065.442.183l2.683 2.684V3.125a.625.625 0 0 1 1.25 0v8.49l2.683-2.683a.62.62 0 0 1 .884 0 .619.619 0 0 1 .183.442.619.619 0 0 1-.183.442l-3.75 3.75a.636.636 0 0 1-.205.136l-.022.009a.605.605 0 0 1-.215.039Z" }),
-          /* @__PURE__ */ jsx41("path", { fill: "currentColor", d: "M3.126 17.5a3.128 3.128 0 0 1-3.125-3.125v-1.25a.625.625 0 1 1 1.25 0v1.25a1.877 1.877 0 0 0 1.875 1.875h13.75a1.876 1.876 0 0 0 1.875-1.875v-1.25a.625.625 0 1 1 1.25 0v1.25a3.128 3.128 0 0 1-3.125 3.125H3.126Z" })
-        ] }),
-        /* @__PURE__ */ jsx41("defs", { children: /* @__PURE__ */ jsx41("clipPath", { id: "boa", children: /* @__PURE__ */ jsx41("path", { fill: "currentColor", d: "M0 0h20v20H0z" }) }) })
-      ]
-    }
-  );
-};
-
-// src/icon/icons/draw_tablet/index.react.tsx
-import { jsx as jsx42, jsxs as jsxs38 } from "react/jsx-runtime";
-var DrawTablet = (props) => {
-  let _ariaLabel = props["aria-label"];
-  _ariaLabel = _ariaLabel && _ariaLabel.trim();
-  const _ariaHidden = !_ariaLabel;
-  const _role = _ariaHidden ? "presentation" : "img";
-  return /* @__PURE__ */ jsxs38(
-    "svg",
-    {
-      className: props.className,
-      viewBox: "0 0 20 20",
-      "aria-label": _ariaLabel,
-      "aria-hidden": _ariaHidden,
-      role: _role,
-      style: {
-        height: "var(--icon-height)",
-        width: "var(--icon-width)",
-        display: "flex",
-        alignItems: "center",
-        justifyContent: "center"
-      },
-      children: [
-        /* @__PURE__ */ jsxs38("g", { "clip-path": "url(#bpa)", children: [
-          /* @__PURE__ */ jsx42("path", { fill: "currentColor", d: "M1.875 20.002A1.877 1.877 0 0 1 0 18.127V6.252a1.877 1.877 0 0 1 1.875-1.875h7.5a.625.625 0 0 1 0 1.25h-7.5a.625.625 0 0 0-.625.625v11.875c0 .345.28.625.625.625h16.25a.625.625 0 0 0 .625-.625v-10a.625.625 0 1 1 1.25 0v10a1.876 1.876 0 0 1-1.875 1.875H1.875Z" }),
-          /* @__PURE__ */ jsx42("path", { fill: "currentColor", d: "M10.29 12.267a.622.622 0 0 1-.624-.663l.19-3.18a.623.623 0 0 1 .127-.342L15.408.947a2.37 2.37 0 0 1 1.905-.945c.524 0 1.02.167 1.438.482a2.398 2.398 0 0 1 .476 3.346l-5.43 7.142a.628.628 0 0 1-.291.211l-3.011 1.05a.632.632 0 0 1-.205.034Zm.68-1.524 1.947-.678 5.31-6.983a1.143 1.143 0 1 0-1.824-1.379l-5.31 6.985-.123 2.055Z" }),
-          /* @__PURE__ */ jsx42("path", { fill: "currentColor", d: "M10.306 18.127a4.033 4.033 0 0 1-3.636-2.258 4.011 4.011 0 0 1-.41-1.479H5.18a1.431 1.431 0 0 1-1.43-1.427V8.944c0-.788.64-1.43 1.43-1.43H7.5a.625.625 0 0 1 0 1.251H5.18a.18.18 0 0 0-.18.18v4.017c0 .098.08.178.178.178h1.175c.155-.683.47-1.34.943-1.962a.62.62 0 0 1 .827-.152.625.625 0 0 1 .167.91c-.837 1.099-1.02 2.33-.5 3.378a2.792 2.792 0 0 0 2.515 1.563 2.798 2.798 0 0 0 1.988-.825 2.793 2.793 0 0 0 .821-1.99 2.8 2.8 0 0 0-.099-.738.625.625 0 0 1 1.206-.329 4.077 4.077 0 0 1-.163 2.62 4.036 4.036 0 0 1-.881 1.318 4.033 4.033 0 0 1-2.871 1.194Z" })
-        ] }),
-        /* @__PURE__ */ jsx42("defs", { children: /* @__PURE__ */ jsx42("clipPath", { id: "bpa", children: /* @__PURE__ */ jsx42("path", { fill: "currentColor", d: "M0 0h20v20H0z" }) }) })
-      ]
-    }
-  );
-};
-
-// src/icon/icons/elements/index.react.tsx
-import { jsx as jsx43, jsxs as jsxs39 } from "react/jsx-runtime";
-var Elements = (props) => {
-  let _ariaLabel = props["aria-label"];
-  _ariaLabel = _ariaLabel && _ariaLabel.trim();
-  const _ariaHidden = !_ariaLabel;
-  const _role = _ariaHidden ? "presentation" : "img";
-  return /* @__PURE__ */ jsxs39(
-    "svg",
-    {
-      className: props.className,
-      viewBox: "0 0 20 20",
-      "aria-label": _ariaLabel,
-      "aria-hidden": _ariaHidden,
-      role: _role,
-      style: {
-        height: "var(--icon-height)",
-        width: "var(--icon-width)",
-        display: "flex",
-        alignItems: "center",
-        justifyContent: "center"
-      },
-      children: [
-        /* @__PURE__ */ jsx43("g", { "clip-path": "url(#bqa)", children: /* @__PURE__ */ jsx43("path", { fill: "currentColor", d: "M1.46 8.75A1.46 1.46 0 0 1 0 7.292V1.458A1.46 1.46 0 0 1 1.459 0h5.834A1.46 1.46 0 0 1 8.75 1.458v5.834A1.46 1.46 0 0 1 7.293 8.75H1.459Zm0-7.5a.208.208 0 0 0-.209.208v5.834c0 .115.093.208.208.208h5.834a.208.208 0 0 0 .208-.208V1.458a.208.208 0 0 0-.208-.208H1.459Zm14.166 7.5a4.38 4.38 0 0 1-4.375-4.375A4.38 4.38 0 0 1 15.626 0 4.38 4.38 0 0 1 20 4.375a4.38 4.38 0 0 1-4.375 4.375Zm0-7.5A3.128 3.128 0 0 0 12.5 4.375 3.128 3.128 0 0 0 15.626 7.5a3.128 3.128 0 0 0 3.125-3.125 3.129 3.129 0 0 0-3.125-3.125ZM11.387 20a1.383 1.383 0 0 1-1.217-.747 1.472 1.472 0 0 1 .013-1.4l3.615-6.509a1.37 1.37 0 0 1 1.612-.656c.345.11.628.346.797.666l3.609 6.496c.242.44.246.967.012 1.41a1.372 1.372 0 0 1-1.205.74h-7.236ZM15 11.875a.116.116 0 0 0-.052.012.108.108 0 0 0-.047.048l-3.623 6.52a.228.228 0 0 0 0 .218.133.133 0 0 0 .117.076h7.22a.126.126 0 0 0 .106-.07.232.232 0 0 0 .002-.226L15.11 11.95c-.032-.06-.077-.075-.11-.075ZM4.376 19.998c-.383 0-.74-.168-.983-.462L.305 15.842a1.315 1.315 0 0 1 0-1.683l3.092-3.698a1.277 1.277 0 0 1 .979-.46c.382 0 .74.168.982.462l3.086 3.695a1.316 1.316 0 0 1 .001 1.683l-3.09 3.698a1.274 1.274 0 0 1-.98.46Zm0-8.747c-.008 0-.013.003-.014.005L1.264 14.96a.064.064 0 0 0 .001.082l3.09 3.696c.006.008.013.011.02.011l.02-.01 3.09-3.698a.064.064 0 0 0 0-.082l-3.09-3.699a.319.319 0 0 0-.02-.01Z" }) }),
-        /* @__PURE__ */ jsx43("defs", { children: /* @__PURE__ */ jsx43("clipPath", { id: "bqa", children: /* @__PURE__ */ jsx43("path", { fill: "currentColor", d: "M0 0h20v20H0z" }) }) })
-      ]
-    }
-  );
-};
-
-// src/icon/icons/envelope/index.react.tsx
-import { jsx as jsx44 } from "react/jsx-runtime";
-var Envelope = (props) => {
-  let _ariaLabel = props["aria-label"];
-  _ariaLabel = _ariaLabel && _ariaLabel.trim();
-  const _ariaHidden = !_ariaLabel;
-  const _role = _ariaHidden ? "presentation" : "img";
-  return /* @__PURE__ */ jsx44(
-    "svg",
-    {
-      className: props.className,
-      viewBox: "0 0 20 20",
-      "aria-label": _ariaLabel,
-      "aria-hidden": _ariaHidden,
-      role: _role,
-      style: {
-        height: "var(--icon-height)",
-        width: "var(--icon-width)",
-        display: "flex",
-        alignItems: "center",
-        justifyContent: "center"
-      },
-      children: /* @__PURE__ */ jsx44("path", { fill: "currentColor", d: "M1.875 16.875A1.876 1.876 0 0 1 0 15V5a1.875 1.875 0 0 1 1.124-1.716 1.88 1.88 0 0 1 .751-.159h16.25A1.878 1.878 0 0 1 20 5v10a1.876 1.876 0 0 1-1.875 1.875H1.875ZM1.25 15c0 .345.28.625.625.625h16.25A.624.624 0 0 0 18.75 15V5.156l-4.68 3.227 2.604 2.408a.627.627 0 0 1-.424 1.084.626.626 0 0 1-.424-.166l-2.811-2.598-1.52 1.049c-.44.303-.961.465-1.495.463a2.616 2.616 0 0 1-1.494-.463L6.985 9.112l-2.81 2.598a.628.628 0 0 1-.884-.035.626.626 0 0 1 .034-.883l2.604-2.408L1.25 5.156V15Zm7.966-5.87c.23.16.5.243.784.243.283 0 .554-.084.784-.242l6.897-4.756H2.319l6.897 4.756Z" })
-    }
-  );
-};
-
-// src/icon/icons/eraser/index.react.tsx
-import { jsx as jsx45 } from "react/jsx-runtime";
-var Eraser = (props) => {
-  let _ariaLabel = props["aria-label"];
-  _ariaLabel = _ariaLabel && _ariaLabel.trim();
-  const _ariaHidden = !_ariaLabel;
-  const _role = _ariaHidden ? "presentation" : "img";
-  return /* @__PURE__ */ jsx45(
-    "svg",
-    {
-      className: props.className,
-      viewBox: "0 0 20 20",
-      "aria-label": _ariaLabel,
-      "aria-hidden": _ariaHidden,
-      role: _role,
-      style: {
-        height: "var(--icon-height)",
-        width: "var(--icon-width)",
-        display: "flex",
-        alignItems: "center",
-        justifyContent: "center"
-      },
-      children: /* @__PURE__ */ jsx45("path", { fill: "currentColor", d: "M4.39 16.043a.62.62 0 0 1-.44-.183L.432 12.343a1.487 1.487 0 0 1 0-2.094l8.473-8.474a1.475 1.475 0 0 1 1.047-.432c.395 0 .767.153 1.046.431l4.54 4.54a1.49 1.49 0 0 1 .002 2.093l-7.408 7.407a.619.619 0 0 1-.883 0 .627.627 0 0 1 0-.884l1.517-1.517L3.9 8.548l-2.584 2.584a.234.234 0 0 0 0 .328l3.516 3.517a.617.617 0 0 1 .182.442.621.621 0 0 1-.624.624Zm5.26-3.513 5.006-5.006a.233.233 0 0 0-.002-.327l-4.537-4.538a.233.233 0 0 0-.329.001L4.785 7.663 9.65 12.53ZM.625 18.542a.625.625 0 1 1 0-1.25h3.75a.625.625 0 0 1 0 1.25H.625Zm10 0a.625.625 0 1 1 0-1.25h1.25a.625.625 0 0 1 0 1.25h-1.25Zm3.75 0a.625.625 0 1 1 0-1.25h1.25a.625.625 0 0 1 0 1.25h-1.25Zm3.75 0a.625.625 0 1 1 0-1.25h1.25a.625.625 0 0 1 0 1.25h-1.25Zm-11.25 0a.625.625 0 1 1 0-1.25h1.25a.625.625 0 0 1 0 1.25h-1.25Z" })
-    }
-  );
-};
-
-// src/icon/icons/expand/index.react.tsx
-import { jsx as jsx46, jsxs as jsxs40 } from "react/jsx-runtime";
-var Expand = (props) => {
-  let _ariaLabel = props["aria-label"];
-  _ariaLabel = _ariaLabel && _ariaLabel.trim();
-  const _ariaHidden = !_ariaLabel;
-  const _role = _ariaHidden ? "presentation" : "img";
-  return /* @__PURE__ */ jsxs40(
-    "svg",
-    {
-      className: props.className,
-      viewBox: "0 0 20 20",
-      "aria-label": _ariaLabel,
-      "aria-hidden": _ariaHidden,
-      role: _role,
-      style: {
-        height: "var(--icon-height)",
-        width: "var(--icon-width)",
-        display: "flex",
-        alignItems: "center",
-        justifyContent: "center"
-      },
-      children: [
-        /* @__PURE__ */ jsx46("g", { "clip-path": "url(#bta)", children: /* @__PURE__ */ jsx46("path", { fill: "currentColor", d: "M19.375 6.248a.625.625 0 0 1-.625-.625v-3.49L15.442 5.44a.622.622 0 0 1-.884-.001.619.619 0 0 1-.183-.442c0-.166.065-.324.183-.441l3.309-3.309h-3.492a.625.625 0 1 1 0-1.25h5a.625.625 0 0 1 .578.387l.008.022A.626.626 0 0 1 20 .622v5a.625.625 0 0 1-.625.625ZM.625 19.998a.629.629 0 0 1-.217-.039l-.018-.006a.629.629 0 0 1-.39-.58v-5a.625.625 0 1 1 1.25 0v3.491l3.308-3.308a.619.619 0 0 1 .884 0 .627.627 0 0 1 0 .884l-3.309 3.308h3.492a.625.625 0 1 1 0 1.25h-5Zm0-13.75A.625.625 0 0 1 0 5.623v-5C0 .55.014.475.04.406L.046.388A.628.628 0 0 1 .387.046L.409.037a.614.614 0 0 1 .217-.039h5a.625.625 0 0 1 0 1.25h-3.49l3.308 3.309a.619.619 0 0 1 .183.441.619.619 0 0 1-.386.578.623.623 0 0 1-.682-.136L1.25 2.132v3.491a.625.625 0 0 1-.625.625Zm13.75 13.75a.625.625 0 1 1 0-1.25h3.49l-3.307-3.308a.627.627 0 0 1 .441-1.067c.167 0 .324.065.442.183l3.308 3.308v-3.49a.626.626 0 0 1 1.251 0v5a.638.638 0 0 1-.038.214l-.005.014a.637.637 0 0 1-.344.35l-.021.008a.61.61 0 0 1-.216.04h-5.001v-.002ZM7.5 13.748c-.69 0-1.25-.56-1.25-1.25v-5c0-.689.56-1.25 1.25-1.25h5c.69 0 1.25.561 1.25 1.25v5c0 .69-.56 1.25-1.25 1.25h-5Zm0-1.25h5v-5h-5v5Z" }) }),
-        /* @__PURE__ */ jsx46("defs", { children: /* @__PURE__ */ jsx46("clipPath", { id: "bta", children: /* @__PURE__ */ jsx46("path", { fill: "currentColor", d: "M0 0h20v20H0z" }) }) })
-      ]
-    }
-  );
-};
-
-// src/icon/icons/expand_resize/index.react.tsx
-import { jsx as jsx47, jsxs as jsxs41 } from "react/jsx-runtime";
-var ExpandResize = (props) => {
-  let _ariaLabel = props["aria-label"];
-  _ariaLabel = _ariaLabel && _ariaLabel.trim();
-  const _ariaHidden = !_ariaLabel;
-  const _role = _ariaHidden ? "presentation" : "img";
-  return /* @__PURE__ */ jsxs41(
-    "svg",
-    {
-      className: props.className,
-      viewBox: "0 0 20 20",
-      "aria-label": _ariaLabel,
-      "aria-hidden": _ariaHidden,
-      role: _role,
-      style: {
-        height: "var(--icon-height)",
-        width: "var(--icon-width)",
-        display: "flex",
-        alignItems: "center",
-        justifyContent: "center"
-      },
-      children: [
-        /* @__PURE__ */ jsx47("g", { "clip-path": "url(#bua)", children: /* @__PURE__ */ jsx47("path", { fill: "currentColor", d: "M.625 19.998a.628.628 0 0 1-.217-.039l-.018-.006a.63.63 0 0 1-.39-.58v-5.625a.625.625 0 0 1 1.25 0v4.116l6.433-6.433a.619.619 0 0 1 .884 0 .62.62 0 0 1 0 .883l-6.434 6.433H6.25a.625.625 0 0 1 0 1.25l-5.625.001Zm11.25-11.25a.619.619 0 0 1-.442-.183.619.619 0 0 1-.183-.442c0-.166.065-.324.183-.441l6.434-6.434H13.75a.625.625 0 1 1 0-1.25h5.625a.625.625 0 0 1 .578.387l.008.022A.59.59 0 0 1 20 .623v5.625a.625.625 0 1 1-1.25 0V2.132l-6.433 6.434a.622.622 0 0 1-.442.182Z" }) }),
-        /* @__PURE__ */ jsx47("defs", { children: /* @__PURE__ */ jsx47("clipPath", { id: "bua", children: /* @__PURE__ */ jsx47("path", { fill: "currentColor", d: "M0 0h20v20H0z" }) }) })
-      ]
-    }
-  );
-};
-
-// src/icon/icons/external_link/index.react.tsx
-import { jsx as jsx48, jsxs as jsxs42 } from "react/jsx-runtime";
-var ExternalLink = (props) => {
-  let _ariaLabel = props["aria-label"];
-  _ariaLabel = _ariaLabel && _ariaLabel.trim();
-  const _ariaHidden = !_ariaLabel;
-  const _role = _ariaHidden ? "presentation" : "img";
-  return /* @__PURE__ */ jsxs42(
-    "svg",
-    {
-      className: props.className,
-      viewBox: "0 0 20 20",
-      "aria-label": _ariaLabel,
-      "aria-hidden": _ariaHidden,
-      role: _role,
-      style: {
-        height: "var(--icon-height)",
-        width: "var(--icon-width)",
-        display: "flex",
-        alignItems: "center",
-        justifyContent: "center"
-      },
-      children: [
-        /* @__PURE__ */ jsxs42("g", { "clip-path": "url(#bva)", children: [
-          /* @__PURE__ */ jsx48("path", { fill: "currentColor", d: "M6.876 13.748a.628.628 0 0 1-.442-1.067L17.867 1.247h-4.115a.625.625 0 0 1 0-1.25h5.625a.625.625 0 0 1 .578.387l.007.022a.61.61 0 0 1 .04.216v5.625a.625.625 0 0 1-1.25 0V2.132L7.318 13.566a.626.626 0 0 1-.442.182Z" }),
-          /* @__PURE__ */ jsx48("path", { fill: "currentColor", d: "M1.876 19.998a1.877 1.877 0 0 1-1.875-1.875v-12.5a1.877 1.877 0 0 1 1.875-1.875h7.5a.625.625 0 0 1 0 1.25h-7.5a.625.625 0 0 0-.625.625v12.5c0 .345.28.625.625.625h12.5a.625.625 0 0 0 .625-.625v-7.5a.625.625 0 1 1 1.25 0v7.5a1.877 1.877 0 0 1-1.875 1.875h-12.5Z" })
-        ] }),
-        /* @__PURE__ */ jsx48("defs", { children: /* @__PURE__ */ jsx48("clipPath", { id: "bva", children: /* @__PURE__ */ jsx48("path", { fill: "currentColor", d: "M0 0h20v20H0z" }) }) })
-      ]
-    }
-  );
-};
-
-// src/icon/icons/external_link_arrow/index.react.tsx
-import { jsx as jsx49, jsxs as jsxs43 } from "react/jsx-runtime";
-var ExternalLinkArrow = (props) => {
-  let _ariaLabel = props["aria-label"];
-  _ariaLabel = _ariaLabel && _ariaLabel.trim();
-  const _ariaHidden = !_ariaLabel;
-  const _role = _ariaHidden ? "presentation" : "img";
-  return /* @__PURE__ */ jsxs43(
-    "svg",
-    {
-      className: props.className,
-      viewBox: "0 0 20 20",
-      "aria-label": _ariaLabel,
-      "aria-hidden": _ariaHidden,
-      role: _role,
-      style: {
-        height: "var(--icon-height)",
-        width: "var(--icon-width)",
-        display: "flex",
-        alignItems: "center",
-        justifyContent: "center"
-      },
-      children: [
-        /* @__PURE__ */ jsx49("g", { "clip-path": "url(#bwclip0_258_9)", children: /* @__PURE__ */ jsx49("path", { d: "M17.0713 16.1874C16.9836 16.2741 16.8722 16.333 16.7512 16.3566C16.6301 16.3803 16.5047 16.3677 16.3908 16.3205C16.2769 16.2732 16.1794 16.1934 16.1106 16.091C16.0418 15.9886 16.0048 15.8682 16.0042 15.7448L16.0042 4.87892L3.81244 17.0707C3.6953 17.1879 3.53644 17.2537 3.37079 17.2537C3.20514 17.2537 3.04627 17.1879 2.92914 17.0707C2.81201 16.9536 2.7462 16.7947 2.7462 16.6291C2.7462 16.4634 2.81201 16.3046 2.92914 16.1874L15.1204 3.99622L4.25444 3.99626C4.1723 3.99657 4.09092 3.98061 4.01499 3.94929C3.93906 3.91796 3.87009 3.8719 3.81207 3.81377C3.75405 3.75564 3.70812 3.68658 3.67694 3.61059C3.64576 3.53461 3.62995 3.4532 3.63041 3.37106C3.62983 3.2043 3.69523 3.04697 3.81309 2.92912C3.93094 2.81127 4.08827 2.74586 4.25503 2.74645L16.6294 2.74639C16.712 2.74624 16.7938 2.76267 16.8699 2.79471L16.8911 2.80414C17.0264 2.86512 17.1347 2.97347 17.1957 3.10878L17.2016 3.12175C17.2361 3.20075 17.2539 3.28599 17.2541 3.37218L17.254 15.7454C17.2544 15.8275 17.2384 15.9089 17.2071 15.9847C17.1757 16.0606 17.1296 16.1295 17.0713 16.1874Z", fill: "currentColor" }) }),
-        /* @__PURE__ */ jsx49("defs", { children: /* @__PURE__ */ jsx49("clipPath", { id: "bwclip0_258_9", children: /* @__PURE__ */ jsx49("rect", { width: "20", height: "20", fill: "currentColor" }) }) })
-      ]
-    }
-  );
-};
-
-// src/icon/icons/female_headset/index.react.tsx
-import { jsx as jsx50, jsxs as jsxs44 } from "react/jsx-runtime";
-var FemaleHeadset = (props) => {
-  let _ariaLabel = props["aria-label"];
-  _ariaLabel = _ariaLabel && _ariaLabel.trim();
-  const _ariaHidden = !_ariaLabel;
-  const _role = _ariaHidden ? "presentation" : "img";
-  return /* @__PURE__ */ jsxs44(
-    "svg",
-    {
-      className: props.className,
-      viewBox: "0 0 20 20",
-      "aria-label": _ariaLabel,
-      "aria-hidden": _ariaHidden,
-      role: _role,
-      style: {
-        height: "var(--icon-height)",
-        width: "var(--icon-width)",
-        display: "flex",
-        alignItems: "center",
-        justifyContent: "center"
-      },
-      children: [
-        /* @__PURE__ */ jsxs44("g", { "clip-path": "url(#bxa)", children: [
-          /* @__PURE__ */ jsx50("path", { fill: "currentColor", d: "M16.853 13.75a.624.624 0 0 1-.271-.062 5.052 5.052 0 0 1-2.653-3.127 5.004 5.004 0 0 1-3.95 1.94 5.003 5.003 0 0 1-3.951-1.94 5.05 5.05 0 0 1-2.652 3.127.628.628 0 0 1-.835-.29.626.626 0 0 1 .29-.835A3.804 3.804 0 0 0 4.883 10H3.728c-.689 0-1.25-.56-1.25-1.25v-2.5c0-.689.561-1.25 1.25-1.25h.132C4.398 2.326 6.547.347 9.283.04c.232-.026.468-.039.7-.039 2.947 0 5.53 2.13 6.118 5h.127c.69 0 1.25.56 1.25 1.25v2.5c0 .688-.56 1.25-1.25 1.25h-1.154a3.805 3.805 0 0 0 2.052 2.562c.31.15.44.525.29.835a.63.63 0 0 1-.563.352ZM6.324 8.333c.383 1.67 1.903 2.918 3.654 2.918 1.752 0 3.27-1.247 3.655-2.917-.273.07-.554.105-.835.104h-.024A3.75 3.75 0 0 1 9.98 6.966a3.73 3.73 0 0 1-2.778 1.472h-.027c-.29 0-.574-.035-.85-.105Zm9.904.417v-2.5h-1.25v2.5h1.25Zm-12.5 0h1.25v-2.5h-1.25v2.5ZM9.978 5c.276 0 .515.177.597.44a2.506 2.506 0 0 0 2.25 1.748c.29 0 .588-.066.866-.194A3.768 3.768 0 0 0 9.978 3.75a3.768 3.768 0 0 0-3.712 3.244c.277.127.579.194.88.194a2.493 2.493 0 0 0 2.236-1.749A.622.622 0 0 1 9.978 5Zm0-2.5a4.979 4.979 0 0 1 4.329 2.502L14.353 5h.465a4.989 4.989 0 0 0-5.397-3.717A4.956 4.956 0 0 0 5.14 5h.463a.37.37 0 0 1 .047.003A4.98 4.98 0 0 1 9.978 2.5Z" }),
-          /* @__PURE__ */ jsx50("path", { fill: "currentColor", d: "M18.154 20a.629.629 0 0 1-.584-.4 8.064 8.064 0 0 0-3.852-4.313A3.755 3.755 0 0 1 9.98 18.75a3.755 3.755 0 0 1-3.739-3.467 8.186 8.186 0 0 0-3.85 4.314.63.63 0 0 1-.81.361.627.627 0 0 1-.36-.806 9.383 9.383 0 0 1 5.415-5.415 9.353 9.353 0 0 1 6.617-.02.57.57 0 0 1 .125.046 9.394 9.394 0 0 1 5.36 5.389.627.627 0 0 1-.583.848ZM7.478 15c0 1.378 1.122 2.5 2.5 2.5 1.379 0 2.5-1.122 2.5-2.5v-.23a8.097 8.097 0 0 0-5-.002V15Z" })
-        ] }),
-        /* @__PURE__ */ jsx50("defs", { children: /* @__PURE__ */ jsx50("clipPath", { id: "bxa", children: /* @__PURE__ */ jsx50("path", { fill: "currentColor", d: "M0 0h20v20H0z" }) }) })
-      ]
-    }
-  );
-};
-
-// src/icon/icons/flip_right/index.react.tsx
-import { jsx as jsx51, jsxs as jsxs45 } from "react/jsx-runtime";
-var FlipRight = (props) => {
-  let _ariaLabel = props["aria-label"];
-  _ariaLabel = _ariaLabel && _ariaLabel.trim();
-  const _ariaHidden = !_ariaLabel;
-  const _role = _ariaHidden ? "presentation" : "img";
-  return /* @__PURE__ */ jsxs45(
-    "svg",
-    {
-      className: props.className,
-      viewBox: "0 0 20 20",
-      "aria-label": _ariaLabel,
-      "aria-hidden": _ariaHidden,
-      role: _role,
-      style: {
-        height: "var(--icon-height)",
-        width: "var(--icon-width)",
-        display: "flex",
-        alignItems: "center",
-        justifyContent: "center"
-      },
-      children: [
-        /* @__PURE__ */ jsxs45("g", { "clip-path": "url(#bya)", children: [
-          /* @__PURE__ */ jsx51("path", { fill: "currentColor", d: "M4.696 13.125a.62.62 0 0 1-.624-.592.621.621 0 0 1 .036-.241 6.209 6.209 0 0 1 3.2-3.557 6.202 6.202 0 0 1 4.777-.253c.689.245 1.33.608 1.892 1.074l-.222-1.863a.62.62 0 0 1 .347-.637.625.625 0 0 1 .893.489l.44 3.67a.622.622 0 0 1-.528.693l-.07.01a.717.717 0 0 1-.058.008l-3.577.535a.628.628 0 0 1-.595-.991.622.622 0 0 1 .41-.246l2.525-.377A5.005 5.005 0 0 0 9.998 9.37c-.745 0-1.47.165-2.151.49a4.966 4.966 0 0 0-2.56 2.846.63.63 0 0 1-.591.418Z" }),
-          /* @__PURE__ */ jsx51("path", { fill: "currentColor", d: "M1.875 20A1.876 1.876 0 0 1 0 18.125V1.875A1.877 1.877 0 0 1 1.875 0h6.25a.625.625 0 0 1 0 1.25h-6.25a.625.625 0 0 0-.625.625v16.25c0 .345.28.625.625.625h6.25a.625.625 0 1 1 0 1.25h-6.25Zm8.75-15A.625.625 0 0 1 10 4.375V.625a.625.625 0 1 1 1.25 0v3.75a.625.625 0 0 1-.625.625Zm0 15a.624.624 0 0 1-.625-.625v-5a.624.624 0 1 1 1.25 0v5a.624.624 0 0 1-.625.625Zm2.5-18.75a.625.625 0 1 1 0-1.25h1.25a.625.625 0 1 1 0 1.25h-1.25Zm0 18.75a.624.624 0 1 1 0-1.25h1.25a.624.624 0 1 1 0 1.25h-1.25Zm6.25-13.75a.625.625 0 0 1-.625-.625v-1.25a.625.625 0 1 1 1.25 0v1.25a.625.625 0 0 1-.625.625Zm0 5a.624.624 0 0 1-.625-.625v-1.25a.625.625 0 1 1 1.25 0v1.25a.624.624 0 0 1-.625.625Zm0 5a.624.624 0 0 1-.625-.625v-1.25a.624.624 0 1 1 1.25 0v1.25a.624.624 0 0 1-.625.625Zm0-13.75a.625.625 0 0 1-.625-.625.625.625 0 0 0-.625-.625.625.625 0 1 1 0-1.25A1.877 1.877 0 0 1 20 1.875a.625.625 0 0 1-.625.625ZM18.125 20a.624.624 0 1 1 0-1.25c.345 0 .625-.28.625-.625a.624.624 0 1 1 1.25 0A1.876 1.876 0 0 1 18.125 20Z" })
-        ] }),
-        /* @__PURE__ */ jsx51("defs", { children: /* @__PURE__ */ jsx51("clipPath", { id: "bya", children: /* @__PURE__ */ jsx51("path", { fill: "currentColor", d: "M0 0h20v20H0z" }) }) })
-      ]
-    }
-  );
-};
-
-// src/icon/icons/folder_bookmark/index.react.tsx
-import { jsx as jsx52, jsxs as jsxs46 } from "react/jsx-runtime";
-var FolderBookmark = (props) => {
-  let _ariaLabel = props["aria-label"];
-  _ariaLabel = _ariaLabel && _ariaLabel.trim();
-  const _ariaHidden = !_ariaLabel;
-  const _role = _ariaHidden ? "presentation" : "img";
-  return /* @__PURE__ */ jsxs46(
-    "svg",
-    {
-      className: props.className,
-      viewBox: "0 0 20 20",
-      "aria-label": _ariaLabel,
-      "aria-hidden": _ariaHidden,
-      role: _role,
-      style: {
-        height: "var(--icon-height)",
-        width: "var(--icon-width)",
-        display: "flex",
-        alignItems: "center",
-        justifyContent: "center"
-      },
-      children: [
-        /* @__PURE__ */ jsx52("g", { "clip-path": "url(#bza)", children: /* @__PURE__ */ jsx52("path", { fill: "currentColor", d: "M2.04 19.375a.605.605 0 0 1-.193-.03A2.054 2.054 0 0 1 0 17.341V2.5A1.877 1.877 0 0 1 1.875.625h3.75C6.658.625 7.5 1.465 7.5 2.5v.625h1.25V2.5A1.877 1.877 0 0 1 10.625.625h3.75A1.877 1.877 0 0 1 16.25 2.5v.625h.625A1.877 1.877 0 0 1 18.75 5v1.982c.262.092.5.242.7.442a1.858 1.858 0 0 1 .487 1.808L18.11 17.95a1.89 1.89 0 0 1-1.82 1.425l-14.252.001Zm14.252-1.25a.625.625 0 0 0 .604-.464l1.826-8.718a.713.713 0 0 0 .028-.192.62.62 0 0 0-.625-.626H16.25v2.5a.621.621 0 0 1-.625.625.628.628 0 0 1-.375-.125l-2.125-1.594L11 11.125a.627.627 0 0 1-.907-.171.622.622 0 0 1-.093-.329v-6.25H6.875a.625.625 0 0 1-.625-.625V2.5a.625.625 0 0 0-.625-.625h-3.75a.625.625 0 0 0-.625.625v14.833a.792.792 0 0 0 1.548.213l1.955-9.272a1.883 1.883 0 0 1 1.813-1.399h1.559a.625.625 0 0 1 0 1.25h-1.56a.63.63 0 0 0-.594.433l-1.954 9.27-.01.038a2 2 0 0 1-.092.26l12.377-.001Zm-5.667-15a.625.625 0 0 1 .625.625v5.625l1.5-1.125a.627.627 0 0 1 .75 0L15 9.375V2.5a.625.625 0 0 0-.625-.625h-3.75A.625.625 0 0 0 10 2.5v.625h.625Zm6.875 3.75V5a.625.625 0 0 0-.625-.625h-.625v2.5h1.25Z" }) }),
-        /* @__PURE__ */ jsx52("defs", { children: /* @__PURE__ */ jsx52("clipPath", { id: "bza", children: /* @__PURE__ */ jsx52("path", { fill: "currentColor", d: "M0 0h20v20H0z" }) }) })
-      ]
-    }
-  );
-};
-
-// src/icon/icons/font_expand/index.react.tsx
-import { jsx as jsx53, jsxs as jsxs47 } from "react/jsx-runtime";
-var FontExpand = (props) => {
-  let _ariaLabel = props["aria-label"];
-  _ariaLabel = _ariaLabel && _ariaLabel.trim();
-  const _ariaHidden = !_ariaLabel;
-  const _role = _ariaHidden ? "presentation" : "img";
-  return /* @__PURE__ */ jsxs47(
-    "svg",
-    {
-      className: props.className,
-      viewBox: "0 0 20 20",
-      "aria-label": _ariaLabel,
-      "aria-hidden": _ariaHidden,
-      role: _role,
-      style: {
-        height: "var(--icon-height)",
-        width: "var(--icon-width)",
-        display: "flex",
-        alignItems: "center",
-        justifyContent: "center"
-      },
-      children: [
-        /* @__PURE__ */ jsx53("g", { "clip-path": "url(#caa)", children: /* @__PURE__ */ jsx53("path", { fill: "currentColor", d: "M3.75 19.377a.62.62 0 0 1-.222-.042l-.014-.005a.617.617 0 0 1-.206-.137l-2.5-2.5a.627.627 0 0 1 .442-1.068c.167 0 .324.065.442.183l1.433 1.434v-2.866a.625.625 0 0 1 1.25 0v2.866l1.433-1.434a.62.62 0 0 1 .884 0 .627.627 0 0 1 0 .884l-2.5 2.5a.305.305 0 0 1-.039.033.59.59 0 0 1-.165.103l-.021.009a.604.604 0 0 1-.217.04Zm0-13.125a.625.625 0 0 1-.625-.625V2.76L1.692 4.193a.619.619 0 0 1-.442.184.619.619 0 0 1-.442-.184.619.619 0 0 1-.183-.441.62.62 0 0 1 .183-.442l2.5-2.5A.618.618 0 0 1 3.71.632a.655.655 0 0 1 .24.03c.014.002.027.007.04.012a.632.632 0 0 1 .202.136l2.5 2.5a.619.619 0 0 1 .183.442.619.619 0 0 1-.183.441.619.619 0 0 1-.442.184.619.619 0 0 1-.442-.184L4.375 2.76v2.866c0 .345-.28.626-.625.626Zm13.95 9.376c-.01 0-.023 0-.035-.002H16.25a.625.625 0 0 1 0-1.25h.574l-1.288-3.75h-3.573l-1.288 3.75h.575a.625.625 0 0 1 0 1.25h-2.5a.625.625 0 0 1 0-1.25h.603l3.661-10.654c.19-.553.611-.596.736-.596s.546.043.736.596l3.66 10.655h.604a.625.625 0 0 1 0 1.25h-1.018l-.031.001Zm-2.593-6.251L13.75 5.428l-1.357 3.949h2.714Z" }) }),
-        /* @__PURE__ */ jsx53("defs", { children: /* @__PURE__ */ jsx53("clipPath", { id: "caa", children: /* @__PURE__ */ jsx53("path", { fill: "currentColor", d: "M0 0h20v20H0z" }) }) })
-      ]
-    }
-  );
-};
-
-// src/icon/icons/form_template/index.react.tsx
-import { jsx as jsx54, jsxs as jsxs48 } from "react/jsx-runtime";
-var FormTemplate = (props) => {
-  let _ariaLabel = props["aria-label"];
-  _ariaLabel = _ariaLabel && _ariaLabel.trim();
-  const _ariaHidden = !_ariaLabel;
-  const _role = _ariaHidden ? "presentation" : "img";
-  return /* @__PURE__ */ jsxs48(
-    "svg",
-    {
-      className: props.className,
-      viewBox: "0 0 20 20",
-      "aria-label": _ariaLabel,
-      "aria-hidden": _ariaHidden,
-      role: _role,
-      style: {
-        height: "var(--icon-height)",
-        width: "var(--icon-width)",
-        display: "flex",
-        alignItems: "center",
-        justifyContent: "center"
-      },
-      children: [
-        /* @__PURE__ */ jsxs48("g", { "clip-path": "url(#cba)", children: [
-          /* @__PURE__ */ jsx54("path", { fill: "currentColor", d: "M10 19.375a.625.625 0 0 1-.625-.625V17.5h-1.25a.625.625 0 1 1 0-1.25h1.25V15a.625.625 0 1 1 1.25 0v1.25h1.25a.624.624 0 1 1 0 1.25h-1.25v1.25a.624.624 0 0 1-.625.625Z" }),
-          /* @__PURE__ */ jsx54("path", { fill: "currentColor", d: "M13.75 20a.624.624 0 1 1 0-1.25h4.375a.624.624 0 0 0 .625-.625V16.25h-5a.624.624 0 1 1 0-1.25h5v-3.75H1.25V15h5a.625.625 0 1 1 0 1.25h-5v1.875c0 .345.28.625.625.625H6.25a.625.625 0 1 1 0 1.25H1.875A1.876 1.876 0 0 1 0 18.125V1.875A1.877 1.877 0 0 1 1.875 0h16.25A1.877 1.877 0 0 1 20 1.875v16.25A1.876 1.876 0 0 1 18.125 20H13.75Zm5-10V6.25H1.25V10h17.5Zm0-5V1.875a.625.625 0 0 0-.625-.625H1.875a.625.625 0 0 0-.625.625V5h17.5Z" }),
-          /* @__PURE__ */ jsx54("path", { fill: "currentColor", d: "M16.25 9.1a.937.937 0 1 0 0-1.876.937.937 0 0 0 0 1.875Zm0 4.96a.937.937 0 1 0 0-1.874.937.937 0 0 0 0 1.875Zm0-10.034a.937.937 0 1 0 0-1.875.937.937 0 0 0 0 1.875Z" })
-        ] }),
-        /* @__PURE__ */ jsx54("defs", { children: /* @__PURE__ */ jsx54("clipPath", { id: "cba", children: /* @__PURE__ */ jsx54("path", { fill: "currentColor", d: "M0 0h20v20H0z" }) }) })
-      ]
-    }
-  );
-};
-
-// src/icon/icons/frame_expand/index.react.tsx
-import { jsx as jsx55, jsxs as jsxs49 } from "react/jsx-runtime";
-var FrameExpand = (props) => {
-  let _ariaLabel = props["aria-label"];
-  _ariaLabel = _ariaLabel && _ariaLabel.trim();
-  const _ariaHidden = !_ariaLabel;
-  const _role = _ariaHidden ? "presentation" : "img";
-  return /* @__PURE__ */ jsxs49(
-    "svg",
-    {
-      className: props.className,
-      viewBox: "0 0 20 20",
-      "aria-label": _ariaLabel,
-      "aria-hidden": _ariaHidden,
-      role: _role,
-      style: {
-        height: "var(--icon-height)",
-        width: "var(--icon-width)",
-        display: "flex",
-        alignItems: "center",
-        justifyContent: "center"
-      },
-      children: [
-        /* @__PURE__ */ jsx55("g", { "clip-path": "url(#cca)", children: /* @__PURE__ */ jsx55("path", { fill: "currentColor", d: "M19.38 6.875a.625.625 0 0 1-.626-.625V1.873a.625.625 0 0 0-.625-.625h-4.375a.625.625 0 0 1 0-1.25h4.375a1.877 1.877 0 0 1 1.875 1.875V6.25a.625.625 0 0 1-.625.625Zm-5.626 13.122a.625.625 0 1 1 0-1.25h4.375a.625.625 0 0 0 .625-.625V13.75a.625.625 0 0 1 1.25 0v4.372a1.877 1.877 0 0 1-1.875 1.875h-4.375ZM1.88 20a1.876 1.876 0 0 1-1.876-1.875V13.75a.625.625 0 1 1 1.25 0v4.375c0 .345.28.625.625.625h4.375a.625.625 0 1 1 0 1.25H1.88ZM.63 6.875a.625.625 0 0 1-.626-.625V1.873A1.877 1.877 0 0 1 1.88-.002h4.375a.625.625 0 0 1 0 1.25H1.88a.625.625 0 0 0-.625.625V6.25a.625.625 0 0 1-.625.625Z" }) }),
-        /* @__PURE__ */ jsx55("defs", { children: /* @__PURE__ */ jsx55("clipPath", { id: "cca", children: /* @__PURE__ */ jsx55("path", { fill: "currentColor", d: "M0 0h20v20H0z" }) }) })
-      ]
-    }
-  );
-};
-
-// src/icon/icons/frame_shrink/index.react.tsx
-import { jsx as jsx56, jsxs as jsxs50 } from "react/jsx-runtime";
-var FrameShrink = (props) => {
-  let _ariaLabel = props["aria-label"];
-  _ariaLabel = _ariaLabel && _ariaLabel.trim();
-  const _ariaHidden = !_ariaLabel;
-  const _role = _ariaHidden ? "presentation" : "img";
-  return /* @__PURE__ */ jsxs50(
-    "svg",
-    {
-      className: props.className,
-      viewBox: "0 0 20 20",
-      "aria-label": _ariaLabel,
-      "aria-hidden": _ariaHidden,
-      role: _role,
-      style: {
-        height: "var(--icon-height)",
-        width: "var(--icon-width)",
-        display: "flex",
-        alignItems: "center",
-        justifyContent: "center"
-      },
-      children: [
-        /* @__PURE__ */ jsx56("g", { "clip-path": "url(#cda)", children: /* @__PURE__ */ jsx56("path", { fill: "currentColor", d: "M15 6.875A1.877 1.877 0 0 1 13.125 5V.623a.625.625 0 1 1 1.25 0V5c0 .345.28.625.625.625h4.375a.625.625 0 1 1 0 1.25H15ZM13.75 20a.624.624 0 0 1-.625-.625V15A1.876 1.876 0 0 1 15 13.125h4.376a.625.625 0 1 1 0 1.25H15a.624.624 0 0 0-.625.625v4.375a.624.624 0 0 1-.625.625Zm-7.5-.003a.625.625 0 0 1-.625-.625V15A.625.625 0 0 0 5 14.375H.626a.625.625 0 1 1 0-1.25H5A1.876 1.876 0 0 1 6.875 15v4.372a.625.625 0 0 1-.625.625ZM.626 6.875a.625.625 0 0 1 0-1.25H5A.625.625 0 0 0 5.625 5V.623a.625.625 0 0 1 1.25 0V5A1.877 1.877 0 0 1 5 6.875H.626Z" }) }),
-        /* @__PURE__ */ jsx56("defs", { children: /* @__PURE__ */ jsx56("clipPath", { id: "cda", children: /* @__PURE__ */ jsx56("path", { fill: "currentColor", d: "M0 0h20v20H0z" }) }) })
-      ]
-    }
-  );
-};
-
-// src/icon/icons/game_controller/index.react.tsx
-import { jsx as jsx57, jsxs as jsxs51 } from "react/jsx-runtime";
-var GameController = (props) => {
-  let _ariaLabel = props["aria-label"];
-  _ariaLabel = _ariaLabel && _ariaLabel.trim();
-  const _ariaHidden = !_ariaLabel;
-  const _role = _ariaHidden ? "presentation" : "img";
-  return /* @__PURE__ */ jsxs51(
-    "svg",
-    {
-      className: props.className,
-      viewBox: "0 0 20 20",
-      "aria-label": _ariaLabel,
-      "aria-hidden": _ariaHidden,
-      role: _role,
-      style: {
-        height: "var(--icon-height)",
-        width: "var(--icon-width)",
-        display: "flex",
-        alignItems: "center",
-        justifyContent: "center"
-      },
-      children: [
-        /* @__PURE__ */ jsxs51("g", { "clip-path": "url(#cea)", children: [
-          /* @__PURE__ */ jsx57("path", { fill: "currentColor", d: "M4.375 12.502a.625.625 0 0 1-.625-.625v-.625h-.625a.625.625 0 1 1 0-1.25h.625v-.625a.625.625 0 0 1 1.25 0v.625h.625a.625.625 0 0 1 0 1.25H5v.625a.625.625 0 0 1-.625.625Zm10-.625a.938.938 0 0 1-.928-1.046.938.938 0 0 1 1.038-.82.938.938 0 0 1 .82 1.038.938.938 0 0 1-.93.828Zm1.875-1.875a.94.94 0 0 1-.928-1.045.938.938 0 0 1 .819-.82.598.598 0 0 1 .218 0c.428.05.768.392.818.82a.58.58 0 0 1 0 .216.937.937 0 0 1-.927.829Z" }),
-          /* @__PURE__ */ jsx57("path", { fill: "currentColor", d: "M3.289 20.003a3.3 3.3 0 0 1-.845-.11A3.42 3.42 0 0 1 0 16.546v-7.17a4.38 4.38 0 0 1 4.375-4.374h5c0-2.112-1.68-3.75-3.125-3.75H5a.625.625 0 0 1 0-1.25h1.25c2.464 0 4.375 2.687 4.375 5h5A4.38 4.38 0 0 1 20 9.377l.002 7.335a3.286 3.286 0 0 1-3.86 3.239 3.416 3.416 0 0 1-2.465-1.975l-1.606-2.974H7.93l-1.671 3.11a3.27 3.27 0 0 1-2.971 1.891ZM4.375 6.252A3.128 3.128 0 0 0 1.25 9.377v7.185a2.165 2.165 0 0 0 1.532 2.129 2.044 2.044 0 0 0 2.357-1.133l1.867-3.476a.624.624 0 0 1 .55-.33h4.887c.23 0 .44.126.55.328l1.797 3.327.026.053a2.153 2.153 0 0 0 1.557 1.263 2.036 2.036 0 0 0 2.38-2.011l-.003-4.21V9.377a3.128 3.128 0 0 0-3.125-3.125H4.375Z" })
-        ] }),
-        /* @__PURE__ */ jsx57("defs", { children: /* @__PURE__ */ jsx57("clipPath", { id: "cea", children: /* @__PURE__ */ jsx57("path", { fill: "currentColor", d: "M0 0h20v20H0z" }) }) })
-      ]
-    }
-  );
-};
-
-// src/icon/icons/glasses/index.react.tsx
-import { jsx as jsx58 } from "react/jsx-runtime";
-var Glasses = (props) => {
-  let _ariaLabel = props["aria-label"];
-  _ariaLabel = _ariaLabel && _ariaLabel.trim();
-  const _ariaHidden = !_ariaLabel;
-  const _role = _ariaHidden ? "presentation" : "img";
-  return /* @__PURE__ */ jsx58(
-    "svg",
-    {
-      className: props.className,
-      viewBox: "0 0 20 20",
-      "aria-label": _ariaLabel,
-      "aria-hidden": _ariaHidden,
-      role: _role,
-      style: {
-        height: "var(--icon-height)",
-        width: "var(--icon-width)",
-        display: "flex",
-        alignItems: "center",
-        justifyContent: "center"
-      },
-      children: /* @__PURE__ */ jsx58("path", { fill: "currentColor", d: "M15.625 18.75c-2.172 0-4.02-1.625-4.327-3.75H8.701c-.307 2.125-2.154 3.75-4.327 3.75A4.38 4.38 0 0 1 0 14.375v-10A3.128 3.128 0 0 1 3.125 1.25h.625a.625.625 0 0 1 0 1.25h-.625A1.877 1.877 0 0 0 1.25 4.375v6.944A4.34 4.34 0 0 1 4.375 10c2.173 0 4.02 1.625 4.327 3.75h2.597c.307-2.125 2.154-3.75 4.327-3.75a4.34 4.34 0 0 1 3.125 1.32V4.374A1.878 1.878 0 0 0 16.875 2.5h-.625a.625.625 0 1 1 0-1.25h.625A3.128 3.128 0 0 1 20 4.375v10a4.38 4.38 0 0 1-4.375 4.375Zm0-7.5a3.128 3.128 0 0 0-3.125 3.125 3.128 3.128 0 0 0 3.125 3.125 3.128 3.128 0 0 0 3.125-3.125 3.128 3.128 0 0 0-3.125-3.125Zm-11.25 0a3.128 3.128 0 0 0-3.125 3.125A3.128 3.128 0 0 0 4.375 17.5 3.128 3.128 0 0 0 7.5 14.375a3.128 3.128 0 0 0-3.125-3.125Z" })
-    }
-  );
-};
-
-// src/icon/icons/graph_stats/index.react.tsx
-import { jsx as jsx59 } from "react/jsx-runtime";
-var GraphStats = (props) => {
-  let _ariaLabel = props["aria-label"];
-  _ariaLabel = _ariaLabel && _ariaLabel.trim();
-  const _ariaHidden = !_ariaLabel;
-  const _role = _ariaHidden ? "presentation" : "img";
-  return /* @__PURE__ */ jsx59(
-    "svg",
-    {
-      className: props.className,
-      viewBox: "0 0 20 20",
-      "aria-label": _ariaLabel,
-      "aria-hidden": _ariaHidden,
-      role: _role,
-      style: {
-        height: "var(--icon-height)",
-        width: "var(--icon-width)",
-        display: "flex",
-        alignItems: "center",
-        justifyContent: "center"
-      },
-      children: /* @__PURE__ */ jsx59("path", { fill: "currentColor", d: "M.625 15.625a.618.618 0 0 1-.442-.183A.619.619 0 0 1 0 15c0-.167.065-.324.183-.442l5.991-5.99a1.862 1.862 0 0 1 1.327-.549c.501 0 .972.195 1.326.549l2.606 2.606a.622.622 0 0 0 .885 0l5.55-5.549H13.75a.625.625 0 1 1 0-1.25h5.625a.625.625 0 0 1 .578.388l.009.021A.626.626 0 0 1 20 5v5.625a.624.624 0 1 1-1.25 0V6.509l-5.55 5.55a1.862 1.862 0 0 1-1.326.548 1.86 1.86 0 0 1-1.325-.549L7.942 9.451a.62.62 0 0 0-.442-.183.622.622 0 0 0-.442.183l-5.991 5.99a.62.62 0 0 1-.442.184Z" })
-    }
-  );
-};
-
-// src/icon/icons/headphones/index.react.tsx
-import { jsx as jsx60, jsxs as jsxs52 } from "react/jsx-runtime";
-var Headphones = (props) => {
-  let _ariaLabel = props["aria-label"];
-  _ariaLabel = _ariaLabel && _ariaLabel.trim();
-  const _ariaHidden = !_ariaLabel;
-  const _role = _ariaHidden ? "presentation" : "img";
-  return /* @__PURE__ */ jsxs52(
-    "svg",
-    {
-      className: props.className,
-      viewBox: "0 0 20 20",
-      "aria-label": _ariaLabel,
-      "aria-hidden": _ariaHidden,
-      role: _role,
-      style: {
-        height: "var(--icon-height)",
-        width: "var(--icon-width)",
-        display: "flex",
-        alignItems: "center",
-        justifyContent: "center"
-      },
-      children: [
-        /* @__PURE__ */ jsxs52("g", { "clip-path": "url(#cha)", children: [
-          /* @__PURE__ */ jsx60("path", { fill: "currentColor", d: "M3.337 20a.626.626 0 0 1-.547-.323C2.677 19.47.017 14.592 0 10.024-.02 4.51 4.45.007 9.964-.013h.037c5.494 0 9.979 4.47 9.999 9.965.016 4.557-2.608 9.466-2.72 9.672a.625.625 0 1 1-1.1-.594c.027-.048 2.586-4.834 2.57-9.074-.018-4.807-3.943-8.719-8.75-8.719L9.968.925l.001.312c-4.825.018-8.736 3.957-8.718 8.782.015 4.238 2.61 9.007 2.635 9.054a.627.627 0 0 1-.548.927Z" }),
-          /* @__PURE__ */ jsx60("path", { fill: "currentColor", d: "M6.077 19.503a1.457 1.457 0 0 1-1.344-.892l-2.025-4.799a1.448 1.448 0 0 1-.007-1.115 1.44 1.44 0 0 1 .784-.794l1.343-.567a1.446 1.446 0 0 1 1.116-.008c.36.147.643.425.794.785l2.025 4.798a1.46 1.46 0 0 1-.776 1.91l-1.345.567a1.438 1.438 0 0 1-.565.115Zm-.681-7.032a.208.208 0 0 0-.081.017l-1.343.566a.202.202 0 0 0-.112.114.206.206 0 0 0 0 .159l2.026 4.798a.202.202 0 0 0 .113.112.198.198 0 0 0 .16-.001l1.343-.567a.207.207 0 0 0 .111-.273l-2.025-4.799a.201.201 0 0 0-.113-.111.221.221 0 0 0-.08-.015Zm8.527 7.034a1.45 1.45 0 0 1-.566-.116l-1.344-.566a1.452 1.452 0 0 1-.892-1.354 1.45 1.45 0 0 1 .116-.557l2.025-4.799a1.456 1.456 0 0 1 1.91-.776l1.343.566c.358.151.637.434.783.794.147.361.145.758-.006 1.116l-2.025 4.799a1.453 1.453 0 0 1-.794.784 1.442 1.442 0 0 1-.55.109Zm.682-7.033a.208.208 0 0 0-.193.127l-2.025 4.799a.207.207 0 0 0 .044.227c.02.02.042.035.067.045l1.344.567a.21.21 0 0 0 .08.016.19.19 0 0 0 .078-.015.204.204 0 0 0 .114-.113l2.025-4.798a.21.21 0 0 0-.11-.274l-1.344-.566a.233.233 0 0 0-.08-.015Z" })
-        ] }),
-        /* @__PURE__ */ jsx60("defs", { children: /* @__PURE__ */ jsx60("clipPath", { id: "cha", children: /* @__PURE__ */ jsx60("path", { fill: "currentColor", d: "M0 0h20v20H0z" }) }) })
-      ]
-    }
-  );
-};
-
-// src/icon/icons/heart/index.react.tsx
-import { jsx as jsx61, jsxs as jsxs53 } from "react/jsx-runtime";
-var Heart = (props) => {
-  let _ariaLabel = props["aria-label"];
-  _ariaLabel = _ariaLabel && _ariaLabel.trim();
-  const _ariaHidden = !_ariaLabel;
-  const _role = _ariaHidden ? "presentation" : "img";
-  return /* @__PURE__ */ jsxs53(
-    "svg",
-    {
-      className: props.className,
-      viewBox: "0 0 20 20",
-      "aria-label": _ariaLabel,
-      "aria-hidden": _ariaHidden,
-      role: _role,
-      style: {
-        height: "var(--icon-height)",
-        width: "var(--icon-width)",
-        display: "flex",
-        alignItems: "center",
-        justifyContent: "center"
-      },
-      children: [
-        /* @__PURE__ */ jsx61("g", { "clip-path": "url(#cia)", children: /* @__PURE__ */ jsx61("path", { fill: "currentColor", d: "M9.998 18.83a.627.627 0 0 1-.45-.192l-7.99-8.333a5.337 5.337 0 0 1-.994-6.172A5.33 5.33 0 0 1 5.36 1.176c.834 0 1.637.19 2.386.564a5.33 5.33 0 0 1 1.39 1.002l.864.863.864-.863a5.315 5.315 0 0 1 3.785-1.565 5.32 5.32 0 0 1 3.785 1.565 5.347 5.347 0 0 1 0 7.569l-7.982 8.326a.63.63 0 0 1-.453.194ZM5.355 2.424c-.44 0-.877.072-1.3.213a4.075 4.075 0 0 0-2.373 2.056A4.096 4.096 0 0 0 2.45 9.43l7.548 7.872 7.54-7.865a4.106 4.106 0 0 0 .009-5.81 4.075 4.075 0 0 0-2.901-1.199 4.075 4.075 0 0 0-2.9 1.2L10.44 4.931a.624.624 0 0 1-.883 0L8.252 3.627a4.108 4.108 0 0 0-1.066-.768 4.06 4.06 0 0 0-1.83-.436Z" }) }),
-        /* @__PURE__ */ jsx61("defs", { children: /* @__PURE__ */ jsx61("clipPath", { id: "cia", children: /* @__PURE__ */ jsx61("path", { fill: "currentColor", d: "M0 0h20v20H0z" }) }) })
-      ]
-    }
-  );
-};
-
-// src/icon/icons/home/index.react.tsx
-import { jsx as jsx62, jsxs as jsxs54 } from "react/jsx-runtime";
-var Home = (props) => {
-  let _ariaLabel = props["aria-label"];
-  _ariaLabel = _ariaLabel && _ariaLabel.trim();
-  const _ariaHidden = !_ariaLabel;
-  const _role = _ariaHidden ? "presentation" : "img";
-  return /* @__PURE__ */ jsxs54(
-    "svg",
-    {
-      className: props.className,
-      viewBox: "0 0 20 20",
-      "aria-label": _ariaLabel,
-      "aria-hidden": _ariaHidden,
-      role: _role,
-      style: {
-        height: "var(--icon-height)",
-        width: "var(--icon-width)",
-        display: "flex",
-        alignItems: "center",
-        justifyContent: "center"
-      },
-      children: [
-        /* @__PURE__ */ jsx62("path", { fill: "currentColor", d: "M2.307 8.959a.57.57 0 0 1-.314-.094.526.526 0 0 1-.2-.246.495.495 0 0 1-.017-.312.518.518 0 0 1 .174-.264l7.692-6.25a.566.566 0 0 1 .715 0l7.692 6.25c.055.044.1.098.133.16a.497.497 0 0 1-.073.574.545.545 0 0 1-.375.18.577.577 0 0 1-.4-.124l-7.335-5.96-7.334 5.96a.565.565 0 0 1-.358.126Z" }),
-        /* @__PURE__ */ jsx62("path", { fill: "currentColor", d: "M11.648 18.333a.565.565 0 0 1-.389-.152.507.507 0 0 1-.16-.369v-4.687H8.9v4.688a.51.51 0 0 1-.16.368.565.565 0 0 1-.39.152H4.123a1.308 1.308 0 0 1-.892-.352 1.175 1.175 0 0 1-.373-.846v-6.151c.015-.357.177-.693.452-.937l6.333-5.13a.566.566 0 0 1 .714 0l6.317 5.116c.283.249.446.573.468.92v6.181a1.175 1.175 0 0 1-.372.846 1.308 1.308 0 0 1-.892.353h-4.23Zm-7.609-7.508a.29.29 0 0 0-.084.19v6.118c0 .042.019.083.05.112.032.03.075.047.119.047h3.678v-4.348c0-.227.096-.446.266-.607a.945.945 0 0 1 .639-.254h2.582c.24.001.47.092.64.253.17.162.267.38.268.608v4.348h3.679a.173.173 0 0 0 .118-.048.155.155 0 0 0 .05-.112v-6.148a.29.29 0 0 0-.101-.173L9.999 5.998l-5.96 4.827Z" })
-      ]
-    }
-  );
-};
-
-// src/icon/icons/hyperlink/index.react.tsx
-import { jsx as jsx63, jsxs as jsxs55 } from "react/jsx-runtime";
-var Hyperlink = (props) => {
-  let _ariaLabel = props["aria-label"];
-  _ariaLabel = _ariaLabel && _ariaLabel.trim();
-  const _ariaHidden = !_ariaLabel;
-  const _role = _ariaHidden ? "presentation" : "img";
-  return /* @__PURE__ */ jsxs55(
-    "svg",
-    {
-      className: props.className,
-      viewBox: "0 0 20 20",
-      "aria-label": _ariaLabel,
-      "aria-hidden": _ariaHidden,
-      role: _role,
-      style: {
-        height: "var(--icon-height)",
-        width: "var(--icon-width)",
-        display: "flex",
-        alignItems: "center",
-        justifyContent: "center"
-      },
-      children: [
-        /* @__PURE__ */ jsxs55("g", { "clip-path": "url(#cka)", children: [
-          /* @__PURE__ */ jsx63("path", { fill: "currentColor", d: "M4.378 17.914a4.345 4.345 0 0 1-3.094-1.279 4.344 4.344 0 0 1-1.279-3.093c0-1.17.454-2.269 1.28-3.094L5.261 6.47a4.346 4.346 0 0 1 3.096-1.282c1.169 0 2.266.453 3.09 1.275a4.354 4.354 0 0 1 .449 5.67.627.627 0 0 1-1.04-.04.621.621 0 0 1-.074-.473.621.621 0 0 1 .103-.222 3.11 3.11 0 0 0-.32-4.05 3.104 3.104 0 0 0-2.208-.911 3.107 3.107 0 0 0-2.212.916L2.168 11.33a3.128 3.128 0 0 0 0 4.419 3.1 3.1 0 0 0 2.21.913c.835 0 1.62-.324 2.21-.913l.776-.777a.619.619 0 0 1 .884 0 .62.62 0 0 1 0 .884l-.777.776a4.34 4.34 0 0 1-3.093 1.281Z" }),
-          /* @__PURE__ */ jsx63("path", { fill: "currentColor", d: "M11.648 15.02a4.346 4.346 0 0 1-3.09-1.28 4.353 4.353 0 0 1-.447-5.668.627.627 0 0 1 1.113.513.621.621 0 0 1-.102.222 3.11 3.11 0 0 0 .319 4.049 3.104 3.104 0 0 0 2.208.913 3.106 3.106 0 0 0 2.212-.917l3.977-3.977c.59-.59.915-1.375.915-2.21s-.325-1.62-.915-2.21a3.104 3.104 0 0 0-2.21-.913c-.835 0-1.62.324-2.21.913l-.625.625a.62.62 0 0 1-.883 0 .619.619 0 0 1-.183-.442c0-.166.065-.324.183-.441l.625-.625a4.347 4.347 0 0 1 3.093-1.28c1.17 0 2.269.454 3.094 1.28a4.345 4.345 0 0 1 1.281 3.093 4.345 4.345 0 0 1-1.281 3.093l-3.978 3.978a4.345 4.345 0 0 1-3.092 1.283l-.004.001Z" })
-        ] }),
-        /* @__PURE__ */ jsx63("defs", { children: /* @__PURE__ */ jsx63("clipPath", { id: "cka", children: /* @__PURE__ */ jsx63("path", { fill: "currentColor", d: "M0 0h20v20H0z" }) }) })
-      ]
-    }
-  );
-};
-
-// src/icon/icons/image/index.react.tsx
-import { jsx as jsx64, jsxs as jsxs56 } from "react/jsx-runtime";
-var Image = (props) => {
-  let _ariaLabel = props["aria-label"];
-  _ariaLabel = _ariaLabel && _ariaLabel.trim();
-  const _ariaHidden = !_ariaLabel;
-  const _role = _ariaHidden ? "presentation" : "img";
-  return /* @__PURE__ */ jsxs56(
-    "svg",
-    {
-      className: props.className,
-      viewBox: "0 0 20 20",
-      "aria-label": _ariaLabel,
-      "aria-hidden": _ariaHidden,
-      role: _role,
-      style: {
-        height: "var(--icon-height)",
-        width: "var(--icon-width)",
-        display: "flex",
-        alignItems: "center",
-        justifyContent: "center"
-      },
-      children: [
-        /* @__PURE__ */ jsxs56("g", { "clip-path": "url(#cla)", children: [
-          /* @__PURE__ */ jsx64("path", { fill: "currentColor", d: "M2.52 20a1.876 1.876 0 0 1-1.876-1.875V1.875A1.877 1.877 0 0 1 2.52 0h12.5c.492 0 .956.19 1.31.532l2.462 2.403c.359.352.565.84.565 1.342v13.848A1.877 1.877 0 0 1 17.48 20H2.519Zm0-18.75a.625.625 0 0 0-.626.625v16.25c0 .345.28.625.625.625h14.962a.625.625 0 0 0 .625-.625V4.277a.628.628 0 0 0-.189-.447l-2.462-2.402a.62.62 0 0 0-.436-.178H2.52Z" }),
-          /* @__PURE__ */ jsx64("path", { fill: "currentColor", d: "M6.894 8.75a2.503 2.503 0 0 1-2.5-2.5c0-1.378 1.122-2.5 2.5-2.5 1.379 0 2.5 1.122 2.5 2.5s-1.121 2.5-2.5 2.5Zm0-3.75a1.251 1.251 0 0 0 0 2.5 1.25 1.25 0 0 0 0-2.5Zm8.784 10.792a.625.625 0 0 1-.521-.279l-3.271-4.906a.22.22 0 0 0-.303-.062.22.22 0 0 0-.057.057l-2.243 3.205a.625.625 0 0 1-.904.129L6.99 12.82a.214.214 0 0 0-.159-.046.213.213 0 0 0-.146.08l-1.77 2.658a.625.625 0 1 1-1.041-.692l1.759-2.643a1.455 1.455 0 0 1 1.226-.655c.334 0 .65.112.913.323l.87.697 1.86-2.657a1.465 1.465 0 0 1 2.425.028l3.27 4.906a.624.624 0 0 1-.519.972Z" })
-        ] }),
-        /* @__PURE__ */ jsx64("defs", { children: /* @__PURE__ */ jsx64("clipPath", { id: "cla", children: /* @__PURE__ */ jsx64("path", { fill: "currentColor", d: "M0 0h20v20H0z" }) }) })
-      ]
-    }
-  );
-};
-
-// src/icon/icons/image_flower/index.react.tsx
-import { jsx as jsx65, jsxs as jsxs57 } from "react/jsx-runtime";
-var ImageFlower = (props) => {
-  let _ariaLabel = props["aria-label"];
-  _ariaLabel = _ariaLabel && _ariaLabel.trim();
-  const _ariaHidden = !_ariaLabel;
-  const _role = _ariaHidden ? "presentation" : "img";
-  return /* @__PURE__ */ jsxs57(
-    "svg",
-    {
-      className: props.className,
-      viewBox: "0 0 20 20",
-      "aria-label": _ariaLabel,
-      "aria-hidden": _ariaHidden,
-      role: _role,
-      style: {
-        height: "var(--icon-height)",
-        width: "var(--icon-width)",
-        display: "flex",
-        alignItems: "center",
-        justifyContent: "center"
-      },
-      children: [
-        /* @__PURE__ */ jsxs57("g", { "clip-path": "url(#cma)", children: [
-          /* @__PURE__ */ jsx65("path", { fill: "currentColor", d: "M1.875 20A1.876 1.876 0 0 1 0 18.125V1.875A1.877 1.877 0 0 1 1.875 0h16.25A1.877 1.877 0 0 1 20 1.875v16.25A1.876 1.876 0 0 1 18.125 20H1.875Zm16.25-1.25a.624.624 0 0 0 .625-.625V1.875a.625.625 0 0 0-.625-.625H1.875a.625.625 0 0 0-.625.625v16.25c0 .345.28.625.625.625h7.5v-5.17c-.634-.294-1.165-1.015-1.388-1.91a2.49 2.49 0 0 1-2.702-.6.628.628 0 0 1-.146-.25 2.49 2.49 0 0 1 .827-2.64 2.492 2.492 0 0 1-.827-2.637.636.636 0 0 1 .146-.253 2.49 2.49 0 0 1 2.692-.605c.312-1.206 1.14-2.05 2.024-2.05.884 0 1.712.844 2.012 2.054a2.487 2.487 0 0 1 2.704.601.625.625 0 0 1 .145.252 2.49 2.49 0 0 1-.826 2.638 2.494 2.494 0 0 1 .827 2.638.63.63 0 0 1-.147.254 2.49 2.49 0 0 1-2.702.599c-.221.893-.752 1.615-1.387 1.91v5.169h7.498Zm-10.14-8.333c.555 0 1.05.371 1.202.903l.012.043c.184.742.613 1.112.801 1.112.188 0 .617-.37.8-1.108l.002-.006a1.25 1.25 0 0 1 1.516-.907l.042.011a.734.734 0 0 1 .07.025 1.23 1.23 0 0 0 1.268-.177 1.227 1.227 0 0 0-.476-1.184.655.655 0 0 1-.059-.049 1.25 1.25 0 0 1-.035-1.767l.03-.03a.565.565 0 0 1 .064-.055c.37-.276.551-.732.476-1.183a1.222 1.222 0 0 0-1.264-.178 1.229 1.229 0 0 1-1.171-.177 1.258 1.258 0 0 1-.451-.65l-.012-.044c-.183-.741-.612-1.111-.8-1.111-.188 0-.617.37-.8 1.108a1.248 1.248 0 0 1-1.514.913l-.038-.01a.622.622 0 0 1-.081-.029 1.228 1.228 0 0 0-1.264.179 1.225 1.225 0 0 0 .476 1.183.68.68 0 0 1 .059.05c.497.479.512 1.272.034 1.768l-.029.03a.691.691 0 0 1-.064.055 1.223 1.223 0 0 0-.476 1.183 1.228 1.228 0 0 0 1.262.178.767.767 0 0 1 .073-.026c.115-.034.232-.05.348-.05Z" }),
-          /* @__PURE__ */ jsx65("path", { fill: "currentColor", d: "M10 10a1.877 1.877 0 0 1-1.875-1.875c0-1.034.84-1.875 1.875-1.875 1.034 0 1.875.84 1.875 1.875C11.875 9.159 11.035 10 10 10Zm0-2.5a.625.625 0 1 0 0 1.25.625.625 0 0 0 0-1.25ZM7.26 17.808a.643.643 0 0 1-.113-.01 4.305 4.305 0 0 1-1.627-.663c-1.584-.97-2.808-3.428-2.86-3.533a.625.625 0 0 1 .52-.9 8.365 8.365 0 0 1 4.683 1.081.623.623 0 0 1-.31 1.168.619.619 0 0 1-.309-.083 7.129 7.129 0 0 0-2.943-.91c.43.684 1.135 1.66 1.894 2.125.37.248.762.409 1.177.485a.625.625 0 0 1-.113 1.24Zm5.533 0a.625.625 0 0 1-.113-1.24 3.054 3.054 0 0 0 1.156-.471c.772-.475 1.482-1.453 1.914-2.139a7.108 7.108 0 0 0-2.943.911.627.627 0 0 1-.803-.925.627.627 0 0 1 .184-.16 8.37 8.37 0 0 1 4.684-1.08.626.626 0 0 1 .52.9c-.051.104-1.275 2.561-2.88 3.546a4.239 4.239 0 0 1-1.604.648.78.78 0 0 1-.115.01Z" })
-        ] }),
-        /* @__PURE__ */ jsx65("defs", { children: /* @__PURE__ */ jsx65("clipPath", { id: "cma", children: /* @__PURE__ */ jsx65("path", { fill: "currentColor", d: "M0 0h20v20H0z" }) }) })
-      ]
-    }
-  );
-};
-
-// src/icon/icons/indent_decrease/index.react.tsx
-import { jsx as jsx66, jsxs as jsxs58 } from "react/jsx-runtime";
-var IndentDecrease = (props) => {
-  let _ariaLabel = props["aria-label"];
-  _ariaLabel = _ariaLabel && _ariaLabel.trim();
-  const _ariaHidden = !_ariaLabel;
-  const _role = _ariaHidden ? "presentation" : "img";
-  return /* @__PURE__ */ jsxs58(
-    "svg",
-    {
-      className: props.className,
-      viewBox: "0 0 20 20",
-      "aria-label": _ariaLabel,
-      "aria-hidden": _ariaHidden,
-      role: _role,
-      style: {
-        height: "var(--icon-height)",
-        width: "var(--icon-width)",
-        display: "flex",
-        alignItems: "center",
-        justifyContent: "center"
-      },
-      children: [
-        /* @__PURE__ */ jsx66("g", { "clip-path": "url(#cna)", children: /* @__PURE__ */ jsx66("path", { fill: "currentColor", d: "M9.376 3.123a.625.625 0 0 1 0-1.25h10a.625.625 0 1 1 0 1.25h-10Zm5 3.75a.625.625 0 0 1 0-1.25h5a.625.625 0 1 1 0 1.25h-5Zm0 3.75a.625.625 0 0 1 0-1.25h5a.625.625 0 1 1 0 1.25h-5Zm0 3.75a.625.625 0 0 1 0-1.25h5a.625.625 0 1 1 0 1.25h-5Zm-5 3.75a.625.625 0 0 1 0-1.25h10a.625.625 0 1 1 0 1.25h-10ZM6.25 16.87c-.333 0-.647-.13-.883-.364L.468 11.235a1.877 1.877 0 0 1-.017-2.456l4.9-5.273a1.263 1.263 0 0 1 .9-.383c.334 0 .648.13.884.367.236.236.366.55.366.884v1.251h3.125A1.877 1.877 0 0 1 12.5 7.5v5a1.877 1.877 0 0 1-1.875 1.875H7.5v1.247c0 .333-.13.646-.366.882a1.24 1.24 0 0 1-.884.367ZM1.384 9.61a.613.613 0 0 0 .018.793l4.846 5.214.004-1.867a.625.625 0 0 1 .625-.625h3.75a.625.625 0 0 0 .625-.625v-5a.625.625 0 0 0-.625-.625h-3.75a.625.625 0 0 1-.625-.625V4.374L1.383 9.611Z" }) }),
-        /* @__PURE__ */ jsx66("defs", { children: /* @__PURE__ */ jsx66("clipPath", { id: "cna", children: /* @__PURE__ */ jsx66("path", { fill: "currentColor", d: "M0 0h20v20H0z" }) }) })
-      ]
-    }
-  );
-};
-
-// src/icon/icons/indent_increase/index.react.tsx
-import { jsx as jsx67, jsxs as jsxs59 } from "react/jsx-runtime";
-var IndentIncrease = (props) => {
-  let _ariaLabel = props["aria-label"];
-  _ariaLabel = _ariaLabel && _ariaLabel.trim();
-  const _ariaHidden = !_ariaLabel;
-  const _role = _ariaHidden ? "presentation" : "img";
-  return /* @__PURE__ */ jsxs59(
-    "svg",
-    {
-      className: props.className,
-      viewBox: "0 0 20 20",
-      "aria-label": _ariaLabel,
-      "aria-hidden": _ariaHidden,
-      role: _role,
-      style: {
-        height: "var(--icon-height)",
-        width: "var(--icon-width)",
-        display: "flex",
-        alignItems: "center",
-        justifyContent: "center"
-      },
-      children: [
-        /* @__PURE__ */ jsx67("g", { "clip-path": "url(#coa)", children: /* @__PURE__ */ jsx67("path", { fill: "currentColor", d: "M.624 3.123a.625.625 0 0 1 0-1.25h10a.625.625 0 1 1 0 1.25h-10Zm0 3.75a.625.625 0 0 1 0-1.25h5a.625.625 0 0 1 0 1.25h-5Zm0 3.75a.625.625 0 0 1 0-1.25h5a.625.625 0 0 1 0 1.25h-5Zm0 3.75a.625.625 0 0 1 0-1.25h5a.625.625 0 1 1 0 1.25h-5Zm0 3.75a.625.625 0 0 1 0-1.25h10a.625.625 0 1 1 0 1.25h-10Zm13.122-1.253c-.333 0-.645-.128-.88-.363a1.24 1.24 0 0 1-.368-.885v-1.247H9.373A1.877 1.877 0 0 1 7.498 12.5v-5a1.877 1.877 0 0 1 1.875-1.875h3.125V4.373a1.244 1.244 0 0 1 .773-1.156 1.248 1.248 0 0 1 1.36.27l4.901 5.272c.617.721.621 1.75.018 2.456l-4.9 5.274a1.265 1.265 0 0 1-.904.382ZM9.374 6.876a.625.625 0 0 0-.625.625v5c0 .345.28.625.625.625h3.75a.625.625 0 0 1 .625.625v1.872l4.868-5.239a.613.613 0 0 0-.017-.792l-4.846-5.213-.005 1.872a.625.625 0 0 1-.625.625h-3.75Z" }) }),
-        /* @__PURE__ */ jsx67("defs", { children: /* @__PURE__ */ jsx67("clipPath", { id: "coa", children: /* @__PURE__ */ jsx67("path", { fill: "currentColor", d: "M0 0h20v20H0z" }) }) })
-      ]
-    }
-  );
-};
-
-// src/icon/icons/information/index.react.tsx
-import { jsx as jsx68, jsxs as jsxs60 } from "react/jsx-runtime";
-var Information = (props) => {
-  let _ariaLabel = props["aria-label"];
-  _ariaLabel = _ariaLabel && _ariaLabel.trim();
-  const _ariaHidden = !_ariaLabel;
-  const _role = _ariaHidden ? "presentation" : "img";
-  return /* @__PURE__ */ jsxs60(
-    "svg",
-    {
-      className: props.className,
-      viewBox: "0 0 20 20",
-      "aria-label": _ariaLabel,
-      "aria-hidden": _ariaHidden,
-      role: _role,
-      style: {
-        height: "var(--icon-height)",
-        width: "var(--icon-width)",
-        display: "flex",
-        alignItems: "center",
-        justifyContent: "center"
-      },
-      children: [
-        /* @__PURE__ */ jsxs60("g", { "clip-path": "url(#cpa)", children: [
-          /* @__PURE__ */ jsx68("path", { fill: "currentColor", d: "M11.25 14.375A1.876 1.876 0 0 1 9.375 12.5V9.375H8.75a.625.625 0 0 1 0-1.25h.625c.69 0 1.25.56 1.25 1.25V12.5c0 .345.28.625.625.625h.625a.624.624 0 1 1 0 1.25h-.625Zm-1.562-7.5a.937.937 0 1 0 0-1.875.937.937 0 0 0 0 1.875Z" }),
-          /* @__PURE__ */ jsx68("path", { fill: "currentColor", d: "M10 20C4.486 20 0 15.514 0 10S4.486 0 10 0s10 4.486 10 10-4.486 10-10 10Zm0-18.75c-4.825 0-8.75 3.925-8.75 8.75s3.925 8.75 8.75 8.75 8.75-3.925 8.75-8.75S14.825 1.25 10 1.25Z" })
-        ] }),
-        /* @__PURE__ */ jsx68("defs", { children: /* @__PURE__ */ jsx68("clipPath", { id: "cpa", children: /* @__PURE__ */ jsx68("path", { fill: "currentColor", d: "M0 0h20v20H0z" }) }) })
-      ]
-    }
-  );
-};
-
-// src/icon/icons/keyboard/index.react.tsx
-import { jsx as jsx69, jsxs as jsxs61 } from "react/jsx-runtime";
-var Keyboard = (props) => {
-  let _ariaLabel = props["aria-label"];
-  _ariaLabel = _ariaLabel && _ariaLabel.trim();
-  const _ariaHidden = !_ariaLabel;
-  const _role = _ariaHidden ? "presentation" : "img";
-  return /* @__PURE__ */ jsxs61(
-    "svg",
-    {
-      className: props.className,
-      viewBox: "0 0 20 20",
-      "aria-label": _ariaLabel,
-      "aria-hidden": _ariaHidden,
-      role: _role,
-      style: {
-        height: "var(--icon-height)",
-        width: "var(--icon-width)",
-        display: "flex",
-        alignItems: "center",
-        justifyContent: "center"
-      },
-      children: [
-        /* @__PURE__ */ jsxs61("g", { "clip-path": "url(#cqa)", children: [
-          /* @__PURE__ */ jsx69("path", { fill: "currentColor", d: "M3.125 20A3.128 3.128 0 0 1 0 16.875v-5A3.128 3.128 0 0 1 3.125 8.75H5v-.625A4.38 4.38 0 0 1 9.375 3.75H12.5A3.128 3.128 0 0 0 15.625.625a.625.625 0 0 1 1.25 0A4.38 4.38 0 0 1 12.5 5H9.375A3.128 3.128 0 0 0 6.25 8.125v.625h10.625A3.128 3.128 0 0 1 20 11.875v5A3.128 3.128 0 0 1 16.875 20H3.125Zm0-10c-1.034 0-1.875.84-1.875 1.875v5c0 1.034.84 1.875 1.875 1.875h13.75c1.034 0 1.875-.84 1.875-1.875v-5c0-1.034-.84-1.875-1.875-1.875H3.125Z" }),
-          /* @__PURE__ */ jsx69("path", { fill: "currentColor", d: "M5.625 12.5a.625.625 0 0 1 0-1.25h1.25c.345 0 .625.28.625.625s-1.875.625-1.875.625Zm7.5 0a.625.625 0 0 1 0-1.25h1.25a.625.625 0 0 1 0 1.25h-1.25Zm-3.75 0a.625.625 0 0 1 0-1.25h1.25a.625.625 0 0 1 0 1.25h-1.25ZM3.75 15a.625.625 0 0 1 0-1.25H5A.625.625 0 0 1 5 15H3.75Zm3.75 0a.625.625 0 0 1 0-1.25h1.25a.625.625 0 0 1 0 1.25H7.5Zm3.75 0a.625.625 0 0 1 0-1.25h1.25a.625.625 0 0 1 0 1.25h-1.25ZM15 15a.625.625 0 0 1 0-1.25h1.25a.625.625 0 0 1 0 1.25H15Zm-9.375 2.5a.625.625 0 0 1 0-1.25h8.75a.625.625 0 0 1 0 1.25h-8.75Z" })
-        ] }),
-        /* @__PURE__ */ jsx69("defs", { children: /* @__PURE__ */ jsx69("clipPath", { id: "cqa", children: /* @__PURE__ */ jsx69("path", { fill: "currentColor", d: "M0 0h20v20H0z" }) }) })
-      ]
-    }
-  );
-};
-
-// src/icon/icons/layers/index.react.tsx
-import { jsx as jsx70, jsxs as jsxs62 } from "react/jsx-runtime";
-var Layers = (props) => {
-  let _ariaLabel = props["aria-label"];
-  _ariaLabel = _ariaLabel && _ariaLabel.trim();
-  const _ariaHidden = !_ariaLabel;
-  const _role = _ariaHidden ? "presentation" : "img";
-  return /* @__PURE__ */ jsxs62(
-    "svg",
-    {
-      className: props.className,
-      viewBox: "0 0 20 20",
-      "aria-label": _ariaLabel,
-      "aria-hidden": _ariaHidden,
-      role: _role,
-      style: {
-        height: "var(--icon-height)",
-        width: "var(--icon-width)",
-        display: "flex",
-        alignItems: "center",
-        justifyContent: "center"
-      },
-      children: [
-        /* @__PURE__ */ jsx70("g", { "clip-path": "url(#cra)", children: /* @__PURE__ */ jsx70("path", { fill: "currentColor", d: "M10 19.998a1.92 1.92 0 0 1-.877-.213L.616 15.383a1.14 1.14 0 0 1 0-2.021l2.274-1.175L.617 11.01a1.133 1.133 0 0 1-.612-.914 1.132 1.132 0 0 1 .264-.831c.096-.115.215-.209.348-.278L2.893 7.81.616 6.633A1.135 1.135 0 0 1 .267 4.89a1.14 1.14 0 0 1 .348-.277l8.507-4.4a1.917 1.917 0 0 1 1.755 0l8.507 4.398a.616.616 0 0 1 .09.057 1.14 1.14 0 0 1-.088 1.969l-2.274 1.175 2.274 1.176a1.138 1.138 0 0 1 0 2.021l-2.275 1.178 2.275 1.178a1.137 1.137 0 0 1 0 2.021l-8.508 4.397a1.903 1.903 0 0 1-.878.215Zm-.303-1.324a.667.667 0 0 0 .606 0l8.316-4.296-2.87-1.486-4.872 2.52a1.92 1.92 0 0 1-1.754 0l-4.873-2.52-2.868 1.482 8.315 4.301Zm0-4.372a.657.657 0 0 0 .605 0L18.618 10 15.75 8.517l-4.873 2.518a1.922 1.922 0 0 1-1.754 0l-4.87-2.52L1.381 10l8.315 4.302Zm0-4.378a.668.668 0 0 0 .606 0l8.32-4.299-8.32-4.302a.672.672 0 0 0-.606 0l-8.315 4.3 8.315 4.301Z" }) }),
-        /* @__PURE__ */ jsx70("defs", { children: /* @__PURE__ */ jsx70("clipPath", { id: "cra", children: /* @__PURE__ */ jsx70("path", { fill: "currentColor", d: "M0 0h20v20H0z" }) }) })
-      ]
-    }
-  );
-};
-
-// src/icon/icons/layers_back/index.react.tsx
-import { jsx as jsx71, jsxs as jsxs63 } from "react/jsx-runtime";
-var LayersBack = (props) => {
-  let _ariaLabel = props["aria-label"];
-  _ariaLabel = _ariaLabel && _ariaLabel.trim();
-  const _ariaHidden = !_ariaLabel;
-  const _role = _ariaHidden ? "presentation" : "img";
-  return /* @__PURE__ */ jsxs63(
-    "svg",
-    {
-      className: props.className,
-      viewBox: "0 0 20 20",
-      "aria-label": _ariaLabel,
-      "aria-hidden": _ariaHidden,
-      role: _role,
-      style: {
-        height: "var(--icon-height)",
-        width: "var(--icon-width)",
-        display: "flex",
-        alignItems: "center",
-        justifyContent: "center"
-      },
-      children: [
-        /* @__PURE__ */ jsxs63("g", { "clip-path": "url(#csa)", children: [
-          /* @__PURE__ */ jsx71("path", { fill: "currentColor", d: "M1.875 16.25A1.876 1.876 0 0 1 0 14.375v-12.5A1.877 1.877 0 0 1 1.875 0h12.5a1.877 1.877 0 0 1 1.875 1.875v12.5a1.876 1.876 0 0 1-1.875 1.875h-12.5Zm0-15a.625.625 0 0 0-.625.625v12.5c0 .345.28.625.625.625h12.5a.624.624 0 0 0 .625-.625v-12.5a.625.625 0 0 0-.625-.625h-12.5Z" }),
-          /* @__PURE__ */ jsx71("path", { fill: "currentColor", d: "M5.625 20a1.876 1.876 0 0 1-1.875-1.875.625.625 0 1 1 1.25 0c0 .345.28.625.625.625h12.5a.624.624 0 0 0 .625-.625v-12.5A.625.625 0 0 0 18.125 5a.625.625 0 1 1 0-1.25A1.877 1.877 0 0 1 20 5.625v12.5A1.876 1.876 0 0 1 18.125 20h-12.5Z" })
-        ] }),
-        /* @__PURE__ */ jsx71("defs", { children: /* @__PURE__ */ jsx71("clipPath", { id: "csa", children: /* @__PURE__ */ jsx71("path", { fill: "currentColor", d: "M0 0h20v20H0z" }) }) })
-      ]
-    }
-  );
-};
-
-// src/icon/icons/layout_left/index.react.tsx
-import { jsx as jsx72, jsxs as jsxs64 } from "react/jsx-runtime";
-var LayoutLeft = (props) => {
-  let _ariaLabel = props["aria-label"];
-  _ariaLabel = _ariaLabel && _ariaLabel.trim();
-  const _ariaHidden = !_ariaLabel;
-  const _role = _ariaHidden ? "presentation" : "img";
-  return /* @__PURE__ */ jsxs64(
-    "svg",
-    {
-      className: props.className,
-      viewBox: "0 0 20 20",
-      "aria-label": _ariaLabel,
-      "aria-hidden": _ariaHidden,
-      role: _role,
-      style: {
-        height: "var(--icon-height)",
-        width: "var(--icon-width)",
-        display: "flex",
-        alignItems: "center",
-        justifyContent: "center"
-      },
-      children: [
-        /* @__PURE__ */ jsx72("g", { "clip-path": "url(#cta)", children: /* @__PURE__ */ jsx72("path", { fill: "currentColor", d: "M2.5 19.373a1.877 1.877 0 0 1-1.875-1.875v-15A1.877 1.877 0 0 1 2.5.623h15a1.877 1.877 0 0 1 1.875 1.875v15a1.876 1.876 0 0 1-1.875 1.875h-15Zm15-1.25a.625.625 0 0 0 .625-.625v-15a.625.625 0 0 0-.625-.625H6.875v16.25H17.5Zm-15-16.25a.625.625 0 0 0-.625.625v15c0 .345.28.625.625.625h3.125V1.873H2.5Z" }) }),
-        /* @__PURE__ */ jsx72("defs", { children: /* @__PURE__ */ jsx72("clipPath", { id: "cta", children: /* @__PURE__ */ jsx72("path", { fill: "currentColor", d: "M0 0h20v20H0z" }) }) })
-      ]
-    }
-  );
-};
-
-// src/icon/icons/layout_module/index.react.tsx
-import { jsx as jsx73, jsxs as jsxs65 } from "react/jsx-runtime";
-var LayoutModule = (props) => {
-  let _ariaLabel = props["aria-label"];
-  _ariaLabel = _ariaLabel && _ariaLabel.trim();
-  const _ariaHidden = !_ariaLabel;
-  const _role = _ariaHidden ? "presentation" : "img";
-  return /* @__PURE__ */ jsxs65(
-    "svg",
-    {
-      className: props.className,
-      viewBox: "0 0 20 20",
-      "aria-label": _ariaLabel,
-      "aria-hidden": _ariaHidden,
-      role: _role,
-      style: {
-        height: "var(--icon-height)",
-        width: "var(--icon-width)",
-        display: "flex",
-        alignItems: "center",
-        justifyContent: "center"
-      },
-      children: [
-        /* @__PURE__ */ jsx73("g", { "clip-path": "url(#cua)", children: /* @__PURE__ */ jsx73("path", { fill: "currentColor", d: "M1.875 8.748A1.877 1.877 0 0 1 0 6.873v-5A1.877 1.877 0 0 1 1.875-.002h5A1.877 1.877 0 0 1 8.75 1.873v5a1.877 1.877 0 0 1-1.875 1.875h-5Zm0-7.5a.625.625 0 0 0-.625.625v5c0 .345.28.625.625.625h5a.625.625 0 0 0 .625-.625v-5a.625.625 0 0 0-.625-.625h-5Zm0 18.75A1.877 1.877 0 0 1 0 18.123v-5a1.877 1.877 0 0 1 1.875-1.875h5a1.877 1.877 0 0 1 1.875 1.875v5a1.877 1.877 0 0 1-1.875 1.875h-5Zm0-7.5a.625.625 0 0 0-.625.625v5c0 .345.28.625.625.625h5a.625.625 0 0 0 .625-.625v-5a.625.625 0 0 0-.625-.625h-5Zm11.25-3.75a1.877 1.877 0 0 1-1.875-1.875v-5a1.876 1.876 0 0 1 1.875-1.875h5A1.877 1.877 0 0 1 20 1.873v5a1.876 1.876 0 0 1-1.875 1.875h-5Zm0-7.5a.625.625 0 0 0-.625.625v5c0 .345.28.625.625.625h5a.625.625 0 0 0 .625-.625v-5a.625.625 0 0 0-.625-.625h-5Zm0 18.75a1.877 1.877 0 0 1-1.875-1.875v-5a1.876 1.876 0 0 1 1.875-1.875h5A1.877 1.877 0 0 1 20 13.123v5a1.876 1.876 0 0 1-1.875 1.875h-5Zm0-7.5a.625.625 0 0 0-.625.625v5c0 .345.28.625.625.625h5a.625.625 0 0 0 .625-.625v-5a.625.625 0 0 0-.625-.625h-5Z" }) }),
-        /* @__PURE__ */ jsx73("defs", { children: /* @__PURE__ */ jsx73("clipPath", { id: "cua", children: /* @__PURE__ */ jsx73("path", { fill: "currentColor", d: "M0 0h20v20H0z" }) }) })
-      ]
-    }
-  );
-};
-
-// src/icon/icons/layout_right/index.react.tsx
-import { jsx as jsx74, jsxs as jsxs66 } from "react/jsx-runtime";
-var LayoutRight = (props) => {
-  let _ariaLabel = props["aria-label"];
-  _ariaLabel = _ariaLabel && _ariaLabel.trim();
-  const _ariaHidden = !_ariaLabel;
-  const _role = _ariaHidden ? "presentation" : "img";
-  return /* @__PURE__ */ jsxs66(
-    "svg",
-    {
-      className: props.className,
-      viewBox: "0 0 20 20",
-      "aria-label": _ariaLabel,
-      "aria-hidden": _ariaHidden,
-      role: _role,
-      style: {
-        height: "var(--icon-height)",
-        width: "var(--icon-width)",
-        display: "flex",
-        alignItems: "center",
-        justifyContent: "center"
-      },
-      children: [
-        /* @__PURE__ */ jsx74("g", { "clip-path": "url(#cva)", children: /* @__PURE__ */ jsx74("path", { fill: "currentColor", d: "M2.5 19.373a1.877 1.877 0 0 1-1.875-1.875v-15A1.877 1.877 0 0 1 2.5.623h15a1.877 1.877 0 0 1 1.875 1.875v15a1.876 1.876 0 0 1-1.875 1.875h-15Zm15-1.25a.625.625 0 0 0 .625-.625v-15a.625.625 0 0 0-.625-.625h-3.125v16.25H17.5Zm-15-16.25a.625.625 0 0 0-.625.625v15c0 .345.28.625.625.625h10.625V1.873H2.5Z" }) }),
-        /* @__PURE__ */ jsx74("defs", { children: /* @__PURE__ */ jsx74("clipPath", { id: "cva", children: /* @__PURE__ */ jsx74("path", { fill: "currentColor", d: "M0 0h20v20H0z" }) }) })
-      ]
-    }
-  );
-};
-
-// src/icon/icons/library/index.react.tsx
-import { jsx as jsx75, jsxs as jsxs67 } from "react/jsx-runtime";
-var Library = (props) => {
-  let _ariaLabel = props["aria-label"];
-  _ariaLabel = _ariaLabel && _ariaLabel.trim();
-  const _ariaHidden = !_ariaLabel;
-  const _role = _ariaHidden ? "presentation" : "img";
-  return /* @__PURE__ */ jsxs67(
-    "svg",
-    {
-      className: props.className,
-      viewBox: "0 0 20 20",
-      "aria-label": _ariaLabel,
-      "aria-hidden": _ariaHidden,
-      role: _role,
-      style: {
-        height: "var(--icon-height)",
-        width: "var(--icon-width)",
-        display: "flex",
-        alignItems: "center",
-        justifyContent: "center"
-      },
-      children: [
-        /* @__PURE__ */ jsx75("g", { "clip-path": "url(#cwa)", children: /* @__PURE__ */ jsx75("path", { fill: "currentColor", d: "M1.972 20a1.876 1.876 0 0 1-1.875-1.875V6.25a1.877 1.877 0 0 1 1.875-1.875h1.682a.625.625 0 0 1 0 1.25H1.972a.625.625 0 0 0-.625.625v11.875c0 .345.28.625.625.625h1.682a.625.625 0 1 1 0 1.25H1.972Zm5 0a1.876 1.876 0 0 1-1.875-1.875V1.875C5.097.841 5.94 0 6.972 0h2.5a1.877 1.877 0 0 1 1.876 1.875v1.403l1.127-.428c.212-.081.438-.123.665-.123.774 0 1.48.486 1.754 1.209l4.888 12.852a1.86 1.86 0 0 1-.042 1.434 1.865 1.865 0 0 1-1.044.985l-1.169.444a1.871 1.871 0 0 1-2.42-1.086l-3.76-9.888v9.448A1.877 1.877 0 0 1 9.472 20h-2.5Zm0-18.75a.625.625 0 0 0-.625.625v16.25c0 .345.28.625.625.625h2.5a.625.625 0 0 0 .626-.625V1.875a.625.625 0 0 0-.626-.625h-2.5Zm6.169 2.727a.614.614 0 0 0-.222.04l-1.168.445a.623.623 0 0 0-.362.807l4.888 12.851a.619.619 0 0 0 .806.36l1.17-.443a.623.623 0 0 0 .362-.807l-4.89-12.85a.629.629 0 0 0-.584-.403Z" }) }),
-        /* @__PURE__ */ jsx75("defs", { children: /* @__PURE__ */ jsx75("clipPath", { id: "cwa", children: /* @__PURE__ */ jsx75("path", { fill: "currentColor", d: "M0 0h20v20H0z" }) }) })
-      ]
-    }
-  );
-};
-
-// src/icon/icons/light_bulb/index.react.tsx
-import { jsx as jsx76, jsxs as jsxs68 } from "react/jsx-runtime";
-var LightBulb = (props) => {
-  let _ariaLabel = props["aria-label"];
-  _ariaLabel = _ariaLabel && _ariaLabel.trim();
-  const _ariaHidden = !_ariaLabel;
-  const _role = _ariaHidden ? "presentation" : "img";
-  return /* @__PURE__ */ jsxs68(
-    "svg",
-    {
-      className: props.className,
-      viewBox: "0 0 20 20",
-      "aria-label": _ariaLabel,
-      "aria-hidden": _ariaHidden,
-      role: _role,
-      style: {
-        height: "var(--icon-height)",
-        width: "var(--icon-width)",
-        display: "flex",
-        alignItems: "center",
-        justifyContent: "center"
-      },
-      children: [
-        /* @__PURE__ */ jsx76("g", { "clip-path": "url(#cxa)", children: /* @__PURE__ */ jsx76("path", { fill: "currentColor", d: "M8.75 19.375a.625.625 0 1 1 0-1.25h2.5a.624.624 0 1 1 0 1.25h-2.5Zm0-2.5a.625.625 0 0 1-.625-.625v-.964a5.59 5.59 0 0 1-3.75-5.324 5.635 5.635 0 0 1 7.788-5.175 5.6 5.6 0 0 1 1.829 1.232 5.59 5.59 0 0 1 1.633 3.983 5.58 5.58 0 0 1-3.75 5.285v.963a.625.625 0 0 1-.625.625h-2.5Zm1.875-1.25v-.802a.625.625 0 0 1 .47-.606 4.363 4.363 0 0 0 3.28-4.219 4.346 4.346 0 0 0-1.27-3.099 4.345 4.345 0 0 0-3.09-1.292l-.017-.313v.313a4.382 4.382 0 0 0-4.372 4.359 4.371 4.371 0 0 0 3.28 4.252.625.625 0 0 1 .469.604v.802l1.25.001ZM10 3.125a.625.625 0 0 1-.625-.625V1.25a.625.625 0 0 1 1.25 0V2.5a.625.625 0 0 1-.625.625Zm7.5 6.25a.625.625 0 1 1 0-1.25h1.25a.625.625 0 1 1 0 1.25H17.5Zm-16.25 1.25a.625.625 0 1 1 0-1.25H2.5a.625.625 0 0 1 0 1.25H1.25Zm14.938 5.563a.62.62 0 0 1-.442-.184l-.884-.884a.627.627 0 0 1 .441-1.067c.167 0 .324.064.442.183l.884.884a.626.626 0 0 1-.442 1.068ZM4.697 4.697a.62.62 0 0 1-.442-.184l-.884-.884a.619.619 0 0 1-.183-.442c0-.166.064-.324.183-.441a.619.619 0 0 1 .442-.183c.166 0 .324.064.441.183l.884.884a.619.619 0 0 1 .184.442.619.619 0 0 1-.386.577.62.62 0 0 1-.24.048Zm10.606 0a.62.62 0 0 1-.441-.184.619.619 0 0 1-.184-.441c0-.167.065-.324.184-.442l.884-.884a.62.62 0 0 1 .883 0 .619.619 0 0 1 .183.442.619.619 0 0 1-.183.441l-.884.884a.62.62 0 0 1-.442.184ZM3.813 16.188a.626.626 0 0 1-.442-1.068l.884-.884a.622.622 0 0 1 1.066.442.627.627 0 0 1-.183.442l-.884.884a.62.62 0 0 1-.441.183Z" }) }),
-        /* @__PURE__ */ jsx76("defs", { children: /* @__PURE__ */ jsx76("clipPath", { id: "cxa", children: /* @__PURE__ */ jsx76("path", { fill: "currentColor", d: "M0 0h20v20H0z" }) }) })
-      ]
-    }
-  );
-};
-
-// src/icon/icons/lock_locked/index.react.tsx
-import { jsx as jsx77, jsxs as jsxs69 } from "react/jsx-runtime";
-var LockLocked = (props) => {
-  let _ariaLabel = props["aria-label"];
-  _ariaLabel = _ariaLabel && _ariaLabel.trim();
-  const _ariaHidden = !_ariaLabel;
-  const _role = _ariaHidden ? "presentation" : "img";
-  return /* @__PURE__ */ jsxs69(
-    "svg",
-    {
-      className: props.className,
-      viewBox: "0 0 20 20",
-      "aria-label": _ariaLabel,
-      "aria-hidden": _ariaHidden,
-      role: _role,
-      style: {
-        height: "var(--icon-height)",
-        width: "var(--icon-width)",
-        display: "flex",
-        alignItems: "center",
-        justifyContent: "center"
-      },
-      children: [
-        /* @__PURE__ */ jsx77("path", { fill: "currentColor", d: "M4.375 20A1.876 1.876 0 0 1 2.5 18.125v-8.75A1.877 1.877 0 0 1 4.375 7.5H5V5c0-2.757 2.243-5 5-5s5 2.243 5 5v2.5h.625A1.877 1.877 0 0 1 17.5 9.375v8.75A1.876 1.876 0 0 1 15.625 20H4.375Zm0-11.25a.625.625 0 0 0-.625.625v8.75c0 .345.28.625.625.625h11.25a.624.624 0 0 0 .625-.625v-8.75a.625.625 0 0 0-.625-.625H4.375ZM13.75 7.5V5A3.755 3.755 0 0 0 10 1.25 3.754 3.754 0 0 0 6.25 5v2.5h7.5Z" }),
-        /* @__PURE__ */ jsx77("path", { fill: "currentColor", d: "M10 15.625A.625.625 0 0 1 9.375 15v-2.5a.625.625 0 1 1 1.25 0V15a.624.624 0 0 1-.625.625Z" })
-      ]
-    }
-  );
-};
-
-// src/icon/icons/lock_unlocked/index.react.tsx
-import { jsx as jsx78, jsxs as jsxs70 } from "react/jsx-runtime";
-var LockUnlocked = (props) => {
-  let _ariaLabel = props["aria-label"];
-  _ariaLabel = _ariaLabel && _ariaLabel.trim();
-  const _ariaHidden = !_ariaLabel;
-  const _role = _ariaHidden ? "presentation" : "img";
-  return /* @__PURE__ */ jsxs70(
-    "svg",
-    {
-      className: props.className,
-      viewBox: "0 0 20 20",
-      "aria-label": _ariaLabel,
-      "aria-hidden": _ariaHidden,
-      role: _role,
-      style: {
-        height: "var(--icon-height)",
-        width: "var(--icon-width)",
-        display: "flex",
-        alignItems: "center",
-        justifyContent: "center"
-      },
-      children: [
-        /* @__PURE__ */ jsxs70("g", { "clip-path": "url(#cza)", children: [
-          /* @__PURE__ */ jsx78("path", { fill: "currentColor", d: "M6.875 20A1.876 1.876 0 0 1 5 18.125v-8.75A1.877 1.877 0 0 1 6.875 7.5H8.75V5A3.755 3.755 0 0 0 5 1.25 3.755 3.755 0 0 0 1.25 5v3.125a.625.625 0 0 1-1.25 0V5c0-2.757 2.243-5 5-5s5 2.243 5 5v2.5h8.125A1.877 1.877 0 0 1 20 9.375v8.75A1.876 1.876 0 0 1 18.125 20H6.875Zm0-11.25a.625.625 0 0 0-.625.625v8.75c0 .345.28.625.625.625h11.25a.624.624 0 0 0 .625-.625v-8.75a.625.625 0 0 0-.625-.625H6.875Z" }),
-          /* @__PURE__ */ jsx78("path", { fill: "currentColor", d: "M12.5 15.625a.624.624 0 0 1-.625-.625v-2.5a.624.624 0 1 1 1.25 0V15a.624.624 0 0 1-.625.625Z" })
-        ] }),
-        /* @__PURE__ */ jsx78("defs", { children: /* @__PURE__ */ jsx78("clipPath", { id: "cza", children: /* @__PURE__ */ jsx78("path", { fill: "currentColor", d: "M0 0h20v20H0z" }) }) })
-      ]
-    }
-  );
-};
-
-// src/icon/icons/login/index.react.tsx
-import { jsx as jsx79, jsxs as jsxs71 } from "react/jsx-runtime";
-var Login = (props) => {
-  let _ariaLabel = props["aria-label"];
-  _ariaLabel = _ariaLabel && _ariaLabel.trim();
-  const _ariaHidden = !_ariaLabel;
-  const _role = _ariaHidden ? "presentation" : "img";
-  return /* @__PURE__ */ jsxs71(
-    "svg",
-    {
-      className: props.className,
-      viewBox: "0 0 20 20",
-      "aria-label": _ariaLabel,
-      "aria-hidden": _ariaHidden,
-      role: _role,
-      style: {
-        height: "var(--icon-height)",
-        width: "var(--icon-width)",
-        display: "flex",
-        alignItems: "center",
-        justifyContent: "center"
-      },
-      children: [
-        /* @__PURE__ */ jsxs71("g", { "clip-path": "url(#daa)", children: [
-          /* @__PURE__ */ jsx79("path", { fill: "currentColor", d: "M11.875 18.756a.625.625 0 0 1 0-1.25h3.75a.625.625 0 0 0 .625-.625v-1.25a.625.625 0 1 1 1.25 0v1.25a1.876 1.876 0 0 1-1.875 1.875h-3.75Zm5-12.5a.625.625 0 0 1-.625-.625V4.38a.625.625 0 0 0-.625-.625h-3.75a.625.625 0 1 1 0-1.25h3.75A1.877 1.877 0 0 1 17.5 4.38v1.25a.625.625 0 0 1-.625.625ZM8.75 20.005c-.06 0-.12-.004-.18-.013L1.074 18.92A1.255 1.255 0 0 1 0 17.683V3.107c0-.586.415-1.1.988-1.223L8.47.038A1.258 1.258 0 0 1 10 1.256v17.5c0 .334-.13.648-.367.884a1.242 1.242 0 0 1-.884.365ZM1.27 3.102l-.02 14.582 7.502 1.071-.002-17.5-7.48 1.847ZM15 13.755a.62.62 0 0 1-.442-.183l-2.5-2.5a.356.356 0 0 1-.031-.037.592.592 0 0 1-.148-.44.627.627 0 0 1 .18-.407l2.5-2.5a.62.62 0 0 1 .883 0 .619.619 0 0 1 .183.442.619.619 0 0 1-.183.442l-1.434 1.433h5.366a.625.625 0 1 1 0 1.25h-5.366l1.434 1.433A.622.622 0 0 1 15 13.755Z" }),
-          /* @__PURE__ */ jsx79("path", { fill: "currentColor", d: "M6.563 10.943a.939.939 0 0 1-.938-.937c0-.478.354-.875.827-.93a.618.618 0 0 1 .11-.01.937.937 0 0 1 .938.938.94.94 0 0 1-.938.94Z" })
-        ] }),
-        /* @__PURE__ */ jsx79("defs", { children: /* @__PURE__ */ jsx79("clipPath", { id: "daa", children: /* @__PURE__ */ jsx79("path", { fill: "currentColor", d: "M0 0h20v20H0z" }) }) })
-      ]
-    }
-  );
-};
-
-// src/icon/icons/logout/index.react.tsx
-import { jsx as jsx80, jsxs as jsxs72 } from "react/jsx-runtime";
-var Logout = (props) => {
-  let _ariaLabel = props["aria-label"];
-  _ariaLabel = _ariaLabel && _ariaLabel.trim();
-  const _ariaHidden = !_ariaLabel;
-  const _role = _ariaHidden ? "presentation" : "img";
-  return /* @__PURE__ */ jsxs72(
-    "svg",
-    {
-      className: props.className,
-      viewBox: "0 0 20 20",
-      "aria-label": _ariaLabel,
-      "aria-hidden": _ariaHidden,
-      role: _role,
-      style: {
-        height: "var(--icon-height)",
-        width: "var(--icon-width)",
-        display: "flex",
-        alignItems: "center",
-        justifyContent: "center"
-      },
-      children: [
-        /* @__PURE__ */ jsxs72("g", { "clip-path": "url(#dba)", children: [
-          /* @__PURE__ */ jsx80("path", { fill: "currentColor", d: "M11.87 18.756a.625.625 0 0 1 0-1.25h3.75a.625.625 0 0 0 .625-.625v-1.25a.625.625 0 1 1 1.25 0v1.25a1.876 1.876 0 0 1-1.875 1.875h-3.75Zm5-12.5a.625.625 0 0 1-.625-.625V4.38a.625.625 0 0 0-.625-.625h-3.75a.625.625 0 1 1 0-1.25h3.75a1.877 1.877 0 0 1 1.875 1.875v1.25c0 .344-.28.625-.625.625ZM8.744 20.005c-.059 0-.119-.004-.178-.013L1.068 18.92a1.257 1.257 0 0 1-1.073-1.237V3.107a1.255 1.255 0 0 1 .988-1.224L8.464.037a1.258 1.258 0 0 1 1.531 1.218v17.5c0 .334-.13.648-.367.884a1.242 1.242 0 0 1-.884.366ZM1.264 3.102l-.02.006.001 14.575 7.502 1.071-.002-17.5-7.48 1.848ZM16.87 13.755a.623.623 0 0 1-.44-1.067l1.433-1.433h-5.368a.625.625 0 1 1 0-1.25h5.366l-1.434-1.433a.619.619 0 0 1-.183-.442.619.619 0 0 1 .386-.577.624.624 0 0 1 .68.135l2.502 2.5a.52.52 0 0 1 .037.045.629.629 0 0 1-.036.84l-2.5 2.5a.63.63 0 0 1-.442.182Z" }),
-          /* @__PURE__ */ jsx80("path", { fill: "currentColor", d: "M6.558 10.943a.939.939 0 0 1-.938-.937c0-.478.354-.875.827-.93a.618.618 0 0 1 .11-.01.94.94 0 0 1 0 1.877Z" })
-        ] }),
-        /* @__PURE__ */ jsx80("defs", { children: /* @__PURE__ */ jsx80("clipPath", { id: "dba", children: /* @__PURE__ */ jsx80("path", { fill: "currentColor", d: "M0 0h20v20H0z" }) }) })
-      ]
-    }
-  );
-};
-
-// src/icon/icons/magnet/index.react.tsx
-import { jsx as jsx81, jsxs as jsxs73 } from "react/jsx-runtime";
-var Magnet = (props) => {
-  let _ariaLabel = props["aria-label"];
-  _ariaLabel = _ariaLabel && _ariaLabel.trim();
-  const _ariaHidden = !_ariaLabel;
-  const _role = _ariaHidden ? "presentation" : "img";
-  return /* @__PURE__ */ jsxs73(
-    "svg",
-    {
-      className: props.className,
-      viewBox: "0 0 20 20",
-      "aria-label": _ariaLabel,
-      "aria-hidden": _ariaHidden,
-      role: _role,
-      style: {
-        height: "var(--icon-height)",
-        width: "var(--icon-width)",
-        display: "flex",
-        alignItems: "center",
-        justifyContent: "center"
-      },
-      children: [
-        /* @__PURE__ */ jsx81("g", { "clip-path": "url(#dca)", children: /* @__PURE__ */ jsx81("path", { fill: "currentColor", d: "M8.134 19.782a7.864 7.864 0 0 1-5.597-2.315c-3.086-3.091-3.084-8.111 0-11.194L5.608 3.2l.019-.02a.394.394 0 0 1 .02-.02L8.203.604a1.35 1.35 0 0 1 .959-.396c.361 0 .701.14.956.394l2.064 2.064a1.355 1.355 0 0 1-.001 1.914L6.51 10.25a2.293 2.293 0 0 0 0 3.24c.432.432 1.007.67 1.62.67.612 0 1.188-.238 1.62-.67l5.67-5.67a1.346 1.346 0 0 1 .958-.397c.363 0 .703.14.958.396l2.066 2.065a1.36 1.36 0 0 1 0 1.915l-5.671 5.669a7.863 7.863 0 0 1-5.597 2.314Zm-4.713-3.2a6.62 6.62 0 0 0 4.713 1.95 6.62 6.62 0 0 0 4.712-1.947l2.652-2.65-2.212-2.212-2.652 2.652a3.521 3.521 0 0 1-2.504 1.036 3.516 3.516 0 0 1-2.504-1.036 3.547 3.547 0 0 1 0-5.008l2.651-2.652L6.07 4.507l-2.648 2.65a6.619 6.619 0 0 0-1.95 4.712c0 1.783.692 3.456 1.949 4.713Zm12.96-3.531 2.137-2.135a.107.107 0 0 0 0-.149l-2.066-2.064a.104.104 0 0 0-.073-.03.1.1 0 0 0-.07.027l-2.139 2.14 2.212 2.21Zm-7.22-7.22 2.136-2.136a.104.104 0 0 0 .001-.146L9.235 1.486a.106.106 0 0 0-.148 0L6.953 3.624 9.162 5.83Z" }) }),
-        /* @__PURE__ */ jsx81("defs", { children: /* @__PURE__ */ jsx81("clipPath", { id: "dca", children: /* @__PURE__ */ jsx81("path", { fill: "currentColor", d: "M0 0h20v20H0z" }) }) })
-      ]
-    }
-  );
-};
-
-// src/icon/icons/menu/index.react.tsx
-import { jsx as jsx82 } from "react/jsx-runtime";
-var Menu = (props) => {
-  let _ariaLabel = props["aria-label"];
-  _ariaLabel = _ariaLabel && _ariaLabel.trim();
-  const _ariaHidden = !_ariaLabel;
-  const _role = _ariaHidden ? "presentation" : "img";
-  return /* @__PURE__ */ jsx82(
-    "svg",
-    {
-      className: props.className,
-      viewBox: "0 0 20 20",
-      "aria-label": _ariaLabel,
-      "aria-hidden": _ariaHidden,
-      role: _role,
-      style: {
-        height: "var(--icon-height)",
-        width: "var(--icon-width)",
-        display: "flex",
-        alignItems: "center",
-        justifyContent: "center"
-      },
-      children: /* @__PURE__ */ jsx82("path", { fill: "currentColor", d: "M1.875 15.627a.625.625 0 0 1 0-1.25h16.25a.625.625 0 0 1 0 1.25H1.875Zm0-5a.625.625 0 0 1 0-1.25h16.25a.625.625 0 0 1 0 1.25H1.875Zm0-5a.625.625 0 0 1 0-1.25h16.25a.625.625 0 0 1 0 1.25H1.875Z" })
-    }
-  );
-};
-
-// src/icon/icons/message/index.react.tsx
-import { jsx as jsx83, jsxs as jsxs74 } from "react/jsx-runtime";
-var Message = (props) => {
-  let _ariaLabel = props["aria-label"];
-  _ariaLabel = _ariaLabel && _ariaLabel.trim();
-  const _ariaHidden = !_ariaLabel;
-  const _role = _ariaHidden ? "presentation" : "img";
-  return /* @__PURE__ */ jsxs74(
-    "svg",
-    {
-      className: props.className,
-      viewBox: "0 0 20 20",
-      "aria-label": _ariaLabel,
-      "aria-hidden": _ariaHidden,
-      role: _role,
-      style: {
-        height: "var(--icon-height)",
-        width: "var(--icon-width)",
-        display: "flex",
-        alignItems: "center",
-        justifyContent: "center"
-      },
-      children: [
-        /* @__PURE__ */ jsxs74("g", { "clip-path": "url(#dea)", children: [
-          /* @__PURE__ */ jsx83("path", { fill: "currentColor", d: "M.623 20a.625.625 0 0 1-.565-.893l2.426-5.095A9.371 9.371 0 0 1 1.248 9.35C1.256 4.194 5.456 0 10.61 0h.028a9.38 9.38 0 0 1 4.95 1.413 9.31 9.31 0 0 1 4.176 5.834 9.312 9.312 0 0 1-1.164 7.078 9.312 9.312 0 0 1-5.832 4.18 9.34 9.34 0 0 1-6.782-.99L.893 19.94a.647.647 0 0 1-.27.06Zm9.988-18.75c-4.467 0-8.106 3.634-8.113 8.1a8.103 8.103 0 0 0 1.225 4.3.62.62 0 0 1 .034.598l-1.812 3.806 3.805-1.812a.622.622 0 0 1 .598.034 8.092 8.092 0 0 0 6.135 1.012 8.075 8.075 0 0 0 5.055-3.623c2.366-3.805 1.195-8.825-2.61-11.19a8.117 8.117 0 0 0-4.29-1.225h-.027Z" }),
-          /* @__PURE__ */ jsx83("path", { fill: "currentColor", d: "M6.872 7.5a.625.625 0 0 1 0-1.25h7.5a.625.625 0 0 1 0 1.25h-7.5ZM5.622 10a.625.625 0 0 1 0-1.25h10a.625.625 0 0 1 0 1.25h-10Zm1.25 2.5a.625.625 0 1 1 0-1.25h7.5a.625.625 0 1 1 0 1.25h-7.5Z" })
-        ] }),
-        /* @__PURE__ */ jsx83("defs", { children: /* @__PURE__ */ jsx83("clipPath", { id: "dea", children: /* @__PURE__ */ jsx83("path", { fill: "currentColor", d: "M0 0h20v20H0z" }) }) })
-      ]
-    }
-  );
-};
-
-// src/icon/icons/message_question/index.react.tsx
-import { jsx as jsx84, jsxs as jsxs75 } from "react/jsx-runtime";
-var MessageQuestion = (props) => {
-  let _ariaLabel = props["aria-label"];
-  _ariaLabel = _ariaLabel && _ariaLabel.trim();
-  const _ariaHidden = !_ariaLabel;
-  const _role = _ariaHidden ? "presentation" : "img";
-  return /* @__PURE__ */ jsxs75(
-    "svg",
-    {
-      className: props.className,
-      viewBox: "0 0 20 20",
-      "aria-label": _ariaLabel,
-      "aria-hidden": _ariaHidden,
-      role: _role,
-      style: {
-        height: "var(--icon-height)",
-        width: "var(--icon-width)",
-        display: "flex",
-        alignItems: "center",
-        justifyContent: "center"
-      },
-      children: [
-        /* @__PURE__ */ jsxs75("g", { "clip-path": "url(#dfa)", children: [
-          /* @__PURE__ */ jsx84("path", { fill: "currentColor", d: "M.63 20a.625.625 0 0 1-.566-.893l2.426-5.095A9.371 9.371 0 0 1 1.253 9.35C1.261 4.194 5.461 0 10.616 0h.029a9.37 9.37 0 0 1 4.952 1.414 9.312 9.312 0 0 1 4.179 5.833 9.315 9.315 0 0 1-1.17 7.08 9.315 9.315 0 0 1-5.833 4.179 9.333 9.333 0 0 1-6.782-.993L.898 19.94a.64.64 0 0 1-.269.06Zm9.986-18.75c-4.467 0-8.106 3.634-8.113 8.1a8.104 8.104 0 0 0 1.225 4.3.62.62 0 0 1 .034.598L1.95 18.053l3.805-1.812a.62.62 0 0 1 .6.034 8.088 8.088 0 0 0 6.134 1.013 8.074 8.074 0 0 0 5.055-3.621 8.074 8.074 0 0 0 1.014-6.136 8.073 8.073 0 0 0-3.622-5.055 8.116 8.116 0 0 0-4.291-1.226h-.03Z" }),
-          /* @__PURE__ */ jsx84("path", { fill: "currentColor", d: "M10.63 13.75a.938.938 0 1 0 0-1.875.938.938 0 0 0 0 1.875Zm0-2.5a.625.625 0 1 1 0-1.25 1.876 1.876 0 0 0-.001-3.75c-1.034 0-1.875.84-1.875 1.875a.625.625 0 0 1-1.25 0A3.128 3.128 0 0 1 10.63 5a3.128 3.128 0 0 1 3.125 3.125 3.129 3.129 0 0 1-3.125 3.125Z" })
-        ] }),
-        /* @__PURE__ */ jsx84("defs", { children: /* @__PURE__ */ jsx84("clipPath", { id: "dfa", children: /* @__PURE__ */ jsx84("path", { fill: "currentColor", d: "M0 0h20v20H0z" }) }) })
-      ]
-    }
-  );
-};
-
-// src/icon/icons/message_star/index.react.tsx
-import { jsx as jsx85, jsxs as jsxs76 } from "react/jsx-runtime";
-var MessageStar = (props) => {
-  let _ariaLabel = props["aria-label"];
-  _ariaLabel = _ariaLabel && _ariaLabel.trim();
-  const _ariaHidden = !_ariaLabel;
-  const _role = _ariaHidden ? "presentation" : "img";
-  return /* @__PURE__ */ jsxs76(
-    "svg",
-    {
-      className: props.className,
-      viewBox: "0 0 20 20",
-      "aria-label": _ariaLabel,
-      "aria-hidden": _ariaHidden,
-      role: _role,
-      style: {
-        height: "var(--icon-height)",
-        width: "var(--icon-width)",
-        display: "flex",
-        alignItems: "center",
-        justifyContent: "center"
-      },
-      children: [
-        /* @__PURE__ */ jsxs76("g", { "clip-path": "url(#dga)", children: [
-          /* @__PURE__ */ jsx85("path", { fill: "currentColor", d: "M11.158 19.998a1.286 1.286 0 0 1-1.109-.638 1.29 1.29 0 0 1-.074-1.135l1.035-2.377-1.862-1.835a1.19 1.19 0 0 1-.39-.831 1.235 1.235 0 0 1 .748-1.21 1.23 1.23 0 0 1 .532-.098h2.078l1.118-2.407c.174-.329.446-.558.776-.66a1.29 1.29 0 0 1 1.52.63l.018.036 1.115 2.402h2.108c.307 0 .603.116.832.328a1.235 1.235 0 0 1-.016 1.832l-1.818 1.812 1.032 2.37a1.283 1.283 0 0 1-1.177 1.78 1.28 1.28 0 0 1-.64-.17l-2.594-1.459-2.598 1.463a1.28 1.28 0 0 1-.634.167Zm-.033-1.283s0 .01.006.018c.007.013.022.014.026.014l2.926-1.643a.631.631 0 0 1 .613 0l2.907 1.634.021.008c.024-.008.033-.028.027-.042l-1.198-2.75a.622.622 0 0 1 .131-.693l2.148-2.138h-2.469a.627.627 0 0 1-.566-.362l-1.274-2.745a.038.038 0 0 0-.014-.012l-.123-.301.092.298a.04.04 0 0 0-.025.02l-1.272 2.74a.627.627 0 0 1-.567.362h-2.487l2.166 2.135a.63.63 0 0 1 .134.695l-1.202 2.762Z" }),
-          /* @__PURE__ */ jsx85("path", { fill: "currentColor", d: "M.625 20a.625.625 0 0 1-.565-.893l2.426-5.095a9.43 9.43 0 0 1-1.215-4.006C.913 4.848 4.817.362 9.975.003c.217-.015.438-.022.655-.022 4.902 0 9.007 3.833 9.347 8.726a.62.62 0 0 1-.58.666.627.627 0 0 1-.667-.58c-.295-4.24-3.853-7.562-8.102-7.562a8.07 8.07 0 0 0-6.13 2.794 8.07 8.07 0 0 0-1.98 5.893 8.158 8.158 0 0 0 1.207 3.73.62.62 0 0 1 .033.6l-1.811 3.806 3.805-1.811a.623.623 0 0 1 .598.033 8.44 8.44 0 0 0 1.376.684.623.623 0 0 1-.226 1.207.63.63 0 0 1-.224-.042 9.619 9.619 0 0 1-1.288-.612L.894 19.94a.63.63 0 0 1-.27.061Z" })
-        ] }),
-        /* @__PURE__ */ jsx85("defs", { children: /* @__PURE__ */ jsx85("clipPath", { id: "dga", children: /* @__PURE__ */ jsx85("path", { fill: "currentColor", d: "M0 0h20v20H0z" }) }) })
-      ]
-    }
-  );
-};
-
-// src/icon/icons/message_typing/index.react.tsx
-import { jsx as jsx86, jsxs as jsxs77 } from "react/jsx-runtime";
-var MessageTyping = (props) => {
-  let _ariaLabel = props["aria-label"];
-  _ariaLabel = _ariaLabel && _ariaLabel.trim();
-  const _ariaHidden = !_ariaLabel;
-  const _role = _ariaHidden ? "presentation" : "img";
-  return /* @__PURE__ */ jsxs77(
-    "svg",
-    {
-      className: props.className,
-      viewBox: "0 0 20 20",
-      "aria-label": _ariaLabel,
-      "aria-hidden": _ariaHidden,
-      role: _role,
-      style: {
-        height: "var(--icon-height)",
-        width: "var(--icon-width)",
-        display: "flex",
-        alignItems: "center",
-        justifyContent: "center"
-      },
-      children: [
-        /* @__PURE__ */ jsxs77("g", { "clip-path": "url(#dha)", children: [
-          /* @__PURE__ */ jsx86("path", { fill: "currentColor", d: "M.63 20a.625.625 0 0 1-.566-.893l2.426-5.095A9.371 9.371 0 0 1 1.253 9.35C1.261 4.194 5.461 0 10.616 0h.029a9.37 9.37 0 0 1 4.952 1.414 9.312 9.312 0 0 1 4.179 5.833 9.315 9.315 0 0 1-1.17 7.08 9.315 9.315 0 0 1-5.833 4.179 9.333 9.333 0 0 1-6.782-.993L.898 19.94a.64.64 0 0 1-.269.06Zm9.986-18.75c-4.467 0-8.106 3.634-8.113 8.1a8.104 8.104 0 0 0 1.225 4.3.62.62 0 0 1 .034.598L1.95 18.053l3.805-1.812a.62.62 0 0 1 .6.034 8.088 8.088 0 0 0 6.134 1.013 8.074 8.074 0 0 0 5.055-3.621 8.074 8.074 0 0 0 1.014-6.136 8.073 8.073 0 0 0-3.622-5.055 8.116 8.116 0 0 0-4.291-1.226h-.03Z" }),
-          /* @__PURE__ */ jsx86("path", { fill: "currentColor", d: "M6.878 10.313a.938.938 0 1 0 0-1.876.938.938 0 0 0 0 1.876Zm3.752 0a.935.935 0 0 1-.05-1.871l.014-.002a.94.94 0 0 1 .973.936.94.94 0 0 1-.938.937Zm3.75 0a.937.937 0 1 0 0-1.875.937.937 0 0 0 0 1.874Z" })
-        ] }),
-        /* @__PURE__ */ jsx86("defs", { children: /* @__PURE__ */ jsx86("clipPath", { id: "dha", children: /* @__PURE__ */ jsx86("path", { fill: "currentColor", d: "M0 0h20v20H0z" }) }) })
-      ]
-    }
-  );
-};
-
-// src/icon/icons/message_warning/index.react.tsx
-import { jsx as jsx87, jsxs as jsxs78 } from "react/jsx-runtime";
-var MessageWarning = (props) => {
-  let _ariaLabel = props["aria-label"];
-  _ariaLabel = _ariaLabel && _ariaLabel.trim();
-  const _ariaHidden = !_ariaLabel;
-  const _role = _ariaHidden ? "presentation" : "img";
-  return /* @__PURE__ */ jsxs78(
-    "svg",
-    {
-      className: props.className,
-      viewBox: "0 0 20 20",
-      "aria-label": _ariaLabel,
-      "aria-hidden": _ariaHidden,
-      role: _role,
-      style: {
-        height: "var(--icon-height)",
-        width: "var(--icon-width)",
-        display: "flex",
-        alignItems: "center",
-        justifyContent: "center"
-      },
-      children: [
-        /* @__PURE__ */ jsxs78("g", { "clip-path": "url(#dia)", children: [
-          /* @__PURE__ */ jsx87("path", { fill: "currentColor", d: "M.626 20.002a.626.626 0 0 1-.565-.894l2.425-5.094a9.32 9.32 0 0 1-1-6.717C2.45 3.074 6.296.01 10.626.01c.7 0 1.402.08 2.088.237a9.315 9.315 0 0 1 5.848 4.156 9.31 9.31 0 0 1 1.203 7.072.621.621 0 0 1-.972.37.618.618 0 0 1-.254-.405.618.618 0 0 1 .008-.244 8.07 8.07 0 0 0-1.04-6.131 8.069 8.069 0 0 0-5.07-3.6 8.126 8.126 0 0 0-1.813-.206c-3.752 0-7.082 2.656-7.919 6.316a8.058 8.058 0 0 0 1.02 6.075.625.625 0 0 1 .034.6l-1.812 3.806 3.805-1.812a.62.62 0 0 1 .596.033c.211.13.425.25.639.357a.622.622 0 0 1-.086 1.151.628.628 0 0 1-.476-.033 9.404 9.404 0 0 1-.437-.235L.894 19.942a.646.646 0 0 1-.268.06Zm13.75-1.875a.939.939 0 0 1-.861-1.306.938.938 0 0 1 .903-.567l.011.001a.93.93 0 0 1 .884.933c0 .519-.42.939-.937.939Z" }),
-          /* @__PURE__ */ jsx87("path", { fill: "currentColor", d: "M14.376 15.627a.625.625 0 0 1-.625-.625v-2.5a.625.625 0 1 1 1.25 0v2.5a.625.625 0 0 1-.625.625Z" }),
-          /* @__PURE__ */ jsx87("path", { fill: "currentColor", d: "M10.854 20.002a2.106 2.106 0 0 1-1.882-3.045l3.522-7.043a2.095 2.095 0 0 1 1.885-1.163 2.11 2.11 0 0 1 1.878 1.163l3.521 7.042A2.104 2.104 0 0 1 17.9 20h-7.045Zm3.524-10a.85.85 0 0 0-.765.472l-3.522 7.043a.854.854 0 0 0 .762 1.236h7.044a.844.844 0 0 0 .81-.585.845.845 0 0 0-.048-.65l-3.52-7.044a.86.86 0 0 0-.76-.472Z" })
-        ] }),
-        /* @__PURE__ */ jsx87("defs", { children: /* @__PURE__ */ jsx87("clipPath", { id: "dia", children: /* @__PURE__ */ jsx87("path", { fill: "currentColor", d: "M0 0h20v20H0z" }) }) })
-      ]
-    }
-  );
-};
-
-// src/icon/icons/module_edit/index.react.tsx
-import { jsx as jsx88, jsxs as jsxs79 } from "react/jsx-runtime";
-var ModuleEdit = (props) => {
-  let _ariaLabel = props["aria-label"];
-  _ariaLabel = _ariaLabel && _ariaLabel.trim();
-  const _ariaHidden = !_ariaLabel;
-  const _role = _ariaHidden ? "presentation" : "img";
-  return /* @__PURE__ */ jsxs79(
-    "svg",
-    {
-      className: props.className,
-      viewBox: "0 0 20 20",
-      "aria-label": _ariaLabel,
-      "aria-hidden": _ariaHidden,
-      role: _role,
-      style: {
-        height: "var(--icon-height)",
-        width: "var(--icon-width)",
-        display: "flex",
-        alignItems: "center",
-        justifyContent: "center"
-      },
-      children: [
-        /* @__PURE__ */ jsxs79("g", { "clip-path": "url(#dja)", children: [
-          /* @__PURE__ */ jsx88("path", { fill: "currentColor", d: "M5.207 15a1.46 1.46 0 0 1-1.458-1.46v-2.082a1.46 1.46 0 0 1 1.458-1.459h2.084a1.46 1.46 0 0 1 1.458 1.458v2.084a1.46 1.46 0 0 1-1.458 1.458H5.207Zm0-3.75a.209.209 0 0 0-.208.207v2.084c0 .115.093.208.208.208h2.084a.208.208 0 0 0 .208-.208v-2.084a.208.208 0 0 0-.208-.208H5.207Zm2.5-2.5A1.46 1.46 0 0 1 6.25 7.29V5.209A1.46 1.46 0 0 1 7.708 3.75h3.334A1.46 1.46 0 0 1 12.5 5.209V7.29a1.46 1.46 0 0 1-1.458 1.458H7.707Zm0-3.75a.208.208 0 0 0-.208.208V7.29c0 .115.093.208.208.208h3.334a.208.208 0 0 0 .208-.208V5.208a.208.208 0 0 0-.208-.209H7.707Zm4.167 10A1.876 1.876 0 0 1 10 13.123v-1.25A1.877 1.877 0 0 1 11.874 10h1.25A1.877 1.877 0 0 1 15 11.874a.625.625 0 1 1-1.25 0 .625.625 0 0 0-.625-.625h-1.25a.625.625 0 0 0-.625.625v1.25c0 .345.28.625.625.625a.625.625 0 0 1 0 1.25Zm1.25 5a.621.621 0 0 1-.598-.445.622.622 0 0 1-.014-.303l.421-2.108a.618.618 0 0 1 .172-.32l3.793-3.792a1.799 1.799 0 0 1 1.285-.532c.485 0 .941.188 1.284.53a1.82 1.82 0 0 1 .001 2.57l-3.794 3.794a.62.62 0 0 1-.32.171l-2.106.422a.63.63 0 0 1-.124.012Zm.797-1.422 1.002-.201 3.66-3.66a.568.568 0 0 0-.802-.801l-3.66 3.659-.2 1.003Z" }),
-          /* @__PURE__ */ jsx88("path", { fill: "currentColor", d: "M1.874 19.999a1.877 1.877 0 0 1-1.875-1.875V1.874A1.877 1.877 0 0 1 1.874-.001h12.5c.49-.001.96.19 1.31.533l2.497 2.399c.362.354.569.843.569 1.346v6.347a.625.625 0 0 1-1.25 0V4.277a.626.626 0 0 0-.189-.448l-2.496-2.4a.625.625 0 0 0-.44-.18h-12.5a.625.625 0 0 0-.625.625v16.25c0 .345.28.625.625.625h8.75a.625.625 0 0 1 0 1.25h-8.75Z" })
-        ] }),
-        /* @__PURE__ */ jsx88("defs", { children: /* @__PURE__ */ jsx88("clipPath", { id: "dja", children: /* @__PURE__ */ jsx88("path", { fill: "currentColor", d: "M0 0h20v20H0z" }) }) })
-      ]
-    }
-  );
-};
-
-// src/icon/icons/module_four/index.react.tsx
-import { jsx as jsx89, jsxs as jsxs80 } from "react/jsx-runtime";
-var ModuleFour = (props) => {
-  let _ariaLabel = props["aria-label"];
-  _ariaLabel = _ariaLabel && _ariaLabel.trim();
-  const _ariaHidden = !_ariaLabel;
-  const _role = _ariaHidden ? "presentation" : "img";
-  return /* @__PURE__ */ jsxs80(
-    "svg",
-    {
-      className: props.className,
-      viewBox: "0 0 20 20",
-      "aria-label": _ariaLabel,
-      "aria-hidden": _ariaHidden,
-      role: _role,
-      style: {
-        height: "var(--icon-height)",
-        width: "var(--icon-width)",
-        display: "flex",
-        alignItems: "center",
-        justifyContent: "center"
-      },
-      children: [
-        /* @__PURE__ */ jsx89("g", { "clip-path": "url(#dka)", children: /* @__PURE__ */ jsx89("path", { fill: "currentColor", d: "M10 19.375a.632.632 0 0 1-.242-.048l-8.754-3.752A.625.625 0 0 1 .625 15V6.236a.633.633 0 0 1 .063-.257l.021-.04a.624.624 0 0 1 .064-.09l.019-.023a.615.615 0 0 1 .211-.15l4.622-1.981V2.5C5.625 1.205 7.822.625 10 .625s4.375.58 4.375 1.875v1.195l4.62 1.98a.604.604 0 0 1 .102.055l.014.01a.634.634 0 0 1 .096.084l.029.034a.694.694 0 0 1 .14.375V15a.623.623 0 0 1-.38.574l-8.754 3.752a.61.61 0 0 1-.242.049Zm.625-1.573 7.5-3.214v-7.39l-7.5 3.214v7.39Zm-1.25 0v-7.39l-7.5-3.214v7.39l7.5 3.214ZM10 9.32l7.163-3.07-2.789-1.195c-.062 1.257-2.227 1.82-4.374 1.82s-4.313-.563-4.373-1.82L2.837 6.25 10 9.32ZM6.875 4.952c.157.173 1.143.673 3.125.673 1.98 0 2.968-.5 3.125-.673v-1.08c-.787.32-1.91.503-3.125.503-1.216 0-2.338-.183-3.125-.502v1.079ZM6.93 2.5c.268.203 1.25.625 3.07.625 1.82 0 2.802-.422 3.07-.625-.27-.203-1.25-.625-3.07-.625-1.82 0-2.802.422-3.07.625Z" }) }),
-        /* @__PURE__ */ jsx89("defs", { children: /* @__PURE__ */ jsx89("clipPath", { id: "dka", children: /* @__PURE__ */ jsx89("path", { fill: "currentColor", d: "M0 0h20v20H0z" }) }) })
-      ]
-    }
-  );
-};
-
-// src/icon/icons/multiple_users/index.react.tsx
-import { jsx as jsx90, jsxs as jsxs81 } from "react/jsx-runtime";
-var MultipleUsers = (props) => {
-  let _ariaLabel = props["aria-label"];
-  _ariaLabel = _ariaLabel && _ariaLabel.trim();
-  const _ariaHidden = !_ariaLabel;
-  const _role = _ariaHidden ? "presentation" : "img";
-  return /* @__PURE__ */ jsxs81(
-    "svg",
-    {
-      className: props.className,
-      viewBox: "0 0 20 20",
-      "aria-label": _ariaLabel,
-      "aria-hidden": _ariaHidden,
-      role: _role,
-      style: {
-        height: "var(--icon-height)",
-        width: "var(--icon-width)",
-        display: "flex",
-        alignItems: "center",
-        justifyContent: "center"
-      },
-      children: [
-        /* @__PURE__ */ jsxs81("g", { "clip-path": "url(#dla)", children: [
-          /* @__PURE__ */ jsx90("path", { fill: "currentColor", d: "M9.583 6.875a3.442 3.442 0 0 1-3.437-3.438c0-.483.1-.954.3-1.4l.015-.029A3.45 3.45 0 0 1 9.583 0a3.442 3.442 0 0 1 3.435 3.49.561.561 0 0 1-.007.14 3.43 3.43 0 0 1-3.428 3.245Zm-2.185-3.53a2.191 2.191 0 0 0 2.185 2.28 2.198 2.198 0 0 0 1.998-1.295 6.263 6.263 0 0 1-4.183-.984Zm.441-1.222a4.978 4.978 0 0 0 3.893.919A2.19 2.19 0 0 0 9.584 1.25c-.686 0-1.332.33-1.744.873Z" }),
-          /* @__PURE__ */ jsx90("path", { fill: "currentColor", d: "M6.207 9.375a.625.625 0 0 1-.42-1.087 5.637 5.637 0 0 1 3.796-1.475 5.64 5.64 0 0 1 3.798 1.475.625.625 0 0 1-.84.925 4.395 4.395 0 0 0-2.958-1.15c-1.09 0-2.141.409-2.956 1.15a.622.622 0 0 1-.42.162Zm9.21 7.5a3.442 3.442 0 0 1-3.438-3.438c0-.482.101-.953.3-1.399l.014-.029A3.453 3.453 0 0 1 15.417 10a3.442 3.442 0 0 1 3.435 3.49.566.566 0 0 1-.008.14 3.431 3.431 0 0 1-3.427 3.245Zm-2.185-3.533a2.19 2.19 0 0 0 2.185 2.284 2.196 2.196 0 0 0 1.997-1.297 6.333 6.333 0 0 1-4.183-.987Zm.44-1.218a5.04 5.04 0 0 0 3.893.918 2.188 2.188 0 0 0-2.148-1.792c-.687 0-1.335.33-1.745.874Z" }),
-          /* @__PURE__ */ jsx90("path", { fill: "currentColor", d: "M11.458 20a.624.624 0 0 1-.492-1.01 5.64 5.64 0 0 1 .977-.976 5.587 5.587 0 0 1 3.469-1.193 5.617 5.617 0 0 1 4.456 2.17.619.619 0 0 1 .128.46.62.62 0 0 1-.62.549.622.622 0 0 1-.494-.24 4.375 4.375 0 0 0-3.47-1.69 4.347 4.347 0 0 0-2.701.93c-.284.22-.54.477-.76.76a.62.62 0 0 1-.493.24Zm-6.875-3.125a3.442 3.442 0 0 1-3.437-3.438c0-.483.1-.954.3-1.399l.015-.028A3.448 3.448 0 0 1 4.583 10a3.442 3.442 0 0 1 3.435 3.49.565.565 0 0 1-.007.14 3.431 3.431 0 0 1-3.428 3.245Zm-2.185-3.533a2.19 2.19 0 0 0 2.185 2.284 2.196 2.196 0 0 0 1.998-1.297 6.333 6.333 0 0 1-4.182-.987Zm.44-1.218a5.043 5.043 0 0 0 3.894.918 2.189 2.189 0 0 0-2.149-1.792c-.686 0-1.334.33-1.745.874Z" }),
-          /* @__PURE__ */ jsx90("path", { fill: "currentColor", d: "M8.542 20a.62.62 0 0 1-.494-.24 4.372 4.372 0 0 0-3.47-1.69 4.343 4.343 0 0 0-2.7.93c-.284.22-.54.477-.76.76a.624.624 0 0 1-1.038-.077.622.622 0 0 1 .052-.693c.284-.364.612-.692.976-.977a5.587 5.587 0 0 1 3.47-1.193 5.615 5.615 0 0 1 4.456 2.17.625.625 0 0 1-.492 1.01Z" })
-        ] }),
-        /* @__PURE__ */ jsx90("defs", { children: /* @__PURE__ */ jsx90("clipPath", { id: "dla", children: /* @__PURE__ */ jsx90("path", { fill: "currentColor", d: "M0 0h20v20H0z" }) }) })
-      ]
-    }
-  );
-};
-
-// src/icon/icons/newspaper/index.react.tsx
-import { jsx as jsx91, jsxs as jsxs82 } from "react/jsx-runtime";
-var Newspaper = (props) => {
-  let _ariaLabel = props["aria-label"];
-  _ariaLabel = _ariaLabel && _ariaLabel.trim();
-  const _ariaHidden = !_ariaLabel;
-  const _role = _ariaHidden ? "presentation" : "img";
-  return /* @__PURE__ */ jsxs82(
-    "svg",
-    {
-      className: props.className,
-      viewBox: "0 0 20 20",
-      "aria-label": _ariaLabel,
-      "aria-hidden": _ariaHidden,
-      role: _role,
-      style: {
-        height: "var(--icon-height)",
-        width: "var(--icon-width)",
-        display: "flex",
-        alignItems: "center",
-        justifyContent: "center"
-      },
-      children: [
-        /* @__PURE__ */ jsxs82("g", { "clip-path": "url(#dma)", children: [
-          /* @__PURE__ */ jsx91("path", { fill: "currentColor", d: "M2.5 19.375a2.503 2.503 0 0 1-2.5-2.5V2.5A1.877 1.877 0 0 1 1.875.625h12.5A1.877 1.877 0 0 1 16.25 2.5v14.375a1.25 1.25 0 0 0 2.5 0V5A.625.625 0 1 1 20 5v11.875c0 1.378-1.122 2.5-2.5 2.5h-15Zm-.625-17.5a.625.625 0 0 0-.625.625v14.375c0 .69.56 1.25 1.25 1.25h12.834a2.477 2.477 0 0 1-.334-1.25V2.5a.625.625 0 0 0-.625-.625h-12.5Z" }),
-          /* @__PURE__ */ jsx91("path", { fill: "currentColor", d: "M3.75 13.125a.625.625 0 1 1 0-1.25h8.75a.624.624 0 1 1 0 1.25H3.75Zm0 2.5a.625.625 0 1 1 0-1.25h5a.625.625 0 1 1 0 1.25h-5ZM4.375 10c-.69 0-1.25-.56-1.25-1.25V5c0-.69.56-1.25 1.25-1.25h7.5c.69 0 1.25.56 1.25 1.25v3.75c0 .69-.56 1.25-1.25 1.25h-7.5Zm0-1.25h7.5V5h-7.5v3.75Z" })
-        ] }),
-        /* @__PURE__ */ jsx91("defs", { children: /* @__PURE__ */ jsx91("clipPath", { id: "dma", children: /* @__PURE__ */ jsx91("path", { fill: "currentColor", d: "M0 0h20v20H0z" }) }) })
-      ]
-    }
-  );
-};
-
-// src/icon/icons/notebook_pencil/index.react.tsx
-import { jsx as jsx92, jsxs as jsxs83 } from "react/jsx-runtime";
-var NotebookPencil = (props) => {
-  let _ariaLabel = props["aria-label"];
-  _ariaLabel = _ariaLabel && _ariaLabel.trim();
-  const _ariaHidden = !_ariaLabel;
-  const _role = _ariaHidden ? "presentation" : "img";
-  return /* @__PURE__ */ jsxs83(
-    "svg",
-    {
-      className: props.className,
-      viewBox: "0 0 20 20",
-      "aria-label": _ariaLabel,
-      "aria-hidden": _ariaHidden,
-      role: _role,
-      style: {
-        height: "var(--icon-height)",
-        width: "var(--icon-width)",
-        display: "flex",
-        alignItems: "center",
-        justifyContent: "center"
-      },
-      children: [
-        /* @__PURE__ */ jsxs83("g", { "clip-path": "url(#dna)", children: [
-          /* @__PURE__ */ jsx92("path", { fill: "currentColor", d: "M1.872 20.017a1.877 1.877 0 0 1-1.875-1.875V1.892A1.877 1.877 0 0 1 1.872.019h10a1.877 1.877 0 0 1 1.875 1.875.625.625 0 1 1-1.25 0 .625.625 0 0 0-.625-.625H3.747v17.5h8.125c.345 0 .625-.28.625-.626a.625.625 0 0 1 1.25 0 1.876 1.876 0 0 1-1.875 1.875h-10Zm0-18.75a.625.625 0 0 0-.625.626v16.25c0 .345.28.624.625.624h.625v-17.5h-.625ZM17.5 20.018a2.503 2.503 0 0 1-2.5-2.5v-1.824a.557.557 0 0 1 0-.103V5.07l-.003-.051a.63.63 0 0 1 .07-.286L16.925.396a.626.626 0 0 1 1.15 0l1.874 4.376c.034.077.051.16.051.246v12.5c0 1.379-1.122 2.5-2.5 2.5Zm-1.25-2.5a1.251 1.251 0 0 0 2.5 0v-1.25h-2.5v1.25Zm2.5-2.5V5.643h-2.5v9.375h2.5Zm-.323-10.625L17.5 2.229l-.927 2.164h1.854Z" }),
-          /* @__PURE__ */ jsx92("path", { fill: "currentColor", d: "M5.622 8.768a.625.625 0 0 1-.625-.625v-3.75a.625.625 0 0 1 .625-.625h5a.625.625 0 0 1 .625.625v3.75a.625.625 0 0 1-.625.625h-5Zm4.375-1.25v-2.5h-3.75v2.5h3.75Z" })
-        ] }),
-        /* @__PURE__ */ jsx92("defs", { children: /* @__PURE__ */ jsx92("clipPath", { id: "dna", children: /* @__PURE__ */ jsx92("path", { fill: "currentColor", d: "M0 0h20v20H0z" }) }) })
-      ]
-    }
-  );
-};
-
-// src/icon/icons/number_list/index.react.tsx
-import { jsx as jsx93, jsxs as jsxs84 } from "react/jsx-runtime";
-var NumberList = (props) => {
-  let _ariaLabel = props["aria-label"];
-  _ariaLabel = _ariaLabel && _ariaLabel.trim();
-  const _ariaHidden = !_ariaLabel;
-  const _role = _ariaHidden ? "presentation" : "img";
-  return /* @__PURE__ */ jsxs84(
-    "svg",
-    {
-      className: props.className,
-      viewBox: "0 0 20 20",
-      "aria-label": _ariaLabel,
-      "aria-hidden": _ariaHidden,
-      role: _role,
-      style: {
-        height: "var(--icon-height)",
-        width: "var(--icon-width)",
-        display: "flex",
-        alignItems: "center",
-        justifyContent: "center"
-      },
-      children: [
-        /* @__PURE__ */ jsx93("g", { "clip-path": "url(#doa)", children: /* @__PURE__ */ jsx93("path", { fill: "currentColor", d: "M6.875 3.748a.625.625 0 0 1 0-1.25h12.5a.625.625 0 1 1 0 1.25h-12.5Zm0 7.5a.625.625 0 0 1 0-1.25h12.5a.625.625 0 1 1 0 1.25h-12.5Zm0 7.5a.625.625 0 0 1 0-1.25h12.5a.625.625 0 1 1 0 1.25h-12.5Zm-4.999 1.25c-.84 0-1.584-.566-1.808-1.375a.626.626 0 0 1 1.205-.332.627.627 0 0 0 1.23-.167.62.62 0 0 0-.388-.577.619.619 0 0 0-.24-.047.626.626 0 0 1 0-1.25.625.625 0 1 0-.605-.787.625.625 0 0 1-1.207-.325 1.874 1.874 0 1 1 3.21 1.733c.309.344.478.784.479 1.25a1.88 1.88 0 0 1-1.876 1.877Zm-1.251-7.5a.622.622 0 0 1-.586-.838.63.63 0 0 1 .243-.31l2.017-1.325c.13-.19.201-.42.201-.652a.625.625 0 0 0-1.25 0 .625.625 0 0 1-1.25 0C0 8.34.84 7.498 1.875 7.498c1.034 0 1.875.841 1.875 1.875 0 .545-.187 1.079-.527 1.504a.622.622 0 0 1-.145.131l-.365.24h.412c.345 0 .625.28.625.625s-3.125.625-3.125.625Zm0-6.25a.625.625 0 0 1 0-1.25h.625v-3.75H.625a.625.625 0 0 1 0-1.25h.625c.69 0 1.25.561 1.25 1.25v3.75h.625a.625.625 0 0 1 0 1.25h-2.5Z" }) }),
-        /* @__PURE__ */ jsx93("defs", { children: /* @__PURE__ */ jsx93("clipPath", { id: "doa", children: /* @__PURE__ */ jsx93("path", { fill: "currentColor", d: "M0 0h20v20H0z" }) }) })
-      ]
-    }
-  );
-};
-
-// src/icon/icons/option/index.react.tsx
-import { jsx as jsx94 } from "react/jsx-runtime";
-var Option = (props) => {
-  let _ariaLabel = props["aria-label"];
-  _ariaLabel = _ariaLabel && _ariaLabel.trim();
-  const _ariaHidden = !_ariaLabel;
-  const _role = _ariaHidden ? "presentation" : "img";
-  return /* @__PURE__ */ jsx94(
-    "svg",
-    {
-      className: props.className,
-      viewBox: "0 0 20 20",
-      "aria-label": _ariaLabel,
-      "aria-hidden": _ariaHidden,
-      role: _role,
-      style: {
-        height: "var(--icon-height)",
-        width: "var(--icon-width)",
-        display: "flex",
-        alignItems: "center",
-        justifyContent: "center"
-      },
-      children: /* @__PURE__ */ jsx94("path", { fill: "currentColor", d: "M2.813 12.832A2.816 2.816 0 0 1 0 10.019a2.816 2.816 0 0 1 2.813-2.812 2.816 2.816 0 0 1 2.812 2.812 2.816 2.816 0 0 1-2.813 2.813Zm0-4.375a1.564 1.564 0 0 0-1.563 1.562c0 .862.7 1.563 1.563 1.563a1.564 1.564 0 0 0 1.104-2.667 1.564 1.564 0 0 0-1.104-.458Zm14.375 4.375a2.816 2.816 0 0 1-2.813-2.813 2.816 2.816 0 0 1 2.813-2.812A2.816 2.816 0 0 1 20 10.019a2.816 2.816 0 0 1-2.813 2.813Zm0-4.375a1.564 1.564 0 0 0-1.563 1.562c0 .862.7 1.563 1.563 1.563a1.564 1.564 0 0 0 1.104-2.667 1.564 1.564 0 0 0-1.105-.458Zm-7.375 4.668A2.816 2.816 0 0 1 7 10.312 2.816 2.816 0 0 1 9.813 7.5a2.816 2.816 0 0 1 2.812 2.813 2.816 2.816 0 0 1-2.813 2.812Zm0-4.375a1.564 1.564 0 0 0-1.563 1.563c0 .861.7 1.562 1.563 1.562a1.564 1.564 0 0 0 1.104-2.667 1.564 1.564 0 0 0-1.104-.458Z" })
-    }
-  );
-};
-
-// src/icon/icons/palette/index.react.tsx
-import { jsx as jsx95, jsxs as jsxs85 } from "react/jsx-runtime";
-var Palette = (props) => {
-  let _ariaLabel = props["aria-label"];
-  _ariaLabel = _ariaLabel && _ariaLabel.trim();
-  const _ariaHidden = !_ariaLabel;
-  const _role = _ariaHidden ? "presentation" : "img";
-  return /* @__PURE__ */ jsxs85(
-    "svg",
-    {
-      className: props.className,
-      viewBox: "0 0 20 20",
-      "aria-label": _ariaLabel,
-      "aria-hidden": _ariaHidden,
-      role: _role,
-      style: {
-        height: "var(--icon-height)",
-        width: "var(--icon-width)",
-        display: "flex",
-        alignItems: "center",
-        justifyContent: "center"
-      },
-      children: [
-        /* @__PURE__ */ jsxs85("g", { "clip-path": "url(#dqa)", children: [
-          /* @__PURE__ */ jsx95("path", { fill: "currentColor", d: "M11.744 19.99c-3.05 0-6.318-1.357-8.324-3.457-3.684-3.755-3.674-9.862.025-13.611A10.024 10.024 0 0 1 10.845 0c2.48.09 4.703 1.133 6.258 2.933 1.78 2.062 2.513 3.747 2.241 5.151-.327 1.692-1.952 2.53-3.386 3.27-1.03.531-2.003 1.033-1.959 1.611.072.92.579 1.263 1.871 1.263.316 0 .653-.02.91-.039.265-.019.495-.035.703-.035.358 0 .758.043 1.009.369.1.129.247.405.112.806-.802 2.378-2.512 3.92-4.946 4.46a8.86 8.86 0 0 1-1.914.201ZM10.488 1.243a8.768 8.768 0 0 0-6.156 2.56c-3.22 3.263-3.227 8.584-.014 11.86 1.789 1.872 4.694 3.08 7.407 3.08.585 0 1.143-.059 1.66-.174 1.285-.285 2.958-1.067 3.85-3.156-.125.007-.254.016-.367.024-.305.021-.645.045-1.006.045-1.934 0-2.98-.811-3.109-2.41v-.011c-.108-1.406 1.285-2.124 2.633-2.819 1.239-.639 2.52-1.299 2.732-2.395.189-.977-.489-2.393-1.96-4.097-1.326-1.535-3.228-2.423-5.357-2.502a9.75 9.75 0 0 0-.313-.005Z" }),
-          /* @__PURE__ */ jsx95("path", { fill: "currentColor", d: "M6.016 12.976a1.893 1.893 0 0 1-1.892-1.892c0-1.043.848-1.891 1.892-1.891 1.043 0 1.891.848 1.891 1.891a1.894 1.894 0 0 1-1.891 1.892Zm0-2.534a.642.642 0 1 0 0 1.284.642.642 0 0 0 0-1.284Zm3.134 6.563a1.893 1.893 0 0 1-1.893-1.892c0-1.043.849-1.891 1.892-1.891s1.892.848 1.892 1.891c0 1.044-.85 1.892-1.892 1.892Zm0-2.533a.642.642 0 1 0 0 1.285.642.642 0 0 0 0-1.285Zm3.03-7.66a1.893 1.893 0 0 1-1.893-1.892c0-1.043.849-1.892 1.892-1.892s1.892.849 1.892 1.892-.85 1.892-1.892 1.892Zm0-2.534a.642.642 0 1 0 0 1.284.642.642 0 0 0 0-1.284ZM6.807 8.155a1.893 1.893 0 0 1-1.891-1.892c0-1.043.848-1.891 1.891-1.891 1.044 0 1.892.848 1.892 1.891 0 1.044-.85 1.892-1.892 1.892Zm0-2.533a.642.642 0 1 0 0 1.283.642.642 0 0 0 0-1.283Z" })
-        ] }),
-        /* @__PURE__ */ jsx95("defs", { children: /* @__PURE__ */ jsx95("clipPath", { id: "dqa", children: /* @__PURE__ */ jsx95("path", { fill: "currentColor", d: "M0 0h20v20H0z" }) }) })
-      ]
-    }
-  );
-};
-
-// src/icon/icons/paragraph/index.react.tsx
-import { jsx as jsx96, jsxs as jsxs86 } from "react/jsx-runtime";
-var Paragraph = (props) => {
-  let _ariaLabel = props["aria-label"];
-  _ariaLabel = _ariaLabel && _ariaLabel.trim();
-  const _ariaHidden = !_ariaLabel;
-  const _role = _ariaHidden ? "presentation" : "img";
-  return /* @__PURE__ */ jsxs86(
-    "svg",
-    {
-      className: props.className,
-      viewBox: "0 0 20 20",
-      "aria-label": _ariaLabel,
-      "aria-hidden": _ariaHidden,
-      role: _role,
-      style: {
-        height: "var(--icon-height)",
-        width: "var(--icon-width)",
-        display: "flex",
-        alignItems: "center",
-        justifyContent: "center"
-      },
-      children: [
-        /* @__PURE__ */ jsx96("g", { "clip-path": "url(#dra)", children: /* @__PURE__ */ jsx96("path", { fill: "currentColor", d: "M14.375 19.998a.625.625 0 0 1-.625-.625V1.248H10v18.125a.625.625 0 0 1-1.25 0v-6.875h-2.5A6.258 6.258 0 0 1 0 6.248a6.258 6.258 0 0 1 6.25-6.25h13.125a.625.625 0 1 1 0 1.25H15v18.125a.625.625 0 0 1-.625.625ZM6.25 1.248c-2.757 0-5 2.244-5 5 0 2.757 2.243 5 5 5h2.5v-10h-2.5Z" }) }),
-        /* @__PURE__ */ jsx96("defs", { children: /* @__PURE__ */ jsx96("clipPath", { id: "dra", children: /* @__PURE__ */ jsx96("path", { fill: "currentColor", d: "M0 0h20v20H0z" }) }) })
-      ]
-    }
-  );
-};
-
-// src/icon/icons/paragraph_center/index.react.tsx
-import { jsx as jsx97, jsxs as jsxs87 } from "react/jsx-runtime";
-var ParagraphCenter = (props) => {
-  let _ariaLabel = props["aria-label"];
-  _ariaLabel = _ariaLabel && _ariaLabel.trim();
-  const _ariaHidden = !_ariaLabel;
-  const _role = _ariaHidden ? "presentation" : "img";
-  return /* @__PURE__ */ jsxs87(
-    "svg",
-    {
-      className: props.className,
-      viewBox: "0 0 20 20",
-      "aria-label": _ariaLabel,
-      "aria-hidden": _ariaHidden,
-      role: _role,
-      style: {
-        height: "var(--icon-height)",
-        width: "var(--icon-width)",
-        display: "flex",
-        alignItems: "center",
-        justifyContent: "center"
-      },
-      children: [
-        /* @__PURE__ */ jsx97("g", { "clip-path": "url(#dsa)", children: /* @__PURE__ */ jsx97("path", { fill: "currentColor", d: "M1.867 3.123a.625.625 0 0 1 0-1.25h16.25a.625.625 0 1 1 0 1.25H1.867Zm2.5 3.75a.625.625 0 0 1 0-1.25h11.25a.625.625 0 1 1 0 1.25H4.367Zm-3.75 3.75a.625.625 0 0 1 0-1.25h18.75a.625.625 0 1 1 0 1.25H.617Zm3.75 3.75a.625.625 0 0 1 0-1.25h11.25a.625.625 0 1 1 0 1.25H4.367Zm-2.5 3.75a.625.625 0 0 1 0-1.25h16.25a.625.625 0 1 1 0 1.25H1.867Z" }) }),
-        /* @__PURE__ */ jsx97("defs", { children: /* @__PURE__ */ jsx97("clipPath", { id: "dsa", children: /* @__PURE__ */ jsx97("path", { fill: "currentColor", d: "M0 0h20v20H0z" }) }) })
-      ]
-    }
-  );
-};
-
-// src/icon/icons/paragraph_left/index.react.tsx
-import { jsx as jsx98 } from "react/jsx-runtime";
-var ParagraphLeft = (props) => {
-  let _ariaLabel = props["aria-label"];
-  _ariaLabel = _ariaLabel && _ariaLabel.trim();
-  const _ariaHidden = !_ariaLabel;
-  const _role = _ariaHidden ? "presentation" : "img";
-  return /* @__PURE__ */ jsx98(
-    "svg",
-    {
-      className: props.className,
-      viewBox: "0 0 20 20",
-      "aria-label": _ariaLabel,
-      "aria-hidden": _ariaHidden,
-      role: _role,
-      style: {
-        height: "var(--icon-height)",
-        width: "var(--icon-width)",
-        display: "flex",
-        alignItems: "center",
-        justifyContent: "center"
-      },
-      children: /* @__PURE__ */ jsx98("path", { fill: "currentColor", d: "M.625 3.123a.625.625 0 0 1 0-1.25h17.5a.625.625 0 1 1 0 1.25H.625Zm0 3.75a.625.625 0 0 1 0-1.25h15a.625.625 0 1 1 0 1.25h-15Zm0 3.75a.625.625 0 0 1 0-1.25h18.75a.625.625 0 1 1 0 1.25H.625Zm0 3.75a.625.625 0 0 1 0-1.25h15a.625.625 0 1 1 0 1.25h-15Zm0 3.75a.625.625 0 0 1 0-1.25h17.5a.625.625 0 1 1 0 1.25H.625Z" })
-    }
-  );
-};
-
-// src/icon/icons/paragraph_right/index.react.tsx
-import { jsx as jsx99, jsxs as jsxs88 } from "react/jsx-runtime";
-var ParagraphRight = (props) => {
-  let _ariaLabel = props["aria-label"];
-  _ariaLabel = _ariaLabel && _ariaLabel.trim();
-  const _ariaHidden = !_ariaLabel;
-  const _role = _ariaHidden ? "presentation" : "img";
-  return /* @__PURE__ */ jsxs88(
-    "svg",
-    {
-      className: props.className,
-      viewBox: "0 0 20 20",
-      "aria-label": _ariaLabel,
-      "aria-hidden": _ariaHidden,
-      role: _role,
-      style: {
-        height: "var(--icon-height)",
-        width: "var(--icon-width)",
-        display: "flex",
-        alignItems: "center",
-        justifyContent: "center"
-      },
-      children: [
-        /* @__PURE__ */ jsx99("g", { "clip-path": "url(#dua)", children: /* @__PURE__ */ jsx99("path", { fill: "currentColor", d: "M1.667 3.123a.625.625 0 0 1 0-1.25h17.5a.625.625 0 0 1 0 1.25h-17.5Zm2.5 3.75a.625.625 0 0 1 0-1.25h15a.625.625 0 0 1 0 1.25h-15Zm-3.75 3.75a.625.625 0 0 1 0-1.25h18.75a.625.625 0 0 1 0 1.25H.417Zm3.75 3.75a.625.625 0 0 1 0-1.25h15a.625.625 0 1 1 0 1.25h-15Zm-2.5 3.75a.625.625 0 0 1 0-1.25h17.5a.625.625 0 1 1 0 1.25h-17.5Z" }) }),
-        /* @__PURE__ */ jsx99("defs", { children: /* @__PURE__ */ jsx99("clipPath", { id: "dua", children: /* @__PURE__ */ jsx99("path", { fill: "currentColor", d: "M0 0h20v20H0z" }) }) })
-      ]
-    }
-  );
-};
-
-// src/icon/icons/party_balloon/index.react.tsx
-import { jsx as jsx100, jsxs as jsxs89 } from "react/jsx-runtime";
-var PartyBalloon = (props) => {
-  let _ariaLabel = props["aria-label"];
-  _ariaLabel = _ariaLabel && _ariaLabel.trim();
-  const _ariaHidden = !_ariaLabel;
-  const _role = _ariaHidden ? "presentation" : "img";
-  return /* @__PURE__ */ jsxs89(
-    "svg",
-    {
-      className: props.className,
-      viewBox: "0 0 20 20",
-      "aria-label": _ariaLabel,
-      "aria-hidden": _ariaHidden,
-      role: _role,
-      style: {
-        height: "var(--icon-height)",
-        width: "var(--icon-width)",
-        display: "flex",
-        alignItems: "center",
-        justifyContent: "center"
-      },
-      children: [
-        /* @__PURE__ */ jsx100("g", { "clip-path": "url(#dva)", children: /* @__PURE__ */ jsx100("path", { fill: "currentColor", d: "M15.625 20a.625.625 0 0 1-.625-.625v-2.5a.625.625 0 0 0-1.25 0 1.877 1.877 0 0 1-1.875 1.875h-1.25a1.877 1.877 0 0 1-1.875-1.875v-.629h-.862a1.246 1.246 0 0 1-1.118-1.808l.26-.521A7.086 7.086 0 0 1 2.5 7.179C2.41 3.325 5.488.1 9.358 0c3.893.1 6.977 3.328 6.891 7.196a7.085 7.085 0 0 1-4.527 6.72l.26.521a1.25 1.25 0 0 1-1.119 1.809H10v.629c0 .345.28.625.625.625h1.25a.625.625 0 0 0 .625-.625c0-1.034.84-1.876 1.875-1.876 1.034 0 1.875.841 1.875 1.875v2.5a.625.625 0 0 1-.625.625ZM9.367 1.25c-3.172.087-5.691 2.742-5.618 5.924a5.88 5.88 0 0 0 4.327 5.774.623.623 0 0 1 .395.882l-.583 1.167h2.976l-.583-1.167a.624.624 0 0 1 .394-.882 5.866 5.866 0 0 0 4.323-5.75v-.024A5.8 5.8 0 0 0 9.367 1.25Z" }) }),
-        /* @__PURE__ */ jsx100("defs", { children: /* @__PURE__ */ jsx100("clipPath", { id: "dva", children: /* @__PURE__ */ jsx100("path", { fill: "currentColor", d: "M0 0h20v20H0z" }) }) })
-      ]
-    }
-  );
-};
-
-// src/icon/icons/pause/index.react.tsx
-import { jsx as jsx101 } from "react/jsx-runtime";
-var Pause = (props) => {
-  let _ariaLabel = props["aria-label"];
-  _ariaLabel = _ariaLabel && _ariaLabel.trim();
-  const _ariaHidden = !_ariaLabel;
-  const _role = _ariaHidden ? "presentation" : "img";
-  return /* @__PURE__ */ jsx101(
-    "svg",
-    {
-      className: props.className,
-      viewBox: "0 0 20 20",
-      "aria-label": _ariaLabel,
-      "aria-hidden": _ariaHidden,
-      role: _role,
-      style: {
-        height: "var(--icon-height)",
-        width: "var(--icon-width)",
-        display: "flex",
-        alignItems: "center",
-        justifyContent: "center"
-      },
-      children: /* @__PURE__ */ jsx101("path", { fill: "currentColor", d: "M5.625 17.5a1.876 1.876 0 0 1-1.875-1.875V4.375A1.877 1.877 0 0 1 5.625 2.5h1.25A1.877 1.877 0 0 1 8.75 4.375v11.25A1.877 1.877 0 0 1 6.875 17.5h-1.25Zm0-13.75A.625.625 0 0 0 5 4.375v11.25c0 .345.28.625.625.625h1.25a.625.625 0 0 0 .625-.625V4.375a.625.625 0 0 0-.625-.625h-1.25Zm7.5 13.75a1.876 1.876 0 0 1-1.875-1.875V4.375A1.876 1.876 0 0 1 13.125 2.5h1.25a1.877 1.877 0 0 1 1.875 1.875v11.25a1.876 1.876 0 0 1-1.875 1.875h-1.25Zm0-13.75a.625.625 0 0 0-.625.625v11.25c0 .345.28.625.625.625h1.25a.624.624 0 0 0 .625-.625V4.375a.625.625 0 0 0-.625-.625h-1.25Z" })
-    }
-  );
-};
-
-// src/icon/icons/pen_tools/index.react.tsx
-import { jsx as jsx102, jsxs as jsxs90 } from "react/jsx-runtime";
-var PenTools = (props) => {
-  let _ariaLabel = props["aria-label"];
-  _ariaLabel = _ariaLabel && _ariaLabel.trim();
-  const _ariaHidden = !_ariaLabel;
-  const _role = _ariaHidden ? "presentation" : "img";
-  return /* @__PURE__ */ jsxs90(
-    "svg",
-    {
-      className: props.className,
-      viewBox: "0 0 20 20",
-      "aria-label": _ariaLabel,
-      "aria-hidden": _ariaHidden,
-      role: _role,
-      style: {
-        height: "var(--icon-height)",
-        width: "var(--icon-width)",
-        display: "flex",
-        alignItems: "center",
-        justifyContent: "center"
-      },
-      children: [
-        /* @__PURE__ */ jsx102("g", { "clip-path": "url(#dxa)", children: /* @__PURE__ */ jsx102("path", { fill: "currentColor", d: "M6.178 20.008a4.38 4.38 0 0 1-4.375-4.375v-7.5a.625.625 0 0 1 .625-.625h.991l-.346-.982A2.483 2.483 0 0 1 1.243 5.1l-.01-.023a5.375 5.375 0 0 1 .01-4.72.63.63 0 0 1 .561-.349c.154 0 .302.057.417.16 1.337 1.2 1.562 1.283 1.82 1.38.192.07.429.159.87.443.39.257.698.62.887 1.047a2.47 2.47 0 0 1 .105 1.779l1.67 2.691h.024L9.135 3.15a.627.627 0 0 1 .155-.241L12.13.167a.621.621 0 0 1 .64-.14c.226.08.383.275.413.512l.49 3.915a.622.622 0 0 1-.031.287l-.976 2.767h2.262a4.38 4.38 0 0 1 4.375 4.375 4.38 4.38 0 0 1-4.375 4.375h-.673c-.307 2.125-2.153 3.75-4.327 3.75h-3.75Zm-3.125-4.375a3.128 3.128 0 0 0 3.125 3.125h3.75a3.129 3.129 0 0 0 3.125-3.125V8.758h-10v6.875Zm11.875-.625a3.129 3.129 0 0 0 3.125-3.125 3.129 3.129 0 0 0-3.125-3.125h-.625v6.25h.625Zm-3.588-7.5.588-1.666a3.606 3.606 0 0 1-2.15-.76l-.855 2.426h2.417Zm-5.237 0-.959-1.544c-.23.2-.496.354-.783.456l.383 1.088h1.359ZM2.38 4.583a1.243 1.243 0 0 0 1.06.73.617.617 0 0 1 .17 0 1.244 1.244 0 0 0 1.002-.645.612.612 0 0 1 .061-.127l.016-.036a1.24 1.24 0 0 0-.032-.956 1.19 1.19 0 0 0-.43-.512 2.66 2.66 0 0 0-.621-.318c-.364-.135-.65-.266-1.534-1.019a4.131 4.131 0 0 0 .301 2.87l.007.013Zm7.851-.785a2.337 2.337 0 0 0 2.15.758l.033-.093-.316-2.528-1.834 1.77-.033.093Z" }) }),
-        /* @__PURE__ */ jsx102("defs", { children: /* @__PURE__ */ jsx102("clipPath", { id: "dxa", children: /* @__PURE__ */ jsx102("path", { fill: "currentColor", d: "M0 0h20v20H0z" }) }) })
-      ]
-    }
-  );
-};
-
-// src/icon/icons/pen_write/index.react.tsx
-import { jsx as jsx103, jsxs as jsxs91 } from "react/jsx-runtime";
-var PenWrite = (props) => {
-  let _ariaLabel = props["aria-label"];
-  _ariaLabel = _ariaLabel && _ariaLabel.trim();
-  const _ariaHidden = !_ariaLabel;
-  const _role = _ariaHidden ? "presentation" : "img";
-  return /* @__PURE__ */ jsxs91(
-    "svg",
-    {
-      className: props.className,
-      viewBox: "0 0 20 20",
-      "aria-label": _ariaLabel,
-      "aria-hidden": _ariaHidden,
-      role: _role,
-      style: {
-        height: "var(--icon-height)",
-        width: "var(--icon-width)",
-        display: "flex",
-        alignItems: "center",
-        justifyContent: "center"
-      },
-      children: [
-        /* @__PURE__ */ jsxs91("g", { "clip-path": "url(#dya)", children: [
-          /* @__PURE__ */ jsx103("path", { fill: "currentColor", d: "M7.46 12.85a.628.628 0 0 1-.617-.714l.441-3.094a.628.628 0 0 1 .177-.354L15.416.733a2.483 2.483 0 0 1 1.767-.731c.669 0 1.297.26 1.768.731a2.502 2.502 0 0 1 0 3.536l-7.955 7.955a.628.628 0 0 1-.353.177l-3.094.442a.79.79 0 0 1-.088.006Zm.738-1.362 2.062-.295 7.807-7.808A1.25 1.25 0 0 0 16.3 1.618L8.492 9.426l-.294 2.062Z" }),
-          /* @__PURE__ */ jsx103("path", { fill: "currentColor", d: "M2.183 20.002a1.877 1.877 0 0 1-1.875-1.875v-12.5a1.877 1.877 0 0 1 1.875-1.875h6.25a.625.625 0 0 1 0 1.25h-6.25a.625.625 0 0 0-.625.625v12.5c0 .345.28.625.625.625h12.5a.625.625 0 0 0 .625-.625v-6.25a.625.625 0 0 1 1.25 0v6.25a1.877 1.877 0 0 1-1.875 1.875h-12.5Z" })
-        ] }),
-        /* @__PURE__ */ jsx103("defs", { children: /* @__PURE__ */ jsx103("clipPath", { id: "dya", children: /* @__PURE__ */ jsx103("path", { fill: "currentColor", d: "M0 0h20v20H0z" }) }) })
-      ]
-    }
-  );
-};
-
-// src/icon/icons/pencil/index.react.tsx
-import { jsx as jsx104, jsxs as jsxs92 } from "react/jsx-runtime";
-var Pencil = (props) => {
-  let _ariaLabel = props["aria-label"];
-  _ariaLabel = _ariaLabel && _ariaLabel.trim();
-  const _ariaHidden = !_ariaLabel;
-  const _role = _ariaHidden ? "presentation" : "img";
-  return /* @__PURE__ */ jsxs92(
-    "svg",
-    {
-      className: props.className,
-      viewBox: "0 0 20 20",
-      "aria-label": _ariaLabel,
-      "aria-hidden": _ariaHidden,
-      role: _role,
-      style: {
-        height: "var(--icon-height)",
-        width: "var(--icon-width)",
-        display: "flex",
-        alignItems: "center",
-        justifyContent: "center"
-      },
-      children: [
-        /* @__PURE__ */ jsx104("g", { "clip-path": "url(#dza)", children: /* @__PURE__ */ jsx104("path", { fill: "currentColor", d: "M.625 20a.627.627 0 0 1-.605-.783l1.517-5.784a.338.338 0 0 1 .012-.038.415.415 0 0 1 .061-.128.695.695 0 0 1 .07-.095l.02-.023L13.775 1.08A3.623 3.623 0 0 1 16.35.005a3.634 3.634 0 0 1 2.584 1.062l.016.016a3.649 3.649 0 0 1-.05 5.171L6.854 18.301a.519.519 0 0 1-.058.047.595.595 0 0 1-.226.116L.784 19.98a.673.673 0 0 1-.159.02Zm.876-1.5 4.053-1.063-.896-2.092-2.095-.894L1.5 18.499Zm5.124-1.74 8.87-8.868-1.25-1.25-8.37 8.368.75 1.75ZM4.99 14.128l8.37-8.368-1.25-1.249-8.872 8.87 1.752.747ZM16.38 7.01l.856-.856L13.85 2.77l-.857.857 3.385 3.382Zm1.738-1.742a2.38 2.38 0 0 0 .632-1.594 2.385 2.385 0 0 0-.682-1.704l-.015-.015a2.388 2.388 0 0 0-1.693-.699l-.007-.312.001.312a2.381 2.381 0 0 0-1.616.632l3.38 3.38Z" }) }),
-        /* @__PURE__ */ jsx104("defs", { children: /* @__PURE__ */ jsx104("clipPath", { id: "dza", children: /* @__PURE__ */ jsx104("path", { fill: "currentColor", d: "M0 0h20v20H0z" }) }) })
-      ]
-    }
-  );
-};
-
-// src/icon/icons/pin/index.react.tsx
-import { jsx as jsx105, jsxs as jsxs93 } from "react/jsx-runtime";
-var Pin = (props) => {
-  let _ariaLabel = props["aria-label"];
-  _ariaLabel = _ariaLabel && _ariaLabel.trim();
-  const _ariaHidden = !_ariaLabel;
-  const _role = _ariaHidden ? "presentation" : "img";
-  return /* @__PURE__ */ jsxs93(
-    "svg",
-    {
-      className: props.className,
-      viewBox: "0 0 20 20",
-      "aria-label": _ariaLabel,
-      "aria-hidden": _ariaHidden,
-      role: _role,
-      style: {
-        height: "var(--icon-height)",
-        width: "var(--icon-width)",
-        display: "flex",
-        alignItems: "center",
-        justifyContent: "center"
-      },
-      children: [
-        /* @__PURE__ */ jsxs93("g", { "clip-path": "url(#eaa)", children: [
-          /* @__PURE__ */ jsx105("path", { fill: "currentColor", d: "M10.208 20a.623.623 0 0 1-.609-.486c-.233-1.018-1.013-3.848-1.817-4.707a52.204 52.204 0 0 0-.692-.721c-1.76-1.814-3.757-3.871-3.757-7.211A6.883 6.883 0 0 1 10.208 0a6.883 6.883 0 0 1 6.875 6.875c0 3.335-1.991 5.39-3.749 7.2-.242.252-.474.49-.696.727-.806.862-1.588 3.694-1.821 4.712a.62.62 0 0 1-.609.486Zm0-18.75a5.632 5.632 0 0 0-5.625 5.625c0 2.833 1.73 4.616 3.404 6.34.246.254.482.497.708.738.602.645 1.123 1.911 1.513 3.108.391-1.198.914-2.467 1.517-3.113.227-.243.463-.486.703-.735 1.679-1.729 3.405-3.51 3.405-6.338a5.632 5.632 0 0 0-5.625-5.625Z" }),
-          /* @__PURE__ */ jsx105("path", { fill: "currentColor", d: "M10.208 10a3.128 3.128 0 0 1-3.125-3.125 3.128 3.128 0 0 1 3.125-3.125 3.128 3.128 0 0 1 3.125 3.125A3.128 3.128 0 0 1 10.208 10Zm0-5c-1.034 0-1.875.84-1.875 1.875 0 1.034.841 1.875 1.875 1.875s1.875-.84 1.875-1.875c0-1.034-.84-1.875-1.875-1.875Z" })
-        ] }),
-        /* @__PURE__ */ jsx105("defs", { children: /* @__PURE__ */ jsx105("clipPath", { id: "eaa", children: /* @__PURE__ */ jsx105("path", { fill: "currentColor", d: "M0 0h20v20H0z" }) }) })
-      ]
-    }
-  );
-};
-
-// src/icon/icons/pin_remove/index.react.tsx
-import { jsx as jsx106, jsxs as jsxs94 } from "react/jsx-runtime";
-var PinRemove = (props) => {
-  let _ariaLabel = props["aria-label"];
-  _ariaLabel = _ariaLabel && _ariaLabel.trim();
-  const _ariaHidden = !_ariaLabel;
-  const _role = _ariaHidden ? "presentation" : "img";
-  return /* @__PURE__ */ jsxs94(
-    "svg",
-    {
-      className: props.className,
-      viewBox: "0 0 20 20",
-      "aria-label": _ariaLabel,
-      "aria-hidden": _ariaHidden,
-      role: _role,
-      style: {
-        height: "var(--icon-height)",
-        width: "var(--icon-width)",
-        display: "flex",
-        alignItems: "center",
-        justifyContent: "center"
-      },
-      children: [
-        /* @__PURE__ */ jsxs94("g", { "clip-path": "url(#eba)", children: [
-          /* @__PURE__ */ jsx106("path", { fill: "currentColor", d: "M9.792 20a.623.623 0 0 1-.61-.486c-.233-1.018-1.013-3.848-1.816-4.707a54.5 54.5 0 0 0-.693-.721c-1.76-1.814-3.756-3.871-3.756-7.211A6.883 6.883 0 0 1 9.792 0a6.883 6.883 0 0 1 6.875 6.875c0 3.335-1.992 5.39-3.75 7.2-.242.252-.474.49-.695.727-.807.862-1.589 3.694-1.822 4.712a.62.62 0 0 1-.608.486Zm0-18.75a5.632 5.632 0 0 0-5.625 5.625c0 2.833 1.73 4.616 3.403 6.34.247.254.482.497.708.738.603.645 1.124 1.911 1.514 3.108.39-1.198.913-2.467 1.516-3.113.227-.243.464-.486.704-.735 1.676-1.729 3.403-3.51 3.403-6.338A5.63 5.63 0 0 0 9.792 1.25Z" }),
-          /* @__PURE__ */ jsx106("path", { fill: "currentColor", d: "M12.002 10.335a.62.62 0 0 1-.442-.183L9.792 8.383l-1.769 1.769a.622.622 0 0 1-1.02-.203.62.62 0 0 1 .137-.68L8.907 7.5 7.14 5.732a.619.619 0 0 1-.183-.442c0-.167.065-.324.183-.442a.62.62 0 0 1 .883 0l1.769 1.769 1.768-1.769a.62.62 0 0 1 .883 0 .622.622 0 0 1 0 .884L10.676 7.5l1.768 1.768a.619.619 0 0 1 .183.442.619.619 0 0 1-.386.578.621.621 0 0 1-.24.047Z" })
-        ] }),
-        /* @__PURE__ */ jsx106("defs", { children: /* @__PURE__ */ jsx106("clipPath", { id: "eba", children: /* @__PURE__ */ jsx106("path", { fill: "currentColor", d: "M0 0h20v20H0z" }) }) })
-      ]
-    }
-  );
-};
-
-// src/icon/icons/play/index.react.tsx
-import { jsx as jsx107 } from "react/jsx-runtime";
-var Play = (props) => {
-  let _ariaLabel = props["aria-label"];
-  _ariaLabel = _ariaLabel && _ariaLabel.trim();
-  const _ariaHidden = !_ariaLabel;
-  const _role = _ariaHidden ? "presentation" : "img";
-  return /* @__PURE__ */ jsx107(
-    "svg",
-    {
-      className: props.className,
-      viewBox: "0 0 20 20",
-      "aria-label": _ariaLabel,
-      "aria-hidden": _ariaHidden,
-      role: _role,
-      style: {
-        height: "var(--icon-height)",
-        width: "var(--icon-width)",
-        display: "flex",
-        alignItems: "center",
-        justifyContent: "center"
-      },
-      children: /* @__PURE__ */ jsx107("path", { fill: "currentColor", d: "M3.2 19.163a1.878 1.878 0 0 1-1.876-1.875V2.712c0-.309.077-.615.223-.887A1.886 1.886 0 0 1 3.2.835c.306 0 .611.078.884.224l13.602 7.288a1.867 1.867 0 0 1 .981 1.466c.025.245 0 .493-.071.729a1.867 1.867 0 0 1-.91 1.11L4.085 18.94a1.89 1.89 0 0 1-.886.223Zm0-17.076a.62.62 0 0 0-.502.253.627.627 0 0 0-.124.372v14.575a.625.625 0 0 0 .92.55l13.603-7.287a.617.617 0 0 0 .303-.37.624.624 0 0 0-.302-.733L3.495 2.162a.613.613 0 0 0-.295-.075Z" })
-    }
-  );
-};
-
-// src/icon/icons/printer/index.react.tsx
-import { jsx as jsx108, jsxs as jsxs95 } from "react/jsx-runtime";
-var Printer = (props) => {
-  let _ariaLabel = props["aria-label"];
-  _ariaLabel = _ariaLabel && _ariaLabel.trim();
-  const _ariaHidden = !_ariaLabel;
-  const _role = _ariaHidden ? "presentation" : "img";
-  return /* @__PURE__ */ jsxs95(
-    "svg",
-    {
-      className: props.className,
-      viewBox: "0 0 20 20",
-      "aria-label": _ariaLabel,
-      "aria-hidden": _ariaHidden,
-      role: _role,
-      style: {
-        height: "var(--icon-height)",
-        width: "var(--icon-width)",
-        display: "flex",
-        alignItems: "center",
-        justifyContent: "center"
-      },
-      children: [
-        /* @__PURE__ */ jsxs95("g", { "clip-path": "url(#eda)", children: [
-          /* @__PURE__ */ jsx108("path", { fill: "currentColor", d: "M4.375 20a.624.624 0 0 1-.612-.747l.6-3.003H1.874A1.876 1.876 0 0 1 0 14.375v-5c0-.043.004-.085.013-.127l.005-.018a.61.61 0 0 1 .05-.14L2.22 4.788A1.864 1.864 0 0 1 3.897 3.75H5V.625A.625.625 0 0 1 5.625 0h8.75A.625.625 0 0 1 15 .625V3.75h1.102c.715 0 1.358.397 1.678 1.037l2.154 4.309c.023.046.04.095.052.146l.003.018c.007.038.01.076.011.115v5a1.876 1.876 0 0 1-1.875 1.875h-2.488l.6 3.003a.625.625 0 0 1-.612.747H4.375Zm10.488-1.25-1-5H6.138l-1 5h9.725ZM18.125 15a.624.624 0 0 0 .625-.625V10H1.25v4.375c0 .345.28.625.625.625h2.737l.4-1.998a.627.627 0 0 1 .613-.502h8.75c.297 0 .555.212.613.502l.4 1.998h2.737Zm.24-6.25-1.702-3.405a.625.625 0 0 0-.56-.345H15v1.875a.625.625 0 0 1-.625.625h-8.75A.625.625 0 0 1 5 6.875V5H3.897a.622.622 0 0 0-.559.346L1.636 8.75h16.728Zm-4.615-2.5v-5h-7.5v5h7.5Z" }),
-          /* @__PURE__ */ jsx108("path", { fill: "currentColor", d: "M17.188 13.125a.937.937 0 1 0 0-1.875.937.937 0 0 0 0 1.875ZM8.125 17.5a.625.625 0 1 1 0-1.25h3.75a.624.624 0 1 1 0 1.25h-3.75Z" })
-        ] }),
-        /* @__PURE__ */ jsx108("defs", { children: /* @__PURE__ */ jsx108("clipPath", { id: "eda", children: /* @__PURE__ */ jsx108("path", { fill: "currentColor", d: "M0 0h20v20H0z" }) }) })
-      ]
-    }
-  );
-};
-
-// src/icon/icons/question_circle/index.react.tsx
-import { jsx as jsx109, jsxs as jsxs96 } from "react/jsx-runtime";
-var QuestionCircle = (props) => {
-  let _ariaLabel = props["aria-label"];
-  _ariaLabel = _ariaLabel && _ariaLabel.trim();
-  const _ariaHidden = !_ariaLabel;
-  const _role = _ariaHidden ? "presentation" : "img";
-  return /* @__PURE__ */ jsxs96(
-    "svg",
-    {
-      className: props.className,
-      viewBox: "0 0 20 20",
-      "aria-label": _ariaLabel,
-      "aria-hidden": _ariaHidden,
-      role: _role,
-      style: {
-        height: "var(--icon-height)",
-        width: "var(--icon-width)",
-        display: "flex",
-        alignItems: "center",
-        justifyContent: "center"
-      },
-      children: [
-        /* @__PURE__ */ jsxs96("g", { "clip-path": "url(#eea)", children: [
-          /* @__PURE__ */ jsx109("path", { fill: "currentColor", d: "M10 12.5a.625.625 0 0 1-.625-.625v-.838a1.876 1.876 0 0 1 1.25-1.769 1.878 1.878 0 0 0 .093-3.5A1.86 1.86 0 0 0 10 5.625c-.5 0-.972.195-1.326.549A1.86 1.86 0 0 0 8.125 7.5a.625.625 0 0 1-1.25 0c0-.835.326-1.62.916-2.21A3.104 3.104 0 0 1 10 4.377 3.13 3.13 0 0 1 13.125 7.5a3.13 3.13 0 0 1-2.083 2.946.627.627 0 0 0-.417.589v.839A.625.625 0 0 1 10 12.5Zm0 3.125a.937.937 0 1 0 0-1.875.937.937 0 0 0 0 1.875Z" }),
-          /* @__PURE__ */ jsx109("path", { fill: "currentColor", d: "M10 20C4.486 20 0 15.514 0 10S4.486 0 10 0s10 4.486 10 10-4.486 10-10 10Zm0-18.75c-4.825 0-8.75 3.925-8.75 8.75s3.925 8.75 8.75 8.75 8.75-3.925 8.75-8.75S14.825 1.25 10 1.25Z" })
-        ] }),
-        /* @__PURE__ */ jsx109("defs", { children: /* @__PURE__ */ jsx109("clipPath", { id: "eea", children: /* @__PURE__ */ jsx109("path", { fill: "currentColor", d: "M0 0h20v20H0z" }) }) })
-      ]
-    }
-  );
-};
-
-// src/icon/icons/ranking_star/index.react.tsx
-import { jsx as jsx110, jsxs as jsxs97 } from "react/jsx-runtime";
-var RankingStar = (props) => {
-  let _ariaLabel = props["aria-label"];
-  _ariaLabel = _ariaLabel && _ariaLabel.trim();
-  const _ariaHidden = !_ariaLabel;
-  const _role = _ariaHidden ? "presentation" : "img";
-  return /* @__PURE__ */ jsxs97(
-    "svg",
-    {
-      className: props.className,
-      viewBox: "0 0 20 20",
-      "aria-label": _ariaLabel,
-      "aria-hidden": _ariaHidden,
-      role: _role,
-      style: {
-        height: "var(--icon-height)",
-        width: "var(--icon-width)",
-        display: "flex",
-        alignItems: "center",
-        justifyContent: "center"
-      },
-      children: [
-        /* @__PURE__ */ jsx110("g", { "clip-path": "url(#efa)", children: /* @__PURE__ */ jsx110("path", { fill: "currentColor", d: "M1.563 20c-.69 0-1.25-.56-1.25-1.25v-5c0-.69.56-1.25 1.25-1.25h3.125c.689 0 1.25.56 1.25 1.25v5c0 .69-.561 1.25-1.25 1.25H1.563Zm0-1.25h3.125v-5H1.563v5ZM15.313 20c-.69 0-1.25-.56-1.25-1.25v-2.5c0-.69.56-1.25 1.25-1.25h3.124c.69 0 1.25.56 1.25 1.25v2.5c0 .69-.56 1.25-1.25 1.25h-3.125Zm0-1.25h3.124v-2.5h-3.125v2.5ZM8.438 20c-.69 0-1.25-.56-1.25-1.25v-7.5c0-.69.56-1.25 1.25-1.25h3.124c.69 0 1.25.56 1.25 1.25v7.5c0 .69-.56 1.25-1.25 1.25H8.438Zm0-1.25h3.124v-7.5H8.438v7.5Zm-.548-10a1.12 1.12 0 0 1-1.115-1.043 1.123 1.123 0 0 1 .083-.504l.732-1.685-1.32-1.3a1.1 1.1 0 0 1-.334-.777 1.105 1.105 0 0 1 .312-.79c.211-.218.49-.338.786-.338l.044.001h1.373L9.318.61A1.117 1.117 0 0 1 10.9.166c.176.108.319.263.412.446l.868 1.705h1.39a1.103 1.103 0 0 1 1.029.671 1.114 1.114 0 0 1-.247 1.237l-1.318 1.293.73 1.677a1.112 1.112 0 0 1-1.097 1.552 1.12 1.12 0 0 1-.489-.148L10.312 7.55 8.44 8.603a1.1 1.1 0 0 1-.55.148Zm.882-3.822a.63.63 0 0 1 .135.694l-.738 1.7 1.838-1.034a.626.626 0 0 1 .612 0l1.837 1.034-.74-1.699a.628.628 0 0 1 .136-.695l1.386-1.36h-1.442a.621.621 0 0 1-.557-.342l-.925-1.818-.923 1.814a.623.623 0 0 1-.557.342H7.388l1.384 1.364Z" }) }),
-        /* @__PURE__ */ jsx110("defs", { children: /* @__PURE__ */ jsx110("clipPath", { id: "efa", children: /* @__PURE__ */ jsx110("path", { fill: "currentColor", d: "M0 0h20v20H0z" }) }) })
-      ]
-    }
-  );
-};
-
-// src/icon/icons/rating_star_check/index.react.tsx
-import { jsx as jsx111, jsxs as jsxs98 } from "react/jsx-runtime";
-var RatingStarCheck = (props) => {
-  let _ariaLabel = props["aria-label"];
-  _ariaLabel = _ariaLabel && _ariaLabel.trim();
-  const _ariaHidden = !_ariaLabel;
-  const _role = _ariaHidden ? "presentation" : "img";
-  return /* @__PURE__ */ jsxs98(
-    "svg",
-    {
-      className: props.className,
-      viewBox: "0 0 20 20",
-      "aria-label": _ariaLabel,
-      "aria-hidden": _ariaHidden,
-      role: _role,
-      style: {
-        height: "var(--icon-height)",
-        width: "var(--icon-width)",
-        display: "flex",
-        alignItems: "center",
-        justifyContent: "center"
-      },
-      children: [
-        /* @__PURE__ */ jsxs98("g", { "clip-path": "url(#ega)", children: [
-          /* @__PURE__ */ jsx111("path", { fill: "currentColor", d: "M4.263 20a1.294 1.294 0 0 1-1.127-.65 1.296 1.296 0 0 1-.133-.988l1.535-5.601L.381 8.64a1.297 1.297 0 0 1 .805-2.212L6.2 5.933 8.834.718a1.306 1.306 0 0 1 2.33-.003l2.63 5.218 5.025.497c.3.025.586.16.803.378.245.246.378.572.376.919a1.289 1.289 0 0 1-.383.915l-.232.23a.621.621 0 0 1-1.019-.208.62.62 0 0 1 .14-.68l.232-.23a.043.043 0 0 0 .013-.032.044.044 0 0 0-.013-.032.047.047 0 0 0-.03-.013l-5.375-.533a.623.623 0 0 1-.497-.34L10.05 1.282a.057.057 0 0 0-.032-.025l-.02-.001c-.012 0-.018.001-.024.005a.056.056 0 0 0-.025.025L7.162 6.803a.62.62 0 0 1-.496.34l-5.367.531c-.026.002-.035.01-.041.017a.045.045 0 0 0-.011.033c0 .011.005.021.013.03l4.417 4.377a.626.626 0 0 1 .162.609l-1.63 5.948a.047.047 0 0 0 .007.035.049.049 0 0 0 .031.024l.015.001a.04.04 0 0 0 .023-.005L7.402 17.2a.622.622 0 0 1 .87.76.618.618 0 0 1-.315.36l-3.114 1.543c-.18.09-.379.138-.58.137Z" }),
-          /* @__PURE__ */ jsx111("path", { fill: "currentColor", d: "M14.375 20a5.631 5.631 0 0 1-5.625-5.626 5.631 5.631 0 0 1 5.625-5.625A5.632 5.632 0 0 1 20 14.374 5.632 5.632 0 0 1 14.375 20Zm0-10A4.38 4.38 0 0 0 10 14.373a4.38 4.38 0 0 0 4.375 4.375 4.38 4.38 0 0 0 4.375-4.375A4.38 4.38 0 0 0 14.375 10Z" }),
-          /* @__PURE__ */ jsx111("path", { fill: "currentColor", d: "M13.68 17.023c-.33.001-.645-.13-.878-.361l-1.254-1.254a.627.627 0 0 1 .442-1.067c.167 0 .324.065.442.183l1.25 1.25 2.421-3.228a.628.628 0 0 1 .655-.23.625.625 0 0 1 .345.98l-2.42 3.227a1.245 1.245 0 0 1-1.003.5Z" })
-        ] }),
-        /* @__PURE__ */ jsx111("defs", { children: /* @__PURE__ */ jsx111("clipPath", { id: "ega", children: /* @__PURE__ */ jsx111("path", { fill: "currentColor", d: "M0 0h20v20H0z" }) }) })
-      ]
-    }
-  );
-};
-
-// src/icon/icons/read/index.react.tsx
-import { jsx as jsx112, jsxs as jsxs99 } from "react/jsx-runtime";
-var Read = (props) => {
-  let _ariaLabel = props["aria-label"];
-  _ariaLabel = _ariaLabel && _ariaLabel.trim();
-  const _ariaHidden = !_ariaLabel;
-  const _role = _ariaHidden ? "presentation" : "img";
-  return /* @__PURE__ */ jsxs99(
-    "svg",
-    {
-      className: props.className,
-      viewBox: "0 0 20 20",
-      "aria-label": _ariaLabel,
-      "aria-hidden": _ariaHidden,
-      role: _role,
-      style: {
-        height: "var(--icon-height)",
-        width: "var(--icon-width)",
-        display: "flex",
-        alignItems: "center",
-        justifyContent: "center"
-      },
-      children: [
-        /* @__PURE__ */ jsx112("g", { "clip-path": "url(#eha)", children: /* @__PURE__ */ jsx112("path", { fill: "currentColor", d: "M10.417 20a.62.62 0 0 1-.183-.027l-5.759-1.76c-.59-.219-.958-.765-.933-1.364v-.284a2.917 2.917 0 0 1-1.014-.463c-.559-.406-.865-.964-.861-1.571v-.747c0-.485.198-.94.575-1.313.33-.33.785-.574 1.3-.701v-.977A1.308 1.308 0 0 1 5.24 9.451l5.177 1.897 5.19-1.899a1.301 1.301 0 0 1 1.042.11 1.3 1.3 0 0 1 .653 1.233v1.212c.259.128.481.284.665.468.376.374.575.827.575 1.311v.75c.002.605-.304 1.163-.863 1.569-.115.083-.24.159-.377.227v.55a1.354 1.354 0 0 1-.899 1.327L10.6 19.972a.603.603 0 0 1-.183.029Zm.625-1.468 4.964-1.51c.016-.006.05-.055.047-.111v-.251a2.89 2.89 0 0 1-.208.007c-.514 0-.978-.176-1.29-.485a1.243 1.243 0 0 1-.382-.883l-.006-2.252a1.24 1.24 0 0 1 .376-.88c.313-.315.786-.497 1.299-.498.07 0 .14.003.211.007v-.911c0-.023.001-.045.004-.068-.006-.033-.014-.044-.028-.051a.055.055 0 0 0-.027-.008h-.004l-4.955 1.814v6.08h-.001Zm-6.25-1.656c-.004.08.028.132.083.153l4.917 1.503v-6.08l-4.944-1.813h-.004a.057.057 0 0 0-.056.062.566.566 0 0 1 .004.065v.947c.34.07.641.227.867.454.244.245.378.558.377.88L6.03 15.3c0 .325-.136.638-.383.883-.236.227-.534.38-.856.442v.252Zm10.642-1.583c.06.059.197.122.409.124.421-.001.814-.117 1.101-.325.225-.164.349-.36.348-.555v-.753c0-.182-.113-.332-.206-.426-.275-.274-.74-.437-1.243-.438-.177.004-.34.054-.415.13a.135.135 0 0 0-.012.013l.006 2.218a.233.233 0 0 1 .012.012ZM4.364 12.92c-.502 0-.967.163-1.24.437-.095.093-.207.244-.207.427v.75c-.001.197.121.394.346.558.286.208.678.323 1.103.325.168 0 .326-.049.402-.124a.154.154 0 0 0 .013-.014l.005-2.217a.134.134 0 0 0-.013-.014c-.078-.078-.238-.129-.407-.129v-.312l-.002.313Zm6.063-4.794a4.067 4.067 0 0 1-4.063-4.063A4.067 4.067 0 0 1 10.427.001a4.067 4.067 0 0 1 4.062 4.062 4.067 4.067 0 0 1-4.062 4.063Zm0-6.875a2.816 2.816 0 0 0-2.813 2.812 2.816 2.816 0 0 0 2.813 2.813 2.816 2.816 0 0 0 2.812-2.813 2.816 2.816 0 0 0-2.812-2.812Z" }) }),
-        /* @__PURE__ */ jsx112("defs", { children: /* @__PURE__ */ jsx112("clipPath", { id: "eha", children: /* @__PURE__ */ jsx112("path", { fill: "currentColor", d: "M0 0h20v20H0z" }) }) })
-      ]
-    }
-  );
-};
-
-// src/icon/icons/read_home/index.react.tsx
-import { jsx as jsx113, jsxs as jsxs100 } from "react/jsx-runtime";
-var ReadHome = (props) => {
-  let _ariaLabel = props["aria-label"];
-  _ariaLabel = _ariaLabel && _ariaLabel.trim();
-  const _ariaHidden = !_ariaLabel;
-  const _role = _ariaHidden ? "presentation" : "img";
-  return /* @__PURE__ */ jsxs100(
-    "svg",
-    {
-      className: props.className,
-      viewBox: "0 0 20 20",
-      "aria-label": _ariaLabel,
-      "aria-hidden": _ariaHidden,
-      role: _role,
-      style: {
-        height: "var(--icon-height)",
-        width: "var(--icon-width)",
-        display: "flex",
-        alignItems: "center",
-        justifyContent: "center"
-      },
-      children: [
-        /* @__PURE__ */ jsxs100("g", { "clip-path": "url(#eia)", children: [
-          /* @__PURE__ */ jsx113("path", { fill: "currentColor", d: "M7.348 17.016a.638.638 0 0 1-.224-.041 15.978 15.978 0 0 0-5.341-1.048A1.884 1.884 0 0 1 0 14.042V2.917A1.864 1.864 0 0 1 .546 1.59a1.862 1.862 0 0 1 1.324-.553l.047.002.09.003c3.172.169 5.186.954 6.316 1.583.708.395 1.28.843 1.676 1.315.397-.471.968-.92 1.677-1.315 1.13-.63 3.145-1.415 6.327-1.583.034-.003.08-.004.123-.005a1.863 1.863 0 0 1 1.327.553c.355.355.548.827.547 1.327v6.855a.625.625 0 1 1-1.25 0V2.917a.624.624 0 0 0-.623-.629l-.034.001c-5.252.278-7.468 2.337-7.468 3.11l.008 4.374a.622.622 0 0 1-.863.579.624.624 0 0 1-.387-.577L9.375 5.4c0-.773-2.214-2.832-7.443-3.108l-.044-.004a.633.633 0 0 0-.591.388.62.62 0 0 0-.047.239v11.13a.625.625 0 0 0 .586.633c1.955.044 3.889.424 5.734 1.129a.626.626 0 0 1-.222 1.209Z" }),
-          /* @__PURE__ */ jsx113("path", { fill: "currentColor", d: "m13.125 19.148-.034-.001h-1.216A1.877 1.877 0 0 1 10 17.272v-3.456l-.213.186a.621.621 0 0 1-.883-.059.626.626 0 0 1 .06-.881l3.764-3.295a2.496 2.496 0 0 1 1.647-.615c.607 0 1.193.218 1.647.615l3.765 3.295a.626.626 0 0 1-.37 1.093.625.625 0 0 1-.453-.153l-.214-.186v3.456a1.877 1.877 0 0 1-1.875 1.875h-1.217a.56.56 0 0 1-.034.001h-2.499Zm3.75-1.251a.625.625 0 0 0 .625-.625v-4.55l-2.302-2.015a1.25 1.25 0 0 0-.824-.307 1.25 1.25 0 0 0-.823.307l-2.301 2.015v4.55c0 .345.28.625.625.625h.625v-1.874c0-1.034.84-1.875 1.875-1.875 1.034 0 1.875.841 1.875 1.875v1.874h.625Zm-1.875 0v-1.874a.625.625 0 1 0-1.25 0v1.874H15Z" })
-        ] }),
-        /* @__PURE__ */ jsx113("defs", { children: /* @__PURE__ */ jsx113("clipPath", { id: "eia", children: /* @__PURE__ */ jsx113("path", { fill: "currentColor", d: "M0 0h20v20H0z" }) }) })
-      ]
-    }
-  );
-};
-
-// src/icon/icons/read_idea/index.react.tsx
-import { jsx as jsx114, jsxs as jsxs101 } from "react/jsx-runtime";
-var ReadIdea = (props) => {
-  let _ariaLabel = props["aria-label"];
-  _ariaLabel = _ariaLabel && _ariaLabel.trim();
-  const _ariaHidden = !_ariaLabel;
-  const _role = _ariaHidden ? "presentation" : "img";
-  return /* @__PURE__ */ jsxs101(
-    "svg",
-    {
-      className: props.className,
-      viewBox: "0 0 20 20",
-      "aria-label": _ariaLabel,
-      "aria-hidden": _ariaHidden,
-      role: _role,
-      style: {
-        height: "var(--icon-height)",
-        width: "var(--icon-width)",
-        display: "flex",
-        alignItems: "center",
-        justifyContent: "center"
-      },
-      children: [
-        /* @__PURE__ */ jsxs101("g", { "clip-path": "url(#eja)", children: [
-          /* @__PURE__ */ jsx114("path", { fill: "currentColor", d: "M10 20a.594.594 0 0 1-.225-.043.58.58 0 0 1-.274-.208l-.034-.05a.454.454 0 0 1-.049-.1l-.013-.034c-.327-1.03-2.912-2.574-7.635-2.787A1.879 1.879 0 0 1 0 14.894V3.772a1.864 1.864 0 0 1 .546-1.325 1.864 1.864 0 0 1 1.324-.553l.03.001c.037 0 .073.002.108.004 3.005.155 5.531.848 7.11 1.949a.625.625 0 1 1-.715 1.025c-1.369-.956-3.728-1.585-6.471-1.726a.616.616 0 0 0-.5.181.622.622 0 0 0-.182.443v11.127a.626.626 0 0 0 .586.633c3.233.146 5.977.922 7.539 2.11v-8.89a.625.625 0 0 1 1.25 0v8.89c1.563-1.188 4.308-1.965 7.55-2.111a.622.622 0 0 0 .575-.627V7.5a.625.625 0 0 1 1.25 0v7.398a1.874 1.874 0 0 1-1.76 1.88c-4.732.214-7.317 1.759-7.645 2.787a.45.45 0 0 1-.03.076l-.027.049a.67.67 0 0 1-.033.052l-.036.044a.707.707 0 0 1-.121.107l-.02.013a.521.521 0 0 1-.104.052l-.038.015A.608.608 0 0 1 10 20Z" }),
-          /* @__PURE__ */ jsx114("path", { fill: "currentColor", d: "M7.785 7.936a.617.617 0 0 1-.213-.038 19.786 19.786 0 0 0-4.501-1.05.625.625 0 0 1 .138-1.241 21.03 21.03 0 0 1 4.788 1.116.62.62 0 0 1 .411.628.625.625 0 0 1-.623.585Zm0 3.179a.624.624 0 0 1-.21-.037 20.279 20.279 0 0 0-4.506-1.05.626.626 0 0 1 .144-1.242c1.63.187 3.24.562 4.784 1.116a.625.625 0 0 1-.212 1.213Zm0 3.37a.624.624 0 0 1-.21-.037 20.467 20.467 0 0 0-4.547-1.05.622.622 0 0 1-.478-.924.625.625 0 0 1 .618-.318 21.66 21.66 0 0 1 4.825 1.116.62.62 0 0 1 .151 1.1.623.623 0 0 1-.359.113Zm6.59-11.984a.625.625 0 0 1-.625-.625V.626a.625.625 0 1 1 1.25 0v1.25a.625.625 0 0 1-.625.625ZM11.25 4.193a.619.619 0 0 1-.442-.183l-.884-.884a.619.619 0 0 1-.183-.442.619.619 0 0 1 .386-.578.619.619 0 0 1 .68.136l.885.885a.619.619 0 0 1 .183.441.624.624 0 0 1-.625.625Zm6.25 0a.619.619 0 0 1-.442-.183.619.619 0 0 1-.183-.442c0-.166.065-.324.183-.441l.884-.884a.62.62 0 0 1 .884 0 .619.619 0 0 1 .183.441.619.619 0 0 1-.183.442l-.884.884a.619.619 0 0 1-.442.183ZM14.375 12.5a.625.625 0 0 1-.625-.624v-.625h-.625a.625.625 0 0 1-.625-.625V9.364a3.12 3.12 0 0 1-.328-4.712 3.132 3.132 0 0 1 2.203-.913 3.132 3.132 0 0 1 2.794 4.53 3.103 3.103 0 0 1-.919 1.095v1.261a.625.625 0 0 1-.625.625H15v.625a.622.622 0 0 1-.625.626ZM15 10v-.97c0-.224.12-.432.314-.543a1.868 1.868 0 0 0 .936-1.613 1.877 1.877 0 0 0-1.866-1.886l-.02-.312v.312a1.874 1.874 0 0 0-.927 3.5.626.626 0 0 1 .313.54V10H15Z" })
-        ] }),
-        /* @__PURE__ */ jsx114("defs", { children: /* @__PURE__ */ jsx114("clipPath", { id: "eja", children: /* @__PURE__ */ jsx114("path", { fill: "currentColor", d: "M0 0h20v20H0z" }) }) })
-      ]
-    }
-  );
-};
-
-// src/icon/icons/redo/index.react.tsx
-import { jsx as jsx115, jsxs as jsxs102 } from "react/jsx-runtime";
-var Redo = (props) => {
-  let _ariaLabel = props["aria-label"];
-  _ariaLabel = _ariaLabel && _ariaLabel.trim();
-  const _ariaHidden = !_ariaLabel;
-  const _role = _ariaHidden ? "presentation" : "img";
-  return /* @__PURE__ */ jsxs102(
-    "svg",
-    {
-      className: props.className,
-      viewBox: "0 0 20 20",
-      "aria-label": _ariaLabel,
-      "aria-hidden": _ariaHidden,
-      role: _role,
-      style: {
-        height: "var(--icon-height)",
-        width: "var(--icon-width)",
-        display: "flex",
-        alignItems: "center",
-        justifyContent: "center"
-      },
-      children: [
-        /* @__PURE__ */ jsx115("g", { "clip-path": "url(#eka)", children: /* @__PURE__ */ jsx115("path", { fill: "currentColor", d: "M9.991 19.998c-5.509 0-9.993-4.483-9.997-9.994a9.936 9.936 0 0 1 2.928-7.07A9.93 9.93 0 0 1 9.985-.002c3.657 0 7.02 2.032 8.76 5.175V.623a.625.625 0 1 1 1.25 0v6.25a.625.625 0 0 1-.626.625h-6.248a.625.625 0 1 1 0-1.25h4.773a8.716 8.716 0 0 0-7.903-5 8.691 8.691 0 0 0-6.185 2.57 8.695 8.695 0 0 0-2.562 6.185c.002 4.822 3.927 8.745 8.747 8.745.35 0 .63.28.63.625a.629.629 0 0 1-.63.625Z" }) }),
-        /* @__PURE__ */ jsx115("defs", { children: /* @__PURE__ */ jsx115("clipPath", { id: "eka", children: /* @__PURE__ */ jsx115("path", { fill: "currentColor", d: "M0 0h20v20H0z" }) }) })
-      ]
-    }
-  );
-};
-
-// src/icon/icons/refresh/index.react.tsx
-import { jsx as jsx116 } from "react/jsx-runtime";
-var Refresh = (props) => {
-  let _ariaLabel = props["aria-label"];
-  _ariaLabel = _ariaLabel && _ariaLabel.trim();
-  const _ariaHidden = !_ariaLabel;
-  const _role = _ariaHidden ? "presentation" : "img";
-  return /* @__PURE__ */ jsx116(
-    "svg",
-    {
-      className: props.className,
-      viewBox: "0 0 20 20",
-      "aria-label": _ariaLabel,
-      "aria-hidden": _ariaHidden,
-      role: _role,
-      style: {
-        height: "var(--icon-height)",
-        width: "var(--icon-width)",
-        display: "flex",
-        alignItems: "center",
-        justifyContent: "center"
-      },
-      children: /* @__PURE__ */ jsx116("path", { fill: "currentColor", d: "M4.098 16.252H.625a.625.625 0 0 1 0-1.25h2.197A8.747 8.747 0 0 1 1.26 9.817a8.697 8.697 0 0 1 2.697-6.132 8.716 8.716 0 0 1 6.055-2.43 8.71 8.71 0 0 1 3.355.67.624.624 0 1 1-.48 1.153 7.457 7.457 0 0 0-2.875-.574 7.472 7.472 0 0 0-5.19 2.083 7.454 7.454 0 0 0-2.313 5.257 7.483 7.483 0 0 0 1.241 4.3v-2.269a.625.625 0 1 1 1.25.001v3.75a.625.625 0 0 1-.625.626h-.198a.313.313 0 0 1-.038.003l-.042-.003Zm5.898 2.502a8.713 8.713 0 0 1-3.365-.673.623.623 0 0 1-.202-1.021.624.624 0 0 1 .682-.134 7.467 7.467 0 0 0 2.885.578 7.466 7.466 0 0 0 5.195-2.09 7.452 7.452 0 0 0 2.306-5.26 7.48 7.48 0 0 0-1.248-4.299v2.27a.625.625 0 0 1-1.249 0v-3.75a.625.625 0 0 1 .625-.625h3.75a.625.625 0 0 1 0 1.25h-2.196a8.755 8.755 0 0 1 1.569 5.18 8.703 8.703 0 0 1-2.69 6.136 8.716 8.716 0 0 1-6.062 2.438Z" })
-    }
-  );
-};
-
-// src/icon/icons/reload/index.react.tsx
-import { jsx as jsx117, jsxs as jsxs103 } from "react/jsx-runtime";
-var Reload = (props) => {
-  let _ariaLabel = props["aria-label"];
-  _ariaLabel = _ariaLabel && _ariaLabel.trim();
-  const _ariaHidden = !_ariaLabel;
-  const _role = _ariaHidden ? "presentation" : "img";
-  return /* @__PURE__ */ jsxs103(
-    "svg",
-    {
-      className: props.className,
-      viewBox: "0 0 20 20",
-      "aria-label": _ariaLabel,
-      "aria-hidden": _ariaHidden,
-      role: _role,
-      style: {
-        height: "var(--icon-height)",
-        width: "var(--icon-width)",
-        display: "flex",
-        alignItems: "center",
-        justifyContent: "center"
-      },
-      children: [
-        /* @__PURE__ */ jsx117("g", { "clip-path": "url(#ema)", children: /* @__PURE__ */ jsx117("path", { fill: "currentColor", d: "M11.26 18.95a8.724 8.724 0 0 1-7.394-4.047.625.625 0 0 1 1.054-.671 7.477 7.477 0 0 0 6.34 3.467 7.475 7.475 0 0 0 4.018-1.172c3.489-2.224 4.518-6.87 2.295-10.359a7.451 7.451 0 0 0-4.704-3.292 7.466 7.466 0 0 0-5.654.998 7.467 7.467 0 0 0-3.47 6.325v.366l1.433-1.433a.62.62 0 0 1 .884 0 .619.619 0 0 1 .183.441.619.619 0 0 1-.183.442l-2.489 2.488a.61.61 0 0 1-.453.198.625.625 0 0 1-.467-.211L.178 10.015a.62.62 0 0 1 0-.883.619.619 0 0 1 .442-.184c.167 0 .324.065.442.184l1.433 1.433v-.367A8.713 8.713 0 0 1 6.543 2.82a8.71 8.71 0 0 1 6.597-1.164 8.695 8.695 0 0 1 5.488 3.84c2.593 4.07 1.392 9.49-2.678 12.084a8.714 8.714 0 0 1-4.69 1.37Z" }) }),
-        /* @__PURE__ */ jsx117("defs", { children: /* @__PURE__ */ jsx117("clipPath", { id: "ema", children: /* @__PURE__ */ jsx117("path", { fill: "currentColor", d: "M0 0h20v20H0z" }) }) })
-      ]
-    }
-  );
-};
-
-// src/icon/icons/remove/index.react.tsx
-import { jsx as jsx118, jsxs as jsxs104 } from "react/jsx-runtime";
-var Remove = (props) => {
-  let _ariaLabel = props["aria-label"];
-  _ariaLabel = _ariaLabel && _ariaLabel.trim();
-  const _ariaHidden = !_ariaLabel;
-  const _role = _ariaHidden ? "presentation" : "img";
-  return /* @__PURE__ */ jsxs104(
-    "svg",
-    {
-      className: props.className,
-      viewBox: "0 0 20 20",
-      "aria-label": _ariaLabel,
-      "aria-hidden": _ariaHidden,
-      role: _role,
-      style: {
-        height: "var(--icon-height)",
-        width: "var(--icon-width)",
-        display: "flex",
-        alignItems: "center",
-        justifyContent: "center"
-      },
-      children: [
-        /* @__PURE__ */ jsxs104("g", { "clip-path": "url(#ena)", children: [
-          /* @__PURE__ */ jsx118("path", { fill: "currentColor", d: "M10 19.999c-5.514 0-10-4.486-10-10s4.486-10 10-10 10 4.486 10 10-4.486 10-10 10Zm0-18.75c-4.825 0-8.75 3.925-8.75 8.75s3.925 8.75 8.75 8.75 8.75-3.925 8.75-8.75-3.925-8.75-8.75-8.75Z" }),
-          /* @__PURE__ */ jsx118("path", { fill: "currentColor", d: "M13.75 14.374a.621.621 0 0 1-.442-.183L10 10.883l-3.308 3.309a.622.622 0 0 1-1.066-.442c0-.166.065-.325.182-.443L9.117 10 5.808 6.691a.619.619 0 0 1-.183-.442c0-.167.065-.324.183-.441a.619.619 0 0 1 .442-.184c.167 0 .324.065.442.184L10 9.116l3.308-3.308a.62.62 0 0 1 .884 0 .619.619 0 0 1 .183.441.619.619 0 0 1-.183.442l-3.309 3.308 3.309 3.308a.627.627 0 0 1-.442 1.067Z" })
-        ] }),
-        /* @__PURE__ */ jsx118("defs", { children: /* @__PURE__ */ jsx118("clipPath", { id: "ena", children: /* @__PURE__ */ jsx118("path", { fill: "currentColor", d: "M0 0h20v20H0z" }) }) })
-      ]
-    }
-  );
-};
-
-// src/icon/icons/retouch_graph/index.react.tsx
-import { jsx as jsx119, jsxs as jsxs105 } from "react/jsx-runtime";
-var RetouchGraph = (props) => {
-  let _ariaLabel = props["aria-label"];
-  _ariaLabel = _ariaLabel && _ariaLabel.trim();
-  const _ariaHidden = !_ariaLabel;
-  const _role = _ariaHidden ? "presentation" : "img";
-  return /* @__PURE__ */ jsxs105(
-    "svg",
-    {
-      className: props.className,
-      viewBox: "0 0 20 20",
-      "aria-label": _ariaLabel,
-      "aria-hidden": _ariaHidden,
-      role: _role,
-      style: {
-        height: "var(--icon-height)",
-        width: "var(--icon-width)",
-        display: "flex",
-        alignItems: "center",
-        justifyContent: "center"
-      },
-      children: [
-        /* @__PURE__ */ jsx119("g", { "clip-path": "url(#eoa)", children: /* @__PURE__ */ jsx119("path", { fill: "currentColor", d: "M17.5 20a.625.625 0 0 1-.442-1.067l.809-.808H2.5a.595.595 0 0 1-.21-.038.617.617 0 0 1-.282-.203l-.021-.028a.588.588 0 0 1-.085-.175l-.008-.032a.493.493 0 0 1-.016-.108l-.003-.041V2.134l-.808.809a.62.62 0 0 1-.884-.001A.619.619 0 0 1 0 2.5c0-.167.065-.324.183-.442L2.058.183a.617.617 0 0 1 .66-.143l.016.006a.61.61 0 0 1 .208.138l1.875 1.874a.622.622 0 0 1 0 .884.62.62 0 0 1-.884 0l-.808-.809v14.479a13.007 13.007 0 0 0 3.556-2.09 2.495 2.495 0 0 1-.431-1.397c0-1.378 1.122-2.5 2.5-2.5.427 0 .838.107 1.207.313.463-.769.843-1.584 1.135-2.432a16.953 16.953 0 0 1 1.596-3.421 2.497 2.497 0 0 1 3.105-3.7A5.63 5.63 0 0 1 19.361.003a.63.63 0 0 1 .598.85.623.623 0 0 1-.569.4 4.377 4.377 0 0 0-2.62.952c.298.42.459.923.46 1.44 0 1.378-1.122 2.5-2.5 2.5-.373 0-.74-.084-1.073-.244a15.73 15.73 0 0 0-1.381 3.007 14.105 14.105 0 0 1-1.395 2.913c.242.393.37.84.37 1.302 0 1.379-1.122 2.5-2.5 2.5-.39 0-.772-.091-1.12-.267a14.351 14.351 0 0 1-2.161 1.517h12.397l-.809-.808a.621.621 0 0 1 .203-1.019.621.621 0 0 1 .68.137l1.876 1.875a.617.617 0 0 1 .136.204l.008.022a.608.608 0 0 1 0 .431l-.004.013a.623.623 0 0 1-.14.214l-1.875 1.875A.62.62 0 0 1 17.5 20Zm-8.75-8.125a1.25 1.25 0 1 0 0 2.5 1.25 1.25 0 0 0 0-2.5Zm5.978-9.48a1.251 1.251 0 0 0 0 2.5 1.251 1.251 0 0 0 0-2.5Z" }) }),
-        /* @__PURE__ */ jsx119("defs", { children: /* @__PURE__ */ jsx119("clipPath", { id: "eoa", children: /* @__PURE__ */ jsx119("path", { fill: "currentColor", d: "M0 0h20v20H0z" }) }) })
-      ]
-    }
-  );
-};
-
-// src/icon/icons/ribbon_star/index.react.tsx
-import { jsx as jsx120, jsxs as jsxs106 } from "react/jsx-runtime";
-var RibbonStar = (props) => {
-  let _ariaLabel = props["aria-label"];
-  _ariaLabel = _ariaLabel && _ariaLabel.trim();
-  const _ariaHidden = !_ariaLabel;
-  const _role = _ariaHidden ? "presentation" : "img";
-  return /* @__PURE__ */ jsxs106(
-    "svg",
-    {
-      className: props.className,
-      viewBox: "0 0 20 20",
-      "aria-label": _ariaLabel,
-      "aria-hidden": _ariaHidden,
-      role: _role,
-      style: {
-        height: "var(--icon-height)",
-        width: "var(--icon-width)",
-        display: "flex",
-        alignItems: "center",
-        justifyContent: "center"
-      },
-      children: [
-        /* @__PURE__ */ jsxs106("g", { "clip-path": "url(#epa)", children: [
-          /* @__PURE__ */ jsx120("path", { fill: "currentColor", d: "M15 20a.623.623 0 0 1-.53-.292l-3.052-4.844c-.937.18-1.9.18-2.836 0l-3.054 4.844a.621.621 0 0 1-.922.151.628.628 0 0 1-.198-.287l-1.136-3.405-2.77-.555a.629.629 0 0 1-.492-.718.63.63 0 0 1 .13-.29l3.123-3.815A7.507 7.507 0 0 1 2.5 7.5C2.5 3.364 5.864 0 10 0s7.5 3.364 7.5 7.5a7.512 7.512 0 0 1-.762 3.29l3.121 3.814a.627.627 0 0 1-.361 1.009l-2.77.554-1.135 3.405A.625.625 0 0 1 15 20Zm-.175-2.075.833-2.498a.625.625 0 0 1 .47-.415l2.105-.42-2.183-2.667a7.552 7.552 0 0 1-3.38 2.583l2.155 3.417ZM3.873 15.012a.624.624 0 0 1 .47.415l.832 2.497 2.153-3.417a7.554 7.554 0 0 1-3.38-2.583l-2.182 2.667 2.107.421ZM10 1.25A6.257 6.257 0 0 0 3.75 7.5 6.258 6.258 0 0 0 10 13.75a6.258 6.258 0 0 0 6.25-6.25A6.258 6.258 0 0 0 10 1.25Z" }),
-          /* @__PURE__ */ jsx120("path", { fill: "currentColor", d: "M7.577 11.433a1.106 1.106 0 0 1-.967-.555 1.123 1.123 0 0 1-.064-.991l.734-1.685-1.322-1.3a1.103 1.103 0 0 1-.335-.778 1.102 1.102 0 0 1 .313-.79 1.092 1.092 0 0 1 .79-.338h1.408l.867-1.704a1.13 1.13 0 0 1 1.58-.447c.175.106.318.257.413.438L11.866 5h1.41a1.113 1.113 0 0 1 .762 1.908l-1.317 1.294.73 1.676a1.12 1.12 0 0 1-1.098 1.553 1.118 1.118 0 0 1-.488-.148L10 10.232l-1.873 1.054a1.105 1.105 0 0 1-.55.146ZM10 8.891c.107 0 .213.027.307.08l1.836 1.033-.739-1.698a.63.63 0 0 1 .135-.695l1.385-1.361h-1.441a.622.622 0 0 1-.558-.342L10 4.086l-.926 1.819a.623.623 0 0 1-.557.342H7.075L8.462 7.61a.63.63 0 0 1 .135.695l-.74 1.698L9.694 8.97A.62.62 0 0 1 10 8.89Z" })
-        ] }),
-        /* @__PURE__ */ jsx120("defs", { children: /* @__PURE__ */ jsx120("clipPath", { id: "epa", children: /* @__PURE__ */ jsx120("path", { fill: "currentColor", d: "M0 0h20v20H0z" }) }) })
-      ]
-    }
-  );
-};
-
-// src/icon/icons/rocket/index.react.tsx
-import { jsx as jsx121, jsxs as jsxs107 } from "react/jsx-runtime";
-var Rocket = (props) => {
-  let _ariaLabel = props["aria-label"];
-  _ariaLabel = _ariaLabel && _ariaLabel.trim();
-  const _ariaHidden = !_ariaLabel;
-  const _role = _ariaHidden ? "presentation" : "img";
-  return /* @__PURE__ */ jsxs107(
-    "svg",
-    {
-      className: props.className,
-      viewBox: "0 0 20 20",
-      "aria-label": _ariaLabel,
-      "aria-hidden": _ariaHidden,
-      role: _role,
-      style: {
-        height: "var(--icon-height)",
-        width: "var(--icon-width)",
-        display: "flex",
-        alignItems: "center",
-        justifyContent: "center"
-      },
-      children: [
-        /* @__PURE__ */ jsxs107("g", { "clip-path": "url(#eqa)", children: [
-          /* @__PURE__ */ jsx121("path", { fill: "currentColor", d: "M10.183 19.077c-.54 0-1.016-.345-1.186-.856l-.96-2.878-3.325-3.326-2.879-.96a1.238 1.238 0 0 1-.721-.626 1.24 1.24 0 0 1-.07-.953c.061-.185.166-.354.304-.491l.192-.193a5.577 5.577 0 0 1 3.966-1.651c.607 0 1.21.098 1.791.292l.68.227 4.806-4.806A6.894 6.894 0 0 1 16.883.884l1.798-.2a.62.62 0 0 1 .512.18c.135.134.2.32.179.51l-.2 1.799A6.9 6.9 0 0 1 17.2 7.275l-4.807 4.806.226.68a5.593 5.593 0 0 1-1.358 5.756l-.195.194a1.25 1.25 0 0 1-.883.366Zm0-1.25.194-.195a4.35 4.35 0 0 0 1.056-4.475l-.028-.084-2.107 2.102.885 2.652Zm-1.6-3.705 7.732-7.731a5.642 5.642 0 0 0 1.613-3.356l.114-1.022-1.022.114a5.646 5.646 0 0 0-3.357 1.613l-7.73 7.73 2.65 2.652ZM5.506 8.393a4.333 4.333 0 0 0-3.083 1.285l-.194.194 2.652.884L6.986 8.65 6.9 8.622a4.408 4.408 0 0 0-1.394-.229Z" }),
-          /* @__PURE__ */ jsx121("path", { fill: "currentColor", d: "M1.517 19.167a.618.618 0 0 1-.442-.184.625.625 0 0 1-.16-.61c.18-.643 1.122-3.887 1.928-4.691a2.497 2.497 0 0 1 1.604-.727h.04a.625.625 0 0 1 .041 1.248 1.24 1.24 0 0 0-.801.363c-.277.276-.784 1.49-1.278 3.042 1.551-.494 2.766-1 3.043-1.276a1.25 1.25 0 0 0 .366-.884.626.626 0 0 1 1.25-.001 2.51 2.51 0 0 1-.734 1.769c-.803.803-4.046 1.746-4.689 1.927a.578.578 0 0 1-.168.024Z" })
-        ] }),
-        /* @__PURE__ */ jsx121("defs", { children: /* @__PURE__ */ jsx121("clipPath", { id: "eqa", children: /* @__PURE__ */ jsx121("path", { fill: "currentColor", d: "M0 0h20v20H0z" }) }) })
-      ]
-    }
-  );
-};
-
-// src/icon/icons/search/index.react.tsx
-import { jsx as jsx122, jsxs as jsxs108 } from "react/jsx-runtime";
-var Search = (props) => {
-  let _ariaLabel = props["aria-label"];
-  _ariaLabel = _ariaLabel && _ariaLabel.trim();
-  const _ariaHidden = !_ariaLabel;
-  const _role = _ariaHidden ? "presentation" : "img";
-  return /* @__PURE__ */ jsxs108(
-    "svg",
-    {
-      className: props.className,
-      viewBox: "0 0 20 20",
-      "aria-label": _ariaLabel,
-      "aria-hidden": _ariaHidden,
-      role: _role,
-      style: {
-        height: "var(--icon-height)",
-        width: "var(--icon-width)",
-        display: "flex",
-        alignItems: "center",
-        justifyContent: "center"
-      },
-      children: [
-        /* @__PURE__ */ jsx122("g", { "clip-path": "url(#era)", children: /* @__PURE__ */ jsx122("path", { fill: "currentColor", d: "M19.37 19.997a.62.62 0 0 1-.44-.184l-5.43-5.43a8.19 8.19 0 0 1-2.127 1.316 8.14 8.14 0 0 1-6.259.058 8.121 8.121 0 0 1-4.466-4.385 8.124 8.124 0 0 1-.057-6.259A8.125 8.125 0 0 1 4.975.647 8.12 8.12 0 0 1 11.233.59 8.12 8.12 0 0 1 15.7 4.975a8.123 8.123 0 0 1 .058 6.258 8.174 8.174 0 0 1-1.374 2.265l5.43 5.432a.624.624 0 1 1-.442 1.067ZM8.173 1.242c-.93 0-1.852.189-2.708.555A6.883 6.883 0 0 0 1.75 5.581a6.883 6.883 0 0 0 .048 5.301 6.884 6.884 0 0 0 3.784 3.716 6.884 6.884 0 0 0 5.301-.05 6.905 6.905 0 0 0 2.162-1.446.303.303 0 0 1 .064-.064 6.871 6.871 0 0 0 1.489-2.273 6.88 6.88 0 0 0-.05-5.302 6.88 6.88 0 0 0-3.783-3.715 6.9 6.9 0 0 0-2.591-.506Z" }) }),
-        /* @__PURE__ */ jsx122("defs", { children: /* @__PURE__ */ jsx122("clipPath", { id: "era", children: /* @__PURE__ */ jsx122("path", { fill: "currentColor", d: "M0 0h20v20H0z" }) }) })
-      ]
-    }
-  );
-};
-
-// src/icon/icons/settings/index.react.tsx
-import { jsx as jsx123, jsxs as jsxs109 } from "react/jsx-runtime";
-var Settings = (props) => {
-  let _ariaLabel = props["aria-label"];
-  _ariaLabel = _ariaLabel && _ariaLabel.trim();
-  const _ariaHidden = !_ariaLabel;
-  const _role = _ariaHidden ? "presentation" : "img";
-  return /* @__PURE__ */ jsxs109(
-    "svg",
-    {
-      className: props.className,
-      viewBox: "0 0 20 20",
-      "aria-label": _ariaLabel,
-      "aria-hidden": _ariaHidden,
-      role: _role,
-      style: {
-        height: "var(--icon-height)",
-        width: "var(--icon-width)",
-        display: "flex",
-        alignItems: "center",
-        justifyContent: "center"
-      },
-      children: [
-        /* @__PURE__ */ jsxs109("g", { "clip-path": "url(#esa)", children: [
-          /* @__PURE__ */ jsx123("path", { fill: "currentColor", d: "M6.797 19.99a2.251 2.251 0 0 1-2.085-1.391 2.252 2.252 0 0 1-.17-.969l.06-1.17a1.008 1.008 0 0 0-1.005-1.06c-.017 0-.034 0-.051.002l-1.168.06a2.249 2.249 0 0 1-1.627-.577 2.244 2.244 0 0 1-.744-1.56 2.248 2.248 0 0 1 .74-1.79l.868-.783a1.006 1.006 0 0 0 0-1.497l-.868-.788a2.24 2.24 0 0 1-.738-1.562 2.257 2.257 0 0 1 2.362-2.364l1.168.059A1.008 1.008 0 0 0 4.6 3.542l-.056-1.171A2.26 2.26 0 0 1 6.799.007a2.26 2.26 0 0 1 1.668.738l.784.866a1 1 0 0 0 .75.334.989.989 0 0 0 .744-.332l.792-.868a2.24 2.24 0 0 1 1.672-.742 2.256 2.256 0 0 1 2.256 2.373l-.06 1.166a1.009 1.009 0 0 0 .956 1.056l.05.002h.048l1.167-.06.115-.003c.56 0 1.096.208 1.512.585.447.405.71.96.74 1.562a2.25 2.25 0 0 1-.741 1.786l-.869.783a1 1 0 0 0-.33.697 1.001 1.001 0 0 0 .33.798l.869.785c.447.404.71.958.741 1.56a2.239 2.239 0 0 1-.579 1.628 2.267 2.267 0 0 1-1.679.743l-.11-.002-1.168-.06-.055-.001a.99.99 0 0 0-.671.259.996.996 0 0 0-.331.8l.06 1.167a2.25 2.25 0 0 1-.59 1.627 2.242 2.242 0 0 1-1.677.735 2.253 2.253 0 0 1-1.655-.733l-.786-.87a1.01 1.01 0 0 0-1.496 0l-.789.868a2.267 2.267 0 0 1-1.67.736Zm-3.2-5.838a2.261 2.261 0 0 1 2.254 2.373l-.06 1.167a1.004 1.004 0 0 0 1.005 1.05l.049-.001a1 1 0 0 0 .698-.328l.786-.867a2.253 2.253 0 0 1 1.673-.744c.639 0 1.25.271 1.677.744l.784.867c.189.206.46.325.747.325l.04-.001a1.008 1.008 0 0 0 .96-1.052l-.058-1.162a2.24 2.24 0 0 1 .741-1.791 2.25 2.25 0 0 1 1.629-.579l1.166.06.057.001a1 1 0 0 0 .74-.332 1.002 1.002 0 0 0 .165-1.109 1 1 0 0 0-.236-.315l-.868-.783a2.262 2.262 0 0 1 0-3.352l.868-.783a1.002 1.002 0 0 0 .235-1.179 1.003 1.003 0 0 0-.91-.576l-.05.002-1.166.059a2.258 2.258 0 0 1-2.365-2.368l.059-1.167a1 1 0 0 0-.331-.799.98.98 0 0 0-.668-.26l-.057.002a.997.997 0 0 0-.698.331l-.79.867a2.252 2.252 0 0 1-2.644.525 2.243 2.243 0 0 1-.705-.527l-.782-.865a1.01 1.01 0 0 0-.745-.328h-.048a1.008 1.008 0 0 0-.958 1.055l.056 1.166a2.257 2.257 0 0 1-2.372 2.37L2.309 5.79a1.008 1.008 0 0 0-1.052 1.053c.014.269.13.516.33.697l.867.787a2.26 2.26 0 0 1-.002 3.351l-.867.784a1.002 1.002 0 0 0-.33.798 1.004 1.004 0 0 0 1.005.954c.018 0 .036 0 .053-.002l1.168-.059.117-.001Z" }),
-          /* @__PURE__ */ jsx123("path", { fill: "currentColor", d: "M10 14.375A4.38 4.38 0 0 1 5.625 10 4.38 4.38 0 0 1 10 5.625 4.38 4.38 0 0 1 14.375 10 4.38 4.38 0 0 1 10 14.375Zm0-7.5A3.128 3.128 0 0 0 6.875 10 3.128 3.128 0 0 0 10 13.125 3.128 3.128 0 0 0 13.125 10 3.128 3.128 0 0 0 10 6.875Z" })
-        ] }),
-        /* @__PURE__ */ jsx123("defs", { children: /* @__PURE__ */ jsx123("clipPath", { id: "esa", children: /* @__PURE__ */ jsx123("path", { fill: "currentColor", d: "M0 0h20v20H0z" }) }) })
-      ]
-    }
-  );
-};
-
-// src/icon/icons/settings_slider/index.react.tsx
-import { jsx as jsx124, jsxs as jsxs110 } from "react/jsx-runtime";
-var SettingsSlider = (props) => {
-  let _ariaLabel = props["aria-label"];
-  _ariaLabel = _ariaLabel && _ariaLabel.trim();
-  const _ariaHidden = !_ariaLabel;
-  const _role = _ariaHidden ? "presentation" : "img";
-  return /* @__PURE__ */ jsxs110(
-    "svg",
-    {
-      className: props.className,
-      viewBox: "0 0 20 20",
-      "aria-label": _ariaLabel,
-      "aria-hidden": _ariaHidden,
-      role: _role,
-      style: {
-        height: "var(--icon-height)",
-        width: "var(--icon-width)",
-        display: "flex",
-        alignItems: "center",
-        justifyContent: "center"
-      },
-      children: [
-        /* @__PURE__ */ jsxs110("g", { "clip-path": "url(#eta)", children: [
-          /* @__PURE__ */ jsx124("path", { fill: "currentColor", d: "M1.876 20a1.876 1.876 0 0 1-1.875-1.875V1.875A1.877 1.877 0 0 1 1.876 0h16.25a1.877 1.877 0 0 1 1.875 1.875v16.25A1.877 1.877 0 0 1 18.126 20H1.876Zm0-18.75a.625.625 0 0 0-.625.625v16.25c0 .345.28.625.625.625h16.25a.625.625 0 0 0 .625-.625V1.875a.625.625 0 0 0-.625-.625H1.876Z" }),
-          /* @__PURE__ */ jsx124("path", { fill: "currentColor", d: "M8.75 15.625a2.514 2.514 0 0 1-2.417-1.875H4.376a.625.625 0 1 1 0-1.25h1.957a2.512 2.512 0 0 1 2.417-1.875c1.143 0 2.137.79 2.42 1.875h4.456a.625.625 0 1 1 0 1.25h-4.457a2.514 2.514 0 0 1-2.418 1.875Zm0-3.75a1.25 1.25 0 0 0 0 2.5 1.25 1.25 0 0 0 0-2.5Zm5-2.5A2.514 2.514 0 0 1 11.333 7.5H4.376a.625.625 0 0 1 0-1.25h6.956a2.513 2.513 0 0 1 2.419-1.875c1.378 0 2.5 1.122 2.5 2.5a2.504 2.504 0 0 1-2.501 2.5Zm0-3.75c-.688 0-1.25.56-1.25 1.25s.562 1.25 1.25 1.25a1.25 1.25 0 0 0 0-2.5Z" })
-        ] }),
-        /* @__PURE__ */ jsx124("defs", { children: /* @__PURE__ */ jsx124("clipPath", { id: "eta", children: /* @__PURE__ */ jsx124("path", { fill: "currentColor", d: "M0 0h20v20H0z" }) }) })
-      ]
-    }
-  );
-};
-
-// src/icon/icons/share/index.react.tsx
-import { jsx as jsx125, jsxs as jsxs111 } from "react/jsx-runtime";
-var Share = (props) => {
-  let _ariaLabel = props["aria-label"];
-  _ariaLabel = _ariaLabel && _ariaLabel.trim();
-  const _ariaHidden = !_ariaLabel;
-  const _role = _ariaHidden ? "presentation" : "img";
-  return /* @__PURE__ */ jsxs111(
-    "svg",
-    {
-      className: props.className,
-      viewBox: "0 0 20 20",
-      "aria-label": _ariaLabel,
-      "aria-hidden": _ariaHidden,
-      role: _role,
-      style: {
-        height: "var(--icon-height)",
-        width: "var(--icon-width)",
-        display: "flex",
-        alignItems: "center",
-        justifyContent: "center"
-      },
-      children: [
-        /* @__PURE__ */ jsxs111("g", { "clip-path": "url(#eua)", children: [
-          /* @__PURE__ */ jsx125("path", { fill: "currentColor", d: "M8.821 10.667A.66.66 0 0 1 8.148 10V6.833c0-1.75 1.43-3.166 3.197-3.166h5.385L14.038 1a.634.634 0 0 1-.168-.417c0-.166.084-.333.168-.416a.646.646 0 0 1 .42-.167c.169 0 .337.083.421.167l3.786 3.75c.084.083.084.166.168.166v.5c0 .084-.084.167-.168.25l-3.786 3.75a.646.646 0 0 1-.42.167.646.646 0 0 1-.421-.167.634.634 0 0 1-.168-.416c0-.167.084-.334.168-.417L16.73 5h-5.385c-1.01 0-1.85.833-1.85 1.833V10c-.085.333-.337.667-.674.667Z" }),
-          /* @__PURE__ */ jsx125("path", { fill: "currentColor", d: "M3.184 20c-1.01 0-1.85-.833-1.85-1.833v-10c0-1 .84-1.834 1.85-1.834h1.851A.66.66 0 0 1 5.708 7c0 .25-.336.5-.673.5h-1.85c-.421 0-.674.25-.674.667v10a.66.66 0 0 0 .673.666h11.358a.66.66 0 0 0 .674-.666v-6.25a.66.66 0 0 1 .673-.667c.42 0 .673.25.673.667v6.25c0 1-.842 1.833-1.851 1.833H3.184Z" })
-        ] }),
-        /* @__PURE__ */ jsx125("defs", { children: /* @__PURE__ */ jsx125("clipPath", { id: "eua", children: /* @__PURE__ */ jsx125("path", { fill: "currentColor", d: "M0 0h20v20H0z" }) }) })
-      ]
-    }
-  );
-};
-
-// src/icon/icons/shrink/index.react.tsx
-import { jsx as jsx126, jsxs as jsxs112 } from "react/jsx-runtime";
-var Shrink = (props) => {
-  let _ariaLabel = props["aria-label"];
-  _ariaLabel = _ariaLabel && _ariaLabel.trim();
-  const _ariaHidden = !_ariaLabel;
-  const _role = _ariaHidden ? "presentation" : "img";
-  return /* @__PURE__ */ jsxs112(
-    "svg",
-    {
-      className: props.className,
-      viewBox: "0 0 20 20",
-      "aria-label": _ariaLabel,
-      "aria-hidden": _ariaHidden,
-      role: _role,
-      style: {
-        height: "var(--icon-height)",
-        width: "var(--icon-width)",
-        display: "flex",
-        alignItems: "center",
-        justifyContent: "center"
-      },
-      children: [
-        /* @__PURE__ */ jsx126("g", { "clip-path": "url(#eva)", children: /* @__PURE__ */ jsx126("path", { fill: "currentColor", d: "M14.375 6.248a.613.613 0 0 1-.217-.04l-.027-.01a.625.625 0 0 1-.334-.336l-.008-.022a.614.614 0 0 1-.039-.217v-5a.625.625 0 1 1 1.25 0v3.491L18.933.181a.62.62 0 0 1 1.02.202.623.623 0 0 1-.136.682l-3.934 3.933h3.491a.625.625 0 0 1 0 1.25h-4.999Zm-8.75 13.75A.625.625 0 0 1 5 19.373v-3.49l-3.933 3.933A.622.622 0 0 1 0 19.374c0-.166.065-.325.182-.442l3.934-3.934H.625a.625.625 0 0 1 0-1.25h5a.63.63 0 0 1 .217.04l.018.006a.63.63 0 0 1 .39.58v5a.625.625 0 0 1-.625.624Zm13.75 0a.62.62 0 0 1-.442-.183L15 15.883v3.49a.624.624 0 1 1-1.25 0v-5a.64.64 0 0 1 .038-.215l.005-.014a.636.636 0 0 1 .344-.349l.021-.008a.612.612 0 0 1 .216-.04h5a.625.625 0 1 1 0 1.25h-3.49l3.933 3.934a.628.628 0 0 1-.442 1.067ZM.625 6.248a.625.625 0 0 1 0-1.25h3.49L.184 1.065a.622.622 0 0 1 0-.883.619.619 0 0 1 .442-.184c.167 0 .324.065.442.184L5 4.115V.623a.625.625 0 0 1 1.25 0v5a.613.613 0 0 1-.04.218l-.006.017a.633.633 0 0 1-.579.39h-5Zm6.875 7.5c-.69 0-1.25-.56-1.25-1.25v-5c0-.689.56-1.25 1.25-1.25h5c.69 0 1.25.561 1.25 1.25v5c0 .69-.56 1.25-1.25 1.25h-5Zm0-1.25h5v-5h-5v5Z" }) }),
-        /* @__PURE__ */ jsx126("defs", { children: /* @__PURE__ */ jsx126("clipPath", { id: "eva", children: /* @__PURE__ */ jsx126("path", { fill: "currentColor", d: "M0 0h20v20H0z" }) }) })
-      ]
-    }
-  );
-};
-
-// src/icon/icons/shuffle/index.react.tsx
-import { jsx as jsx127, jsxs as jsxs113 } from "react/jsx-runtime";
-var Shuffle = (props) => {
-  let _ariaLabel = props["aria-label"];
-  _ariaLabel = _ariaLabel && _ariaLabel.trim();
-  const _ariaHidden = !_ariaLabel;
-  const _role = _ariaHidden ? "presentation" : "img";
-  return /* @__PURE__ */ jsxs113(
-    "svg",
-    {
-      className: props.className,
-      viewBox: "0 0 20 20",
-      "aria-label": _ariaLabel,
-      "aria-hidden": _ariaHidden,
-      role: _role,
-      style: {
-        height: "var(--icon-height)",
-        width: "var(--icon-width)",
-        display: "flex",
-        alignItems: "center",
-        justifyContent: "center"
-      },
-      children: [
-        /* @__PURE__ */ jsxs113("g", { "clip-path": "url(#ewa)", children: [
-          /* @__PURE__ */ jsx127("path", { fill: "currentColor", d: "M8.417 8.125a.626.626 0 0 1-.502-.253 4.052 4.052 0 0 0-2.938-1.624H.625a.625.625 0 0 1 0-1.25H5a5.303 5.303 0 0 1 3.918 2.13.622.622 0 0 1-.501.997Z" }),
-          /* @__PURE__ */ jsx127("path", { fill: "currentColor", d: "M.625 14.998a.625.625 0 1 1 0-1.25H5c2.663 0 4.408-3.956 4.426-3.996.083-.194 2.079-4.754 5.574-4.754h2.866l-1.433-1.433a.619.619 0 0 1-.184-.442.619.619 0 0 1 .386-.577.624.624 0 0 1 .68.136l2.5 2.5a.617.617 0 0 1 .137.204l.008.021a.59.59 0 0 1 0 .431l-.004.014a.636.636 0 0 1-.14.214l-2.5 2.499a.62.62 0 0 1-.883 0 .619.619 0 0 1-.183-.442c0-.166.065-.324.183-.441l1.434-1.434H15c-2.667 0-4.408 3.957-4.426 3.997-.083.194-2.082 4.753-5.574 4.753H.625Z" }),
-          /* @__PURE__ */ jsx127("path", { fill: "currentColor", d: "M16.875 17.498a.625.625 0 0 1-.442-1.066l1.434-1.434H15l-.047-.001a5.292 5.292 0 0 1-3.871-2.124.627.627 0 0 1 .501-.997c.197 0 .384.094.502.252a4.048 4.048 0 0 0 2.938 1.621h2.844l-1.434-1.433a.622.622 0 0 1 .442-1.067.625.625 0 0 1 .442.183l2.5 2.5a.617.617 0 0 1 .136.205l.008.021a.6.6 0 0 1 0 .43l-.004.014a.623.623 0 0 1-.14.214l-2.5 2.5a.624.624 0 0 1-.442.182Z" })
-        ] }),
-        /* @__PURE__ */ jsx127("defs", { children: /* @__PURE__ */ jsx127("clipPath", { id: "ewa", children: /* @__PURE__ */ jsx127("path", { fill: "currentColor", d: "M0 0h20v20H0z" }) }) })
-      ]
-    }
-  );
-};
-
-// src/icon/icons/social_instagram/index.react.tsx
-import { jsx as jsx128, jsxs as jsxs114 } from "react/jsx-runtime";
-var SocialInstagram = (props) => {
-  let _ariaLabel = props["aria-label"];
-  _ariaLabel = _ariaLabel && _ariaLabel.trim();
-  const _ariaHidden = !_ariaLabel;
-  const _role = _ariaHidden ? "presentation" : "img";
-  return /* @__PURE__ */ jsxs114(
-    "svg",
-    {
-      className: props.className,
-      viewBox: "0 0 20 20",
-      "aria-label": _ariaLabel,
-      "aria-hidden": _ariaHidden,
-      role: _role,
-      style: {
-        height: "var(--icon-height)",
-        width: "var(--icon-width)",
-        display: "flex",
-        alignItems: "center",
-        justifyContent: "center"
-      },
-      children: [
-        /* @__PURE__ */ jsxs114("g", { "clip-path": "url(#exclip0_307_9)", children: [
-          /* @__PURE__ */ jsx128("path", { d: "M5.61667 20H14.3667C15.8593 20 17.2909 19.4076 18.3471 18.353C19.4033 17.2983 19.9978 15.8676 20 14.375V5.625C19.9978 4.13239 19.4033 2.70167 18.3471 1.64702C17.2909 0.592365 15.8593 -1.63314e-06 14.3667 0H5.61667C4.12695 0.00440586 2.69962 0.598718 1.64701 1.65289C0.594392 2.70707 0.00219722 4.13527 0 5.625V14.375C0.00219722 15.8647 0.594392 17.2929 1.64701 18.3471C2.69962 19.4013 4.12695 19.9956 5.61667 20ZM1.24167 5.625C1.24387 4.46535 1.70551 3.35384 2.52551 2.53384C3.3455 1.71385 4.45702 1.2522 5.61667 1.25H14.3667C15.5263 1.2522 16.6378 1.71385 17.4578 2.53384C18.2778 3.35384 18.7395 4.46535 18.7417 5.625V14.375C18.7395 15.5346 18.2778 16.6462 17.4578 17.4662C16.6378 18.2862 15.5263 18.7478 14.3667 18.75H5.61667C4.45702 18.7478 3.3455 18.2862 2.52551 17.4662C1.70551 16.6462 1.24387 15.5346 1.24167 14.375V5.625Z", fill: "currentColor" }),
-          /* @__PURE__ */ jsx128("path", { d: "M10 14.7917C10.9473 14.79 11.8729 14.5076 12.6598 13.9801C13.4467 13.4526 14.0596 12.7036 14.421 11.8279C14.7824 10.9522 14.8761 9.98904 14.6902 9.0601C14.5044 8.13116 14.0474 7.27814 13.3769 6.60885C12.7065 5.93956 11.8527 5.48402 10.9234 5.29981C9.99415 5.1156 9.03113 5.21097 8.15605 5.57389C7.28098 5.9368 6.53311 6.55096 6.00696 7.33877C5.48082 8.12657 5.2 9.05266 5.2 10C5.2022 11.2716 5.70889 12.4904 6.60883 13.3887C7.50876 14.2871 8.7284 14.7917 10 14.7917ZM10 6.45834C10.7001 6.45999 11.384 6.6691 11.9654 7.05926C12.5467 7.44943 12.9993 8.00313 13.2661 8.65043C13.5329 9.29773 13.6019 10.0096 13.4643 10.696C13.3267 11.3825 12.9887 12.0128 12.4931 12.5073C11.9974 13.0017 11.3663 13.3382 10.6796 13.4742C9.99276 13.6102 9.28108 13.5396 8.63441 13.2713C7.98774 13.003 7.4351 12.549 7.0463 11.9668C6.65751 11.3845 6.45 10.7001 6.45 10C6.4522 9.05993 6.82719 8.15911 7.49271 7.49516C8.15823 6.83121 9.05992 6.45834 10 6.45834Z", fill: "currentColor" }),
-          /* @__PURE__ */ jsx128("path", { d: "M15.7 5.33333C16.2753 5.33333 16.7417 4.86696 16.7417 4.29167C16.7417 3.71637 16.2753 3.25 15.7 3.25C15.1247 3.25 14.6583 3.71637 14.6583 4.29167C14.6583 4.86696 15.1247 5.33333 15.7 5.33333Z", fill: "currentColor" })
-        ] }),
-        /* @__PURE__ */ jsx128("defs", { children: /* @__PURE__ */ jsx128("clipPath", { id: "exclip0_307_9", children: /* @__PURE__ */ jsx128("rect", { width: "20", height: "20", fill: "currentColor" }) }) })
-      ]
-    }
-  );
-};
-
-// src/icon/icons/study_math/index.react.tsx
-import { jsx as jsx129, jsxs as jsxs115 } from "react/jsx-runtime";
-var StudyMath = (props) => {
-  let _ariaLabel = props["aria-label"];
-  _ariaLabel = _ariaLabel && _ariaLabel.trim();
-  const _ariaHidden = !_ariaLabel;
-  const _role = _ariaHidden ? "presentation" : "img";
-  return /* @__PURE__ */ jsxs115(
-    "svg",
-    {
-      className: props.className,
-      viewBox: "0 0 20 20",
-      "aria-label": _ariaLabel,
-      "aria-hidden": _ariaHidden,
-      role: _role,
-      style: {
-        height: "var(--icon-height)",
-        width: "var(--icon-width)",
-        display: "flex",
-        alignItems: "center",
-        justifyContent: "center"
-      },
-      children: [
-        /* @__PURE__ */ jsxs115("g", { "clip-path": "url(#eya)", children: [
-          /* @__PURE__ */ jsx129("path", { fill: "currentColor", d: "M6.25 8.75a.625.625 0 0 1-.625-.624V6.25H3.75a.625.625 0 1 1 0-1.25h1.875V3.126a.625.625 0 0 1 1.25 0V5H8.75a.625.625 0 0 1 0 1.25H6.875v1.875a.625.625 0 0 1-.625.625Zm9.375 4.376a.62.62 0 0 1-.442-.183l-1.433-1.434-1.433 1.434a.622.622 0 0 1-1.066-.443c0-.165.065-.324.182-.442l1.434-1.433-1.434-1.433a.619.619 0 0 1-.183-.442.619.619 0 0 1 .386-.577.625.625 0 0 1 .68.135l1.434 1.434 1.433-1.434a.62.62 0 0 1 .884 0 .619.619 0 0 1 .183.442.619.619 0 0 1-.183.442l-1.434 1.433 1.434 1.433a.628.628 0 0 1-.442 1.068ZM4.375 15a.625.625 0 0 1 0-1.25H8.75a.625.625 0 1 1 0 1.25H4.375Zm2.187-1.875a.933.933 0 0 1-.93-.842v-.005a.584.584 0 0 1-.008-.086v-.004c0-.476.355-.872.83-.928a.94.94 0 0 1 1.036.82.68.68 0 0 1 .002.218.933.933 0 0 1-.929.827h-.001Zm.001 4.375a.939.939 0 0 1-.938-.933v-.004c0-.472.352-.868.822-.927a.625.625 0 0 1 .115-.011.942.942 0 0 1 .937.912l.001.026v.015a.93.93 0 0 1-.272.647.931.931 0 0 1-.662.276h-.003Z" }),
-          /* @__PURE__ */ jsx129("path", { fill: "currentColor", d: "M2.5 20.001a1.877 1.877 0 0 1-1.875-1.875V1.876A1.877 1.877 0 0 1 2.5.001h12.53c.493 0 .96.19 1.31.534l2.468 2.401c.36.35.567.84.567 1.343v13.848a1.876 1.876 0 0 1-1.875 1.875h-15V20Zm0-18.75a.625.625 0 0 0-.625.625v16.25c0 .345.28.625.625.625h15a.625.625 0 0 0 .625-.625V4.278a.627.627 0 0 0-.19-.447l-2.467-2.4a.621.621 0 0 0-.437-.18H2.5Z" })
-        ] }),
-        /* @__PURE__ */ jsx129("defs", { children: /* @__PURE__ */ jsx129("clipPath", { id: "eya", children: /* @__PURE__ */ jsx129("path", { fill: "currentColor", d: "M0 0h20v20H0z" }) }) })
-      ]
-    }
-  );
-};
-
-// src/icon/icons/subscript/index.react.tsx
-import { jsx as jsx130, jsxs as jsxs116 } from "react/jsx-runtime";
-var Subscript = (props) => {
-  let _ariaLabel = props["aria-label"];
-  _ariaLabel = _ariaLabel && _ariaLabel.trim();
-  const _ariaHidden = !_ariaLabel;
-  const _role = _ariaHidden ? "presentation" : "img";
-  return /* @__PURE__ */ jsxs116(
-    "svg",
-    {
-      className: props.className,
-      viewBox: "0 0 20 20",
-      "aria-label": _ariaLabel,
-      "aria-hidden": _ariaHidden,
-      role: _role,
-      style: {
-        height: "var(--icon-height)",
-        width: "var(--icon-width)",
-        display: "flex",
-        alignItems: "center",
-        justifyContent: "center"
-      },
-      children: [
-        /* @__PURE__ */ jsx130("g", { "clip-path": "url(#eza)", children: /* @__PURE__ */ jsx130("path", { fill: "currentColor", d: "M13.75 14.373a.619.619 0 0 1-.442-.183l-6.12-6.12-6.121 6.12a.62.62 0 0 1-.884 0 .627.627 0 0 1 0-.884l6.121-6.121-6.12-6.12A.619.619 0 0 1 0 .623.62.62 0 0 1 .183.182a.619.619 0 0 1 .442-.184c.167 0 .324.065.442.184l6.12 6.12 6.121-6.12a.62.62 0 0 1 .884 0 .619.619 0 0 1 .183.441.619.619 0 0 1-.183.442l-6.12 6.12 6.12 6.122a.627.627 0 0 1-.442 1.066Zm3.125 5.625a.627.627 0 0 1-.488-1.015l2.11-2.638c.164-.204.253-.46.253-.722a.625.625 0 0 0-1.25 0 .625.625 0 0 1-1.25 0c0-1.034.84-1.875 1.875-1.875 1.034 0 1.875.841 1.875 1.875 0 .545-.188 1.078-.527 1.503l-1.297 1.622h1.199a.625.625 0 0 1 0 1.25h-2.5Z" }) }),
-        /* @__PURE__ */ jsx130("defs", { children: /* @__PURE__ */ jsx130("clipPath", { id: "eza", children: /* @__PURE__ */ jsx130("path", { fill: "currentColor", d: "M0 0h20v20H0z" }) }) })
-      ]
-    }
-  );
-};
-
-// src/icon/icons/subtract/index.react.tsx
-import { jsx as jsx131, jsxs as jsxs117 } from "react/jsx-runtime";
-var Subtract = (props) => {
-  let _ariaLabel = props["aria-label"];
-  _ariaLabel = _ariaLabel && _ariaLabel.trim();
-  const _ariaHidden = !_ariaLabel;
-  const _role = _ariaHidden ? "presentation" : "img";
-  return /* @__PURE__ */ jsxs117(
-    "svg",
-    {
-      className: props.className,
-      viewBox: "0 0 20 20",
-      "aria-label": _ariaLabel,
-      "aria-hidden": _ariaHidden,
-      role: _role,
-      style: {
-        height: "var(--icon-height)",
-        width: "var(--icon-width)",
-        display: "flex",
-        alignItems: "center",
-        justifyContent: "center"
-      },
-      children: [
-        /* @__PURE__ */ jsx131("g", { "clip-path": "url(#faa)", children: /* @__PURE__ */ jsx131("path", { fill: "currentColor", d: "M.625 10.657a.625.625 0 1 1 0-1.25h18.75a.625.625 0 1 1 0 1.25H.625Z" }) }),
-        /* @__PURE__ */ jsx131("defs", { children: /* @__PURE__ */ jsx131("clipPath", { id: "faa", children: /* @__PURE__ */ jsx131("path", { fill: "currentColor", d: "M0 0h20v20H0z" }) }) })
-      ]
-    }
-  );
-};
-
-// src/icon/icons/superscript/index.react.tsx
-import { jsx as jsx132, jsxs as jsxs118 } from "react/jsx-runtime";
-var Superscript = (props) => {
-  let _ariaLabel = props["aria-label"];
-  _ariaLabel = _ariaLabel && _ariaLabel.trim();
-  const _ariaHidden = !_ariaLabel;
-  const _role = _ariaHidden ? "presentation" : "img";
-  return /* @__PURE__ */ jsxs118(
-    "svg",
-    {
-      className: props.className,
-      viewBox: "0 0 20 20",
-      "aria-label": _ariaLabel,
-      "aria-hidden": _ariaHidden,
-      role: _role,
-      style: {
-        height: "var(--icon-height)",
-        width: "var(--icon-width)",
-        display: "flex",
-        alignItems: "center",
-        justifyContent: "center"
-      },
-      children: [
-        /* @__PURE__ */ jsx132("g", { "clip-path": "url(#fba)", children: /* @__PURE__ */ jsx132("path", { fill: "currentColor", d: "M13.75 19.998a.619.619 0 0 1-.442-.183l-6.12-6.12-6.121 6.12a.62.62 0 0 1-.884 0 .627.627 0 0 1 0-.884l6.121-6.121-6.12-6.12A.619.619 0 0 1 0 6.248c0-.166.065-.324.183-.441a.619.619 0 0 1 .442-.184c.167 0 .324.065.442.184l6.12 6.12 6.121-6.12a.62.62 0 0 1 .884 0 .619.619 0 0 1 .183.441.619.619 0 0 1-.183.442L8.07 12.81l6.12 6.122a.627.627 0 0 1-.441 1.066Zm3.125-13.75a.626.626 0 0 1-.488-1.015l2.11-2.637c.163-.204.253-.461.253-.723a.625.625 0 0 0-1.25 0 .625.625 0 0 1-1.25 0c0-1.034.84-1.875 1.875-1.875C19.159-.002 20 .84 20 1.873c0 .545-.188 1.078-.527 1.503l-1.297 1.622h1.199a.625.625 0 0 1 0 1.25h-2.5Z" }) }),
-        /* @__PURE__ */ jsx132("defs", { children: /* @__PURE__ */ jsx132("clipPath", { id: "fba", children: /* @__PURE__ */ jsx132("path", { fill: "currentColor", d: "M0 0h20v20H0z" }) }) })
-      ]
-    }
-  );
-};
-
-// src/icon/icons/synchronize__arrows__square/index.react.tsx
-import { jsx as jsx133, jsxs as jsxs119 } from "react/jsx-runtime";
-var SynchronizeArrowsSquare = (props) => {
-  let _ariaLabel = props["aria-label"];
-  _ariaLabel = _ariaLabel && _ariaLabel.trim();
-  const _ariaHidden = !_ariaLabel;
-  const _role = _ariaHidden ? "presentation" : "img";
-  return /* @__PURE__ */ jsxs119(
-    "svg",
-    {
-      className: props.className,
-      viewBox: "0 0 20 20",
-      "aria-label": _ariaLabel,
-      "aria-hidden": _ariaHidden,
-      role: _role,
-      style: {
-        height: "var(--icon-height)",
-        width: "var(--icon-width)",
-        display: "flex",
-        alignItems: "center",
-        justifyContent: "center"
-      },
-      children: [
-        /* @__PURE__ */ jsx133("g", { "clip-path": "url(#fca)", children: /* @__PURE__ */ jsx133("path", { fill: "currentColor", d: "M6.251 17.708a3.128 3.128 0 0 1-3.125-3.125v-4.118l-2.058 2.058a.619.619 0 0 1-.884 0 .626.626 0 0 1 0-.884L3.31 8.514a.384.384 0 0 1 .048-.039.613.613 0 0 1 .788 0 .318.318 0 0 1 .047.038l3.125 3.125a.627.627 0 0 1-.681 1.02.618.618 0 0 1-.202-.135l-2.059-2.059v4.12a1.877 1.877 0 0 0 1.875 1.874h6.875a.625.625 0 0 1 0 1.25H6.251Zm10-5.625a.626.626 0 0 1-.466-.21l-3.1-3.1a.619.619 0 0 1 0-.884.624.624 0 0 1 .883 0l2.058 2.059V5.833a1.877 1.877 0 0 0-1.875-1.875H6.876a.625.625 0 1 1 0-1.25h6.875a3.129 3.129 0 0 1 3.125 3.125v4.114l2.059-2.059a.62.62 0 0 1 .883 0 .619.619 0 0 1 .183.442.619.619 0 0 1-.183.442l-3.113 3.113a.611.611 0 0 1-.454.198Z" }) }),
-        /* @__PURE__ */ jsx133("defs", { children: /* @__PURE__ */ jsx133("clipPath", { id: "fca", children: /* @__PURE__ */ jsx133("path", { fill: "currentColor", d: "M0 0h20v20H0z" }) }) })
-      ]
-    }
-  );
-};
-
-// src/icon/icons/synchronize_arrow_clock/index.react.tsx
-import { jsx as jsx134, jsxs as jsxs120 } from "react/jsx-runtime";
-var SynchronizeArrowClock = (props) => {
-  let _ariaLabel = props["aria-label"];
-  _ariaLabel = _ariaLabel && _ariaLabel.trim();
-  const _ariaHidden = !_ariaLabel;
-  const _role = _ariaHidden ? "presentation" : "img";
-  return /* @__PURE__ */ jsxs120(
-    "svg",
-    {
-      className: props.className,
-      viewBox: "0 0 20 20",
-      "aria-label": _ariaLabel,
-      "aria-hidden": _ariaHidden,
-      role: _role,
-      style: {
-        height: "var(--icon-height)",
-        width: "var(--icon-width)",
-        display: "flex",
-        alignItems: "center",
-        justifyContent: "center"
-      },
-      children: [
-        /* @__PURE__ */ jsxs120("g", { "clip-path": "url(#fda)", children: [
-          /* @__PURE__ */ jsx134("path", { fill: "currentColor", d: "M11.25 18.958a.625.625 0 0 1 0-1.25c4.136 0 7.5-3.364 7.5-7.5 0-4.135-3.364-7.5-7.5-7.5-3.787 0-6.966 2.871-7.438 6.551l1.371-1.37a.62.62 0 0 1 .884 0 .619.619 0 0 1 .183.442.619.619 0 0 1-.183.442L3.578 11.26a.61.61 0 0 1-.453.197.625.625 0 0 1-.467-.21L.183 8.773A.62.62 0 0 1 0 8.33c0-.167.065-.324.183-.442a.619.619 0 0 1 .442-.183c.167 0 .324.065.442.183l1.476 1.476c.429-4.42 4.194-7.907 8.707-7.907 4.825 0 8.75 3.925 8.75 8.75s-3.925 8.75-8.75 8.75Z" }),
-          /* @__PURE__ */ jsx134("path", { fill: "currentColor", d: "M10.625 11.456A.625.625 0 0 1 10 10.83V5.206a.625.625 0 1 1 1.25 0v5H15a.625.625 0 0 1 0 1.25h-4.375Z" })
-        ] }),
-        /* @__PURE__ */ jsx134("defs", { children: /* @__PURE__ */ jsx134("clipPath", { id: "fda", children: /* @__PURE__ */ jsx134("path", { fill: "currentColor", d: "M0 0h20v20H0z" }) }) })
-      ]
-    }
-  );
-};
-
-// src/icon/icons/synchronize_arrows_warning/index.react.tsx
-import { jsx as jsx135, jsxs as jsxs121 } from "react/jsx-runtime";
-var SynchronizeArrowsWarning = (props) => {
-  let _ariaLabel = props["aria-label"];
-  _ariaLabel = _ariaLabel && _ariaLabel.trim();
-  const _ariaHidden = !_ariaLabel;
-  const _role = _ariaHidden ? "presentation" : "img";
-  return /* @__PURE__ */ jsxs121(
-    "svg",
-    {
-      className: props.className,
-      viewBox: "0 0 20 20",
-      "aria-label": _ariaLabel,
-      "aria-hidden": _ariaHidden,
-      role: _role,
-      style: {
-        height: "var(--icon-height)",
-        width: "var(--icon-width)",
-        display: "flex",
-        alignItems: "center",
-        justifyContent: "center"
-      },
-      children: [
-        /* @__PURE__ */ jsxs121("g", { "clip-path": "url(#fea)", children: [
-          /* @__PURE__ */ jsx135("path", { fill: "currentColor", d: "M10.004 18.323a8.14 8.14 0 0 1-6.059-2.704.62.62 0 0 1-.034-.793.626.626 0 0 1 .966-.041 6.886 6.886 0 0 0 9.708.54 6.897 6.897 0 0 0 2.222-4.172l-.74.739A.622.622 0 0 1 15 11.45c0-.166.066-.325.182-.442l1.875-1.875a.397.397 0 0 1 .049-.04.645.645 0 0 1 .446-.138.665.665 0 0 1 .216.058.65.65 0 0 1 .17.116l1.878 1.878a.619.619 0 0 1 0 .883.622.622 0 0 1-.884 0l-.854-.854a8.147 8.147 0 0 1-2.661 5.217 8.097 8.097 0 0 1-5.414 2.07ZM2.5 11.453a.626.626 0 0 1-.468-.213L.183 9.393a.62.62 0 0 1 0-.885.62.62 0 0 1 .884 0l.854.855a8.15 8.15 0 0 1 2.66-5.217 8.105 8.105 0 0 1 5.87-2.058 8.073 8.073 0 0 1 5.605 2.695.625.625 0 0 1-.932.833 6.886 6.886 0 0 0-9.707-.54 6.895 6.895 0 0 0-2.222 4.17l.738-.738a.62.62 0 0 1 .884 0 .62.62 0 0 1 0 .884l-1.863 1.862a.609.609 0 0 1-.454.199Z" }),
-          /* @__PURE__ */ jsx135("path", { fill: "currentColor", d: "M10 11.45a.625.625 0 0 1-.625-.625v-5a.625.625 0 0 1 1.25 0v5a.625.625 0 0 1-.625.625Zm0 3.128a.939.939 0 0 1-.938-.938c0-.478.356-.873.83-.928a.633.633 0 0 1 .225-.002l.027.007c.454.07.793.463.793.923a.938.938 0 0 1-.937.938Z" })
-        ] }),
-        /* @__PURE__ */ jsx135("defs", { children: /* @__PURE__ */ jsx135("clipPath", { id: "fea", children: /* @__PURE__ */ jsx135("path", { fill: "currentColor", d: "M0 0h20v20H0z" }) }) })
-      ]
-    }
-  );
-};
-
-// src/icon/icons/tablet_touch/index.react.tsx
-import { jsx as jsx136, jsxs as jsxs122 } from "react/jsx-runtime";
-var TabletTouch = (props) => {
-  let _ariaLabel = props["aria-label"];
-  _ariaLabel = _ariaLabel && _ariaLabel.trim();
-  const _ariaHidden = !_ariaLabel;
-  const _role = _ariaHidden ? "presentation" : "img";
-  return /* @__PURE__ */ jsxs122(
-    "svg",
-    {
-      className: props.className,
-      viewBox: "0 0 20 20",
-      "aria-label": _ariaLabel,
-      "aria-hidden": _ariaHidden,
-      role: _role,
-      style: {
-        height: "var(--icon-height)",
-        width: "var(--icon-width)",
-        display: "flex",
-        alignItems: "center",
-        justifyContent: "center"
-      },
-      children: [
-        /* @__PURE__ */ jsxs122("g", { "clip-path": "url(#ffa)", children: [
-          /* @__PURE__ */ jsx136("path", { fill: "currentColor", d: "M19.375 20a.625.625 0 0 1-.625-.625v-2.883a1.869 1.869 0 0 0-1.567-1.849l-1.66-.276A.624.624 0 0 1 15 13.75v-2.5a.625.625 0 0 0-1.25 0v4.375a.63.63 0 0 1-.296.532.629.629 0 0 1-.608.027l-.98-.489a.74.74 0 0 0-.792 1.238l2.442 1.954a.625.625 0 1 1-.781.976l-2.44-1.953a1.994 1.994 0 0 1-.29-2.825 1.986 1.986 0 0 1 2.422-.508l.073.037V11.25c0-1.034.84-1.875 1.875-1.875 1.034 0 1.875.84 1.875 1.875v1.97l1.138.19A3.115 3.115 0 0 1 20 16.493v2.883a.625.625 0 0 1-.625.625Z" }),
-          /* @__PURE__ */ jsx136("path", { fill: "currentColor", d: "M1.875 18.75A1.878 1.878 0 0 1 0 16.875v-2.458a.286.286 0 0 1 0-.083V1.875C0 .841.842 0 1.875 0h10a1.877 1.877 0 0 1 1.875 1.875V7.5a.625.625 0 1 1-1.25 0V1.875a.625.625 0 0 0-.625-.625h-10a.625.625 0 0 0-.625.625V13.75h6.875a.625.625 0 1 1 0 1.25H1.25v1.875c0 .345.28.625.625.625h6.25a.625.625 0 1 1 0 1.25h-6.25Z" }),
-          /* @__PURE__ */ jsx136("path", { fill: "currentColor", d: "M3.125 7.5a.625.625 0 0 1-.625-.625v-2.5a.625.625 0 0 1 .625-.625h2.5a.625.625 0 0 1 .625.625v2.5a.625.625 0 0 1-.625.625h-2.5ZM5 6.25V5H3.75v1.25H5ZM8.125 7.5a.625.625 0 0 1-.625-.625v-2.5a.625.625 0 0 1 .625-.625h2.5a.625.625 0 0 1 .625.625v2.5a.625.625 0 0 1-.625.625h-2.5ZM10 6.25V5H8.75v1.25H10ZM3.125 12.5a.625.625 0 0 1-.625-.625v-2.5a.625.625 0 0 1 .625-.625h2.5a.625.625 0 0 1 .625.625v2.5c0 .345-3.125.625-3.125.625ZM5 11.25V10H3.75v1.25H5Z" })
-        ] }),
-        /* @__PURE__ */ jsx136("defs", { children: /* @__PURE__ */ jsx136("clipPath", { id: "ffa", children: /* @__PURE__ */ jsx136("path", { fill: "currentColor", d: "M0 0h20v20H0z" }) }) })
-      ]
-    }
-  );
-};
-
-// src/icon/icons/task_list/index.react.tsx
-import { jsx as jsx137, jsxs as jsxs123 } from "react/jsx-runtime";
-var TaskList = (props) => {
-  let _ariaLabel = props["aria-label"];
-  _ariaLabel = _ariaLabel && _ariaLabel.trim();
-  const _ariaHidden = !_ariaLabel;
-  const _role = _ariaHidden ? "presentation" : "img";
-  return /* @__PURE__ */ jsxs123(
-    "svg",
-    {
-      className: props.className,
-      viewBox: "0 0 20 20",
-      "aria-label": _ariaLabel,
-      "aria-hidden": _ariaHidden,
-      role: _role,
-      style: {
-        height: "var(--icon-height)",
-        width: "var(--icon-width)",
-        display: "flex",
-        alignItems: "center",
-        justifyContent: "center"
-      },
-      children: [
-        /* @__PURE__ */ jsxs123("g", { "clip-path": "url(#fga)", children: [
-          /* @__PURE__ */ jsx137("path", { fill: "currentColor", d: "M6.25 9.383a.625.625 0 0 1 0-1.25h7.5a.625.625 0 1 1 0 1.25h-7.5Zm0 3.125a.625.625 0 1 1 0-1.25h7.5a.624.624 0 1 1 0 1.25h-7.5Zm0 3.125a.625.625 0 1 1 0-1.25h7.5a.624.624 0 1 1 0 1.25h-7.5Z" }),
-          /* @__PURE__ */ jsx137("path", { fill: "currentColor", d: "M3.75 20.008a1.876 1.876 0 0 1-1.875-1.875V5.008A1.877 1.877 0 0 1 3.75 3.133h2.553C6.603 1.36 8.175.008 10 .008s3.397 1.351 3.697 3.125h2.553a1.877 1.877 0 0 1 1.875 1.875v13.125a1.876 1.876 0 0 1-1.875 1.875H3.75Zm0-15.625a.625.625 0 0 0-.625.625v13.125c0 .345.28.625.625.625h12.5a.624.624 0 0 0 .625-.625V5.008a.625.625 0 0 0-.625-.625h-3.125a.625.625 0 0 1-.625-.625c0-1.378-1.122-2.5-2.5-2.5a2.503 2.503 0 0 0-2.5 2.5.625.625 0 0 1-.625.625H3.75Z" }),
-          /* @__PURE__ */ jsx137("path", { fill: "currentColor", d: "M10 4.383a.937.937 0 1 0 0-1.875.937.937 0 0 0 0 1.875Z" })
-        ] }),
-        /* @__PURE__ */ jsx137("defs", { children: /* @__PURE__ */ jsx137("clipPath", { id: "fga", children: /* @__PURE__ */ jsx137("path", { fill: "currentColor", d: "M0 0h20v20H0z" }) }) })
-      ]
-    }
-  );
-};
-
-// src/icon/icons/task_list_clock/index.react.tsx
-import { jsx as jsx138, jsxs as jsxs124 } from "react/jsx-runtime";
-var TaskListClock = (props) => {
-  let _ariaLabel = props["aria-label"];
-  _ariaLabel = _ariaLabel && _ariaLabel.trim();
-  const _ariaHidden = !_ariaLabel;
-  const _role = _ariaHidden ? "presentation" : "img";
-  return /* @__PURE__ */ jsxs124(
-    "svg",
-    {
-      className: props.className,
-      viewBox: "0 0 20 20",
-      "aria-label": _ariaLabel,
-      "aria-hidden": _ariaHidden,
-      role: _role,
-      style: {
-        height: "var(--icon-height)",
-        width: "var(--icon-width)",
-        display: "flex",
-        alignItems: "center",
-        justifyContent: "center"
-      },
-      children: [
-        /* @__PURE__ */ jsxs124("g", { "clip-path": "url(#fha)", children: [
-          /* @__PURE__ */ jsx138("path", { fill: "currentColor", d: "M14.375 20.01a5.631 5.631 0 0 1-5.625-5.626 5.631 5.631 0 0 1 5.625-5.625A5.631 5.631 0 0 1 20 14.384a5.631 5.631 0 0 1-5.625 5.625Zm0-10A4.38 4.38 0 0 0 10 14.383a4.38 4.38 0 0 0 4.375 4.375 4.38 4.38 0 0 0 4.375-4.375 4.38 4.38 0 0 0-4.375-4.375Z" }),
-          /* @__PURE__ */ jsx138("path", { fill: "currentColor", d: "M14.375 15.01a.625.625 0 0 1-.625-.626v-2.21a.624.624 0 1 1 1.25 0v1.585h1.585a.624.624 0 1 1 0 1.25h-2.21Zm-10-5.626a.625.625 0 0 1 0-1.25H8.75a.625.625 0 1 1 0 1.25H4.375Zm0 3.126a.625.625 0 0 1 0-1.25h2.5a.625.625 0 1 1 0 1.25h-2.5Zm0 3.124a.625.625 0 0 1 0-1.25h2.5a.625.625 0 1 1 0 1.25h-2.5Z" }),
-          /* @__PURE__ */ jsx138("path", { fill: "currentColor", d: "M1.875 20.01A1.877 1.877 0 0 1 0 18.133V5.01a1.877 1.877 0 0 1 1.875-1.875h2.553C4.729 1.36 6.3.01 8.125.01s3.397 1.351 3.697 3.125h2.553A1.877 1.877 0 0 1 16.25 5.01v1.875a.625.625 0 1 1-1.25 0V5.01a.625.625 0 0 0-.625-.625H11.25a.625.625 0 0 1-.625-.625c0-1.378-1.122-2.5-2.5-2.5a2.503 2.503 0 0 0-2.5 2.5.625.625 0 0 1-.625.625H1.875a.625.625 0 0 0-.625.625v13.125c0 .345.28.625.625.625h6.25a.625.625 0 1 1 0 1.25h-6.25Z" }),
-          /* @__PURE__ */ jsx138("path", { fill: "currentColor", d: "M8.125 4.384a.938.938 0 1 0 0-1.875.938.938 0 0 0 0 1.875Z" })
-        ] }),
-        /* @__PURE__ */ jsx138("defs", { children: /* @__PURE__ */ jsx138("clipPath", { id: "fha", children: /* @__PURE__ */ jsx138("path", { fill: "currentColor", d: "M0 0h20v20H0z" }) }) })
-      ]
-    }
-  );
-};
-
-// src/icon/icons/teacher/index.react.tsx
-import { jsx as jsx139, jsxs as jsxs125 } from "react/jsx-runtime";
-var Teacher = (props) => {
-  let _ariaLabel = props["aria-label"];
-  _ariaLabel = _ariaLabel && _ariaLabel.trim();
-  const _ariaHidden = !_ariaLabel;
-  const _role = _ariaHidden ? "presentation" : "img";
-  return /* @__PURE__ */ jsxs125(
-    "svg",
-    {
-      className: props.className,
-      viewBox: "0 0 20 20",
-      "aria-label": _ariaLabel,
-      "aria-hidden": _ariaHidden,
-      role: _role,
-      style: {
-        height: "var(--icon-height)",
-        width: "var(--icon-width)",
-        display: "flex",
-        alignItems: "center",
-        justifyContent: "center"
-      },
-      children: [
-        /* @__PURE__ */ jsx139("g", { "clip-path": "url(#fia)", children: /* @__PURE__ */ jsx139("path", { fill: "currentColor", d: "M4.375 6.25A3.128 3.128 0 0 1 1.25 3.125 3.128 3.128 0 0 1 4.375 0 3.128 3.128 0 0 1 7.5 3.125 3.128 3.128 0 0 1 4.375 6.25Zm0-5c-1.034 0-1.875.84-1.875 1.875C2.5 4.159 3.34 5 4.375 5 5.409 5 6.25 4.16 6.25 3.125c0-1.034-.84-1.875-1.875-1.875ZM2.5 20a.623.623 0 0 1-.622-.563L1.31 13.75H.625A.625.625 0 0 1 0 13.125v-2.5A4.38 4.38 0 0 1 4.375 6.25a4.38 4.38 0 0 1 4.375 4.375v2.5a.625.625 0 0 1-.625.625h-.684l-.569 5.688A.624.624 0 0 1 6.25 20H2.5Zm3.184-1.25.569-5.688a.624.624 0 0 1 .622-.562H7.5v-1.875A3.15 3.15 0 0 0 5 7.565v4.31a.625.625 0 1 1-1.25 0v-4.31a3.15 3.15 0 0 0-2.5 3.06V12.5h.625c.322 0 .59.242.622.563l.569 5.687h2.618Zm4.941-4.375a.624.624 0 1 1 0-1.25h7.5a.624.624 0 0 0 .625-.625V1.875a.625.625 0 0 0-.625-.625h-8.75a.625.625 0 0 1 0-1.25h8.75A1.877 1.877 0 0 1 20 1.875V12.5a1.876 1.876 0 0 1-1.875 1.875h-7.5Z" }) }),
-        /* @__PURE__ */ jsx139("defs", { children: /* @__PURE__ */ jsx139("clipPath", { id: "fia", children: /* @__PURE__ */ jsx139("path", { fill: "currentColor", d: "M0 0h20v20H0z" }) }) })
-      ]
-    }
-  );
-};
-
-// src/icon/icons/teacher_correct/index.react.tsx
-import { jsx as jsx140, jsxs as jsxs126 } from "react/jsx-runtime";
-var TeacherCorrect = (props) => {
-  let _ariaLabel = props["aria-label"];
-  _ariaLabel = _ariaLabel && _ariaLabel.trim();
-  const _ariaHidden = !_ariaLabel;
-  const _role = _ariaHidden ? "presentation" : "img";
-  return /* @__PURE__ */ jsxs126(
-    "svg",
-    {
-      className: props.className,
-      viewBox: "0 0 20 20",
-      "aria-label": _ariaLabel,
-      "aria-hidden": _ariaHidden,
-      role: _role,
-      style: {
-        height: "var(--icon-height)",
-        width: "var(--icon-width)",
-        display: "flex",
-        alignItems: "center",
-        justifyContent: "center"
-      },
-      children: [
-        /* @__PURE__ */ jsx140("g", { "clip-path": "url(#fja)", children: /* @__PURE__ */ jsx140("path", { fill: "currentColor", d: "M4.375 6.25A3.128 3.128 0 0 1 1.25 3.125 3.128 3.128 0 0 1 4.375 0 3.128 3.128 0 0 1 7.5 3.125 3.128 3.128 0 0 1 4.375 6.25Zm0-5c-1.034 0-1.875.84-1.875 1.875C2.5 4.159 3.34 5 4.375 5 5.409 5 6.25 4.16 6.25 3.125c0-1.034-.84-1.875-1.875-1.875ZM2.5 20a.622.622 0 0 1-.622-.563L1.31 13.75H.625A.625.625 0 0 1 0 13.125v-2.5A4.38 4.38 0 0 1 4.375 6.25a4.38 4.38 0 0 1 4.375 4.375v2.5c0 .345-.28.625-.625.625h-.684l-.569 5.688A.624.624 0 0 1 6.25 20H2.5Zm3.184-1.25.569-5.688c.032-.32.3-.562.622-.562H7.5v-1.875A3.15 3.15 0 0 0 5 7.565v4.31a.625.625 0 0 1-1.25 0v-4.31a3.15 3.15 0 0 0-2.5 3.06V12.5h.625c.322 0 .59.242.622.563l.569 5.687h2.618Zm3.691-17.5a.625.625 0 0 1 0-1.25h10a.625.625 0 0 1 0 1.25h-10Zm1.25 13.125a.625.625 0 0 1 0-1.25h8.75a.625.625 0 0 1 0 1.25h-8.75ZM13.75 10a.619.619 0 0 1-.442-.183l-2.5-2.5a.619.619 0 0 1-.183-.442.624.624 0 0 1 1.067-.442l1.99 1.99L17 4a.629.629 0 0 1 .874-.126c.276.208.332.6.125.876l-3.75 5a.623.623 0 0 1-.499.25Z" }) }),
-        /* @__PURE__ */ jsx140("defs", { children: /* @__PURE__ */ jsx140("clipPath", { id: "fja", children: /* @__PURE__ */ jsx140("path", { fill: "currentColor", d: "M0 0h20v20H0z" }) }) })
-      ]
-    }
-  );
-};
-
-// src/icon/icons/text_bold/index.react.tsx
-import { jsx as jsx141, jsxs as jsxs127 } from "react/jsx-runtime";
-var TextBold = (props) => {
-  let _ariaLabel = props["aria-label"];
-  _ariaLabel = _ariaLabel && _ariaLabel.trim();
-  const _ariaHidden = !_ariaLabel;
-  const _role = _ariaHidden ? "presentation" : "img";
-  return /* @__PURE__ */ jsxs127(
-    "svg",
-    {
-      className: props.className,
-      viewBox: "0 0 20 20",
-      "aria-label": _ariaLabel,
-      "aria-hidden": _ariaHidden,
-      role: _role,
-      style: {
-        height: "var(--icon-height)",
-        width: "var(--icon-width)",
-        display: "flex",
-        alignItems: "center",
-        justifyContent: "center"
-      },
-      children: [
-        /* @__PURE__ */ jsx141("g", { "clip-path": "url(#fka)", children: /* @__PURE__ */ jsx141("path", { fill: "currentColor", d: "M3.125 19.998a.625.625 0 0 1 0-1.25H5v-17.5H3.125a.625.625 0 0 1 0-1.25h8.75a4.38 4.38 0 0 1 4.375 4.375 4.4 4.4 0 0 1-2.194 3.791 6.283 6.283 0 0 1 3.444 5.584 6.258 6.258 0 0 1-6.25 6.25H3.125Zm8.125-1.25c2.757 0 5-2.243 5-5 0-2.756-2.243-5-5-5h-5v10h5Zm.625-11.25A3.128 3.128 0 0 0 15 4.373a3.128 3.128 0 0 0-3.125-3.125H6.25v6.25h5.625Z" }) }),
-        /* @__PURE__ */ jsx141("defs", { children: /* @__PURE__ */ jsx141("clipPath", { id: "fka", children: /* @__PURE__ */ jsx141("path", { fill: "currentColor", d: "M0 0h20v20H0z" }) }) })
-      ]
-    }
-  );
-};
-
-// src/icon/icons/text_format/index.react.tsx
-import { jsx as jsx142, jsxs as jsxs128 } from "react/jsx-runtime";
-var TextFormat = (props) => {
-  let _ariaLabel = props["aria-label"];
-  _ariaLabel = _ariaLabel && _ariaLabel.trim();
-  const _ariaHidden = !_ariaLabel;
-  const _role = _ariaHidden ? "presentation" : "img";
-  return /* @__PURE__ */ jsxs128(
-    "svg",
-    {
-      className: props.className,
-      viewBox: "0 0 20 20",
-      "aria-label": _ariaLabel,
-      "aria-hidden": _ariaHidden,
-      role: _role,
-      style: {
-        height: "var(--icon-height)",
-        width: "var(--icon-width)",
-        display: "flex",
-        alignItems: "center",
-        justifyContent: "center"
-      },
-      children: [
-        /* @__PURE__ */ jsx142("g", { "clip-path": "url(#fla)", children: /* @__PURE__ */ jsx142("path", { fill: "currentColor", d: "M9.063 14.373a.625.625 0 0 1 0-1.25h.312V6.875H7.5V7.5a.625.625 0 0 1-1.25 0v-.94a.943.943 0 0 1 .936-.935h2.782L10 5.623l.033.002h2.78a.942.942 0 0 1 .937.936V7.5a.625.625 0 1 1-1.25 0v-.625h-1.875v6.248h.313a.625.625 0 1 1 0 1.25H9.062ZM2.5 13.748a.625.625 0 0 1-.625-.625v-6.25a.625.625 0 0 1 1.25 0v6.25a.625.625 0 0 1-.625.625Zm-1.406-8.75A1.095 1.095 0 0 1 0 3.904V1.093C0 .489.49-.002 1.094-.002h2.812C4.509-.002 5 .49 5 1.093v2.811c0 .604-.49 1.094-1.094 1.094H1.094Zm2.656-1.25v-2.5h-2.5v2.5h2.5Zm-2.656 16.25A1.095 1.095 0 0 1 0 18.904v-2.811c0-.604.49-1.095 1.094-1.095h2.812c.603 0 1.094.491 1.094 1.095v2.811c0 .604-.49 1.094-1.094 1.094H1.094Zm2.656-1.25v-2.5h-2.5v2.5h2.5Zm13.75-5a.625.625 0 0 1-.625-.625v-6.25a.625.625 0 1 1 1.25 0v6.25a.625.625 0 0 1-.625.625ZM6.875 3.123a.625.625 0 0 1 0-1.25h6.25a.625.625 0 1 1 0 1.25h-6.25Zm0 15a.625.625 0 0 1 0-1.25h6.25a.625.625 0 1 1 0 1.25h-6.25Zm9.219-13.125A1.095 1.095 0 0 1 15 3.904V1.093c0-.604.49-1.095 1.094-1.095h2.812C19.509-.002 20 .49 20 1.093v2.811c0 .604-.49 1.094-1.094 1.094h-2.812Zm2.656-1.25v-2.5h-2.5v2.5h2.5Zm-2.656 16.25A1.095 1.095 0 0 1 15 18.904v-2.811c0-.604.49-1.095 1.094-1.095h2.812c.603 0 1.094.491 1.094 1.095v2.811c0 .604-.49 1.094-1.094 1.094h-2.812Zm2.656-1.25v-2.5h-2.5v2.5h2.5Z" }) }),
-        /* @__PURE__ */ jsx142("defs", { children: /* @__PURE__ */ jsx142("clipPath", { id: "fla", children: /* @__PURE__ */ jsx142("path", { fill: "currentColor", d: "M0 0h20v20H0z" }) }) })
-      ]
-    }
-  );
-};
-
-// src/icon/icons/text_italic/index.react.tsx
-import { jsx as jsx143, jsxs as jsxs129 } from "react/jsx-runtime";
-var TextItalic = (props) => {
-  let _ariaLabel = props["aria-label"];
-  _ariaLabel = _ariaLabel && _ariaLabel.trim();
-  const _ariaHidden = !_ariaLabel;
-  const _role = _ariaHidden ? "presentation" : "img";
-  return /* @__PURE__ */ jsxs129(
-    "svg",
-    {
-      className: props.className,
-      viewBox: "0 0 20 20",
-      "aria-label": _ariaLabel,
-      "aria-hidden": _ariaHidden,
-      role: _role,
-      style: {
-        height: "var(--icon-height)",
-        width: "var(--icon-width)",
-        display: "flex",
-        alignItems: "center",
-        justifyContent: "center"
-      },
-      children: [
-        /* @__PURE__ */ jsx143("g", { "clip-path": "url(#fma)", children: /* @__PURE__ */ jsx143("path", { fill: "currentColor", d: "M.625 19.998a.625.625 0 0 1 0-1.25h4l9.333-17.5h-3.333a.625.625 0 1 1 0-1.25h8.75a.625.625 0 1 1 0 1.25h-4l-9.333 17.5h3.333a.625.625 0 1 1 0 1.25H.625Z" }) }),
-        /* @__PURE__ */ jsx143("defs", { children: /* @__PURE__ */ jsx143("clipPath", { id: "fma", children: /* @__PURE__ */ jsx143("path", { fill: "currentColor", d: "M0 0h20v20H0z" }) }) })
-      ]
-    }
-  );
-};
-
-// src/icon/icons/text_style/index.react.tsx
-import { jsx as jsx144, jsxs as jsxs130 } from "react/jsx-runtime";
-var TextStyle = (props) => {
-  let _ariaLabel = props["aria-label"];
-  _ariaLabel = _ariaLabel && _ariaLabel.trim();
-  const _ariaHidden = !_ariaLabel;
-  const _role = _ariaHidden ? "presentation" : "img";
-  return /* @__PURE__ */ jsxs130(
-    "svg",
-    {
-      className: props.className,
-      viewBox: "0 0 20 20",
-      "aria-label": _ariaLabel,
-      "aria-hidden": _ariaHidden,
-      role: _role,
-      style: {
-        height: "var(--icon-height)",
-        width: "var(--icon-width)",
-        display: "flex",
-        alignItems: "center",
-        justifyContent: "center"
-      },
-      children: [
-        /* @__PURE__ */ jsx144("g", { "clip-path": "url(#fna)", children: /* @__PURE__ */ jsx144("path", { fill: "currentColor", d: "M6.25 19.998a.625.625 0 0 1 0-1.25h3.125v-17.5h-6.25A1.245 1.245 0 0 0 1.875 2.5v.624a.625.625 0 0 1-1.25 0V2.5a2.482 2.482 0 0 1 .73-1.767 2.485 2.485 0 0 1 1.768-.735h13.752c1.378 0 2.5 1.122 2.5 2.5v.627a.625.625 0 1 1-1.25 0V2.5c0-.69-.56-1.252-1.25-1.252h-6.25v17.5h3.125a.625.625 0 1 1 0 1.25h-7.5Z" }) }),
-        /* @__PURE__ */ jsx144("defs", { children: /* @__PURE__ */ jsx144("clipPath", { id: "fna", children: /* @__PURE__ */ jsx144("path", { fill: "currentColor", d: "M0 0h20v20H0z" }) }) })
-      ]
-    }
-  );
-};
-
-// src/icon/icons/text_sync/index.react.tsx
-import { jsx as jsx145, jsxs as jsxs131 } from "react/jsx-runtime";
-var TextSync = (props) => {
-  let _ariaLabel = props["aria-label"];
-  _ariaLabel = _ariaLabel && _ariaLabel.trim();
-  const _ariaHidden = !_ariaLabel;
-  const _role = _ariaHidden ? "presentation" : "img";
-  return /* @__PURE__ */ jsxs131(
-    "svg",
-    {
-      className: props.className,
-      viewBox: "0 0 20 20",
-      "aria-label": _ariaLabel,
-      "aria-hidden": _ariaHidden,
-      role: _role,
-      style: {
-        height: "var(--icon-height)",
-        width: "var(--icon-width)",
-        display: "flex",
-        alignItems: "center",
-        justifyContent: "center"
-      },
-      children: [
-        /* @__PURE__ */ jsxs131("g", { "clip-path": "url(#foa)", children: [
-          /* @__PURE__ */ jsx145("path", { fill: "currentColor", d: "M14.5 20.005c-.582 0-1.16-.097-1.71-.287A5.244 5.244 0 0 1 10 17.48v1.27a.625.625 0 0 1-1.25 0v-3.124A.625.625 0 0 1 9.375 15H12.5a.625.625 0 1 1 0 1.25h-1.723a4.021 4.021 0 0 0 3.722 2.502c.613 0 1.207-.139 1.764-.412a3.993 3.993 0 0 0 2.034-2.308.625.625 0 0 1 1.182.404 5.234 5.234 0 0 1-2.666 3.025 5.178 5.178 0 0 1-2.313.543Zm1.75-6.253a.625.625 0 0 1 0-1.25h1.723A4.021 4.021 0 0 0 14.258 10a4.023 4.023 0 0 0-3.805 2.717.623.623 0 0 1-1.012.279.625.625 0 0 1-.171-.683 5.273 5.273 0 0 1 6.69-3.279 5.244 5.244 0 0 1 2.79 2.238v-1.27a.625.625 0 1 1 1.25-.001v3.125a.625.625 0 0 1-.625.625H16.25Zm-13.125-7.5a.625.625 0 0 1 0-1.25h8.75a.625.625 0 1 1 0 1.25h-8.75Zm0 3.75a.625.625 0 1 1 0-1.25h5a.625.625 0 0 1 0 1.25h-5Zm0 3.75a.625.625 0 1 1 0-1.25H6.25a.625.625 0 0 1 0 1.25H3.125Z" }),
-          /* @__PURE__ */ jsx145("path", { fill: "currentColor", d: "M1.875 17.502A1.876 1.876 0 0 1 0 15.627V1.877A1.877 1.877 0 0 1 1.875.002h8.857c.501 0 .971.195 1.325.548l2.394 2.393c.354.355.549.826.549 1.326v1.983a.625.625 0 1 1-1.25 0V4.269a.632.632 0 0 0-.182-.441l-2.394-2.394a.623.623 0 0 0-.441-.182H1.874a.625.625 0 0 0-.625.625v13.75c0 .345.28.625.625.625H6.25a.625.625 0 0 1 0 1.25H1.875Z" })
-        ] }),
-        /* @__PURE__ */ jsx145("defs", { children: /* @__PURE__ */ jsx145("clipPath", { id: "foa", children: /* @__PURE__ */ jsx145("path", { fill: "currentColor", d: "M0 0h20v20H0z" }) }) })
-      ]
-    }
-  );
-};
-
-// src/icon/icons/text_underline/index.react.tsx
-import { jsx as jsx146, jsxs as jsxs132 } from "react/jsx-runtime";
-var TextUnderline = (props) => {
-  let _ariaLabel = props["aria-label"];
-  _ariaLabel = _ariaLabel && _ariaLabel.trim();
-  const _ariaHidden = !_ariaLabel;
-  const _role = _ariaHidden ? "presentation" : "img";
-  return /* @__PURE__ */ jsxs132(
-    "svg",
-    {
-      className: props.className,
-      viewBox: "0 0 20 20",
-      "aria-label": _ariaLabel,
-      "aria-hidden": _ariaHidden,
-      role: _role,
-      style: {
-        height: "var(--icon-height)",
-        width: "var(--icon-width)",
-        display: "flex",
-        alignItems: "center",
-        justifyContent: "center"
-      },
-      children: [
-        /* @__PURE__ */ jsx146("g", { "clip-path": "url(#fpa)", children: /* @__PURE__ */ jsx146("path", { fill: "currentColor", d: "M.625 19.998a.625.625 0 0 1 0-1.25h18.75a.625.625 0 1 1 0 1.25H.625ZM10 16.25A6.258 6.258 0 0 1 3.75 10V1.248h-.625a.625.625 0 0 1 0-1.25h2.5a.625.625 0 0 1 0 1.25H5V10c0 2.757 2.243 5 5 5s5-2.243 5-5V1.248h-.625a.625.625 0 1 1 0-1.25h2.5a.625.625 0 1 1 0 1.25h-.625V10A6.258 6.258 0 0 1 10 16.25Z" }) }),
-        /* @__PURE__ */ jsx146("defs", { children: /* @__PURE__ */ jsx146("clipPath", { id: "fpa", children: /* @__PURE__ */ jsx146("path", { fill: "currentColor", d: "M0 0h20v20H0z" }) }) })
-      ]
-    }
-  );
-};
-
-// src/icon/icons/transfer/index.react.tsx
-import { jsx as jsx147, jsxs as jsxs133 } from "react/jsx-runtime";
-var Transfer = (props) => {
-  let _ariaLabel = props["aria-label"];
-  _ariaLabel = _ariaLabel && _ariaLabel.trim();
-  const _ariaHidden = !_ariaLabel;
-  const _role = _ariaHidden ? "presentation" : "img";
-  return /* @__PURE__ */ jsxs133(
-    "svg",
-    {
-      className: props.className,
-      viewBox: "0 0 20 20",
-      "aria-label": _ariaLabel,
-      "aria-hidden": _ariaHidden,
-      role: _role,
-      style: {
-        height: "var(--icon-height)",
-        width: "var(--icon-width)",
-        display: "flex",
-        alignItems: "center",
-        justifyContent: "center"
-      },
-      children: [
-        /* @__PURE__ */ jsx147("g", { "clip-path": "url(#fqa)", children: /* @__PURE__ */ jsx147("path", { fill: "currentColor", d: "M16.875 15.625a.626.626 0 0 1-.442-1.068l1.434-1.433H8.75a.625.625 0 0 1 0-1.25h9.116l-1.433-1.433A.618.618 0 0 1 16.25 10a.619.619 0 0 1 .386-.577.625.625 0 0 1 .68.135l2.5 2.5a.617.617 0 0 1 .137.204l.008.022a.608.608 0 0 1 0 .43l-.004.015a.623.623 0 0 1-.14.213l-2.5 2.5a.625.625 0 0 1-.442.183Zm-13.75-5a.621.621 0 0 1-.442-.183l-2.5-2.5A.617.617 0 0 1 0 7.5a.61.61 0 0 1 .04-.218l.006-.017a.617.617 0 0 1 .138-.208l2.5-2.499a.62.62 0 0 1 .883 0A.619.619 0 0 1 3.75 5a.619.619 0 0 1-.183.442L2.133 6.875h9.117a.625.625 0 1 1 0 1.25H2.134l1.433 1.433a.623.623 0 0 1-.442 1.067Z" }) }),
-        /* @__PURE__ */ jsx147("defs", { children: /* @__PURE__ */ jsx147("clipPath", { id: "fqa", children: /* @__PURE__ */ jsx147("path", { fill: "currentColor", d: "M0 0h20v20H0z" }) }) })
-      ]
-    }
-  );
-};
-
-// src/icon/icons/trash/index.react.tsx
-import { jsx as jsx148, jsxs as jsxs134 } from "react/jsx-runtime";
-var Trash = (props) => {
-  let _ariaLabel = props["aria-label"];
-  _ariaLabel = _ariaLabel && _ariaLabel.trim();
-  const _ariaHidden = !_ariaLabel;
-  const _role = _ariaHidden ? "presentation" : "img";
-  return /* @__PURE__ */ jsxs134(
-    "svg",
-    {
-      className: props.className,
-      viewBox: "0 0 20 20",
-      "aria-label": _ariaLabel,
-      "aria-hidden": _ariaHidden,
-      role: _role,
-      style: {
-        height: "var(--icon-height)",
-        width: "var(--icon-width)",
-        display: "flex",
-        alignItems: "center",
-        justifyContent: "center"
-      },
-      children: [
-        /* @__PURE__ */ jsxs134("g", { "clip-path": "url(#fra)", children: [
-          /* @__PURE__ */ jsx148("path", { fill: "currentColor", d: "M5.526 19.375a1.886 1.886 0 0 1-1.868-1.72L2.55 4.375h-1.3a.625.625 0 0 1 0-1.25h5V2.5A1.877 1.877 0 0 1 8.125.625h3.75A1.877 1.877 0 0 1 13.75 2.5v.625h5a.625.625 0 1 1 0 1.25h-1.3l-1.107 13.28a1.885 1.885 0 0 1-1.868 1.72h-8.95Zm-.623-1.823c.026.321.3.573.623.573h8.949a.627.627 0 0 0 .623-.573l1.098-13.177H3.804l1.099 13.177ZM12.5 3.125V2.5a.625.625 0 0 0-.625-.625h-3.75A.625.625 0 0 0 7.5 2.5v.625h5Z" }),
-          /* @__PURE__ */ jsx148("path", { fill: "currentColor", d: "M8.125 15a.625.625 0 0 1-.625-.625v-6.25a.625.625 0 0 1 1.25 0v6.25a.625.625 0 0 1-.625.625Zm3.75 0a.624.624 0 0 1-.625-.625v-6.25a.625.625 0 1 1 1.25 0v6.25a.624.624 0 0 1-.625.625Z" })
-        ] }),
-        /* @__PURE__ */ jsx148("defs", { children: /* @__PURE__ */ jsx148("clipPath", { id: "fra", children: /* @__PURE__ */ jsx148("path", { fill: "currentColor", d: "M0 0h20v20H0z" }) }) })
-      ]
-    }
-  );
-};
-
-// src/icon/icons/undo/index.react.tsx
-import { jsx as jsx149, jsxs as jsxs135 } from "react/jsx-runtime";
-var Undo = (props) => {
-  let _ariaLabel = props["aria-label"];
-  _ariaLabel = _ariaLabel && _ariaLabel.trim();
-  const _ariaHidden = !_ariaLabel;
-  const _role = _ariaHidden ? "presentation" : "img";
-  return /* @__PURE__ */ jsxs135(
-    "svg",
-    {
-      className: props.className,
-      viewBox: "0 0 20 20",
-      "aria-label": _ariaLabel,
-      "aria-hidden": _ariaHidden,
-      role: _role,
-      style: {
-        height: "var(--icon-height)",
-        width: "var(--icon-width)",
-        display: "flex",
-        alignItems: "center",
-        justifyContent: "center"
-      },
-      children: [
-        /* @__PURE__ */ jsx149("g", { "clip-path": "url(#fsa)", children: /* @__PURE__ */ jsx149("path", { fill: "currentColor", d: "M9.997 19.998a.633.633 0 0 1-.625-.625c0-.344.281-.625.625-.625h.006c4.82 0 8.744-3.923 8.747-8.745.003-4.825-3.918-8.752-8.742-8.755a8.715 8.715 0 0 0-7.908 5h4.773a.625.625 0 0 1 0 1.25H.625A.625.625 0 0 1 0 6.873V.623a.625.625 0 1 1 1.25 0v4.551c1.74-3.143 5.1-5.176 8.753-5.176 5.518.004 10 4.492 9.997 10.006-.003 5.511-4.488 9.994-9.997 9.994h-.006Z" }) }),
-        /* @__PURE__ */ jsx149("defs", { children: /* @__PURE__ */ jsx149("clipPath", { id: "fsa", children: /* @__PURE__ */ jsx149("path", { fill: "currentColor", d: "M0 0h20v20H0z" }) }) })
-      ]
-    }
-  );
-};
-
-// src/icon/icons/user/index.react.tsx
-import { jsx as jsx150 } from "react/jsx-runtime";
-var User = (props) => {
-  let _ariaLabel = props["aria-label"];
-  _ariaLabel = _ariaLabel && _ariaLabel.trim();
-  const _ariaHidden = !_ariaLabel;
-  const _role = _ariaHidden ? "presentation" : "img";
-  return /* @__PURE__ */ jsx150(
-    "svg",
-    {
-      className: props.className,
-      viewBox: "0 0 20 20",
-      "aria-label": _ariaLabel,
-      "aria-hidden": _ariaHidden,
-      role: _role,
-      style: {
-        height: "var(--icon-height)",
-        width: "var(--icon-width)",
-        display: "flex",
-        alignItems: "center",
-        justifyContent: "center"
-      },
-      children: /* @__PURE__ */ jsx150("path", { fill: "currentColor", d: "M10 10c-2.757 0-5-2.243-5-5a5 5 0 0 1 .67-2.498l.008-.014A5.024 5.024 0 0 1 10 0c2.757 0 5 2.243 5 5s-2.243 5-5 5ZM6.442 3.822C6.316 4.202 6.25 4.6 6.25 5A3.755 3.755 0 0 0 10 8.75a3.756 3.756 0 0 0 3.74-3.462 7.873 7.873 0 0 1-2.273.337h-.017a7.99 7.99 0 0 1-5.007-1.803Zm.604-1.129a6.702 6.702 0 0 0 4.41 1.682c.74 0 1.465-.122 2.16-.362A3.759 3.759 0 0 0 10 1.25a3.747 3.747 0 0 0-2.954 1.443ZM18.125 20a.624.624 0 0 1-.625-.625c0-4.136-3.364-7.5-7.5-7.5s-7.5 3.364-7.5 7.5a.625.625 0 1 1-1.25 0c0-4.825 3.925-8.75 8.75-8.75s8.75 3.925 8.75 8.75a.624.624 0 0 1-.625.625Z" })
-    }
-  );
-};
-
-// src/icon/icons/user_edit/index.react.tsx
-import { jsx as jsx151, jsxs as jsxs136 } from "react/jsx-runtime";
-var UserEdit = (props) => {
-  let _ariaLabel = props["aria-label"];
-  _ariaLabel = _ariaLabel && _ariaLabel.trim();
-  const _ariaHidden = !_ariaLabel;
-  const _role = _ariaHidden ? "presentation" : "img";
-  return /* @__PURE__ */ jsxs136(
-    "svg",
-    {
-      className: props.className,
-      viewBox: "0 0 20 20",
-      "aria-label": _ariaLabel,
-      "aria-hidden": _ariaHidden,
-      role: _role,
-      style: {
-        height: "var(--icon-height)",
-        width: "var(--icon-width)",
-        display: "flex",
-        alignItems: "center",
-        justifyContent: "center"
-      },
-      children: [
-        /* @__PURE__ */ jsxs136("g", { "clip-path": "url(#fua)", children: [
-          /* @__PURE__ */ jsx151("path", { fill: "currentColor", d: "M6.24 8.125a4.067 4.067 0 0 1-4.062-4.063c0-.699.185-1.39.533-2.003.014-.029.029-.05.038-.064A4.082 4.082 0 0 1 6.241 0c1.99 0 3.71 1.482 4.011 3.448a.583.583 0 0 1 .019.128c.021.18.032.336.032.486a4.067 4.067 0 0 1-4.062 4.063Zm-2.724-4.76a2.816 2.816 0 0 0 2.725 3.51 2.817 2.817 0 0 0 2.782-2.402 6.272 6.272 0 0 1-1.632.215h-.006a6.324 6.324 0 0 1-3.87-1.324Zm3.894.073c.505 0 1.014-.079 1.508-.236A2.83 2.83 0 0 0 6.24 1.25c-.823 0-1.602.364-2.134.987a5.14 5.14 0 0 0 3.27 1.2h.01l.023.001ZM9.366 20a.619.619 0 0 1-.442-.183.62.62 0 0 1-.17-.564l.624-3.125a.623.623 0 0 1 .171-.32l6.359-6.36c.45-.45 1.05-.698 1.688-.698.638 0 1.238.248 1.688.698l.014.014c.931.915.96 2.411.063 3.34a.748.748 0 0 1-.17.131l-6.258 6.259a.62.62 0 0 1-.32.17l-3.125.626a.614.614 0 0 1-.122.012Zm.797-1.422 2.02-.404 6.224-6.225a.624.624 0 0 1 .127-.098c.359-.45.311-1.086-.12-1.503a1.106 1.106 0 0 0-.817-.348c-.304 0-.591.118-.806.333l-6.224 6.225-.404 2.02Z" }),
-          /* @__PURE__ */ jsx151("path", { fill: "currentColor", d: "M.616 15a.626.626 0 0 1-.625-.623 6.238 6.238 0 0 1 6.224-6.255c.564 0 1.13.077 1.678.23a6.188 6.188 0 0 1 3.757 2.94.623.623 0 0 1-.543.932.628.628 0 0 1-.545-.317 4.996 4.996 0 0 0-4.348-2.531c-.852 0-1.696.222-2.441.643a4.989 4.989 0 0 0-2.533 4.354.624.624 0 0 1-.622.627H.616Z" })
-        ] }),
-        /* @__PURE__ */ jsx151("defs", { children: /* @__PURE__ */ jsx151("clipPath", { id: "fua", children: /* @__PURE__ */ jsx151("path", { fill: "currentColor", d: "M0 0h20v20H0z" }) }) })
-      ]
-    }
-  );
-};
-
-// src/icon/icons/user_question/index.react.tsx
-import { jsx as jsx152, jsxs as jsxs137 } from "react/jsx-runtime";
-var UserQuestion = (props) => {
-  let _ariaLabel = props["aria-label"];
-  _ariaLabel = _ariaLabel && _ariaLabel.trim();
-  const _ariaHidden = !_ariaLabel;
-  const _role = _ariaHidden ? "presentation" : "img";
-  return /* @__PURE__ */ jsxs137(
-    "svg",
-    {
-      className: props.className,
-      viewBox: "0 0 20 20",
-      "aria-label": _ariaLabel,
-      "aria-hidden": _ariaHidden,
-      role: _role,
-      style: {
-        height: "var(--icon-height)",
-        width: "var(--icon-width)",
-        display: "flex",
-        alignItems: "center",
-        justifyContent: "center"
-      },
-      children: [
-        /* @__PURE__ */ jsxs137("g", { "clip-path": "url(#fva)", children: [
-          /* @__PURE__ */ jsx152("path", { fill: "currentColor", d: "M12.918 20a.625.625 0 0 1-.625-.625v-3.116a.623.623 0 0 1 .727-.617l.84.14a1.863 1.863 0 0 0 1.398-.323 1.863 1.863 0 0 0 .785-1.527v-2.048a.625.625 0 0 1 .474-.607l.779-.194a.622.622 0 0 0 .368-.954l-1.516-2.274a.622.622 0 0 1-.105-.347v-.625c0-1.417-.63-2.772-1.775-3.815-1.286-1.171-3.041-1.816-4.941-1.816-.26 0-.523.012-.783.036a6.878 6.878 0 0 0-6.222 7.464A6.879 6.879 0 0 0 5.73 14.08a.626.626 0 0 1 .313.54v4.75a.625.625 0 0 1-1.25 0v-4.398a8.153 8.153 0 0 1-3.716-6.108 8.062 8.062 0 0 1 1.851-5.93 8.066 8.066 0 0 1 5.5-2.89 9.27 9.27 0 0 1 .897-.043c2.213 0 4.267.76 5.783 2.143 1.408 1.283 2.184 2.965 2.184 4.74v.435l1.41 2.117c.12.179.207.376.259.586a1.863 1.863 0 0 1-.211 1.419 1.858 1.858 0 0 1-1.153.854l-.306.077v1.56a3.125 3.125 0 0 1-3.64 3.083l-.111-.018v2.378a.62.62 0 0 1-.622.624Z" }),
-          /* @__PURE__ */ jsx152("path", { fill: "currentColor", d: "M9.644 11.25a.625.625 0 0 1-.625-.625v-.213a1.876 1.876 0 0 1 1.25-1.769 1.878 1.878 0 0 0 1.25-1.767A1.877 1.877 0 0 0 9.644 5 1.878 1.878 0 0 0 7.77 6.876a.625.625 0 0 1-1.25-.001 3.128 3.128 0 0 1 3.125-3.124 3.13 3.13 0 0 1 3.125 3.126 3.13 3.13 0 0 1-2.083 2.946.627.627 0 0 0-.417.589v.214a.625.625 0 0 1-.625.624Zm0 3.125a.938.938 0 1 0 0-1.875.938.938 0 0 0 0 1.875Z" })
-        ] }),
-        /* @__PURE__ */ jsx152("defs", { children: /* @__PURE__ */ jsx152("clipPath", { id: "fva", children: /* @__PURE__ */ jsx152("path", { fill: "currentColor", d: "M0 0h20v20H0z" }) }) })
-      ]
-    }
-  );
-};
-
-// src/icon/icons/users/index.react.tsx
-import { jsx as jsx153, jsxs as jsxs138 } from "react/jsx-runtime";
-var Users = (props) => {
-  let _ariaLabel = props["aria-label"];
-  _ariaLabel = _ariaLabel && _ariaLabel.trim();
-  const _ariaHidden = !_ariaLabel;
-  const _role = _ariaHidden ? "presentation" : "img";
-  return /* @__PURE__ */ jsxs138(
-    "svg",
-    {
-      className: props.className,
-      viewBox: "0 0 20 20",
-      "aria-label": _ariaLabel,
-      "aria-hidden": _ariaHidden,
-      role: _role,
-      style: {
-        height: "var(--icon-height)",
-        width: "var(--icon-width)",
-        display: "flex",
-        alignItems: "center",
-        justifyContent: "center"
-      },
-      children: [
-        /* @__PURE__ */ jsx153("path", { fill: "currentColor", d: "M19.375 17.5a.625.625 0 0 1-.625-.625c0-.37-.042-.74-.123-1.1a4.974 4.974 0 0 0-2.212-3.13 4.985 4.985 0 0 0-3.776-.647.623.623 0 0 1-.667-.943.624.624 0 0 1 .391-.275 6.234 6.234 0 0 1 4.72.81 6.215 6.215 0 0 1 2.764 3.908c.1.45.153.914.153 1.375 0 .344-.28.625-.625.626Z" }),
-        /* @__PURE__ */ jsx153("path", { fill: "currentColor", d: "M13.747 10.623c-.502 0-1-.093-1.467-.275a.617.617 0 0 1-.346-.33.619.619 0 0 1 .32-.822.63.63 0 0 1 .479-.013 2.809 2.809 0 0 0 3.794-2.211 6.3 6.3 0 0 1-4.986-.75.615.615 0 0 1-.277-.389.622.622 0 0 1 .943-.668 5.044 5.044 0 0 0 2.681.772h.027a5.04 5.04 0 0 0 1.51-.235 2.795 2.795 0 0 0-2.09-1.892 2.798 2.798 0 0 0-2.118.392.625.625 0 1 1-.68-1.048 4.04 4.04 0 0 1 3.057-.567 4.036 4.036 0 0 1 2.563 1.76c.313.483.516 1.019.604 1.594a.61.61 0 0 1 .02.155 4.04 4.04 0 0 1-1.818 3.871 4.04 4.04 0 0 1-2.216.655ZM11.875 17.5a.625.625 0 0 1-.625-.626c0-2.756-2.243-5-5-5s-5 2.243-5 5a.625.625 0 1 1-1.25 0 6.258 6.258 0 0 1 6.25-6.25 6.258 6.258 0 0 1 6.25 6.25.625.625 0 0 1-.625.625Z" }),
-        /* @__PURE__ */ jsx153("path", { fill: "currentColor", d: "M10.402 10.838a.632.632 0 0 1-.534-.299 5.493 5.493 0 0 1-.575-1.293 4.028 4.028 0 0 1-3.044 1.377 4.025 4.025 0 0 1-3.044-1.379 5.505 5.505 0 0 1-.574 1.289.622.622 0 0 1-1.04.042.623.623 0 0 1-.028-.693 4.28 4.28 0 0 0 .625-2.38l.01-.695v-.02a2.822 2.822 0 0 1-.01-.225A4.067 4.067 0 0 1 6.25 2.499a4.067 4.067 0 0 1 4.06 4.124v.864a4.31 4.31 0 0 0 .626 2.401.62.62 0 0 1-.208.86.632.632 0 0 1-.325.09Zm-6.835-3.44a2.81 2.81 0 0 0 5.365.004 5.197 5.197 0 0 1-2.683-1.26 5.193 5.193 0 0 1-2.682 1.256Zm2.683-2.77c.182 0 .354.08.473.217A3.916 3.916 0 0 0 9.03 6.152a2.827 2.827 0 0 0-2.78-2.404 2.826 2.826 0 0 0-2.78 2.4 3.917 3.917 0 0 0 2.307-1.304.622.622 0 0 1 .473-.216Z" })
-      ]
-    }
-  );
-};
-
-// src/icon/icons/video_file_add/index.react.tsx
-import { jsx as jsx154, jsxs as jsxs139 } from "react/jsx-runtime";
-var VideoFileAdd = (props) => {
-  let _ariaLabel = props["aria-label"];
-  _ariaLabel = _ariaLabel && _ariaLabel.trim();
-  const _ariaHidden = !_ariaLabel;
-  const _role = _ariaHidden ? "presentation" : "img";
-  return /* @__PURE__ */ jsxs139(
-    "svg",
-    {
-      className: props.className,
-      viewBox: "0 0 20 20",
-      "aria-label": _ariaLabel,
-      "aria-hidden": _ariaHidden,
-      role: _role,
-      style: {
-        height: "var(--icon-height)",
-        width: "var(--icon-width)",
-        display: "flex",
-        alignItems: "center",
-        justifyContent: "center"
-      },
-      children: [
-        /* @__PURE__ */ jsxs139("g", { "clip-path": "url(#fxa)", children: [
-          /* @__PURE__ */ jsx154("path", { fill: "currentColor", d: "M14.375 20a5.631 5.631 0 0 1-5.625-5.625 5.631 5.631 0 0 1 5.625-5.625A5.631 5.631 0 0 1 20 14.375 5.631 5.631 0 0 1 14.375 20Zm0-10A4.38 4.38 0 0 0 10 14.375a4.38 4.38 0 0 0 4.375 4.375 4.38 4.38 0 0 0 4.375-4.375A4.38 4.38 0 0 0 14.375 10Z" }),
-          /* @__PURE__ */ jsx154("path", { fill: "currentColor", d: "M14.375 17.5a.624.624 0 0 1-.625-.625V15h-1.875a.624.624 0 1 1 0-1.25h1.875v-1.875a.624.624 0 1 1 1.25 0v1.875h1.875a.624.624 0 1 1 0 1.25H15v1.875a.624.624 0 0 1-.625.625Zm-12.5 0A1.876 1.876 0 0 1 0 15.625V1.875A1.877 1.877 0 0 1 1.875 0h8.857c.501 0 .971.195 1.325.548l2.394 2.394c.354.354.549.825.549 1.325v2.608a.625.625 0 1 1-1.25 0V4.268a.632.632 0 0 0-.182-.442l-2.394-2.393a.623.623 0 0 0-.441-.183H1.874a.625.625 0 0 0-.625.625v13.75c0 .345.28.625.625.625h5a.625.625 0 1 1 0 1.25h-5Z" }),
-          /* @__PURE__ */ jsx154("path", { fill: "currentColor", d: "M4.947 11.25a1.205 1.205 0 0 1-.971-.496 1.197 1.197 0 0 1-.226-.7V6.196a1.19 1.19 0 0 1 1.202-1.193c.185 0 .367.042.528.123l3.86 1.93c.284.142.497.387.598.69a1.198 1.198 0 0 1-.6 1.449l-3.858 1.93a1.19 1.19 0 0 1-.533.126ZM5 9.968l3.683-1.842L5 6.283v3.684Z" })
-        ] }),
-        /* @__PURE__ */ jsx154("defs", { children: /* @__PURE__ */ jsx154("clipPath", { id: "fxa", children: /* @__PURE__ */ jsx154("path", { fill: "currentColor", d: "M0 0h20v20H0z" }) }) })
-      ]
-    }
-  );
-};
-
-// src/icon/icons/video_player/index.react.tsx
-import { jsx as jsx155, jsxs as jsxs140 } from "react/jsx-runtime";
-var VideoPlayer = (props) => {
-  let _ariaLabel = props["aria-label"];
-  _ariaLabel = _ariaLabel && _ariaLabel.trim();
-  const _ariaHidden = !_ariaLabel;
-  const _role = _ariaHidden ? "presentation" : "img";
-  return /* @__PURE__ */ jsxs140(
-    "svg",
-    {
-      className: props.className,
-      viewBox: "0 0 20 20",
-      "aria-label": _ariaLabel,
-      "aria-hidden": _ariaHidden,
-      role: _role,
-      style: {
-        height: "var(--icon-height)",
-        width: "var(--icon-width)",
-        display: "flex",
-        alignItems: "center",
-        justifyContent: "center"
-      },
-      children: [
-        /* @__PURE__ */ jsxs140("g", { "clip-path": "url(#fya)", children: [
-          /* @__PURE__ */ jsx155("path", { fill: "currentColor", d: "M1.875 20A1.876 1.876 0 0 1 0 18.125V1.875A1.877 1.877 0 0 1 1.875 0h16.25A1.877 1.877 0 0 1 20 1.875v16.25A1.876 1.876 0 0 1 18.125 20H1.875Zm0-18.75a.625.625 0 0 0-.625.625v16.25c0 .345.28.625.625.625h16.25a.624.624 0 0 0 .625-.625V1.875a.625.625 0 0 0-.625-.625H1.875Z" }),
-          /* @__PURE__ */ jsx155("path", { fill: "currentColor", d: "M8.214 13.75c-.144 0-.287-.023-.426-.07a1.325 1.325 0 0 1-.772-.672 1.347 1.347 0 0 1-.141-.596V7.588A1.34 1.34 0 0 1 8.81 6.391l4.825 2.413a1.339 1.339 0 0 1 0 2.393L8.812 13.61c-.186.092-.39.14-.598.14Zm0-6.25a.088.088 0 0 0-.089.088v4.824a.09.09 0 0 0 .06.084.093.093 0 0 0 .067-.005l4.824-2.413a.087.087 0 0 0 .048-.078.087.087 0 0 0-.048-.078L8.25 7.509a.078.078 0 0 0-.038-.009Z" })
-        ] }),
-        /* @__PURE__ */ jsx155("defs", { children: /* @__PURE__ */ jsx155("clipPath", { id: "fya", children: /* @__PURE__ */ jsx155("path", { fill: "currentColor", d: "M0 0h20v20H0z" }) }) })
-      ]
-    }
-  );
-};
-
-// src/icon/icons/view/index.react.tsx
-import { jsx as jsx156, jsxs as jsxs141 } from "react/jsx-runtime";
-var View = (props) => {
-  let _ariaLabel = props["aria-label"];
-  _ariaLabel = _ariaLabel && _ariaLabel.trim();
-  const _ariaHidden = !_ariaLabel;
-  const _role = _ariaHidden ? "presentation" : "img";
-  return /* @__PURE__ */ jsxs141(
-    "svg",
-    {
-      className: props.className,
-      viewBox: "0 0 20 20",
-      "aria-label": _ariaLabel,
-      "aria-hidden": _ariaHidden,
-      role: _role,
-      style: {
-        height: "var(--icon-height)",
-        width: "var(--icon-width)",
-        display: "flex",
-        alignItems: "center",
-        justifyContent: "center"
-      },
-      children: [
-        /* @__PURE__ */ jsxs141("g", { "clip-path": "url(#fza)", children: [
-          /* @__PURE__ */ jsx156("path", { fill: "currentColor", d: "M9.868 16.25c-3.61 0-7.196-2.534-9.349-4.904a2.007 2.007 0 0 1-.002-2.694C2.671 6.282 6.261 3.75 9.88 3.75h.253c3.607 0 7.195 2.532 9.35 4.9.692.77.693 1.928.001 2.695-2.155 2.372-5.747 4.905-9.37 4.905h-.245ZM9.878 5c-3.21 0-6.461 2.32-8.434 4.49a.758.758 0 0 0 .002 1.018c1.971 2.17 5.222 4.491 8.433 4.491h.222c3.223 0 6.48-2.32 8.455-4.493a.758.758 0 0 0-.002-1.017C16.581 7.32 13.327 5 10.118 5h-.24Z" }),
-          /* @__PURE__ */ jsx156("path", { fill: "currentColor", d: "M9.998 13.748a3.726 3.726 0 0 1-2.651-1.099 3.721 3.721 0 0 1-1.098-2.652A3.754 3.754 0 0 1 10 6.248c1.002 0 1.944.39 2.652 1.098a3.723 3.723 0 0 1 1.098 2.65 3.726 3.726 0 0 1-1.099 2.655 3.727 3.727 0 0 1-2.652 1.097ZM10 7.498a2.503 2.503 0 0 0-2.5 2.5c0 .667.26 1.295.732 1.767a2.484 2.484 0 0 0 1.767.732c.668 0 1.296-.26 1.768-.731.471-.472.732-1.1.732-1.768 0-.67-.26-1.297-.732-1.769A2.485 2.485 0 0 0 10 7.5Z" })
-        ] }),
-        /* @__PURE__ */ jsx156("defs", { children: /* @__PURE__ */ jsx156("clipPath", { id: "fza", children: /* @__PURE__ */ jsx156("path", { fill: "currentColor", d: "M0 0h20v20H0z" }) }) })
-      ]
-    }
-  );
-};
-
-// src/icon/icons/view_off/index.react.tsx
-import { jsx as jsx157, jsxs as jsxs142 } from "react/jsx-runtime";
-var ViewOff = (props) => {
-  let _ariaLabel = props["aria-label"];
-  _ariaLabel = _ariaLabel && _ariaLabel.trim();
-  const _ariaHidden = !_ariaLabel;
-  const _role = _ariaHidden ? "presentation" : "img";
-  return /* @__PURE__ */ jsxs142(
-    "svg",
-    {
-      className: props.className,
-      viewBox: "0 0 20 20",
-      "aria-label": _ariaLabel,
-      "aria-hidden": _ariaHidden,
-      role: _role,
-      style: {
-        height: "var(--icon-height)",
-        width: "var(--icon-width)",
-        display: "flex",
-        alignItems: "center",
-        justifyContent: "center"
-      },
-      children: [
-        /* @__PURE__ */ jsxs142("g", { "clip-path": "url(#gaa)", children: [
-          /* @__PURE__ */ jsx157("path", { fill: "currentColor", d: "M2.32 18.125a.63.63 0 0 1-.583-.397.62.62 0 0 1 .15-.679l15.625-15a.62.62 0 0 1 .883.019.623.623 0 0 1-.017.884l-15.626 15a.628.628 0 0 1-.433.173Zm7.67-1.25h-.092a8.837 8.837 0 0 1-2.614-.402.626.626 0 0 1 .371-1.193c.74.23 1.5.347 2.257.347h.208c3.21 0 6.463-2.321 8.438-4.494a.758.758 0 0 0-.002-1.016 16.485 16.485 0 0 0-2.466-2.211.619.619 0 0 1-.197-.767.627.627 0 0 1 .934-.243 17.79 17.79 0 0 1 2.655 2.382c.694.769.694 1.926.003 2.694-2.155 2.371-5.746 4.905-9.364 4.905l-.131-.002Zm-6.842-2.699a.625.625 0 0 1-.383-.132 18.021 18.021 0 0 1-2.244-2.071 2.008 2.008 0 0 1-.003-2.695c2.155-2.371 5.747-4.904 9.364-4.904l.197.001c.705 0 1.406.085 2.084.254a.625.625 0 1 1-.303 1.214 7.34 7.34 0 0 0-1.774-.217h-.19c-3.22 0-6.477 2.32-8.45 4.492a.758.758 0 0 0 0 1.017c.64.7 1.338 1.344 2.087 1.925a.622.622 0 0 1-.385 1.116Z" }),
-          /* @__PURE__ */ jsx157("path", { fill: "currentColor", d: "M6.875 11.25a.625.625 0 0 1-.625-.625c0-1 .39-1.94 1.098-2.65a3.72 3.72 0 0 1 2.65-1.1c.172 0 .311.14.312.312l.002.625a.315.315 0 0 1-.091.221.309.309 0 0 1-.221.092 2.483 2.483 0 0 0-1.768.733 2.483 2.483 0 0 0-.732 1.767.623.623 0 0 1-.625.625ZM10 14.375a.311.311 0 0 1-.312-.313v-.624a.31.31 0 0 1 .312-.313 2.483 2.483 0 0 0 1.768-.732 2.488 2.488 0 0 0 .732-1.769.625.625 0 1 1 1.25 0 3.726 3.726 0 0 1-1.098 2.652 3.726 3.726 0 0 1-2.651 1.099Z" })
-        ] }),
-        /* @__PURE__ */ jsx157("defs", { children: /* @__PURE__ */ jsx157("clipPath", { id: "gaa", children: /* @__PURE__ */ jsx157("path", { fill: "currentColor", d: "M0 0h20v20H0z" }) }) })
-      ]
-    }
-  );
-};
-
-// src/icon/icons/volume_off/index.react.tsx
-import { jsx as jsx158 } from "react/jsx-runtime";
-var VolumeOff = (props) => {
-  let _ariaLabel = props["aria-label"];
-  _ariaLabel = _ariaLabel && _ariaLabel.trim();
-  const _ariaHidden = !_ariaLabel;
-  const _role = _ariaHidden ? "presentation" : "img";
-  return /* @__PURE__ */ jsx158(
-    "svg",
-    {
-      className: props.className,
-      viewBox: "0 0 20 20",
-      "aria-label": _ariaLabel,
-      "aria-hidden": _ariaHidden,
-      role: _role,
-      style: {
-        height: "var(--icon-height)",
-        width: "var(--icon-width)",
-        display: "flex",
-        alignItems: "center",
-        justifyContent: "center"
-      },
-      children: /* @__PURE__ */ jsx158("path", { fill: "currentColor", d: "M12.497 16.563a1.86 1.86 0 0 1-1.101-.36l-2.719-1.73a.625.625 0 0 1 .671-1.054l2.737 1.742.034.023a.621.621 0 0 0 .88-.12.628.628 0 0 0 .124-.376v-4.375a.625.625 0 1 1 1.25 0v4.375a1.887 1.887 0 0 1-1.036 1.679c-.261.13-.549.197-.84.195Zm-8.747-3.75a1.877 1.877 0 0 1-1.875-1.876v-2.5A1.877 1.877 0 0 1 3.75 6.562h2.318l5.328-3.39a1.876 1.876 0 0 1 2.98 1.515.625.625 0 1 1-1.25 0 .625.625 0 0 0-1-.5.675.675 0 0 1-.037.026L6.586 7.715a.623.623 0 0 1-.336.097h-2.5a.625.625 0 0 0-.625.625v2.5c0 .346.28.626.625.626h.521a.625.625 0 1 1 0 1.25h-.52ZM2.5 17.188a.622.622 0 0 1-.597-.44.622.622 0 0 1 .197-.666l15-12.5a.625.625 0 0 1 .8.96l-15 12.5a.628.628 0 0 1-.4.145Z" })
-    }
-  );
-};
-
-// src/icon/icons/volume_on/index.react.tsx
-import { jsx as jsx159 } from "react/jsx-runtime";
-var VolumeOn = (props) => {
-  let _ariaLabel = props["aria-label"];
-  _ariaLabel = _ariaLabel && _ariaLabel.trim();
-  const _ariaHidden = !_ariaLabel;
-  const _role = _ariaHidden ? "presentation" : "img";
-  return /* @__PURE__ */ jsx159(
-    "svg",
-    {
-      className: props.className,
-      viewBox: "0 0 20 20",
-      "aria-label": _ariaLabel,
-      "aria-hidden": _ariaHidden,
-      role: _role,
-      style: {
-        height: "var(--icon-height)",
-        width: "var(--icon-width)",
-        display: "flex",
-        alignItems: "center",
-        justifyContent: "center"
-      },
-      children: /* @__PURE__ */ jsx159("path", { fill: "currentColor", d: "M11.242 16.87a1.85 1.85 0 0 1-1.101-.358l-5.329-3.39H2.494A1.877 1.877 0 0 1 .62 11.247v-2.5a1.877 1.877 0 0 1 1.875-1.875h2.318L10.14 3.48a1.854 1.854 0 0 1 1.942-.161 1.864 1.864 0 0 1 1.036 1.677v10c0 .403-.133.802-.375 1.125a1.882 1.882 0 0 1-1.5.749Zm-.412-1.4.034.022a.622.622 0 0 0 .822-.054.628.628 0 0 0 .184-.442v-10a.623.623 0 0 0-1-.5.9.9 0 0 1-.036.026L5.62 7.84v4.314l5.211 3.315ZM2.494 8.121a.625.625 0 0 0-.625.625v2.5c0 .345.28.625.625.625H4.37v-3.75H2.494Zm15.386 5.626a.621.621 0 0 1-.616-.734.619.619 0 0 1 .09-.227 4.89 4.89 0 0 0 0-5.244.626.626 0 0 1 1.053-.674 6.147 6.147 0 0 1 0 6.59.623.623 0 0 1-.527.289Zm-2.417-1.091a.626.626 0 0 1-.486-1.02 2.296 2.296 0 0 0 0-2.882.625.625 0 0 1 .971-.788 3.553 3.553 0 0 1 0 4.459.625.625 0 0 1-.486.23Z" })
-    }
-  );
-};
-
-// src/icon/icons/warning/index.react.tsx
-import { jsx as jsx160, jsxs as jsxs143 } from "react/jsx-runtime";
-var Warning = (props) => {
-  let _ariaLabel = props["aria-label"];
-  _ariaLabel = _ariaLabel && _ariaLabel.trim();
-  const _ariaHidden = !_ariaLabel;
-  const _role = _ariaHidden ? "presentation" : "img";
-  return /* @__PURE__ */ jsxs143(
-    "svg",
-    {
-      className: props.className,
-      viewBox: "0 0 20 20",
-      "aria-label": _ariaLabel,
-      "aria-hidden": _ariaHidden,
-      role: _role,
-      style: {
-        height: "var(--icon-height)",
-        width: "var(--icon-width)",
-        display: "flex",
-        alignItems: "center",
-        justifyContent: "center"
-      },
-      children: [
-        /* @__PURE__ */ jsxs143("g", { "clip-path": "url(#gda)", children: [
-          /* @__PURE__ */ jsx160("path", { fill: "currentColor", d: "M10 15a.937.937 0 1 0 0-1.875A.937.937 0 0 0 10 15Zm0-3.125a.625.625 0 0 1-.625-.625V4.375a.625.625 0 0 1 1.25 0v6.875a.624.624 0 0 1-.625.625Z" }),
-          /* @__PURE__ */ jsx160("path", { fill: "currentColor", d: "M10 20C4.486 20 0 15.514 0 10S4.486 0 10 0s10 4.486 10 10-4.486 10-10 10Zm0-18.75c-4.825 0-8.75 3.925-8.75 8.75s3.925 8.75 8.75 8.75 8.75-3.925 8.75-8.75S14.825 1.25 10 1.25Z" })
-        ] }),
-        /* @__PURE__ */ jsx160("defs", { children: /* @__PURE__ */ jsx160("clipPath", { id: "gda", children: /* @__PURE__ */ jsx160("path", { fill: "currentColor", d: "M0 0h20v20H0z" }) }) })
-      ]
-    }
-  );
-};
-
-// src/icon/icons/youtube/index.react.tsx
-import { jsx as jsx161, jsxs as jsxs144 } from "react/jsx-runtime";
-var Youtube = (props) => {
-  let _ariaLabel = props["aria-label"];
-  _ariaLabel = _ariaLabel && _ariaLabel.trim();
-  const _ariaHidden = !_ariaLabel;
-  const _role = _ariaHidden ? "presentation" : "img";
-  return /* @__PURE__ */ jsxs144(
-    "svg",
-    {
-      className: props.className,
-      viewBox: "0 0 20 20",
-      "aria-label": _ariaLabel,
-      "aria-hidden": _ariaHidden,
-      role: _role,
-      style: {
-        height: "var(--icon-height)",
-        width: "var(--icon-width)",
-        display: "flex",
-        alignItems: "center",
-        justifyContent: "center"
-      },
-      children: [
-        /* @__PURE__ */ jsxs144("g", { "clip-path": "url(#gea)", children: [
-          /* @__PURE__ */ jsx161("path", { fill: "currentColor", d: "M16.508 2.708H3.492A3.492 3.492 0 0 0 0 6.2v7.6a3.492 3.492 0 0 0 3.492 3.492h13.016A3.493 3.493 0 0 0 20 13.8V6.2a3.491 3.491 0 0 0-3.492-3.492ZM18.75 13.8a2.242 2.242 0 0 1-2.242 2.242H3.492A2.242 2.242 0 0 1 1.25 13.8V6.2a2.242 2.242 0 0 1 2.242-2.242h13.016A2.242 2.242 0 0 1 18.75 6.2v7.6Z" }),
-          /* @__PURE__ */ jsx161("path", { fill: "currentColor", d: "M8.25 5.542a.625.625 0 0 0-.958.525v7.216a.625.625 0 0 0 .958.525l5.642-3.608a.626.626 0 0 0 0-1.05L8.25 5.542Zm.292 6.6V7.208l3.858 2.5-3.858 2.434Z" })
-        ] }),
-        /* @__PURE__ */ jsx161("defs", { children: /* @__PURE__ */ jsx161("clipPath", { id: "gea", children: /* @__PURE__ */ jsx161("path", { fill: "currentColor", d: "M0 0h20v20H0z" }) }) })
-      ]
-    }
-  );
-};
-export {
-  Add,
-  AddFigure,
-  AddSquare,
-  AlarmBell,
-  AlarmClock,
-  AlarmTimer,
-  AnalyticBars,
-  ArrowBack,
-  ArrowDown,
-  ArrowLeft,
-  ArrowRight,
-  ArrowUp,
-  Assets,
-  AudioRecording,
-  Backspace,
-  Banned,
-  Binocular,
-  BookFlipPage,
-  BookOpen,
-  Bookmark,
-  BulletList,
-  Calendar,
-  Camera,
-  Chat,
-  Check,
-  CheckCircle,
-  Checklist,
-  ChevronDown,
-  ChevronUp,
-  Circle,
-  CircleShapeAdd,
-  Clock,
-  ClockHand,
-  Close,
-  ColorBrush,
-  CopyPaste,
-  CursorPointer,
-  CursorPointerCircle,
-  Dice,
-  Dictionary,
-  Download,
-  DrawTablet,
-  Elements,
-  Envelope,
-  Eraser,
-  Expand,
-  ExpandResize,
-  ExternalLink,
-  ExternalLinkArrow,
-  FemaleHeadset,
-  FlipRight,
-  FolderBookmark,
-  FontExpand,
-  FormTemplate,
-  FrameExpand,
-  FrameShrink,
-  GameController,
-  Glasses,
-  GraphStats,
-  Headphones,
-  Heart,
-  Home,
-  Hyperlink,
-  Image,
-  ImageFlower,
-  IndentDecrease,
-  IndentIncrease,
-  Information,
-  Keyboard,
-  Layers,
-  LayersBack,
-  LayoutLeft,
-  LayoutModule,
-  LayoutRight,
-  Library,
-  LightBulb,
-  LockLocked,
-  LockUnlocked,
-  Login,
-  Logout,
-  Magnet,
-  Menu,
-  Message,
-  MessageQuestion,
-  MessageStar,
-  MessageTyping,
-  MessageWarning,
-  ModuleEdit,
-  ModuleFour,
-  MultipleUsers,
-  Newspaper,
-  NotebookPencil,
-  NumberList,
-  Option,
-  Palette,
-  Paragraph,
-  ParagraphCenter,
-  ParagraphLeft,
-  ParagraphRight,
-  PartyBalloon,
-  Pause,
-  PenTools,
-  PenWrite,
-  Pencil,
-  Pin,
-  PinRemove,
-  Play,
-  Printer,
-  QuestionCircle,
-  RankingStar,
-  RatingStarCheck,
-  Read,
-  ReadHome,
-  ReadIdea,
-  Redo,
-  Refresh,
-  Reload,
-  Remove,
-  RetouchGraph,
-  RibbonStar,
-  Rocket,
-  Search,
-  Settings,
-  SettingsSlider,
-  Share,
-  Shrink,
-  Shuffle,
-  SocialInstagram,
-  StudyMath,
-  Subscript,
-  Subtract,
-  Superscript,
-  SynchronizeArrowClock,
-  SynchronizeArrowsSquare,
-  SynchronizeArrowsWarning,
-  TabletTouch,
-  TaskList,
-  TaskListClock,
-  Teacher,
-  TeacherCorrect,
-  TextBold,
-  TextFormat,
-  TextItalic,
-  TextStyle,
-  TextSync,
-  TextUnderline,
-  Transfer,
-  Trash,
-  Undo,
-  User,
-  UserEdit,
-  UserQuestion,
-  Users,
-  VideoFileAdd,
-  VideoPlayer,
-  View,
-  ViewOff,
-  VolumeOff,
-  VolumeOn,
-  Warning,
-  Youtube
-};
