+++ packages/kobber-icons/src/icon/types/kobber-icons-types.ts
@@ -161,167 +161,3 @@ export type IconType =
   | "kobber-warning"
   | "kobber-youtube"
 ; 
-
-export const iconsList = [
- "kobber-add", 
- "kobber-add_figure", 
- "kobber-add_square", 
- "kobber-alarm_bell", 
- "kobber-alarm_clock", 
- "kobber-alarm_timer", 
- "kobber-analytic_bars", 
- "kobber-arrow_back", 
- "kobber-arrow_down", 
- "kobber-arrow_left", 
- "kobber-arrow_right", 
- "kobber-arrow_up", 
- "kobber-assets", 
- "kobber-audio_recording", 
- "kobber-backspace", 
- "kobber-banned", 
- "kobber-binocular", 
- "kobber-book_flip_page", 
- "kobber-book_open", 
- "kobber-bookmark", 
- "kobber-bullet_list", 
- "kobber-calendar", 
- "kobber-camera", 
- "kobber-chat", 
- "kobber-check", 
- "kobber-check_circle", 
- "kobber-checklist", 
- "kobber-chevron_down", 
- "kobber-chevron_up", 
- "kobber-circle", 
- "kobber-circle_shape_add", 
- "kobber-clock", 
- "kobber-clock_hand", 
- "kobber-close", 
- "kobber-color_brush", 
- "kobber-copy_paste", 
- "kobber-cursor_pointer", 
- "kobber-cursor_pointer_circle", 
- "kobber-dice", 
- "kobber-dictionary", 
- "kobber-download", 
- "kobber-draw_tablet", 
- "kobber-elements", 
- "kobber-envelope", 
- "kobber-eraser", 
- "kobber-expand", 
- "kobber-expand_resize", 
- "kobber-external_link", 
- "kobber-external_link_arrow", 
- "kobber-female_headset", 
- "kobber-flip_right", 
- "kobber-folder_bookmark", 
- "kobber-font_expand", 
- "kobber-form_template", 
- "kobber-frame_expand", 
- "kobber-frame_shrink", 
- "kobber-game_controller", 
- "kobber-glasses", 
- "kobber-graph_stats", 
- "kobber-headphones", 
- "kobber-heart", 
- "kobber-home", 
- "kobber-hyperlink", 
- "kobber-image", 
- "kobber-image_flower", 
- "kobber-indent_decrease", 
- "kobber-indent_increase", 
- "kobber-information", 
- "kobber-keyboard", 
- "kobber-layers", 
- "kobber-layers_back", 
- "kobber-layout_left", 
- "kobber-layout_module", 
- "kobber-layout_right", 
- "kobber-library", 
- "kobber-light_bulb", 
- "kobber-lock_locked", 
- "kobber-lock_unlocked", 
- "kobber-login", 
- "kobber-logout", 
- "kobber-magnet", 
- "kobber-menu", 
- "kobber-message", 
- "kobber-message_question", 
- "kobber-message_star", 
- "kobber-message_typing", 
- "kobber-message_warning", 
- "kobber-module_edit", 
- "kobber-module_four", 
- "kobber-multiple_users", 
- "kobber-newspaper", 
- "kobber-notebook_pencil", 
- "kobber-number_list", 
- "kobber-option", 
- "kobber-palette", 
- "kobber-paragraph", 
- "kobber-paragraph_center", 
- "kobber-paragraph_left", 
- "kobber-paragraph_right", 
- "kobber-party_balloon", 
- "kobber-pause", 
- "kobber-pen_tools", 
- "kobber-pen_write", 
- "kobber-pencil", 
- "kobber-pin", 
- "kobber-pin_remove", 
- "kobber-play", 
- "kobber-printer", 
- "kobber-question_circle", 
- "kobber-ranking_star", 
- "kobber-rating_star_check", 
- "kobber-read", 
- "kobber-read_home", 
- "kobber-read_idea", 
- "kobber-redo", 
- "kobber-refresh", 
- "kobber-reload", 
- "kobber-remove", 
- "kobber-retouch_graph", 
- "kobber-ribbon_star", 
- "kobber-rocket", 
- "kobber-search", 
- "kobber-settings", 
- "kobber-settings_slider", 
- "kobber-share", 
- "kobber-shrink", 
- "kobber-shuffle", 
- "kobber-social_instagram", 
- "kobber-study_math", 
- "kobber-subscript", 
- "kobber-subtract", 
- "kobber-superscript", 
- "kobber-synchronize__arrows__square", 
- "kobber-synchronize_arrow_clock", 
- "kobber-synchronize_arrows_warning", 
- "kobber-tablet_touch", 
- "kobber-task_list", 
- "kobber-task_list_clock", 
- "kobber-teacher", 
- "kobber-teacher_correct", 
- "kobber-text_bold", 
- "kobber-text_format", 
- "kobber-text_italic", 
- "kobber-text_style", 
- "kobber-text_sync", 
- "kobber-text_underline", 
- "kobber-transfer", 
- "kobber-trash", 
- "kobber-undo", 
- "kobber-user", 
- "kobber-user_edit", 
- "kobber-user_question", 
- "kobber-users", 
- "kobber-video_file_add", 
- "kobber-video_player", 
- "kobber-view", 
- "kobber-view_off", 
- "kobber-volume_off", 
- "kobber-volume_on", 
- "kobber-warning", 
- "kobber-youtube"
-];
