+++ packages/kobber-icons/tsup-scripts/list-icons-and-components.ts
@@ -1,5 +1,17 @@
 import { getIconNames } from ".";
 
+export const listIcons = (symbols: NodeListOf<SVGSymbolElement>) => {
+  let iconsListString = "export const iconsList = [\n";
+  symbols.forEach((symbol, index) => {
+    if (index > 0) {
+      iconsListString += `, \n`;
+    }
+    iconsListString += ` "${symbol.id}"`;
+  });
+  iconsListString += `\n];`;
+  return iconsListString;
+};
+
 export const listWebComponents = (symbols: NodeListOf<SVGSymbolElement>) => {
   let webComponentExports = "";
