+++ packages/kobber-icons/tsup-scripts/make-types.ts
@@ -1,0 +1,9 @@
+export const makeIconTypes = (symbols: NodeListOf<SVGSymbolElement>) => {
+  let iconTypeString = "export type IconType = \n";
+
+  symbols.forEach(symbol => {
+    iconTypeString += `  | "${symbol.id}"\n`;
+  });
+  iconTypeString += `; `;
+  return iconTypeString;
+};
